% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanAnnuels.r
\docType{methods}
\name{barplot,BilanAnnuels-method}
\alias{barplot,BilanAnnuels-method}
\alias{barplot.BilanAnnuels}
\alias{barplot.bilA}
\title{barplot method for object \link{BilanAnnuels}}
\usage{
\S4method{barplot}{BilanAnnuels}(height, legend.tex = NULL...)
}
\arguments{
\item{height}{An object of class BilanAnnuels}

\item{legend.tex}{See barplot help}

\item{...}{additional arguments passed to barplot}
}
\description{
barplot method for object \link{BilanAnnuels}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
\link{BilanAnnuels-class} for examples
}

