% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_annual.R
\name{barplot,report_annual-method}
\alias{barplot,report_annual-method}
\alias{barplot.report_annual}
\title{barplot method for object \link{report_annual-class}}
\usage{
\S4method{barplot}{report_annual}(height, legend.text = NULL, ...)
}
\arguments{
\item{height}{An object of class report_annual}

\item{legend.text}{See barplot help}

\item{...}{additional arguments passed to barplot}
}
\value{
No return value, called for side effects
}
\description{
barplot method for object \link{report_annual-class}
}
\seealso{
\link{report_annual-class} for examples
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
