% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bilA}
\alias{bilA}
\title{An object of class \link{BilaAnnuels-class} with data loaded}
\format{An object of class \link{BilanAnnuels-class} with data slot loaded.}
\usage{
bilA
}
\description{
The dataset corresponds to the three fishways located on the Arzal dam, filled with annual data
}
\keyword{data}

