% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bilAM}
\alias{bilAM}
\title{Annual migration of salmon by migradour}
\format{An object of class \link{BilaAnnuels-class} with data slot loaded.}
\usage{
bilAM
}
\description{
The dataset corresponds to the fishways of the Adour for adult migrant salmons
}
\keyword{data}

