% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_carlot.r
\docType{class}
\name{Bilan_carlot-class}
\alias{Bilan_carlot-class}
\title{Class "Bilan_carlot"}
\description{
Bilan_carlot Bilan class calls the content of the postgres view vue_lot_ope_car, it displays the
results of a categorical variable, or quantitative variable attached for lot, for instance,
it can be used to analyse size or sex
}
\note{
This class is displayed by interface_bilan_lot
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("Bilan_carlot", ...)}
}
\examples{

showClass("Bilan_carlot")
object=new("Bilan_carlot")

}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\references{
\url{http://w3.eptb-vilaine.fr:8080/tracstacomi/wiki/Recette\%20BilanLot}
}
\seealso{
Other Bilan Class \code{\linkS4class{Bilan_carlot}}
\code{\linkS4class{Bilan_poids_moyen}}
\code{\linkS4class{Bilan_stades_pigm}} 
\code{\linkS4class{Bilan_taille}}
\code{\linkS4class{BilanConditionEnv}} 
\code{\linkS4class{BilanEspeces}}
\code{\linkS4class{BilanFonctionnementDC}}
\code{\linkS4class{BilanFonctionnementDF}}
\code{\linkS4class{BilanMigration}}
\code{\linkS4class{BilanMigrationConditionEnv}}
\code{\linkS4class{BilanMigrationInterAnnuelle}}
\code{\linkS4class{BilanMigrationPar}}
}

