% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_poids_moyen.r
\docType{class}
\name{Bilan_poids_moyen-class}
\alias{Bilan_poids_moyen-class}
\title{Class "Bilan_poids_moyen"}
\description{
In trapping ladders, glass eel are seldom counted, as they are too tiny to handle and too numerous to count.
The usuall operation is to weight them, or to use a bucket to measure their volume. These weights or volumes will later
need to be converted to numbers. The glass eel weight may follow a seasonal pattern. It's the case for Anguilla anguilla 
glass eel in the Bay of Biscay. Weights can be modelled using sine wave curves, or more complex gam models. 
This class has a model method to try those models, which can also be used to extact coefficients manually
to manually test more complex models.
Some plots are provided to display the coefficients stored in the database, and the model results. A parameter provided in 
the graphical interface or in the command line (slot liste) takes values "1", ">1", "tous" wich mean respectively use
only individual sample of glass eels, or use "group weights" which can be more close to the real weight of glass eel
during counts as glass eel are not completely drained from their water during handling to preserve their muccus. The list choice
"tous" means that both individual and group weights are selected.
}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{"data.frame"} data for bilan lot}

\item{\code{calcdata}}{A list containing two processed data frames, data and coe}

\item{\code{dc}}{Object of class \code{\link{RefDC-class}}, the counting device}

\item{\code{anneedebut}}{Object of class \code{\link{RefAnnee-class}}. RefAnnee allows to choose the year of beginning}

\item{\code{anneefin}}{Object of class \code{\link{RefAnnee-class}}
refAnnee allows to choose last year of the Bilan}

\item{\code{coe}}{Object of class \code{\link{RefCoe-class}} class loading coefficient
of conversion between quantity (weights or volumes of glass eel) and numbers}

\item{\code{liste}}{Object of class \code{\link{RefListe-class}} RefListe referential
class choose within a list, here the choice is wether subsamples or not. Subsamples
in the stacomi database are samples with a non null value for parent sample. Migration
counts are never made on subsamples but those can be integrated to calculate mean weights.}
}}
\note{
In this class some tools are avalaible to import glass eel measurement from
experimental fishing in the estuary. For the charge method dates for the
request are from august to august (a glass eel season)
}
\examples{
require(stacomiR)
# launching stacomi without selecting the scheme or interface
stacomi(gr_interface=FALSE,
		login_window=FALSE,
		database_expected=FALSE)
\dontrun{
	#create an instance of the class
	bilPM<-new("Bilan_poids_moyen")
	baseODBC<-get("baseODBC",envir=envir_stacomi)
	baseODBC[c(2,3)]<-rep("iav",2)
	assign("baseODBC",baseODBC,envir_stacomi)
	sch<-get("sch",envir=envir_stacomi)
	assign("sch","iav.",envir_stacomi)
	bilPM@liste<-charge(object=bilPM@liste,listechoice=c("=1",">1","tous"),label="")
	# here I'm using weights when number are larger than 1 ie wet weight
	# always choose a date from one year to the next eg 2010 to 2011
	# as the dates are from august to august
	bilPM<-choice_c(bilPM,
			dc=c(6),			
			anneedebut="2009",
			anneefin="2015",
			selectedvalue=">1",
			silent=FALSE)
	bilPM<-connect(bilPM)	
	bilPM<-calcule(bilPM)
}	
# load the dataset generated by previous lines
data("bilPM")
# the calculation will fill the slot calcdata

# A ggplot showing the trend in weight
plot(bilPM, plot.type=1)
# A plot showing both the data and the trend as recorded in the database
plot(bilPM, plot.type=2)
# Same as plot.type=1 but with size according to size of the sample,
# usefull for wet weights where weight are recorded on a number of glass eel
plot(bilPM, plot.type=3)
\dontrun{
# First model with nls, see Guerault and Desaunay (1993) 
model(bilPM,model.type="seasonal")
model(bilPM,model.type="seasonal1")
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other Bilan.Objects: \code{\link{BilanAgedemer-class}},
  \code{\link{BilanAnnuels-class}},
  \code{\link{BilanArgentee-class}},
  \code{\link{BilanConditionEnv-class}},
  \code{\link{BilanEspeces-class}},
  \code{\link{BilanFonctionnementDC-class}},
  \code{\link{BilanFonctionnementDF-class}},
  \code{\link{BilanMigration-class}},
  \code{\link{BilanMigrationCar-class}},
  \code{\link{BilanMigrationInterAnnuelle-class}},
  \code{\link{BilanMigrationMult-class}},
  \code{\link{BilanMigrationMultConditionEnv-class}},
  \code{\link{Bilan_carlot-class}}
}
\keyword{classes}

