% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bilanArg}
\alias{bilanArg}
\title{Silver eel migration in the Somme}
\format{An object of class \link{BilanAnnuels-class} with data slot loaded.}
\usage{
bilanArg
}
\description{
The dataset corresponds to the silver eel traps ("anguillères) for 2015-2016.
This dataset has been kindly provided by the Fédération de Pêche de la Somme,
given the upstream location of the trap, all individuals are female
}
\keyword{data}

