% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanArgentee.r
\docType{class}
\name{BilanArgentee-class}
\alias{BilanArgentee-class}
\title{Class "BilanArgentee"}
\description{
the BilanArgentee class is used to calculate various statistics about the silver eel run
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame with data generated from the database}

\item{\code{calcdata}}{A list of dc with processed data. Each dc contains a data frame with 
\itemize{
\item (1) qualitative data on body contrast (CONT), presence of punctuation on the lateral line (LINP)
\item (2) quantitative data "BL" Body length,"W" weight,"Dv" vertical eye diameter,"Dh" horizontal eye diameter,"FL" pectoral fin length
\item (3) calculated durif stages, Pankhurst's index, Fulton's body weight coefficient K_ful
\item (4) other columns containing data pertaining to the sample and the control operation:  lot_identifiant,ope_identifiant,
ope_dic_identifiant,ope_date_debut,ope_date_fin,dev_code (destination code of fish),
dev_libelle (text for destination of fish)
}}

\item{\code{dc}}{Object of class \link{RefDC-class}: the control devices}

\item{\code{taxons}}{Object of class \link{RefTaxon-class}: the speciess}

\item{\code{stades}}{Object of class \link{RefStades-class} : the stages of the fish}

\item{\code{par}}{Object of class \link{Refpar-class}: the parameters used}

\item{\code{horodatedebut}}{An object of class \code{RefHorodate-class}}

\item{\code{horodatefin}}{An object of class \code{RefHorodate-class}}
}}
\note{
This class is displayed by interface_bilanArgentee
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("BilanArgentee", ...)}
}
\examples{
require(stacomiR)
# launching stacomi without selecting the scheme or interface
stacomi(gr_interface=FALSE,
		login_window=FALSE,
		database_expected=FALSE)
# the following script will load data from the two Anguillère monitored in the Somme

\dontrun{
	#create an instance of the class
	bilanArg<-new("BilanArgentee")
	baseODBC<-get("baseODBC",envir=envir_stacomi)
	baseODBC[c(2,3)]<-rep("fd80",2)
	assign("baseODBC",baseODBC,envir_stacomi)
	sch<-get("sch",envir=envir_stacomi)
	assign("sch","fd80.",envir_stacomi)
	bilanArg<-choice_c(bilanArg,
			dc=c(2,6),			
			horodatedebut="2010-09-01",
			horodatefin="2016-10-04",
			silent=FALSE)
	bilanArg<-connect(bilanArg)
	
}	
# load the dataset generated by previous lines
data("bilanArg")
# the calculation will fill the slot calcdata
bilanArg<-calcule(bilanArg)
# stages are in bilanArg@calcdata[["6"]][,"stage"] 
#look at data structure using str(bilanArg@calcdata[["6"]])

# standard plot as drawn by Laurent Beaulaton (Analyse des données d'argenture acquises en France)
# showing Durif's stage according to size and eye diameter
plot(bilanArg, plot.type=1)

# number per month or year and Durif's stage (year if number of dc >1)

plot(bilanArg, plot.type=2)

# plot showing fulton's coefficient, and size weight graphs
# inspired from Acou et al., 2009 
# Differential Production and Condition Indices of Premigrant
# Eels in Two Small Atlantic Coastal Catchments
# of France
plot(bilanArg, plot.type=3)
# get a list of summary data and print output to screen

plot(bilanArg, plot.type=3)
# print a summary statistic, and save the output in a list for later use
stats<-summary(bilanArg)
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other Bilan Objects: \code{\link{BilanAnnuels-class}},
  \code{\link{BilanEspeces-class}},
  \code{\link{BilanFonctionnementDC-class}},
  \code{\link{BilanFonctionnementDF-class}},
  \code{\link{BilanMigration-class}},
  \code{\link{BilanMigrationConditionEnv-class}},
  \code{\link{BilanMigrationInterAnnuelle-class}},
  \code{\link{BilanMigrationMult-class}},
  \code{\link{Bilan_carlot-class}},
  \code{\link{Bilan_poids_moyen-class}},
  \code{\link{Bilan_stades_pigm-class}},
  \code{\link{Bilan_taille-class}}
}
\keyword{classes}

