% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanEspeces.r
\docType{class}
\name{BilanEspeces-class}
\alias{BilanEspeces-class}
\alias{bilanEspeces}
\title{Class "BilanEspeces" simple migration / number report}
\description{
This class is used to make the assessment of all species, and their number. It is intended
as a simple way to check what fishes are present (taxa + development stage). Unlike the BilanAnnuels,
it is not restricted on chosen taxa or stages but gives counts for all species present. The taxa is reported unless 
a taxa has several case, in which case the different stages for the taxa will be reported
Using the split arguments
the calc method of the class will count numbers, subsamples are not accounted for in the Overview.
The split argument currently takes values year or month. The class is intended to be used over long periods
e.g years. The plot method writes either an histogram or a pie chart of number per
year/week/month.
}
\section{Slots}{

\describe{
\item{\code{dc}}{an object of class \link{RefDC-class}}

\item{\code{anneedebut}}{Object of class \code{\link{RefAnnee-class}}}

\item{\code{anneefin}}{Object of class \code{\link{RefAnnee-class}}}

\item{\code{data}}{\code{data.frame}}

\item{\code{calcdata}}{\code{data.frame} with data processed by the calc method}

\item{\code{split}}{Object of class \code{\link{RefListe-class}} RefListe referential class choose within a list}
}}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other Bilan Classes 
\code{\linkS4class{Bilan_carlot}},
\code{\linkS4class{Bilan_poids_moyen}},
\code{\linkS4class{BilanConditionEnv}}, \code{\linkS4class{BilanEspeces}},
\code{\linkS4class{BilanFonctionnementDC}},
\code{\linkS4class{BilanFonctionnementDF}},
\code{\linkS4class{BilanMigration}},
\code{\linkS4class{BilanMigrationMultConditionEnv}},
\code{\linkS4class{BilanMigrationInterAnnuelle}},
\code{\linkS4class{BilanMigrationCar}}

Other Bilan.Objects: \code{\link{BilanAgedemer-class}},
  \code{\link{BilanAnnuels-class}},
  \code{\link{BilanArgentee-class}},
  \code{\link{BilanConditionEnv-class}},
  \code{\link{BilanFonctionnementDC-class}},
  \code{\link{BilanFonctionnementDF-class}},
  \code{\link{BilanMigration-class}},
  \code{\link{BilanMigrationCar-class}},
  \code{\link{BilanMigrationInterAnnuelle-class}},
  \code{\link{BilanMigrationMult-class}},
  \code{\link{BilanMigrationMultConditionEnv-class}},
  \code{\link{Bilan_carlot-class}},
  \code{\link{Bilan_poids_moyen-class}}
}
\keyword{classes}

