% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanFonctionnementDC.r
\docType{class}
\name{BilanFonctionnementDC-class}
\alias{BilanFonctionnementDC-class}
\title{Class "BilanFonctionnementDC" Bilan du fonctionnement du dispositif de
comptage}
\description{
The counting device is not always working. It may me stopped either
following a monitoring protocol, or due to misfunction of the device, this
class allows to draw graphics allowing an overview of the device operation
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame}

\item{\code{dc}}{An object of class \code{RefDC-class}}

\item{\code{horodatedebut}}{An object of class \code{RefHorodate-class}}

\item{\code{horodatefin}}{An object of class \code{RefHorodate-class}}
}}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("BilanFonctionnementDC", ...)}.
}
\examples{
require(stacomiR)
stacomi(gr_interface=FALSE,
		login_window=FALSE,
		database_expected=FALSE)
###########################################################
## An example that will work only if the database is present 
## and the program installed and comprises the schema iav
###########################################################"
\dontrun{
	bfDC=new("BilanFonctionnementDC")
	bfDC<-choice_c(bfDC,
			5,
			horodatedebut="2000-01-01",
			horodatefin="2015-12-31",
			silent=TRUE)
	Sys.setenv(TZ='GMT')
	# This dataset formating is GMT. If this option is not set
	# the dataset is tranformed from timestamp to date
	bfDC<-charge(bfDC)
	bfDC<-connect(bfDC)
	# this dataset has been loaded by the previous lines
###########################################################	
# Without connexion to the database (use dataset bfDC)
##########################################################
	data("bfDC")
	plot(bfDC,plot.type="1")
	plot(bfDC,plot.type="2")
	plot(bfDC,plot.type="3",main="trial title")
	plot(bfDC,plot.type="4",main="trial title")
# the following will write in the datawd folder
 summary(bfDC)
	}
##

  
	

}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other Bilan.Objects: \code{\link{BilanAgedemer-class}},
  \code{\link{BilanAnnuels-class}},
  \code{\link{BilanArgentee-class}},
  \code{\link{BilanConditionEnv-class}},
  \code{\link{BilanEspeces-class}},
  \code{\link{BilanFonctionnementDF-class}},
  \code{\link{BilanMigration-class}},
  \code{\link{BilanMigrationCar-class}},
  \code{\link{BilanMigrationInterAnnuelle-class}},
  \code{\link{BilanMigrationMult-class}},
  \code{\link{BilanMigrationMultConditionEnv-class}},
  \code{\link{Bilan_carlot-class}},
  \code{\link{Bilan_poids_moyen-class}}
}
\keyword{classes}

