% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bilanFonctionnementDC_bM}
\alias{bilanFonctionnementDC_bM}
\title{An object of class \link{BilanFonctionnementDC-class} with data loaded}
\format{An object of class BilanFonctionnementDC}
\usage{
bilanFonctionnementDC_bM
}
\description{
This dataset corresponds to the data collected at three different control devices
it is loaded along with \link{bM_Arzal}
}
\keyword{data}
