% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanFonctionnementDF.r
\docType{class}
\name{BilanFonctionnementDF-class}
\alias{BilanFonctionnementDF-class}
\title{Class "BilanFonctionnementDF" Report fishway work}
\description{
The DF (Dispositif de Franchissement) is a fishway. It may be automated and
be operated only during certain periods. This report allows to see the detail of its work.
In the database four types of operation are set,  "1"=normal operation,
"2"=Device stopped in nomral operation (ie lift ascending, high tide...),
"3"="Stopped for maintenance or other problem",
"4"="Works but not fully operational, ie flow problem, flood, clogged with debris...",
"5"="Not known")
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("BilanFonctionnementDF")}.
}
\examples{
inst/examples/bilanFonctionnementDF_example.R
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other Bilan Class \code{\linkS4class{Bilan_carlot}}
\code{\linkS4class{Bilan_poids_moyen}}
\code{\linkS4class{Bilan_stades_pigm}} 
\code{\linkS4class{Bilan_taille}}
\code{\linkS4class{BilanConditionEnv}} 
\code{\linkS4class{BilanEspeces}}
\code{\linkS4class{BilanFonctionnementDC}}
\code{\linkS4class{BilanFonctionnementDF}}
\code{\linkS4class{BilanMigration}}
\code{\linkS4class{BilanMigrationConditionEnv}}
\code{\linkS4class{BilanMigrationInterAnnuelle}}
\code{\linkS4class{BilanMigrationPar}}
}
\concept{
Bilan Object
}

