% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bilanFonctionnementDF_bM}
\alias{bilanFonctionnementDF_bM}
\title{Fishway operation for the vertical slot fishway (Arzal dam, Vilaine, France).}
\format{An object of class \link{BilanFonctionnementDF-class}}
\usage{
bilanFonctionnementDF_bM
}
\description{
This dataset corresponds to the data collected at in the vertical slot fishway
it is loaded along with \link{bM_Arzal} and used to demonstrate the \link{BilanMigration-class}
when the database is not installed.
}
\keyword{data}

