\name{BilanMigration_functions}
\alias{funstat}
\alias{funstatJournalier}
\alias{funtable}
\alias{funtraitement_poids}

\title{Function to calculate statistics per month or per day within class \code{\linkS4class{BilanMigration}} }
\usage{
  funstat(tableau, duree, taxon, stade, DC)
  funstatJournalier(tableau, duree, taxon, stade, DC)
  funtable(tableau, duree, taxon, stade, DC, resum)
  funtraitement_poids(tableau, duree)
}
\arguments{
  \item{tableau}{the table generated by \code{funSousListeBilanMigration} in class \code{\linkS4class{BilanMigration}}}

  \item{duree}{a POSIXt sequence generated for the range of the period selected}

  \item{taxon}{the latin name of the selected taxa}

  \item{stade}{the stage selected}

  \item{DC}{the code of the counting device}
  
  \item{resum}{A monthly summary of data generated by \code{funstat}}
}
\description{
  \code{funstat} will calculate statistics per month
  \code{funstatJournalier} calculates daily values
  \code{funtable}  function to print and save statistics in .csv and .html  formats
  \code{funtraitement_poids}  returns a table where weights and number are calculated from number and weights respectively. Performs a query to
  collect the conversion coefficients, this function is used for glass eel which are most often weighted, not counted
}
\author{
  Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

