% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationCar.r
\docType{class}
\name{BilanMigrationCar-class}
\alias{BilanMigrationCar-class}
\alias{BilanMigrationMult}
\alias{bilanMigrationMult}
\title{Migration report along with quantitative and
qualitative characteristics}
\description{
Migration along with qualitative or quantitative characteristics or both
(e.g.) weight of eels according to the size class per period of time, weight
of fish according to gender, number of fish per age class. This class does not split migration evenly over 
time period. So, unlike calculations made in class BilanMigration and BilanMigrationMult
the whole time span of the migration operation is not considered, only  the date of beginning of 
the operation is used to perform calculation.
}
\section{Slots}{

\describe{
\item{\code{calcdata}}{A "list" of calculated data, filled in by the calcule method}

\item{\code{parquan}}{An object of class \link{Refparquan-class}, quantitative parameter}

\item{\code{parqual}}{An object of class \link{Refparqual-class}, quanlitative parameter}

\item{\code{echantillon}}{An object of class \link{RefChoix-class}, vector of choice}

\item{\code{valeurs_possibles}}{A \code{data.frame} choice among possible choice of a qualitative parameter (discrete)}

\item{\code{dc}}{an object of class \link{RefDC-class} inherited from \link{BilanMigration-class}}

\item{\code{taxons}}{An object of class \link{RefTaxon-class} inherited from \link{BilanMigration-class}}

\item{\code{stades}}{An object of class \link{RefStades-class} inherited from \link{BilanMigration-class}}

\item{\code{pasDeTemps}}{An object of class \link{PasDeTempsJournalier-class} inherited from \link{BilanMigration-class}}

\item{\code{data}}{A \code{data.frame} inherited from \link{BilanMigration-class}, stores the results}

\item{\code{time.sequence}}{An object of class "POSIXct" inherited from \link{BilanMigration-class}}
}}
\note{
The main difference between this class and \link{Bilan_carlot-class} is that this class allows to
select (or not) the samples, and that it handles quantitative and qualitative parameters separately.

program : default two parameter choice, checking box "none" will allow the program to ignore the parameter
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("BilanMigrationCar", ...)}.  they are loaded by the interface
using interface_BilanMigrationCar function.
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other Bilan.Objects: \code{\link{BilanAgedemer-class}},
  \code{\link{BilanAnnuels-class}},
  \code{\link{BilanArgentee-class}},
  \code{\link{BilanConditionEnv-class}},
  \code{\link{BilanEspeces-class}},
  \code{\link{BilanFonctionnementDC-class}},
  \code{\link{BilanFonctionnementDF-class}},
  \code{\link{BilanMigration-class}},
  \code{\link{BilanMigrationInterAnnuelle-class}},
  \code{\link{BilanMigrationMult-class}},
  \code{\link{BilanMigrationMultConditionEnv-class}},
  \code{\link{Bilan_carlot-class}},
  \code{\link{Bilan_poids_moyen-class}}
}
\concept{
Bilan Object
}
\keyword{classes}

