% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMult.r
\docType{class}
\name{BilanMigrationMult-class}
\alias{BilanMigrationMult-class}
\title{class BilanMigrationMult}
\description{
Migration counts for several Fish counting devices (DC), several taxa and several stages.
This migration count can be built either by the graphical interface or in command line
\code{new("BilanMigrationMult",
dc=new("RefDC"),taxons=("RefTaxon"),stades=new("RefStades"),pasDeTemps=new("PasDeTempsJournalier"),data=data.frame(),calcdata=list(),
coef_conversion=data.frame())}.
}
\section{Slots}{

\describe{
\item{\code{dc}}{An object of class \code{RefDC}}

\item{\code{taxons}}{An object of class \code{\link{RefTaxon}}}

\item{\code{stades}}{An object of class \code{\link{RefStades}}}

\item{\code{pasDeTemps}}{An object of class \code{\link{PasDeTempsJournalier}}}

\item{\code{data}}{A data.frame containing raw data filled by the connect method}

\item{\code{calcdata}}{A "list" of calculated daily data, one per dc, filled in by the calcule method}

\item{\code{coef_conversion}}{A data.frame of daily weight to number conversion coefficients, filled in by the connect
method if any weight are found in the data slot.}
}}
\note{
A Migration Bilan comes from a migration monitoring : the fishes are monitored in a section of river, this section is
called a control station (station). Most often, there is a dam, one or several fishways (DF) which comprise one or several counting devices (DC).
On each counting device, the migration is recorded. It can be either an instant recording (video control) or the use of traps,
Operations are monitoring operations during a period. For each operation, several species of fishes can be recorded (samples). In the case of migratory 
fishes the stage of development is important as it may indicate generic migrations, to and fro, between the river and the sea.

Hence a Multiple Migration Bilan is built from several one or several counting devices (DC), one or several Taxa (Taxon), one or several stages
(stage). The migration can be also recorded not as numbers, but in the case of glass eels, as weight, which will be later transformed to number, 
from daily conversion coefficients. The methods in this class test whether the counts are numbers or another type of quantity.
This class makes different calculations than BilanMigration, it does not handle escapement coefficients,
it uses quantities other than numbers if necessary (only used for glass eel in the connect method).
}
\examples{
library(stacomiR)

stacomi(gr_interface=FALSE,) ## launches the application in the command line
# here is the structure of the class
showClass("BilanMigrationMult")
data("bMM_Arzal")
# use the following to get the raw data loaded by the connect method
# not shown there as the database and program might not be installed
#bMM_Arzal<-connect(bMM_Arzal)
str(bMM_Arzal@data)
# calculations
bMM_Arzal<-calcule(bMM_Arzal)
#Individual plot for all DC, taxon and stage where data present
plot(bMM_Arzal)
#cumulated migration at the station (all stages and DC grouped)
cumplot(bMM_Arzal)
# combined plot for ggplot2
plot1(bMM_Arzal)
# data will be written in the datadirectory specified in the stacomi/calcmig.csv file
summary(bMM_Arzal)
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

