% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMultConditionEnv.r
\docType{class}
\name{BilanMigrationMultConditionEnv-class}
\alias{BilanMigrationMultConditionEnv}
\alias{BilanMigrationMultConditionEnv-class}
\alias{bilanmigrationmultconditionenv}
\title{Class "BilanMigrationMultConditionEnv"}
\description{
Enables to compute an annual overview of fish migration and environmental
conditions in the same chart. Environmental conditions may trigger migration events, variation in flow
or temperatures can  be plotted along migration to check graphically for a possible relation. To enable this, 
environmental conditions are loaded from an "environmental monitoring station", which records environmental
parameters and is attached to a migratory station in the database.
This class enables both continuous output (temperature -flow) as well as discrete parameters (qualitative = moon 
phase, type of operation of a gate, opening of a gate...) which will be displayed on the graph. Values are scaled so that 
single plot can display migration numbers and environmental parameters. Environmental parameters when stored at a 
time scale lower that a day are averaged per day, unless they are qualitative parameters, in which case only the first
event of the day is displayed on the annual plot.
}
\section{Slots}{

\describe{
\item{\code{bilanMigrationMult}}{\link{BilanMigrationMult-class}}

\item{\code{bilanConditionEnv}}{\link{BilanConditionEnv-class}}
}}
\examples{
require(stacomiR)
stacomi(gr_interface=FALSE,
		login_window=FALSE,
		database_expected=FALSE)
# the following will load the data provided the user has access to the database
# with data in the iav example scheme.
\dontrun{
	bmmCE<-new("BilanMigrationMultConditionEnv")
	bmmCE<-choice_c(bmmCE,
			dc=c(5,6,12),
			taxon=c("Anguilla anguilla"),
			stade=c("AGJ","AGG","CIV"),
			stationMesure=c("temp_gabion","coef_maree","phases_lune"),
			datedebut="2008-01-01",
			datefin="2008-12-31",
			silent=FALSE)	
	bmmCE<-charge(bmmCE)
	bmmCE<-connect(bmmCE)
	
}	

data("bmmCE")
bmmCE<-calcule(bmmCE,silent=TRUE)
# An example of plot with custom colors.
plot(bmmCE,
		color_station=c("temp_gabion"="red","coef_maree"="blue","phases_lune"="pink"),
 		color_dc=c("5"="yellow","6"="orange","12"="purple")
)
}
\author{
cedric.briand"at"eptb-vilaine.fr marion.legrand"at"logrami.fr

Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other Bilan.Objects: \code{\link{BilanAgedemer-class}},
  \code{\link{BilanAnnuels-class}},
  \code{\link{BilanArgentee-class}},
  \code{\link{BilanConditionEnv-class}},
  \code{\link{BilanEspeces-class}},
  \code{\link{BilanFonctionnementDC-class}},
  \code{\link{BilanFonctionnementDF-class}},
  \code{\link{BilanMigration-class}},
  \code{\link{BilanMigrationCar-class}},
  \code{\link{BilanMigrationInterAnnuelle-class}},
  \code{\link{BilanMigrationMult-class}},
  \code{\link{Bilan_carlot-class}},
  \code{\link{Bilan_poids_moyen-class}}

Other Bilan.Objects: \code{\link{BilanAgedemer-class}},
  \code{\link{BilanAnnuels-class}},
  \code{\link{BilanArgentee-class}},
  \code{\link{BilanConditionEnv-class}},
  \code{\link{BilanEspeces-class}},
  \code{\link{BilanFonctionnementDC-class}},
  \code{\link{BilanFonctionnementDF-class}},
  \code{\link{BilanMigration-class}},
  \code{\link{BilanMigrationCar-class}},
  \code{\link{BilanMigrationInterAnnuelle-class}},
  \code{\link{BilanMigrationMult-class}},
  \code{\link{Bilan_carlot-class}},
  \code{\link{Bilan_poids_moyen-class}}
}
\keyword{classes}

