% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanOperation.r
\docType{class}
\name{BilanOperation-class}
\alias{BilanOperation}
\alias{BilanOperation-class}
\alias{bilanOperation}
\alias{bilanoperation}
\title{Report on operations}
\description{
Operations are monitoring operations generated between two dates. In the case of video monitoring
or similar, they can be instantaneous
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("BilanOperation")}.
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other Bilan Class \code{\linkS4class{Bilan_carlot}}
\code{\linkS4class{Bilan_poids_moyen}}
\code{\linkS4class{BilanConditionEnv}} 
\code{\linkS4class{BilanEspeces}}
\code{\linkS4class{BilanFonctionnementDC}}
\code{\linkS4class{BilanFonctionnementDF}}
\code{\linkS4class{BilanMigration}}
\code{\linkS4class{BilanMigrationMultConditionEnv}}
\code{\linkS4class{BilanMigrationInterAnnuelle}}
\code{\linkS4class{BilanMigrationCar}}
}
\concept{
Bilan Object
}
\keyword{classes}

