% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bilanOperation}
\alias{bilanOperation}
\title{An object of class \link{BilanOperation-class} with data loaded}
\format{An object of class BilanOperation}
\usage{
bilanOperation
}
\description{
This dataset corresponds to the data collected at three different control devices
it is loaded along with \link{bM_Arzal}
}
\keyword{data}

