% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bM_Arzal}
\alias{bM_Arzal}
\title{Video counting of thin lipped mullet (Liza ramada) in 2015 on the Vilaine (France)}
\format{An object of class bilanMigration with 8 slots:
\describe{
  \item{dc}{the \code{RefDC} object with 4 slots filled with data corresponding to the iav postgres schema}
  \item{taxons}{the \code{RefTaxon} the taxa selected}
  \item{stades}{the \code{RefStades} the stage selected}
  \item{pasDeTemps}{the \code{PasDeTempsJournalier} calculated for all 2015}
  \item{data}{ A dataframe with 10304 rows and 11 variables
			\describe{
  			\item{ope_identifiant}{operation id}
  			\item{lot_identifiant}{sample id}
  			\item{lot_identifiant}{sample id}
  			\item{ope_dic_identifiant}{dc id}
  			\item{lot_tax_code}{species id}
  			\item{lot_std_code}{stage id}
  			\item{value}{the value}
  			\item{type_de_quantite}{either effectif (number) or poids (weights)}
  			\item{lot_dev_code}{destination of the fishes}
  			\item{lot_methode_obtention}{method of data collection, measured, calculated...} 
  			}
  }
  \item{coef_conversion}{A data frame with 0 observations : no quantity are reported for video recording of mullets, only numbers}
  \item{time.sequence}{A time sequence generated for the bilan, used internally}
}}
\usage{
bM_Arzal
}
\description{
This data corresponds to the data collected at the vertical slot fishway
in 2015, video recording of the thin lipped mullet Liza ramada migration
}
\keyword{data}
