% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bmi}
\alias{bmi}
\title{An object of class \link{BilanMigrationInterAnnuelle-class} with data loaded}
\format{An object of class BilanMigrationInterAnnuelle-class with data slot loaded.}
\usage{
bmi
}
\description{
The first eel trapping ladder in France was built by Antoine Legault and the team from Rennes
in the Sevre Niortaise, Marais Poitevin. Also refurbished several times since 1984 it has been 
operational at the same location and provides one of the longest series of eel migration. For this reason,
the dataset has been loaded as an example for the BilanMigrationInterAnnuelle-class
}
\keyword{data}

