% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bmi_vichy}
\alias{bmi_vichy}
\title{Seasonality of salmon migration at the Vichy counting station (Loire)}
\format{An object of class \link{BilanMigrationInterAnnuelle-class} with 7 slots:
\describe{
  \item{dc}{the \code{RefDC} object with 4 slots filled with data corresponding to the iav postgres schema}
  \item{taxons}{the \code{RefTaxon} the taxa selected}
  \item{stades}{the \code{RefStades} the stage selected}
  \item{anneeDebut}{the \code{PasDeTempsJournalier} calculated for all 2015}
  \item{anneeFin}{the \code{PasDeTempsJournalier} calculated for all 2015}
  \item{data}{ A dataframe with 7138 rows and 10 variables
			\describe{
  			\item{bjo_identifiant}{sample id}
  			\item{bjo_dis_identifiant}{dc id}
  			\item{bjo_tax_code}{species id}
  			\item{bjo_std_code}{stage id}
  			\item{bjo_annee}{year}
  			\item{bjo_jour}{date}
  			\item{bjo_labelquantite}{method of data collection, measured, calculated...}
  			\item{bjo_horodateexport}{date with special format for export}
  			\item{bjo_org_code}{organisme provided the data}
  			}
  }
}}
\usage{
bmi_vichy
}
\description{
This data corresponds to the data collected at the Vichy fishway
between 1997 and 2012, video recording of the Salmo salar upstream migration.
This dataset has been kindly provided by Loire Grands Migrateurs.
}
\keyword{data}

