% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bMM_Arzal}
\alias{bMM_Arzal}
\title{Anguilla migration at the Arzal station (BilanMigrationMult-class)}
\format{An object of class bilanMigrationMult with slots:
\describe{
  \item{dc}{the \code{RefDC} object filled with data}
  \item{taxons}{the \code{RefTaxon} object filled in with data corresponding to dc}
  \item{stades}{the \code{RefStades} object filled in with data corresponding to dc, and taxon}
  \item{pasDeTemps}{the \code{PasDeTempsJournalier} calculated for all 2011}
  \item{data}{ A dataframe with 400 rows and 11 variables
			\describe{
  			\item{ope_identifiant}{operation id}
  			\item{lot_identifiant}{sample id}
  			\item{lot_identifiant}{sample id}
  			\item{ope_dic_identifiant}{dc id}
  			\item{lot_tax_code}{species id}
  			\item{lot_std_code}{stage id}
  			\item{value}{the value}
  			\item{type_de_quantite}{either effectif (number) or poids (weights)}
  			\item{lot_dev_code}{destination of the fishes}
  			\item{lot_methode_obtention}{method of data collection, measured, calculated...} 
  			}
  }
  \item{calcdata}{slot to be filled with the calcule method} 
  \item{coef_conversion}{A data frame with 364 observations with daily coefficients to convert from weight to numbers}
  \item{time.sequence}{A time sequence generated for the bilan, used internally by the object}
}}
\usage{
bMM_Arzal
}
\description{
This data corresponds to the data collected from three fishways
and correspond to the migration station at Arzal in 2011 for all
stages of eel (Anguilla anguilla)
}
\keyword{data}
