% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bmmCE}
\alias{bmmCE}
\title{An object of class BilanMigrationMultConditionEnv with data loaded}
\format{An object of class \link{BilanConditionEnv-class} with data slot loaded:
\describe{
  \item{bilanMigrationMult}{An object of class \link{BilanMigrationMult-class}}
  \item{bilanConditionEnv}{ An object of class \link{BilanConditionEnv-class}}#' 	
  }}
\usage{
bmmCE
}
\description{
The dataset correspond to data loaded for the Arzal dam (Vilaine) in 2008, two quantitative
parameters (temperature and tide coefficient and a qualitative parameter moon phase are loaded.
}
\keyword{data}

