% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{calcmig}
\alias{calcmig}
\title{dataframe containing the default connection arguments
the program will use a file installed in c:/program files/stacomi but
if not found will switch to the default}
\format{An object of class \code{data.frame} with 1 rows and 11 columns.}
\usage{
calcmig
}
\description{
dataframe containing the default connection arguments
the program will use a file installed in c:/program files/stacomi but
if not found will switch to the default
}
\keyword{datasets}
