% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_carlot.r
\docType{methods}
\name{calcule,Bilan_carlot-method}
\alias{calcule,Bilan_carlot-method}
\title{Calcule method for Bilan_carlot}
\usage{
\S4method{calcule}{Bilan_carlot}(object, h)
}
\arguments{
\item{object}{An object of class \code{\link{Bilan_carlot}}}

\item{...}{Additional parameters}
}
\description{
Calcule method for Bilan_carlot
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

