% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_poids_moyen.r
\docType{methods}
\name{calcule,Bilan_poids_moyen-method}
\alias{calcule,Bilan_poids_moyen-method}
\title{Calcule method for Bilan_poids_moyen}
\usage{
\S4method{calcule}{Bilan_poids_moyen}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{Bilan_poids_moyen-class}}}

\item{silent}{Boolean, if TRUE, information messages are not displayed, only warnings and errors}
}
\description{
Calcule method for Bilan_poids_moyen
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
