% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_taille.r
\docType{methods}
\name{calcule,Bilan_taille-method}
\alias{calcule,Bilan_taille-method}
\title{Calcule method for BilanTaille}
\usage{
\S4method{calcule}{Bilan_taille}(object)
}
\arguments{
\item{object}{An object of class \code{\link{Bilan_taille}}}
}
\description{
Calcule method for BilanTaille
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

