% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanAgedemer.r
\docType{methods}
\name{calcule,BilanAgedemer-method}
\alias{calcule,BilanAgedemer-method}
\title{Split data according to the limits
set in the limit1hm, and limit2hm arguments of the \link{BilanAgedemer-class}.}
\usage{
\S4method{calcule}{BilanAgedemer}(object, silent)
}
\arguments{
\item{object}{An object of class \code{\link{BilanAgedemer-class}}}

\item{silent}{Default FALSE, if TRUE the program should no display messages}
}
\value{
An object of class \link{BilanAgedemer-class} with calculated data in slot calcdata
}
\description{
If no value are provided in the limit1hm slot, an error is returned, if
no value is provided in the limit2hm slot a default upper value for salmon
size is taken to ensure all salmon are either of age 1 or 2, but no age 3 are
returned
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

