% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanArgentee.r
\docType{methods}
\name{calcule,BilanArgentee-method}
\alias{calcule,BilanArgentee-method}
\title{Calcule method for BilanArgentee, this method will pass the data from long to wide format 
( one line per individual) and calculate Durif silvering index and Pankhurst and Fulton's K.}
\usage{
\S4method{calcule}{BilanArgentee}(object, silent)
}
\arguments{
\item{object}{An object of class \code{\link{BilanArgentee-class}}}

\item{silent}{Boolean, if TRUE, information messages are not displayed, only warnings and errors}
}
\description{
Calcule method for BilanArgentee, this method will pass the data from long to wide format 
( one line per individual) and calculate Durif silvering index and Pankhurst and Fulton's K.
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
