% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigration.r
\docType{methods}
\name{calcule,BilanMigration-method}
\alias{calcule,BilanMigration-method}
\title{calcule method for BilanMigration}
\usage{
\S4method{calcule}{BilanMigration}(object)
}
\arguments{
\item{object}{An object of class \code{\link{BilanMigration-class}}}
}
\value{
BilanMigration with slots filled by user choice
}
\description{
calcule method for BilanMigration
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

