% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationCar.r
\docType{methods}
\name{calcule,BilanMigrationCar-method}
\alias{calcule,BilanMigrationCar-method}
\title{Computes data to a standard format for the summary and plot methods.}
\usage{
\S4method{calcule}{BilanMigrationCar}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{BilanMigrationCar-class}}}

\item{silent}{Boolean default FALSE, if TRUE information messages not displayed}
}
\description{
Merges the content of the list elements "parqual" and "parquan" in the data slot, and creates
a single dataframe with one line per qualitative and quantitative pair. This methods allow to 
cross one quantity (e.g. length) with a qualitative parameter (e.g. sex).
}

