% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationConditionEnv.r
\docType{methods}
\name{calcule,BilanMigrationConditionEnv-method}
\alias{calcule,BilanMigrationConditionEnv-method}
\title{Performs the calculations of environment conditions attached to a migration monitoring station}
\usage{
\S4method{calcule}{BilanMigrationConditionEnv}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{BilanMigrationConditionEnv-class}}}

\item{...}{additional parameters}
}
\value{
\code{\link{BilanMigrationEnv-class}}
}
\description{
Performs the calculations of environment conditions attached to a migration monitoring station
}

