% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMult.r
\docType{methods}
\name{calcule,BilanMigrationMult-method}
\alias{calcule,BilanMigrationMult-method}
\title{calcule method for BilanMigrationMult}
\usage{
\S4method{calcule}{BilanMigrationMult}(object, negative = FALSE,
  silent = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{BilanMigrationMult-class}}}

\item{negative}{a boolean indicating if a separate sum must be done for positive and negative values, if true, positive and negative counts return 
different rows}

\item{silent}{Default FALSE, should messages be stopped}
}
\value{
BilanMigrationMult with a list in calcdata, one for each triplet (dc/taxa/stage) with data
}
\description{
does the calculation once data are filled.
}
\note{
The class BilanMigrationMult does not handle  escapement rates. Use class BilanMigration if you want to handle them. The class does not handler
'devenir' i.e. the destination of the fishes.
}
