% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMultConditionEnv.r
\docType{methods}
\name{calcule,BilanMigrationMultConditionEnv-method}
\alias{calcule,BilanMigrationMultConditionEnv-method}
\title{Calculations for migration in the class \link{BilanMigrationMultConditionEnv-class}}
\usage{
\S4method{calcule}{BilanMigrationMultConditionEnv}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{BilanMigrationMultConditionEnv-class}}

\item{silent}{Boolean default FALSE, if TRUE information messages not displayed}
}
\value{
\link{BilanMigrationMultConditionEnv-class} with data in slot bmmCE@bilanMigrationMult@calcdata
}
\description{
Runs the calcule method in \link{BilanMigrationMult-class}
}

