% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationPar.r
\docType{methods}
\name{calcule,BilanMigrationPar-method}
\alias{calcule,BilanMigrationPar-method}
\title{calcule methode}
\usage{
\S4method{calcule}{BilanMigrationPar}(object)
}
\arguments{
\item{object}{An object of class \code{\link{BilanMigrationPar-class}}}
}
\description{
calcule methode
}

