% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig.R
\name{calcule,report_mig-method}
\alias{calcule,report_mig-method}
\alias{calcule.report_mig}
\title{Transforms migration per period to daily migrations, and performs the conversion from weights to number is data
are stored as weights (glass eel).}
\usage{
\S4method{calcule}{report_mig}(object, negative = FALSE, silent = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{report_mig-class}}}

\item{negative}{a boolean indicating if a separate sum must be done for positive and negative values, if true, positive and negative counts return
different rows}

\item{silent}{Boolean, if TRUE, information messages are not displayed, only warnings and errors}
}
\value{
report_mig with calcdata slot filled. It is a list with one element per counting device containing
\describe{
\item{method}{In the case of instantaneous periods (video counting) the sum of daily values is done by the \link{fun_report_mig_mult} method and the value indicated in method is "sum".
 If any migration monitoring period is longer than a day, then the migration is split using the \link{fun_report_mig_mult_overlaps} function and the value indicated in the
method is "overlaps" as the latter method uses the overlap package to split migration period.}
\item{data}{the calculated data.}
\item{contient_poids}{A boolean which indicates, in the case of glass eel, that the function \link{fun_weight_conversion} has been run to convert the weights to numbers using the weight
to number coefficients in the database (see \link{report_ge_weight}).}
\item{negative}{A parameter indicating if negative migration (downstream in the case of upstream migration devices) have been converted to positive numbers,
not developed yet}}
}
\description{
The calculation must be launched once data are filled by the connect method. Currently the negative argument
has no effect.
}
\note{
The class report_mig does not handle escapement rates nor
'devenir' i.e. the destination of the fishes.
}
