% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_char.R
\name{calcule,report_mig_char-method}
\alias{calcule,report_mig_char-method}
\alias{calcule.report_mig_char}
\title{Computes data to a standard format for the summary and plot methods.}
\usage{
\S4method{calcule}{report_mig_char}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_mig_char-class}}

\item{silent}{Boolean default FALSE, if TRUE information messages not displayed}
}
\value{
An object of class \link{report_mig_char-class} with slot \code{@calcdata} filled
}
\description{
Merges the content of the list elements 'parqual' and 'parquan' in the data slot, and creates
a single dataframe with one line per qualitative and quantitative pair. This methods allow to 
cross one quantity (e.g. length) with a qualitative parameter (e.g. sex).
}
