% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_env.R
\name{calcule,report_mig_env-method}
\alias{calcule,report_mig_env-method}
\alias{calcule.report_mig_env}
\title{Internal handler function}
\usage{
\S4method{calcule}{report_mig_env}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_mig_env-class}}

\item{silent}{Boolean default FALSE, if TRUE information messages not displayed}

\item{h}{a handler}

\item{...}{Additional parameters}
}
\value{
\link{report_mig_env-class} with data in slot r_mig_env@report_mig_mult@calcdata
}
\description{
Internal handler function
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{Calculations}
\keyword{Runs}
\keyword{\link{report_mig_env-class}}
\keyword{\link{report_mig_mult-class}}
\keyword{calcule}
\keyword{class}
\keyword{for}
\keyword{in}
\keyword{internal}
\keyword{method}
\keyword{migration}
\keyword{the}
