% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_sample_char.R
\docType{methods}
\name{calcule,report_sample_char-method}
\alias{calcule,report_sample_char-method}
\alias{calcule.report_sample_char}
\title{Calculation for report_sample_char}
\usage{
\S4method{calcule}{report_sample_char}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{report_sample_char-class}}}

\item{silent}{Boolean, if TRUE, information messages are not displayed, only warnings and errors}
}
\description{
In that class, most treatments are done in the query, this method checks that data are available and fills information for year, month, two weeks, week, doy
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
