% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_silver_eel.R
\name{calcule,report_silver_eel-method}
\alias{calcule,report_silver_eel-method}
\alias{calcule.report_silver_eel}
\title{Calculate individual silver eel parameters.}
\usage{
\S4method{calcule}{report_silver_eel}(object, silent)
}
\arguments{
\item{object}{An object of class \link{report_silver_eel-class}}

\item{silent}{Boolean, if TRUE, information messages are not displayed, only warnings and errors}
}
\value{
An object of class \link{report_silver_eel-class} with slot calcdata filled, as a list
for each counting device
}
\description{
This calcule method for report_silver_eel, will transform data from long (one line per size characteristic,
size, weight, eye diameter, pectoral fin measurement, lateral line and constrast) to wide format (one
line per silver eel). It will also calculate Durif silvering index and Pankhurst and Fulton's K.
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
