% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generic.R
\name{calcule}
\alias{calcule}
\title{Generic method for calculations}
\usage{
calcule(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Additional parms}
}
\description{
Generic method for calculations
}
\author{
cedric.briand
}
