% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitaires.r
\name{ceil.POSIXt}
\alias{ceil.POSIXt}
\title{ceil.POSIXT function imported from depuis Hmisc that cannot be loaded}
\usage{
ceil.POSIXt(x, digits = c("secs", "mins", "hours", "days", "months", "years"),
  ...)
}
\arguments{
\item{x}{A POSIXt}

\item{digits}{}
}
\value{
a truncated time value
}
\description{
ceil.POSIXT function imported from depuis Hmisc that cannot be loaded
}
\author{
Charles Dupont
}

