% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_carlot.r
\docType{methods}
\name{charge,Bilan_carlot-method}
\alias{charge,Bilan_carlot-method}
\title{charge method for Bilan_carlot class}
\usage{
\S4method{charge}{Bilan_carlot}(object, h)
}
\arguments{
\item{object}{An object of class \link{Bilan_carlot-class}}

\item{h}{a handler}
}
\value{
An object of class \link{Bilan_carlot-class} with slots filled with user choice

An object of the class
}
\description{
this method verifies that boxes have been clicked in the user interface and gets the objects pasted in 
envir_stacomi
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

