% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_poids_moyen.r
\docType{methods}
\name{charge,Bilan_poids_moyen-method}
\alias{charge,Bilan_poids_moyen-method}
\title{charge method for Bilan_poids_moyen class}
\usage{
\S4method{charge}{Bilan_poids_moyen}(object)
}
\arguments{
\item{object}{An object of class \link{Bilan_poids_moyen-class}}
}
\value{
Bilan_poids_moyen with slots filled with user choice
}
\description{
charge method for Bilan_poids_moyen class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

