% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_stades_pigm.r
\docType{methods}
\name{charge,Bilan_stades_pigm-method}
\alias{charge,Bilan_stades_pigm-method}
\title{Charge : method used by the graphical interface to buid object from values
stored in envir_stacomi}
\usage{
\S4method{charge}{Bilan_stades_pigm}(object, h)
}
\value{
Bilan_stades_pigm with slots filled with user choice
}
\description{
Charge : method used by the graphical interface to buid object from values
stored in envir_stacomi
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

