% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_taille.r
\docType{methods}
\name{charge,Bilan_taille-method}
\alias{charge,Bilan_taille-method}
\title{charge method for Bilan_taille}
\usage{
\S4method{charge}{Bilan_taille}(object)
}
\arguments{
\item{object}{An object of class \code{\link{Bilan_taille-class}}}
}
\description{
charge method for Bilan_taille
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

