% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanAgedemer.r
\docType{methods}
\name{charge,BilanAgedemer-method}
\alias{charge,BilanAgedemer-method}
\alias{charge.BilanAgedemer}
\alias{charge.bilanagedemer}
\title{Loads data and check that all choices in the graphical interface have been made.}
\usage{
\S4method{charge}{BilanAgedemer}(object, h)
}
\arguments{
\item{object}{An object of class \link{BilanAgedemer-class}}

\item{h}{a handler}
}
\value{
An object of class \link{BilanAgedemer-class} with slots filled with user choice
}
\description{
It is not necessary to run this method if the choice_c method has been run.  
This method verifies that boxes have been clicked in the user interface and gets the objects pasted in 
envir_stacomi
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

