% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanAnnuels.r
\docType{methods}
\name{charge,BilanAnnuels-method}
\alias{charge,BilanAnnuels-method}
\title{charge method for BilanAnnuels class}
\usage{
\S4method{charge}{BilanAnnuels}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{BilanAnnuels-class}}

\item{silent}{Stops messages from being displayed if silent=TRUE, default FALSE}
}
\description{
Method used by the graphical interface to load data and check that all choices have
been made by the user
}
\keyword{internal}

