% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanArgentee.r
\docType{methods}
\name{charge,BilanArgentee-method}
\alias{charge,BilanArgentee-method}
\title{charge method for BilanArgentee class}
\usage{
\S4method{charge}{BilanArgentee}(object, h)
}
\arguments{
\item{object}{An object of class \link{BilanArgentee-class}}

\item{h}{a handler}
}
\value{
An object of class \link{BilanArgentee-class} with slots filled with user choice

An object of the class
}
\description{
this method verifies that boxes have been clicked in the user interface and gets the objects pasted in 
envir_stacomi
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

