% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanEspeces.r
\docType{methods}
\name{charge,BilanEspeces-method}
\alias{charge,BilanEspeces-method}
\alias{charge.BilanEspeces}
\alias{charge.bilanespeces}
\title{charge method for BilanEspeces}
\usage{
\S4method{charge}{BilanEspeces}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{BilanEspeces-class}}

\item{silent}{Stops displaying the messages.}
}
\value{
BilanEspeces with slots filled by user choice
}
\description{
Verifies the content of objects when the graphical interface is used, it is not necessary
to call the charge method if the choice_c method has been used
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

