% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanFonctionnementDF.r
\docType{methods}
\name{charge,BilanFonctionnementDF-method}
\alias{charge,BilanFonctionnementDF-method}
\title{charge method for BilanFonctionnementDF}
\usage{
\S4method{charge}{BilanFonctionnementDF}(object, h)
}
\value{
An object of class \ref{BilanFonctionnementDF}
}
\description{
used by the graphical interface to retreive the objects of Referential classes
assigned to envir_stacomi
}
\author{
cedric.briand
}

