% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationInterAnnuelle.r
\docType{methods}
\name{charge,BilanMigrationInterAnnuelle-method}
\alias{charge,BilanMigrationInterAnnuelle-method}
\title{loading method for BilanMigrationInterannuelle class}
\usage{
\S4method{charge}{BilanMigrationInterAnnuelle}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{BilanMigrationInterAnnuelle-class}}

\item{silent}{Boolean, if TRUE, information messages are not displayed}
}
\value{
An object of class  \link{BilanMigrationInterAnnuelle-class}
}
\description{
loading method for BilanMigrationInterannuelle class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
