% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMultConditionEnv.r
\docType{methods}
\name{charge,BilanMigrationMultConditionEnv-method}
\alias{charge,BilanMigrationMultConditionEnv-method}
\title{charge method for BilanMigrationMultConditionEnv class}
\usage{
\S4method{charge}{BilanMigrationMultConditionEnv}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{BilanMigrationMultConditionEnv-class}}

\item{silent}{Should the function remain silent (boolean)}
}
\description{
#' Unique the other Bilan classes where the charge method is only used by the graphical interface 
to collect and test objects in the environment envir_stacomi, and see if the right choices have
been made in the graphical interface, this methods runs the \link{charge,BilanMigrationMult-method}
and needs to be called from the command line (see examples)
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

