% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanOperation.r
\docType{methods}
\name{charge,BilanOperation-method}
\alias{charge,BilanOperation-method}
\title{charge method for BilanOperation}
\usage{
\S4method{charge}{BilanOperation}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{BilanOperation-class}}

\item{silent}{Keeps program silent}
}
\value{
An object of class \link{BilanOperation-class}
}
\description{
used by the graphical interface to retrieve Referential classes
assigned to envir_stacomi
}
\author{
cedric.briand
}
