\name{charge-methods}
\docType{methods}
\alias{charge-methods}
\alias{charge,Bilan_lot-method}
\alias{charge,Bilan_poids_moyen-method}
\alias{charge,Bilan_stades_pigm-method}
\alias{charge,Bilan_taille-method}
\alias{charge,BilanConditionEnv-method}
\alias{charge,BilanEspeces-method}
\alias{charge,BilanFonctionnementDC-method}
\alias{charge,BilanFonctionnementDF-method}
\alias{charge,BilanMigrationInterAnnuelle-method}
\alias{charge,RefAnnee-method}
\alias{charge,RefCheckBox-method}
\alias{charge,RefChoix-method}
\alias{charge,RefCoe-method}
\alias{charge,RefDC-method}
\alias{charge,RefDF-method}
\alias{charge,RefListe-method}
\alias{charge,Refpar-method}
\alias{charge,Refparqual-method}
\alias{charge,Refparquan-method}
\alias{charge,RefPoidsMoyenPeche-method}
\alias{charge,RefStades-method}
\alias{charge,RefStationMesure-method}
\alias{charge,RefTaxon-method}
\alias{charge,RefTextBox-method}
\title{Methods for Function charge}
\description{
 This methods loads the data, it is in fact sometimes redundant with connect which should be
 reserved for instances of RequeteODBC and daughter classes. See individual .r files for help and examples
}
\section{Methods}{
\describe{

\item{\code{signature(objet = "Bilan_lot")}}{ ~~describe this method here }

\item{\code{signature(objet = "Bilan_poids_moyen")}}{ ~~describe this method here }

\item{\code{signature(objet = "Bilan_stades_pigm")}}{ ~~describe this method here }

\item{\code{signature(objet = "Bilan_taille")}}{ ~~describe this method here }

\item{\code{signature(objet = "BilanConditionEnv")}}{ ~~describe this method here }

\item{\code{signature(objet = "BilanEspeces")}}{ ~~describe this method here }

\item{\code{signature(objet = "BilanFonctionnementDC")}}{ ~~describe this method here }

\item{\code{signature(objet = "BilanFonctionnementDF")}}{ ~~describe this method here }

\item{\code{signature(objet = "BilanMigrationInterAnnuelle")}}{ ~~describe this method here }

\item{\code{signature(objet = "RefAnnee")}}{ ~~describe this method here }

\item{\code{signature(objet = "RefCheckBox")}}{ ~~describe this method here }

\item{\code{signature(objet = "RefChoix")}}{ ~~describe this method here }

\item{\code{signature(objet = "RefCoe")}}{ ~~describe this method here }

\item{\code{signature(objet = "RefDC")}}{ ~~describe this method here }

\item{\code{signature(objet = "RefDF")}}{ ~~describe this method here }

\item{\code{signature(objet = "RefListe")}}{ ~~describe this method here }

\item{\code{signature(objet = "Refpar")}}{ ~~describe this method here }

\item{\code{signature(objet = "Refparqual")}}{ ~~describe this method here }

\item{\code{signature(objet = "Refparquan")}}{ ~~describe this method here }

\item{\code{signature(objet = "RefPoidsMoyenPeche")}}{ ~~describe this method here }

\item{\code{signature(objet = "RefStades")}}{ ~~describe this method here }

\item{\code{signature(objet = "RefStationMesure")}}{ ~~describe this method here }

\item{\code{signature(objet = "RefTaxon")}}{ ~~describe this method here }

\item{\code{signature(objet = "RefTextBox")}}{ ~~describe this method here }
}}
\keyword{methods}
