% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_checkbox.R
\docType{methods}
\name{charge,ref_checkbox-method}
\alias{charge,ref_checkbox-method}
\title{Loading method for ref_checkbox referential objects}
\usage{
\S4method{charge}{ref_checkbox}(object, title, labels, checked)
}
\arguments{
\item{object}{An object of class \link{ref_checkbox-class}}

\item{title}{Title of the frame}

\item{labels}{Labels for checked}

\item{checked}{Vector of boolean indicating if ref_checkbox are checked}
}
\value{
An object of class \link{ref_checkbox-class}
}
\description{
Loading method for ref_checkbox referential objects
}
\examples{
\dontrun{
 object=new("ref_checkbox")
charge(object,title="essai",labels=c("par1","par2","par3"),checked=c(TRUE,TRUE,TRUE))
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

