% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_dc.R
\name{charge,ref_dc-method}
\alias{charge,ref_dc-method}
\title{Method to load the counting devices of the control station}
\usage{
\S4method{charge}{ref_dc}(object)
}
\arguments{
\item{object}{An object of class \link{ref_dc-class}}
}
\value{
an object of class ref_dc with data loaded
}
\description{
Method to load the counting devices of the control station
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
