% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_list.R
\name{charge,ref_list-method}
\alias{charge,ref_list-method}
\alias{charge.ref_list}
\title{Loading method for ref_list referential objects}
\usage{
\S4method{charge}{ref_list}(object, listechoice, label)
}
\arguments{
\item{object}{An object of class \link{ref_list-class}}

\item{listechoice}{A character vector setting the possible values in which the user can select}

\item{label}{A label for refliste}
}
\value{
An S4 object of class \link{ref_list-class}

An S4 object of class \link{ref_list-class}
}
\description{
Loading method for ref_list referential objects
}
\examples{
\dontrun{
 object=new('ref_list')
charge(object)
}
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
