% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_par.R
\docType{methods}
\name{charge,ref_par-method}
\alias{charge,ref_par-method}
\title{Loading method for Repar referential objects}
\usage{
\S4method{charge}{ref_par}(object)
}
\arguments{
\item{object}{An object of class \link{ref_par-class}}
}
\value{
An S4 object of class ref_par
}
\description{
Loading method for Repar referential objects
}
\examples{
\dontrun{
 object=new("ref_par")
charge(object)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
