% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_parqual.R
\name{charge,ref_parqual-method}
\alias{charge,ref_parqual-method}
\title{Loading method for Reparqual referential objects}
\usage{
\S4method{charge}{ref_parqual}(object)
}
\arguments{
\item{object}{An object of class \link{ref_parqual-class}}
}
\value{
An S4 object of class ref_parqual
}
\description{
Loading method for Reparqual referential objects
}
\examples{
\dontrun{
 object=new('ref_parqual')
 charge(object)
}
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
