% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_parquan.R
\name{charge,ref_parquan-method}
\alias{charge,ref_parquan-method}
\title{Loading method for Reparquan referential objects}
\usage{
\S4method{charge}{ref_parquan}(object)
}
\arguments{
\item{object}{An object of class \link{ref_parquan-class}}
}
\value{
An S4 object of class \link{ref_parquan-class} with data loaded
}
\description{
Loading method for Reparquan referential objects
}
\examples{
\dontrun{
 object=new('ref_parquan')
 charge(object)
}
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
