% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_textbox.R
\name{charge,ref_textbox-method}
\alias{charge,ref_textbox-method}
\title{Loading method for ref_textbox referential objects}
\usage{
\S4method{charge}{ref_textbox}(object, title, label)
}
\arguments{
\item{object}{An object of class \link{ref_textbox-class}}

\item{title}{A title for the frame}

\item{label}{A label for the TextBox}
}
\value{
An S4 object of class \link{ref_textbox-class}
}
\description{
Loading method for ref_textbox referential objects
}
\examples{
\dontrun{
 object=new('ref_textbox')
charge(object,title='un titre',label='20')
}
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
