% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_year.R
\name{charge,ref_year-method}
\alias{charge,ref_year-method}
\title{Loading method for ref_year referential objects}
\usage{
\S4method{charge}{ref_year}(object, objectreport = "report_ge_weight")
}
\arguments{
\item{object}{An object of class \link{ref_year-class}}

\item{objectreport}{The object report, default \code{report_ge_weight} other possible value report_mig_interannual}
}
\value{
object An object of class \link{ref_year-class} with slot data filled with the available years for the corresponding report
}
\description{
Selects year available either in the bjo table if report_object==report_mig_interannual) or in the t_operation_ope table
}
\examples{
\dontrun{
object=new("ref_year")
charge(object)
 validObject(annee)
showMethods("charge")
}
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
