% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefCheckBox.r
\docType{methods}
\name{charge,RefCheckBox-method}
\alias{charge,RefCheckBox-method}
\title{Loading method for ReCheckBox referential objects}
\usage{
\S4method{charge}{RefCheckBox}(object, title, labels, checked)
}
\value{
An S4 object of class RefCheckBox
}
\description{
Loading method for ReCheckBox referential objects
}
\examples{
\dontrun{
 object=new("RefCheckBox")
charge(object,title="essai",labels=c("par1","par2","par3"),checked=c(TRUE,TRUE,TRUE))
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

