% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefChoix.r
\docType{methods}
\name{charge,RefChoix-method}
\alias{charge,RefChoix-method}
\title{Loading method for Rechoix referential objects}
\usage{
\S4method{charge}{RefChoix}(object, vecteur, label, selected)
}
\arguments{
\item{object}{An object of class RefChoix}

\item{vecteur}{A vector of name, see example code.}

\item{selected}{An integer indicating which object is selected at launch}
}
\value{
An S4 object of class RefChoix
}
\description{
Loading method for Rechoix referential objects
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
#' @examples 
 object=new("RefChoix")
charge(object,vecteur=c("oui","non"),label="essai",selected=as.integer(1))
}
\seealso{
Other Referential objects: \code{\link{RefAnnee-class}},
  \code{\link{RefCheckBox-class}},
  \code{\link{RefChoix-class}}, \code{\link{RefCoe-class}},
  \code{\link{RefDC-class}}, \code{\link{RefDF-class}},
  \code{\link{RefHorodate-class}},
  \code{\link{RefListe-class}},
  \code{\link{RefPoidsMoyenPeche}},
  \code{\link{RefStades-class}},
  \code{\link{Refpar-class}},
  \code{\link{Refparqual-class}},
  \code{\link{Refparquan-class}},
  \code{\link{Refperiode-class}},
  \code{\link{charge,RefMsg-method}}
}

