% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefCoe.r
\docType{methods}
\name{charge,RefCoe-method}
\alias{charge,RefCoe-method}
\title{loads the coefficients for the period defined in class}
\usage{
\S4method{charge}{RefCoe}(object)
}
\arguments{
\item{object}{An object of class \link{RefCoe-class}}
}
\value{
Object of class RefCoe
}
\description{
The slots datedebut and datefin have to be filled before using charge
}
\examples{
\dontrun{
object<- new("RefCoe")
object@datedebut<-strptime("01/01/1996",format="\%d/\%m/\%Y")
object@datefin<-strptime("01/01/1997",format="\%d/\%m/\%Y")
charge(object) 
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

