% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefDC.r
\docType{methods}
\name{charge,RefDC-method}
\alias{charge,RefDC-method}
\title{Method to load the counting devices of the control station}
\usage{
\S4method{charge}{RefDC}(object)
}
\arguments{
\item{object}{An object of class \link{RefDC-class}}
}
\value{
Object of class RefDC
}
\description{
Method to load the counting devices of the control station
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
