% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefDF.r
\docType{methods}
\name{charge,RefDF-method}
\alias{charge,RefDF-method}
\title{Loading method for DF referential objects}
\usage{
\S4method{charge}{RefDF}(object)
}
\value{
An object of class RefDF
}
\description{
Loading method for DF referential objects
}
\section{Slots}{

\describe{
\item{\code{object}}{An object of class \link{RefDF-class}}
}}
\examples{
\dontrun{
 object=new("RefDF")
charge(object)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

