% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefListe.r
\docType{methods}
\name{charge,RefListe-method}
\alias{charge,RefListe-method}
\title{Loading method for Refliste referential objects}
\usage{
\S4method{charge}{RefListe}(object, vecteur, label)
}
\value{
An object of class RefListe
}
\description{
Loading method for Refliste referential objects
}
\examples{
 object=new("RefListe")
charge(object)
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

