% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefMsg.r
\docType{methods}
\name{charge,RefMsg-method}
\alias{charge,RefMsg-method}
\title{Loading method for RefMsg referential objects}
\usage{
\S4method{charge}{RefMsg}(object)
}
\value{
An S4 object of class RefMsg
}
\description{
loads the common table ts_messager_msr
}
\examples{
 object=new("RefMsg")
 charge(object)
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other Referential objects: \code{\link{RefAnnee-class}},
  \code{\link{RefCheckBox-class}},
  \code{\link{RefChoix-class}}, \code{\link{RefCoe-class}},
  \code{\link{RefDC-class}}, \code{\link{RefDF-class}},
  \code{\link{RefHorodate-class}},
  \code{\link{RefListe-class}},
  \code{\link{RefPoidsMoyenPeche-class}},
  \code{\link{RefStades-class}},
  \code{\link{Refpar-class}},
  \code{\link{Refparqual-class}},
  \code{\link{Refparquan-class}},
  \code{\link{Refperiode-class}},
  \code{\link{charge,RefChoix-method}}
}

