% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refpar.r
\docType{methods}
\name{charge,Refpar-method}
\alias{charge,Refpar-method}
\title{Loading method for Repar referential objects}
\usage{
\S4method{charge}{Refpar}(object)
}
\value{
An S4 object of class Refpar
}
\description{
Loading method for Repar referential objects
}
\examples{
 object=new("Refpar")
charge(object)
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

