% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refparquan.r
\docType{methods}
\name{charge,Refparquan-method}
\alias{charge,Refparquan-method}
\title{Loading method for Reparquan referential objects}
\usage{
\S4method{charge}{Refparquan}(object)
}
\value{
An S4 object of class Refparquan
}
\description{
Loading method for Reparquan referential objects
}
\examples{
\dontrun{
 object=new("Refparquan")
 charge(object)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

