% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefPoidsMoyenPeche.r
\docType{methods}
\name{charge,RefPoidsMoyenPeche-method}
\alias{charge,RefPoidsMoyenPeche-method}
\title{Charge method for RefPoidsMoyenPeche}
\usage{
\S4method{charge}{RefPoidsMoyenPeche}(object)
}
\value{
An object of class \code{\linkS4class{RefPoidsMoyenPeche}}
}
\description{
Charge method for RefPoidsMoyenPeche
}
\author{
cedric.briand
}

