% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefStades.r
\docType{methods}
\name{charge,RefStades-method}
\alias{charge,RefStades-method}
\title{Loading method for RefStades referential objects}
\usage{
\S4method{charge}{RefStades}(object)
}
\value{
An S4 object of class RefStades
}
\description{
Loading method for RefStades referential objects
}
\examples{
\dontrun{
 object=new("RefStades")
 charge(object)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

