% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefStationMesure.r
\docType{methods}
\name{charge,RefStationMesure-method}
\alias{charge,RefStationMesure-method}
\title{Loading method for RefStationMesure referential object}
\usage{
\S4method{charge}{RefStationMesure}(object)
}
\arguments{
\item{object}{An object of class \link{RefStationMesure-class}}
}
\value{
An S4 object of class RefStationMesure
}
\description{
Loading method for RefStationMesure referential object
}
\examples{
\dontrun{
 object=new("RefStationMesure")
 charge(object)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

