% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefTaxon.r
\docType{methods}
\name{charge,RefTaxon-method}
\alias{charge,RefTaxon-method}
\title{Loading method for RefTaxon referential objects}
\usage{
\S4method{charge}{RefTaxon}(object)
}
\value{
An S4 object of class RefTaxon
}
\description{
Loading method for RefTaxon referential objects
}
\examples{
\dontrun {
 object=new("RefTaxon")
 charge(object)}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

