% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_annual.R
\name{charge,report_annual-method}
\alias{charge,report_annual-method}
\alias{charge.report_annual}
\title{charge method for report_annual class}
\usage{
\S4method{charge}{report_annual}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_annual-class}}

\item{silent}{Stops messages from being displayed if silent=TRUE, default FALSE}
}
\value{
object An object of class \link{report_annual-class} with data set from values assigned in \code{envir_stacomi} environment
}
\description{
Method used by the graphical interface to load data and check that all choices have
been made by the user
}
\keyword{internal}
