% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_dc.R
\name{charge,report_dc-method}
\alias{charge,report_dc-method}
\alias{charge.report_dc}
\title{charge method for report_dc}
\usage{
\S4method{charge}{report_dc}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_dc-class}}

\item{silent}{boolean, default FALSE, if TRUE messages are not displayed.}
}
\value{
object An object of class \link{report_dc-class} with data set from values assigned in \code{envir_stacomi} environment
}
\description{
used by the graphical interface to retrieve the objects of referential classes
assigned to envir_stacomi
}
\keyword{internal}
