% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_df.R
\name{charge,report_df-method}
\alias{charge,report_df-method}
\alias{charge.report_df}
\title{charge method for report_df}
\usage{
\S4method{charge}{report_df}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_df-class}}

\item{silent}{Keeps program silent}
}
\value{
An object of class \link{report_df-class}  with data set from values assigned in \code{envir_stacomi} environment
}
\description{
used by the graphical interface to retrieve the objects of referential classes
assigned to envir_stacomi
}
\keyword{internal}
