% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_env.R
\name{charge,report_env-method}
\alias{charge,report_env-method}
\alias{charge.report_env}
\title{charge method for report_env class}
\usage{
\S4method{charge}{report_env}(object, silent)
}
\arguments{
\item{object}{An object of class \link{report_env-class}}

\item{silent}{Default FALSE, if TRUE the program should no display messages}
}
\value{
An object of class \link{report_env-class} with data set from values assigned in \code{envir_stacomi} environment
}
\description{
charge method for report_env class
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
\keyword{internal}
