% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_ge_weight.R
\docType{methods}
\name{charge,report_ge_weight-method}
\alias{charge,report_ge_weight-method}
\alias{charge.report_ge_weight}
\title{charge method for report_ge_weight class}
\usage{
\S4method{charge}{report_ge_weight}(object)
}
\arguments{
\item{object}{An object of class \link{report_ge_weight-class}}
}
\value{
report_ge_weight with slots filled with user choice
}
\description{
charge method for report_ge_weight class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

