% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig.R
\name{charge,report_mig-method}
\alias{charge,report_mig-method}
\alias{charge.report_mig}
\title{Loads additional data on migration control operations, df (fishway) dc (counting device).}
\usage{
\S4method{charge}{report_mig}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{report_mig-class}}}

\item{silent}{Should the program be returning messages}
}
\value{
An object of class \link{report_mig-class} with slots filled from values assigned in \code{envir_stacomi} environment
}
\description{
this method creates additional classes in envir_stacomi for later use in plot (operations,
DF operation, DC operation). So unlike in most report classes where the charge method is only
used by the graphical interface, it is necessary to run charge for report_mig.
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
