% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_char.R
\name{charge,report_mig_char-method}
\alias{charge,report_mig_char-method}
\alias{charge.report_mig_char}
\title{charge method for report_mig_char}
\usage{
\S4method{charge}{report_mig_char}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_mig_char-class}}

\item{silent}{Default FALSE, if TRUE the program should not display messages}
}
\value{
\link{report_mig_char-class}  with slot filled from values assigned in \code{envir_stacomi} environment
}
\description{
Used by the graphical interface to collect and test objects in the environment envir_stacomi, 
fills also the data slot by the connect method. It is not necessary to run the charge method
if the choice is made from the command line using the choice_c method.
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
\keyword{internal}
