% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_env.R
\name{charge,report_mig_env-method}
\alias{charge,report_mig_env-method}
\alias{charge.report_mig_env}
\title{charge method for report_mig_env class}
\usage{
\S4method{charge}{report_mig_env}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_mig_env-class}}

\item{silent}{Should the function remain silent (boolean)}
}
\value{
An object of class \link{report_mig_env-class} with data set from values assigned in \code{envir_stacomi} environment
}
\description{
#' Unique the other report classes where the charge method is only used by the graphical interface
to collect and test objects in the environment envir_stacomi, and see if the right choices have
been made in the graphical interface, this methods runs the \link{charge,report_mig_mult-method}
and needs to be called from the command line (see examples)
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
