% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_interannual.R
\name{charge,report_mig_interannual-method}
\alias{charge,report_mig_interannual-method}
\alias{charge.report_mig_interannual}
\title{loading method for report_mig_interannual class}
\usage{
\S4method{charge}{report_mig_interannual}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_mig_interannual-class}}

\item{silent}{Boolean, if TRUE, information messages are not displayed}
}
\value{
An object of class  \link{report_mig_interannual-class} with slots set from values assigned in \code{envir_stacomi} environment
}
\description{
loading method for report_mig_interannual class
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
\keyword{internal}
