% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_ope.R
\name{charge,report_ope-method}
\alias{charge,report_ope-method}
\alias{charge.report_ope}
\title{charge method for report_ope}
\usage{
\S4method{charge}{report_ope}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_ope-class}}

\item{silent}{Keeps program silent}
}
\value{
An object of class \link{report_ope-class}  with slots filled from values assigned in \code{envir_stacomi} environment
}
\description{
used by the graphical interface to retrieve referential classes
assigned to envir_stacomi
}
\author{
cedric.briand
}
\keyword{internal}
