% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_sample_char.R
\name{charge,report_sample_char-method}
\alias{charge,report_sample_char-method}
\alias{charge.report_sample_char}
\title{charge method for report_sample_char class}
\usage{
\S4method{charge}{report_sample_char}(object)
}
\arguments{
\item{object}{An object of class \link{report_sample_char-class}}
}
\value{
An object of class \link{report_sample_char-class} with slots filled with user choice

An object of the class \link{report_sample_char-class} with slots filled from values assigned in \code{envir_stacomi} environment
}
\description{
this method verifies that boxes have been clicked in the user interface and gets the objects pasted in 
envir_stacomi
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
\keyword{internal}
