% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_silver_eel.R
\name{charge,report_silver_eel-method}
\alias{charge,report_silver_eel-method}
\alias{charge.report_silver_eel}
\title{charge method for report_silver_eel class}
\usage{
\S4method{charge}{report_silver_eel}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_silver_eel-class}}

\item{silent}{Boolean, if TRUE, information messages are not displayed}
}
\value{
An object of class \link{report_silver_eel-class}  with slots filled from values assigned in \code{envir_stacomi} environment

An object of the class
}
\description{
this method is used by the graphical interface load the user's choice and get the objects pasted in
envir_stacomi. It is not necessary to run this method when loading from the command line using the
choice_c method
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
\keyword{internal}
