% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_species.R
\name{charge,report_species-method}
\alias{charge,report_species-method}
\alias{charge.report_species}
\title{handler for calculation}
\usage{
\S4method{charge}{report_species}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_species-class}}

\item{silent}{Stops displaying the messages.}

\item{h}{a handler}

\item{...}{Additional parameters}
}
\value{
report_species with slots filled by user choice
}
\description{
internal use
}
\author{
Cedric Briand \email{cedric.briand'at'eptb-vilaine.fr}

Cedric Briand \email{cedric.briand'at'eptb-vilaine.fr}
}
\keyword{Verifies}
\keyword{been}
\keyword{call}
\keyword{charge}
\keyword{choice_c}
\keyword{content}
\keyword{for}
\keyword{graphical}
\keyword{has}
\keyword{if}
\keyword{interface}
\keyword{internal}
\keyword{is}
\keyword{it}
\keyword{method}
\keyword{necessary}
\keyword{not}
\keyword{objects}
\keyword{of}
\keyword{report_species}
\keyword{the}
\keyword{to}
\keyword{used}
\keyword{used,}
\keyword{when}
