% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generic.r
\name{charge}
\alias{charge}
\title{Generic method to load referentials}
\usage{
charge(object, ...)

charge(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Additional parms}

\item{object}{Object}

\item{...}{Additional parms}
}
\description{
Generic method to load referentials

Generic method to load referentials
}
\author{
cedric.briand

cedric.briand
}

