% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefMsg.r
\docType{methods}
\name{charge_avec_filtre,RefMsg-method}
\alias{charge_avec_filtre,RefMsg-method}
\title{Loading method for RefMsg referential objects}
\usage{
\S4method{charge_avec_filtre}{RefMsg}(object, lang)
}
\value{
An S4 object of class RefMsg
}
\description{
searching ref.ts_messagerlang_mrl for the lines corresponding to lang
}
\examples{
 object=new("RefMsg")
 charge_avec_filtre(object,lang='French')
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

