% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refpar.r
\docType{methods}
\name{charge_avec_filtre,Refpar-method}
\alias{charge_avec_filtre,Refpar-method}
\title{Loading method for Repar referential objects searching only those parameters existing for a DC, a Taxon, and a stade}
\usage{
\S4method{charge_avec_filtre}{Refpar}(object, dc_selectionne, taxon_selectionne,
  stade_selectionne)
}
\value{
An S4 object of class Refpar
}
\description{
Loading method for Repar referential objects searching only those parameters existing for a DC, a Taxon, and a stade
}
\examples{
 object=new("Refpar")
charge_avec_filtre(object,dc_selectionne=6,taxon_selectionne=2038,stade_selectionne="CIV")
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

