% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refparqual.r
\docType{methods}
\name{charge_avec_filtre,Refparqual-method}
\alias{charge_avec_filtre,Refparqual-method}
\title{Loading method for Reparqual referential objects searching only those parameters existing for a DC, a Taxon, and a stade}
\usage{
\S4method{charge_avec_filtre}{Refparqual}(object, dc_selectionne,
  taxon_selectionne, stade_selectionne)
}
\value{
An S4 object of class Refparqual
}
\description{
Loading method for Reparqual referential objects searching only those parameters existing for a DC, a Taxon, and a stade
}
\examples{
 dc_selectionne=6
taxon_selectionne=2038
 stade_selectionne="AGJ"
 object=new("Refparqual")
 charge_avec_filtre(object,dc_selectionne,taxon_selectionne,stade_selectionne)
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

