% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refparquan.r
\docType{methods}
\name{charge_avec_filtre,Refparquan-method}
\alias{charge_avec_filtre,Refparquan-method}
\title{Loading method for Reparquan referential objects searching only those parameters existing for a DC, a Taxon, and a stade}
\usage{
\S4method{charge_avec_filtre}{Refparquan}(object, dc_selectionne,
  taxon_selectionne, stade_selectionne)
}
\arguments{
\item{object}{An object of class \link{Refparquan-class}}

\item{dc_selectionne}{The dc set in the Bilan object}

\item{taxon_selectionne}{The taxa set in the Bilan object}

\item{stade_selectionne}{The stage set in the Bilan object}
}
\value{
An S4 object of class Refparqualn
}
\description{
Loading method for Reparquan referential objects searching only those parameters existing for a DC, a Taxon, and a stade
}
\examples{
\dontrun{
 dc_selectionne=6
taxon_selectionne=2038
 stade_selectionne="AGJ"
 object=new("Refparquan")
 charge_avec_filtre(object,dc_selectionne,taxon_selectionne,stade_selectionne)
}		
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
