% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefTaxon.r
\docType{methods}
\name{charge_avec_filtre,RefTaxon-method}
\alias{charge_avec_filtre,RefTaxon-method}
\title{Loading method for RefTaxon referential objects searching only those stages existing for a DC and a Taxon}
\usage{
\S4method{charge_avec_filtre}{RefTaxon}(object, dc_selectionne)
}
\value{
An S4 object of class RefTaxon
}
\description{
Loading method for RefTaxon referential objects searching only those stages existing for a DC and a Taxon
}
\examples{
\dontrun {
 dc_selectionne=6
 object=new("RefTaxon")
 charge_avec_filtre(object,dc_selectionne=dc_selectionne)}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

