% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generic.r
\name{charge_avec_filtre}
\alias{charge_avec_filtre}
\title{Generic method to load referentials, with filters from the parent object in the database}
\usage{
charge_avec_filtre(object, ...)

charge_avec_filtre(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Additional parms}

\item{object}{Object}

\item{...}{Additional parms}
}
\description{
Generic method to load referentials, with filters from the parent object in the database

Generic method to load referentials, with filters from the parent object in the database
}
\author{
cedric.briand

cedric.briand
}

