% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_parqual.R
\name{charge_complement,ref_parqual-method}
\alias{charge_complement,ref_parqual-method}
\title{Loads an additional dataset
this method is loaded to obtain the possible values of a qualitative parameter. Here data only contains one line}
\usage{
\S4method{charge_complement}{ref_parqual}(object)
}
\arguments{
\item{object}{An object of class \link{ref_parqual-class}}
}
\value{
An S4 object of class \link{ref_parqual-class} with the valqual slot filled
}
\description{
Loads an additional dataset
this method is loaded to obtain the possible values of a qualitative parameter. Here data only contains one line
}
\examples{
\dontrun{
 dc_selected=6
taxa_selected=2038
 stage_selected='AGJ'
 object=new('ref_parqual')
 object<-charge(object)
 charge_complement(object)
}  
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
