% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generic.R
\name{charge_with_filter}
\alias{charge_with_filter}
\title{Generic method to load referentials, with filters from the parent object in the database}
\usage{
charge_with_filter(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Additional parms}
}
\description{
Generic method to load referentials, with filters from the parent object in the database
}
\author{
cedric.briand
}
\keyword{internal}
