% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refparqual.r
\docType{methods}
\name{chargecomplement,Refparqual-method}
\alias{chargecomplement,Refparqual-method}
\title{method chargecomplement
cette methode est appellee apres la selection de l'object (data ne contient plus qu'une seule ligne)
et permet de lancer une requete pour obtenir un complement, ici les valeurs possibles d'un parametre qualitatif}
\usage{
\S4method{chargecomplement}{Refparqual}(object)
}
\value{
An S4 object of class Refparqual with the valqual slot filled
}
\description{
method chargecomplement
cette methode est appellee apres la selection de l'object (data ne contient plus qu'une seule ligne)
et permet de lancer une requete pour obtenir un complement, ici les valeurs possibles d'un parametre qualitatif
}
\examples{
 dc_selectionne=6
taxon_selectionne=2038
 stade_selectionne="AGJ"
 object=new("Refparqual")
 object<-charge(object)
 chargecomplement(object)		
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

