\name{chargexml}
\alias{chargexml}
\title{chargexml loads the informations stored in c:/program files/stacomi/calcmig...}
\description{chargexml loads the informations stored in c:/program files/stacomi/calcmig.xml file}
\value{a list with the datawd place and the baseODBC vector}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
