\name{chnames}
\alias{chnames}
\title{This function replaces the variable names in a data...}
\usage{chnames(objet, old_variable_name, new_variable_name)}
\description{This function replaces the variable names in a data.frame}
\value{objet}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{objet}{a data frame}
\item{old_variable_name}{}
\item{new_variable_name}{}}
