% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasdeTemps.r
\docType{methods}
\name{choice,PasDeTemps-method}
\alias{choice,PasDeTemps-method}
\title{Method to select timesteps from the graphical interface}
\usage{
\S4method{choice}{PasDeTemps}(object)
}
\arguments{
\item{object}{An object of class \link{PasDeTemps-class}}
}
\description{
Method to select timesteps from the graphical interface
}
