% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasDeTempsJournalier.r
\docType{methods}
\name{choice,PasDeTempsJournalier-method}
\alias{choice,PasDeTempsJournalier-method}
\title{Method to select timesteps from the graphical interface}
\usage{
\S4method{choice}{PasDeTempsJournalier}(object)
}
\arguments{
\item{object}{An object of class \link{PasDeTempsJournalier-class}}
}
\description{
Method to select timesteps from the graphical interface
}
\keyword{internal}

