% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_checkbox.R
\docType{methods}
\name{choice,ref_checkbox-method}
\alias{choice,ref_checkbox-method}
\title{Choice method for ref_checkbox referential objects}
\usage{
\S4method{choice}{ref_checkbox}(object)
}
\arguments{
\item{object}{An object of class \link{ref_checkbox-class}}
}
\description{
Choice method for ref_checkbox referential objects
}
\examples{
\dontrun{
object=new("ref_checkbox")
object<- charge(object,title="essai",labels=c("par1","par2","par3"),checked=c(TRUE,TRUE,TRUE))
win=gwindow(title="test ref_checkbox")
group=ggroup(container=win,horizontal=FALSE)
choice(object) 
dispose(win)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

