% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_choice.R
\docType{methods}
\name{choice,ref_choice-method}
\alias{choice,ref_choice-method}
\title{Choice method for Rechoice referential objects}
\usage{
\S4method{choice}{ref_choice}(object)
}
\arguments{
\item{object}{An object of class \link{ref_choice-class}}
}
\description{
Used by the graphical interface.
}
\examples{
\dontrun{
 object=new("ref_choice")
 object<-charge(object,vecteur=c("oui","non"),label="essai",selected=as.integer(1))
win=gwindow(title="test ref_choice")
group=ggroup(container=win,horizontal=FALSE)
choice(object) 
dispose(win)}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

