% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_df.R
\docType{methods}
\name{choice,ref_df-method}
\alias{choice,ref_df-method}
\title{Graphical method to choose a fishway through the interface}
\usage{
\S4method{choice}{ref_df}(object)
}
\arguments{
\item{object}{An object of class \link{ref_df-class}}
}
\description{
Graphical method to choose a fishway through the interface
}
\note{
the choice method assigns an object of class ref_df in the environment envir_stacomi
}
\examples{
 
 \dontrun{ 
win=gwindow()
group=ggroup(container=win,horizontal=FALSE)
object=new("ref_df")
object<-charge(object)
choice(object)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}
