% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_env.R
\docType{methods}
\name{choice,ref_env-method}
\alias{choice,ref_env-method}
\title{Choice method for ref_env referential object}
\usage{
\S4method{choice}{ref_env}(object, is.enabled = TRUE,
  title = gettext("Monitoring stations selection", domain =
  "R-stacomiR"))
}
\arguments{
\item{object}{An object of class \link{ref_env-class}}

\item{is.enabled}{A boolean parameter, if TRUE the frame is enabled when first displayed}

\item{title}{The title of the frame, defaut to "Monitoring stations selection"}
}
\description{
Choice method for ref_env referential object
}
\examples{
 
\dontrun{
object=new("ref_env")
win=gwindow()
group=ggroup(container=win,horizontal=FALSE)
object<-charge(object)
choice(object)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}
