% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_list.R
\docType{methods}
\name{choice,ref_list-method}
\alias{choice,ref_list-method}
\title{Choice method for ref_list referential objects}
\usage{
\S4method{choice}{ref_list}(object, is.enabled = TRUE)
}
\arguments{
\item{object}{An object of class \link{ref_list-class}}

\item{is.enabled}{A boolean indicating whether the frame is enabled when first displayed, default to TRUE}
}
\description{
Choice method for ref_list referential objects
}
\note{
the choice method assigns an object of class refList named ref_list in the environment envir_stacomi
}
\examples{
 
\dontrun{
 object=new("ref_list")
win=gwindow()
group=ggroup(container=win,horizontal=FALSE)
object<-charge(object,vecteur=c("choice1","choice2"),label="please choose")
choice(object)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

