% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_stage.R
\docType{methods}
\name{choice,ref_stage-method}
\alias{choice,ref_stage-method}
\title{Choice method for ref_stage referential objects}
\usage{
\S4method{choice}{ref_stage}(object, objectreport = NULL, is.enabled = TRUE)
}
\arguments{
\item{object}{An object of class \link{ref_stage-class}}

\item{objectreport}{An object report which includes the \link{ref_stage-class}, default NULL}

\item{is.enabled}{Sets if the frame is enabled at launch, defaut TRUE}
}
\description{
Choice method for ref_stage referential objects
}
\note{
the method tests if the load is called from within a "report" object, and loads par, parqual, or parquan objects accordingly
}
\examples{
\dontrun{
 object=new("ref_stage")
win=gwindow()
group=ggroup(container=win,horizontal=FALSE)
object<-charge(object)
report_migPar=new(report_migPar)
objectreport=report_taille # for another test
choice(object,objectreport=report_migPar)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

