% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_textbox.R
\docType{methods}
\name{choice,ref_textbox-method}
\alias{choice,ref_textbox-method}
\title{Choice method for ReTextBox referential objects}
\usage{
\S4method{choice}{ref_textbox}(object, nomassign = "ref_textbox")
}
\arguments{
\item{object}{An object of class \link{ref_textbox-class}}

\item{nomassign}{The name with which the object will be assigned in envir_stacomi}
}
\description{
Choice method for ReTextBox referential objects
}
\examples{
\dontrun{
object=new("ref_textbox")
object<- charge(object,title="le titre",label="20")
win=gwindow(title="test ref_textbox")
group=ggroup(container=win,horizontal=FALSE)
choice(object) 
dispose(win)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}
