% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_timestep_daily.R
\docType{methods}
\name{choice,ref_timestep_daily-method}
\alias{choice,ref_timestep_daily-method}
\title{Method to select timesteps from the graphical interface}
\usage{
\S4method{choice}{ref_timestep_daily}(object)
}
\arguments{
\item{object}{An object of class \link{ref_timestep_daily-class}}
}
\description{
Method to select timesteps from the graphical interface
}
\keyword{internal}

