% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefChoix.r
\docType{methods}
\name{choice,RefChoix-method}
\alias{choice,RefChoix-method}
\title{Choice method for Rechoice referential objects}
\usage{
\S4method{choice}{RefChoix}(object)
}
\arguments{
\item{object}{An object of class \link{RefChoix-class}}
}
\description{
Used by the graphical interface.
}
\examples{
\dontrun{
 object=new("RefChoix")
 object<-charge(object,vecteur=c("oui","non"),label="essai",selected=as.integer(1))
win=gwindow(title="test refChoix")
group=ggroup(container=win,horizontal=FALSE)
choice(object) 
dispose(win)}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

