% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefDF.r
\docType{methods}
\name{choice,RefDF-method}
\alias{choice,RefDF-method}
\title{Graphical method to choose a fishway through the interface}
\usage{
\S4method{choice}{RefDF}(object)
}
\arguments{
\item{object}{An object of class \link{RefDF-class}}
}
\description{
Graphical method to choose a fishway through the interface
}
\note{
the choice method assigns an object of class refDF in the environment envir_stacomi
}
\examples{
 
 \dontrun{ 
win=gwindow()
group=ggroup(container=win,horizontal=FALSE)
object=new("RefDF")
object<-charge(object)
choice(object)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
