% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefHorodate.r
\docType{methods}
\name{choice,RefHorodate-method}
\alias{choice,RefHorodate-method}
\title{Graphical interface}
\usage{
\S4method{choice}{RefHorodate}(object, label = "date",
  nomassign = "horodate",
  funoutlabel = "nous avons le choix dans la date\\n", decal = 0)
}
\arguments{
\item{object}{An object of class \link{RefHorodate-class}}

\item{label}{Label for the gframe}

\item{nomassign}{The name assigned in environment envir_stacomi}

\item{funoutlabel, }{text displayed by the interface}

\item{decal}{Default 0, number of years to shift forward or backward}
}
\value{
Selects the date in the graphical interface, and assigns an object of class POSIXt with name nomassign in envir_stacomi
}
\description{
Graphical interface
}
\keyword{internal}

