% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefListe.r
\docType{methods}
\name{choice,RefListe-method}
\alias{choice,RefListe-method}
\title{Choice method for RefListe referential objects}
\usage{
\S4method{choice}{RefListe}(object, is.enabled = TRUE)
}
\description{
Choice method for RefListe referential objects
}
\note{
the choice method assigns an object of class refList named refListe in the environment envir_stacomi
}
\examples{
 
 object=new("RefListe")
win=gwindow()
group=ggroup(container=win,horizontal=FALSE)
object<-charge(object)
choice(object)
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

