% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refpar.r
\docType{methods}
\name{choice,Refpar-method}
\alias{choice,Refpar-method}
\title{Choice method for Refpar referential objects}
\usage{
\S4method{choice}{Refpar}(object,
  label = "Choix d'une caracteristique de lot", nomassign = "refpar",
  frameassign = "frame_par", is.enabled = TRUE)
}
\description{
Choice method for Refpar referential objects
}
\note{
the choice method assigns an object of class Refpar named refpar in the environment envir_stacomi

this method choice is also on sons objects Refparquan, hence the parameters,however it was redefined in refparqual

to load the possible values of qualitative parameters
}
\examples{
 
 object=new("Refpar")
win=gwindow()
group=ggroup(container=win,horizontal=FALSE)
object<-charge(object)
choice(object)
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

