% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refparqual.r
\docType{methods}
\name{choice,Refparqual-method}
\alias{choice,Refparqual-method}
\title{Choice method for Refparqual referential objects, internal use}
\usage{
\S4method{choice}{Refparqual}(object,
  label = "Choix d'une caracteristique qualitative de lot",
  nomassign = "refpar", frameassign = "frame_par", is.enabled = TRUE)
}
\arguments{
\item{object}{An object of class \link{Refparqual-class}}

\item{label}{The label to set in the frame}

\item{nomassign}{The name used when assigning the object Refparqual to the \code{envir_stacomi} environment}

\item{frameassign}{The name of the gframe assigned to the user environment}

\item{is.enabled}{Is the frame enabled at launch.}
}
\description{
Choice method for Refparqual referential objects, internal use
}
\note{
the choice method assigns an object of class Refparqual named refparqual in the environment envir_stacomi
this method rewrites the method from Refpar, as it integrates a request of the possible values of qualitative parameters, hence the parameters,however it was redefined in refparqual
to load the possible values of qualitative parameters
}
\examples{
 
\dontrun{
 object=new("Refparqual")
win=gwindow()
group=ggroup(container=win,horizontal=FALSE)
object<-charge(object)
choice(object)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

