% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefStades.r
\docType{methods}
\name{choice,RefStades-method}
\alias{choice,RefStades-method}
\title{Choice method for RefStades referential objects}
\usage{
\S4method{choice}{RefStades}(object, objectBilan = NULL, is.enabled = TRUE)
}
\description{
Choice method for RefStades referential objects
}
\note{
the method tests if the load is called from within a "bilan" object, and loads par, parqual, or parquan objects accordingly
}
\examples{
 object=new("RefStades")
win=gwindow()
group=ggroup(container=win,horizontal=FALSE)
object<-charge(object)
bilanMigrationPar=new(BilanMigrationPar)
# objectBilan=bilan_taille # pour autre test
choice(object,objectBilan=bilanMigrationPar)
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

