% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefStationMesure.r
\docType{methods}
\name{choice,RefStationMesure-method}
\alias{choice,RefStationMesure-method}
\title{Choice method for RefStationMesure referential object}
\usage{
\S4method{choice}{RefStationMesure}(object, is.enabled = TRUE,
  title = get("msg", envir = envir_stacomi)$RefStationMesure.3)
}
\description{
Choice method for RefStationMesure referential object
}
\examples{
 
object=new("RefStationMesure")
win=gwindow()
group=ggroup(container=win,horizontal=FALSE)
object<-charge(object)
choice(object)
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

