% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefTaxon.r
\docType{methods}
\name{choice,RefTaxon-method}
\alias{choice,RefTaxon-method}
\title{Choice method for Reftaxon referential objects with only one taxa selected}
\usage{
\S4method{choice}{RefTaxon}(object, objectBilan = NULL, is.enabled = TRUE)
}
\description{
Choice method for Reftaxon referential objects with only one taxa selected
}
\examples{
 \dontrun{
 object=new("RefTaxon")
win=gwindow()
group<-ggroup(container=win,horizontal=FALSE)
object<-charge(object)
bilanMigration<-new(BilanMigration)
choice(object,objectBilan=bilanMigration)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

