% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_poids_moyen.r
\docType{methods}
\name{choice_c,Bilan_poids_moyen-method}
\alias{choice_c,Bilan_poids_moyen-method}
\title{command line interface for \link{Bilan_poids_moyen-class}}
\usage{
\S4method{choice_c}{Bilan_poids_moyen}(object, dc, anneedebut, anneefin,
  selectedvalue, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{Bilan_poids_moyen-class}}

\item{dc}{A numeric or integer, the code of the dc, coerced to integer,see \link{choice_c,RefDC-method}}

\item{anneedebut}{The starting the first year, passed as charcter or integer}

\item{anneefin}{the finishing year, must be > anneedebut (minimum one year in august to the next in august)}

\item{selectedvalue}{A character to select and object in the \link{RefListe-class}}

\item{silent}{Boolean, if TRUE, information messages are not displayed}
}
\value{
An object of class \link{Bilan_poids_moyen-class}
The choice_c method fills in the data slot for classes \link{RefDC-class} \link{Ref_year-class}
\link{RefCoe-class} \link{RefListe-class}
}
\description{
command line interface for \link{Bilan_poids_moyen-class}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

