% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanAnnuels.r
\docType{methods}
\name{choice_c,BilanAnnuels-method}
\alias{choice_c,BilanAnnuels-method}
\title{command line interface for \link{BilanAnnuels-class}}
\usage{
\S4method{choice_c}{BilanAnnuels}(object, dc, taxons, stades, anneedebut,
  anneefin, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{BilanAnnuels-class}}

\item{dc}{A numeric or integer, the code of the dc, coerced to integer,see \link{choice_c,RefDC-method}}

\item{taxons}{Either a species name in latin or the SANDRE code for species (ie 2038=Anguilla anguilla),
it should match the ref.tr_taxon_tax referential table in the stacomi database, see \link{choice_c,RefTaxon-method}}

\item{stades}{A stage code matching the ref.tr_stadedeveloppement_std table in the stacomi database, see \link{choice_c,RefStades-method}}

\item{anneedebut}{The starting the first year, passed as charcter or integer}

\item{anneefin}{the finishing year}

\item{silent}{Boolean, if TRUE, information messages are not displayed}
}
\value{
An object of class \link{BilanAnnuels-class}
The choice_c method fills in the data slot for classes \link{RefDC-class}, \link{RefTaxon-class}, \link{RefStades-class} and two slots of \link{Ref_year-class}
}
\description{
command line interface for \link{BilanAnnuels-class}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

