% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanConditionEnv.r
\docType{methods}
\name{choice_c,BilanConditionEnv-method}
\alias{choice_c,BilanConditionEnv-method}
\title{command line interface for BilanConditionEnv class}
\usage{
\S4method{choice_c}{BilanConditionEnv}(object, stationMesure, datedebut,
  datefin, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{BilanConditionEnv-class}}

\item{datedebut}{The starting date as a character, formats like \code{\%Y-\%m-\%d} or \code{\%d-\%m-\%Y} can be used as input}

\item{datefin}{The finishing date of the Bilan, for this class this will be used to calculate the number of daily steps.}

\item{silent}{Boolean default FALSE, if TRUE information messages not displayed.}

\item{stationmesure}{A character, the code of the monitoring station, which records environmental parameters \link{choice_c,RefStationMesure-method}}
}
\value{
An object of class \link{BilanConditionEnv-class}
The choice_c method fills in the data slot for RefStationMesure and  and then 
uses the choice_c methods of these object to select the data.
}
\description{
command line interface for BilanConditionEnv class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
