% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanEspeces.r
\docType{methods}
\name{choice_c,BilanEspeces-method}
\alias{choice_c,BilanEspeces-method}
\title{command line interface for \link{BilanEspeces-class}}
\usage{
\S4method{choice_c}{BilanEspeces}(object, dc, split = "none", anneedebut,
  anneefin, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{BilanEspeces-class}}

\item{dc}{A numeric or integer, the code of the dc, coerced to integer,see \link{choice_c,RefDC-method}}

\item{split}{one of c("none","week","month","year")}

\item{anneedebut}{The starting the first year, passed as character or integer}

\item{anneefin}{the finishing year}

\item{silent}{Boolean, if TRUE, information messages are not displayed}
}
\value{
An object of class \link{BilanEspeces-class}
}
\description{
command line interface for \link{BilanEspeces-class}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

