% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanFonctionnementDF.r
\docType{methods}
\name{choice_c,BilanFonctionnementDF-method}
\alias{choice_c,BilanFonctionnementDF-method}
\title{command line interface for BilanFonctionnementDF class}
\usage{
\S4method{choice_c}{BilanFonctionnementDF}(object, df, horodatedebut,
  horodatefin, ...)
}
\arguments{
\item{object}{An object of class \link{RefDC-class}}

\item{df}{The df to set}

\item{horodatedebut}{A POSIXt or Date or character to fix the date of beginning of the Bilan}

\item{horodatefin}{A POSIXt or Date or character to fix the last date of the Bilan}
}
\value{
An object of class \link{RefDC-class} with slots filled
}
\description{
The choice_c method fills in the data slot for RefDC, and then 
uses the choice_c methods of these object to "select" the data.
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

