% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigration.r
\docType{methods}
\name{choice_c,BilanMigration-method}
\alias{choice_c,BilanMigration-method}
\title{command line interface for BilanMigration class}
\usage{
\S4method{choice_c}{BilanMigration}(object, dc, taxons, stades, datedebut,
  datefin)
}
\arguments{
\item{object}{An object of class \link{BilanMigration-class}}

\item{dc}{A numeric or integer, the code of the dc, coerced to integer,see \link{choice_c,RefDC-method}}

\item{taxons}{Either a species name in latin or the SANDRE code for species (ie 2038=Anguilla anguilla),
these should match the ref.tr_taxon_tax referential table in the stacomi database, see \link{choice_c,RefTaxon-method}}

\item{stades}{A stage code matching the ref.tr_stadedeveloppement_std table in the stacomi database see \link{choice_c,RefStades-method}}

\item{datedebut}{The starting date as a character, formats like \code{\%Y-\%m-\%d} or \code{\%d-\%m-\%Y} can be used as input}

\item{datefin}{The finishing date of the Bilan, for this class this will be used to calculate the number of daily steps.
The choice_c method fills in the data slot for RefDC, RefTaxon, RefStades, and RefPasDeTempsJournalier and then 
uses the choice_c methods of these object to select the data.}
}
\description{
command line interface for BilanMigration class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

