% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationCar.r
\docType{methods}
\name{choice_c,BilanMigrationCar-method}
\alias{choice_c,BilanMigrationCar-method}
\title{command line interface for BilanMigrationCar class}
\usage{
\S4method{choice_c}{BilanMigrationCar}(object, dc, taxons, stades,
  parquan = NULL, parqual = NULL, horodatedebut, horodatefin,
  echantillon = TRUE, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{BilanMigrationCar-class}}

\item{dc}{A numeric or integer, the code of the dc, coerced to integer,see \link{choice_c,RefDC-method}}

\item{taxons}{'2220=Salmo salar',
these should match the ref.tr_taxon_tax referential table in the stacomi database, see \link{choice_c,RefTaxon-method}}

\item{stades}{TODO}

\item{horodatedebut}{The starting date as a character, formats like \code{\%Y-\%m-\%d} or \code{\%d-\%m-\%Y} can be used as input}

\item{horodatefin}{The finishing date of the Bilan, for this class this will be used to calculate the number of daily steps.}

\item{echantillon}{Default TRUE,}

\item{silent}{Default FALSE, if TRUE the program should no display messages}

\item{car}{Sample TODO}
}
\value{
An object of class \link{BilanAgedemer-class}
The choice_c method fills in the data slot for classes \link{RefDC-class}, \link{RefTaxon-class}, \link{RefStades-class}, \link{Refpar-class} and two slots of \link{RefHorodate-class} and then 
uses the choice_c methods of these object to select the data.
}
\description{
command line interface for BilanMigrationCar class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
