% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasDeTempsJournalier.r
\docType{methods}
\name{choice_c,PasDeTempsJournalier-method}
\alias{choice_c,PasDeTempsJournalier-method}
\title{choice_c method for class PasDeTempsJournalier}
\usage{
\S4method{choice_c}{PasDeTempsJournalier}(object, datedebut, datefin)
}
\arguments{
\item{datedebut}{a character (format \code{"15/01/1996"} or \code{"1996-01-15"} or \code{"15-01-1996"}), or POSIXct object}

\item{datefin}{a character of POSIXct object}
}
\description{
the choice_c method is intented to have the same behaviour as choice (which creates a
widget in the graphical interface) but from the command line.
}
\examples{
\dontrun{
 object=new("RefDC")
 object<-charge(object)
 choice_c(object=object,datedebut="2012-01-01",datefin="2013-01-01")
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

