% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_horodate.R
\name{choice_c,ref_horodate-method}
\alias{choice_c,ref_horodate-method}
\title{Graphical interface}
\usage{
\S4method{choice_c}{ref_horodate}(
  object,
  nomassign = "horodate",
  funoutlabel = "nous avons le choix dans la date\\n",
  horodate,
  silent = FALSE
)
}
\arguments{
\item{object}{An object of class \link{ref_horodate-class}}

\item{nomassign}{The name assigned in environment envir_stacomi}

\item{funoutlabel, }{text displayed by the interface}

\item{horodate}{The horodate to set, formats "\%d/\%m/\%Y \%H:\%M:\%s", "\%d/\%m/\%y \%H:\%M:\%s", "\%Y-\%m-\%d  \%H:\%M:\%s" formats
can also be passed with the date set to the minute \%d/\%m/\%Y \%H:\%M or the day  \%d/\%m/\%Y
\dots are accepted. The choice_c method assigns and}

\item{silent}{Default FALSE, should messages be displayed}

\item{label}{Label for the gframe}

\item{decal}{Default 0, number of years to shift forward or backward}
}
\value{
Selects the date in the graphical interface, and assigns an object of class POSIXt with name nomassign in envir_stacomi
Command line

An object of class \link{ref_horodate-class} with slot \emph{horodate} set,
 and assigns an object of class POSIXt with name nomassign in envir_stacomi
}
\description{
Graphical interface
}
\keyword{internal}
