% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_stage.R
\docType{methods}
\name{choice_c,ref_stage-method}
\alias{choice_c,ref_stage-method}
\title{choice_c method for ref_stage}
\usage{
\S4method{choice_c}{ref_stage}(object, stage, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{ref_stage-class}}

\item{stage}{the vector of stages chosen}

\item{silent}{Boolean, if TRUE, information messages are not displayed}
}
\description{
the choice_c method is intendedto have the same behaviour as choice (which creates a
widget in the graphical interface) but from the command line. The values passed to the choice_c method
for stage is the code.  Any numeric value will be discarded
}
\examples{
\dontrun{
object=new("ref_taxa")
object<-charge(object)
objectreport=new("report_mig_mult")
choice_c(object=object,objectreport=objectreport,"Anguilla anguilla")
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
