% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_taxa.R
\docType{methods}
\name{choice_c,ref_taxa-method}
\alias{choice_c,ref_taxa-method}
\title{choice_c method for ref_taxa}
\usage{
\S4method{choice_c}{ref_taxa}(object, taxa)
}
\arguments{
\item{object}{An object from the class ref_taxa}

\item{taxa}{The vector of taxa, can be either code (numeric) or latin name}
}
\description{
the choice_cc method is intendedto have the same behaviour as choice (which creates a
widget in the graphical interface) but from the command line. The values passed to the choice_c method
for taxa can be either numeric (2038 = Anguilla anguilla) or character.
}
\examples{
\dontrun{
object=new("ref_taxa")
object<-charge(object)
objectreport=new("report_mig_mult")
choice_c(object=object,objectreport=objectreport,"Anguilla anguilla")
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

