% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_timestep_daily.R
\name{choice_c,ref_timestep_daily-method}
\alias{choice_c,ref_timestep_daily-method}
\title{choice_c method for class ref_timestep_daily}
\usage{
\S4method{choice_c}{ref_timestep_daily}(object, datedebut, datefin)
}
\arguments{
\item{object}{An object of class \link{ref_timestep_daily-class}}

\item{datedebut}{A character (format \code{'15/01/1996'} or \code{'1996-01-15'} or \code{'15-01-1996'}), or POSIXct object}

\item{datefin}{A character}
}
\value{
An S4 object of class \link{ref_timestep_daily-class} with date selected
}
\description{
the choice_c method is intended to have the same behaviour as choice (which creates a
widget in the graphical interface) but from the command line.
}
\examples{
\dontrun{
 object=new('ref_dc')
 object<-charge(object)
 choice_c(object=object,datedebut='2012-01-01',datefin='2013-01-01')
}
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
