% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefChoix.r
\docType{methods}
\name{choice_c,RefChoix-method}
\alias{choice_c,RefChoix-method}
\title{Choice_c method for Refchoix referential objects}
\usage{
\S4method{choice_c}{RefChoix}(object, selectedvalue)
}
\arguments{
\item{object}{An object of class \link{RefListe-class}}
}
\description{
Choice_c method for Refchoix referential objects
}
\note{
the choice method assigns an object of class refList named refListe in the environment envir_stacomi
}
\examples{
 
\dontrun{
object=new("RefListe")
object<-charge(object,vecteur=c("1","2"),label="please choose")
object<-choice_c(object)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
