% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefDC.r
\docType{methods}
\name{choice_c,RefDC-method}
\alias{choice_c,RefDC-method}
\title{Command line interface to select a counting device}
\usage{
\S4method{choice_c}{RefDC}(object, dc)
}
\arguments{
\item{object}{an object of class RefDC}

\item{dc}{a character vector of dc chosen}
}
\description{
the choice_c method is intented to have the same behaviour as choice (which creates a
widget in the graphical interface) but from the command line.  The parameters for dc are transformed to integer as the RefDC only 
takes integer in the dc slots. The method also loads the stations and ouvrages (dams) associated with the counting device (dc).
The values passed to the choice_c method are then checked with the setValidty method.
Finally, if an objectBilan is passed as a parameter, the method will do a charge_avec_filtre to select only the taxa present in the counting devices
}
\examples{
\dontrun{
win=gwindow()
group=ggroup(container=win,horizontal=FALSE)
object=new("RefDC")
object<-charge(object)
objectBilan=new("BilanMigrationMult")
choice_c(object=object,objectBilan=objectBilan,dc=1)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
