% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefDF.r
\docType{methods}
\name{choice_c,RefDF-method}
\alias{choice_c,RefDF-method}
\title{Command line interface to choose a fishway}
\usage{
\S4method{choice_c}{RefDF}(object, df)
}
\arguments{
\item{object}{an object of class \link{RefDF-class}}

\item{df}{a character vector of df chosen}
}
\description{
the choice_c method is intented to have the same behaviour as choice (which creates a
widget in the graphical interface) but from the command line.  The parameters for dF are transformed to integer as the RefDF only 
takes integer in the df slots. 
DF are third in hierarchy in the stacomi database Station>ouvrage>DF>DC>operation. This class is only used in the
BilanFonctionnementDF class.
}
\examples{
\dontrun{
win=gwindow()
group=ggroup(container=win,horizontal=FALSE)
object=new("RefDF")
object<-charge(object)
objectBilan=new("BilanMigrationMult")
choice_c(object=object,objectBilan=objectBilan,dc=1)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

