% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefHorodate.r
\docType{methods}
\name{choice_c,RefHorodate-method}
\alias{choice_c,RefHorodate-method}
\title{Command line}
\usage{
\S4method{choice_c}{RefHorodate}(object, nomassign = "horodate",
  funoutlabel = "nous avons le choix dans la date\\n", decal = 0, horodate)
}
\arguments{
\item{object}{An object of class \link{RefHorodate-class}}

\item{funoutlabel, }{text displayed by the interface}

\item{horodate}{The horodate to set, formats "%d/%m/%Y %H:%M:%s", "%d/%m/%y %H:%M:%s", "%Y-%m-%d  %H:%M:%s"
\dots are accepted}

\item{label}{Label for the gframe}

\item{nomassing}{The name assigned in environment envir_stacomi}

\item{decale}{Default 0, number of years to shift forward or backward, the date will be set to the first day of the year}

\item{affichecal}{Default TRUE, should the calendar be displayed}
}
\value{
An object of class \link{RefHorodate-class} with slot \emph{horodate} set
}
\description{
Command line
}

