% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refpar.r
\docType{methods}
\name{choice_c,Refpar-method}
\alias{choice_c,Refpar-method}
\title{Command line interface to select a parameter}
\usage{
\S4method{choice_c}{Refpar}(object, par, silent = FALSE)
}
\arguments{
\item{object}{an object of class  \link{Refpar-class}}

\item{par}{A character vector of par}

\item{silent}{Default FALSE but not used there}
}
\value{
An object of class \link{Refpar-class}
}
\description{
the choice_c method is intented to have the same behaviour as choice (which creates a
widget in the graphical interface) but from the command line. 
If an objectBilan is passed as a parameter, the method will do a charge_avec_filtre to select only the taxa present in the counting devices
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
