% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefStades.r
\docType{methods}
\name{choice_c,RefStades-method}
\alias{choice_c,RefStades-method}
\title{choice_c method for RefStades}
\usage{
\S4method{choice_c}{RefStades}(object, stades)
}
\arguments{
\item{stades}{the vector of stages chosen}
}
\description{
the choice_c method is intented to have the same behaviour as choice (which creates a
widget in the graphical interface) but from the command line. The values passed to the choice_c method
for stades is the code.  Any numeric value will be discarded
}
\examples{
object=new("RefTaxon")
object<-charge(object)
objectBilan=new("BilanMigrationMult")
choice_c(object=object,objectBilan=objectBilan,"Anguilla anguilla")
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

