% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefStationMesure.r
\docType{methods}
\name{choice_c,RefStationMesure-method}
\alias{choice_c,RefStationMesure-method}
\title{Command line interface to select a monitoring  station}
\usage{
\S4method{choice_c}{RefStationMesure}(object, stationMesure)
}
\arguments{
\item{object}{an object of class RefStationMesure}

\item{stationmesure}{a character vector of the monitoring station code (corresponds to stm_libelle in the tj_stationmesure_stm table)}
}
\description{
the choice_c method is intented to have the same behaviour as choice (which creates a
widget in the graphical interface) but from the command line.
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
