% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reftextBox.r, R/ReftextBox.r
\docType{methods}
\name{choice_c,RefTextBox-method}
\alias{choice_c,RefTextBox-method}
\title{Choice_c method for ReTextBox referential objects}
\usage{
\S4method{choice_c}{RefTextBox}(object, value)

\S4method{choice_c}{RefTextBox}(object, value)
}
\arguments{
\item{object}{An object of class \link{RefTextBox-class}}

\item{value}{The value to set}

\item{object}{An object of class \link{RefTextBox-class}}

\item{value}{The value to set}
}
\description{
Choice_c method for ReTextBox referential objects

Choice_c method for ReTextBox referential objects
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}

Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

