% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_annual.R
\docType{methods}
\name{choice_c,report_annual-method}
\alias{choice_c,report_annual-method}
\alias{choice_c.report_annual}
\title{command line interface for \link{report_annual-class}}
\usage{
\S4method{choice_c}{report_annual}(object, dc, taxa, stage, anneedebut,
  anneefin, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_annual-class}}

\item{dc}{A numeric or integer, the code of the dc, coerced to integer,see \link{choice_c,ref_dc-method}}

\item{taxa}{Either a species name in latin or the SANDRE code for species (ie 2038=Anguilla anguilla),
it should match the ref.tr_taxon_tax referential table in the stacomi database, see \link{choice_c,ref_taxa-method}}

\item{stage}{A stage code matching the ref.tr_stadedeveloppement_std table in the stacomi database, see \link{choice_c,ref_stage-method}}

\item{anneedebut}{The starting the first year, passed as charcter or integer}

\item{anneefin}{the finishing year}

\item{silent}{Boolean, if TRUE, information messages are not displayed}
}
\value{
An object of class \link{report_annual-class}
The choice_c method fills in the data slot for classes \link{ref_dc-class}, \link{ref_taxa-class}, \link{ref_stage-class} and two slots of \link{ref_year-class}
}
\description{
command line interface for \link{report_annual-class}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
