% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_df.R
\name{choice_c,report_df-method}
\alias{choice_c,report_df-method}
\alias{choice_c.report_df}
\title{command line interface for report_df class}
\usage{
\S4method{choice_c}{report_df}(object, df, horodatedebut, horodatefin, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{ref_df-class}}

\item{df}{The df to set}

\item{horodatedebut}{A POSIXt or Date or character to fix the date of beginning of the report}

\item{horodatefin}{A POSIXt or Date or character to fix the last date of the report}

\item{silent}{Should program be silent or display messages}
}
\value{
An object of class \link{ref_df-class}  with data selected
}
\description{
The choice_c method fills in the data slot for ref_df, and then
uses the choice_c methods of these object to "select" the data.
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
