% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_env.R
\name{choice_c,report_env-method}
\alias{choice_c,report_env-method}
\alias{choice_c.report_env}
\title{command line interface for report_env class}
\usage{
\S4method{choice_c}{report_env}(object, stationMesure, datedebut, datefin, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_env-class}}

\item{stationMesure}{A character, the code of the monitoring station, which records environmental parameters \link{choice_c,ref_env-method}}

\item{datedebut}{The starting date as a character, formats like \code{\%Y-\%m-\%d} or \code{\%d-\%m-\%Y} can be used as input}

\item{datefin}{The finishing date of the report, for this class this will be used to calculate the number of daily steps.}

\item{silent}{Boolean default FALSE, if TRUE information messages not displayed.}
}
\value{
An object of class \link{report_env-class} with data selected
}
\description{
The choice_c method fills in the data slot for \link{ref_env-class} by runnning the charge method of this object.
It then runs the choice method on this object. It also applies the choice method for objects of class \link{ref_horodate-class}
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
