% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_char.R
\name{choice_c,report_mig_char-method}
\alias{choice_c,report_mig_char-method}
\alias{choice_c.report_mig_char}
\title{command line interface for report_mig_char class}
\usage{
\S4method{choice_c}{report_mig_char}(
  object,
  dc,
  taxa,
  stage,
  parquan = NULL,
  parqual = NULL,
  horodatedebut,
  horodatefin,
  echantillon = c("with", "without"),
  silent = FALSE
)
}
\arguments{
\item{object}{An object of class \link{report_mig_char-class}}

\item{dc}{A numeric or integer, the code of the dc, coerced to integer,see \link{choice_c,ref_dc-method}}

\item{taxa}{'2220=Salmo salar', can be a vector with several values
these should match the ref.tr_taxon_tax referential table in the stacomi database, see \link{choice_c,ref_taxa-method}}

\item{stage}{The stages selected, can be a vector with several values}

\item{parquan}{Quantitative parameter}

\item{parqual}{Qualitative parameter}

\item{horodatedebut}{The starting date as a character, formats like \code{\%Y-\%m-\%d} or \code{\%d-\%m-\%Y} can be used as input}

\item{horodatefin}{The finishing date of the report, for this class this will be used to calculate the number of daily steps}

\item{echantillon}{'with' can be 'without', checking without modifies the query
in the connect method so that subsamples are not allowed}

\item{silent}{Default FALSE, if TRUE the program should no display messages}
}
\value{
An object of class \link{report_sea_age-class}
The choice_c method fills in the data slot for classes \link{ref_dc-class}, \link{ref_taxa-class}, \link{ref_stage-class}, \link{ref_par-class} and two slots of \link{ref_horodate-class} and then 
uses the choice_c methods of these object to select the data.
}
\description{
command line interface for report_mig_char class
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
