% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_interannual.R
\name{choice_c,report_mig_interannual-method}
\alias{choice_c,report_mig_interannual-method}
\alias{choice_c.report_mig_interannual}
\title{command line interface for report_mig_interannual class}
\usage{
\S4method{choice_c}{report_mig_interannual}(object, dc, taxa, stage, start_year, end_year, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_mig_interannual-class}}

\item{dc}{A numeric or integer, the code of the dc, coerced to integer,see \link{choice_c,ref_dc-method}}

\item{taxa}{Either a species name in latin or the SANDRE code for species (ie 2038=Anguilla anguilla),
it should match the ref.tr_taxon_tax referential table in the stacomi database, see \link{choice_c,ref_taxa-method}}

\item{stage}{A stage code matching the ref.tr_stadedeveloppement_std table in the stacomi database, see \link{choice_c,ref_stage-method}}

\item{start_year}{The starting the first year, passed as character or integer}

\item{end_year}{the finishing year}

\item{silent}{Boolean, if TRUE, information messages are not displayed}
}
\value{
An object of class \link{report_mig_interannual-class} with data selected
The choice_c method fills in the data slot for classes \link{ref_dc-class}, \link{ref_taxa-class}, \link{ref_stage-class} and two slots of \link{ref_year-class}
}
\description{
command line interface for report_mig_interannual class
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
