% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_mult.R
\name{choice_c,report_mig_mult-method}
\alias{choice_c,report_mig_mult-method}
\alias{choice_c.report_mig_mult}
\title{command line used to build report_mig_mult class}
\usage{
\S4method{choice_c}{report_mig_mult}(object, dc, taxa, stage, datedebut, datefin, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_mig-class}}

\item{dc}{A numeric or integer, the code of the dc, coerced to integer,see \link{choice_c,ref_dc-method}}

\item{taxa}{Either a species name in latin or the SANDRE code for species (ie 2038=Anguilla anguilla),
these should match the ref.tr_taxon_tax referential table in the stacomi database, see \link{choice_c,ref_taxa-method}}

\item{stage}{A stage code matching the ref.tr_stadedeveloppement_std table in the stacomi database see \link{choice_c,ref_stage-method}}

\item{datedebut}{The starting date as a character, formats like \code{\%Y-\%m-\%d} or \code{\%d-\%m-\%Y} can be used as input}

\item{datefin}{The finishing date of the report, for this class this will be used to calculate the number of daily steps.}

\item{silent}{Should messages be hided default FALSE}
}
\value{
An object of class \link{report_mig_mult-class} with data selected
}
\description{
The choice_c method fills in the data slot for ref_dc, ref_taxa, ref_stage and then 
uses the choice_c methods of these object to 'select' the data.
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
