% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_silver_eel.R
\name{choice_c,report_silver_eel-method}
\alias{choice_c,report_silver_eel-method}
\alias{choice_c.report_silver_eel}
\title{command line interface for report_silver_eel class}
\usage{
\S4method{choice_c}{report_silver_eel}(
  object,
  dc,
  taxa = 2038,
  stage = "AGG",
  par = c("1786", "CCCC", "BBBB", "CONT", "LINP", "A111", "PECT"),
  horodatedebut,
  horodatefin,
  silent = FALSE
)
}
\arguments{
\item{object}{An object of class \link{report_silver_eel-class}}

\item{dc}{A numeric or integer, the code of the dc, coerced to integer,see \link{choice_c,ref_dc-method}}

\item{taxa}{'2038=Anguilla anguilla',
these should match the ref.tr_taxon_tax referential table in the stacomi database, see \link{choice_c,ref_taxa-method}}

\item{stage}{'AGG'}

\item{par}{Parameters chosen for the report are body size (1786), vertical eye diameter (BBBB), horizontal eye diameter (CCCC),
body contrast (CONT), presence of punctuation on the lateral line (LINP), length of the pectoral fin (PECT)}

\item{horodatedebut}{The starting date as a character, formats like \code{\%Y-\%m-\%d} or \code{\%d-\%m-\%Y} can be used as input}

\item{horodatefin}{The finishing date of the report, for this class this will be used to calculate the number of daily steps.}

\item{silent}{Boolean, if TRUE, information messages are not displayed}
}
\value{
An object of class \link{report_mig-class}
}
\description{
#' The choice_c method fills in the data slot for classes \link{ref_dc-class}, \link{ref_taxa-class}, \link{ref_stage-class}, \link{ref_par-class} and two slots of \link{ref_horodate-class} and then
uses the choice_c methods of these object to select the data.
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
