% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasdeTemps.r
\docType{methods}
\name{choicemult,PasDeTemps-method}
\alias{choicemult,PasDeTemps-method}
\title{choice method for PasdeTemps
this method differs from choice as it is called within a notebook,
it does not allow for multiple choice to be made}
\usage{
\S4method{choicemult}{PasDeTemps}(object)
}
\description{
choice method for PasdeTemps
this method differs from choice as it is called within a notebook,
it does not allow for multiple choice to be made
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

