% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_choice.R
\docType{methods}
\name{choicemult,ref_choice-method}
\alias{choicemult,ref_choice-method}
\title{Multiple Choice method for ref_choice referential objects, to put together with notebook widgets}
\usage{
\S4method{choicemult}{ref_choice}(object, selected_value)
}
\arguments{
\item{object}{An object of class \link{ref_choice-class}}

\item{selected_value}{the value selected in the combo}
}
\description{
Multiple Choice method for ref_choice referential objects, to put together with notebook widgets
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

