% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_horodate.R
\docType{methods}
\name{choicemult,ref_horodate-method}
\alias{choicemult,ref_horodate-method}
\title{Multiple Choice method for ref_horodate referential objects, to put together with notebook widgets}
\usage{
\S4method{choicemult}{ref_horodate}(object, label = "date",
  nomassign = "horodate",
  funoutlabel = "nous avons le choix dans la date\\n", decal = 0)
}
\arguments{
\item{object}{An object of class \link{ref_horodate-class}}

\item{label}{the name to write in the frame}

\item{nomassign}{the name with which the frame will be assigned to envir_stacomi}

\item{funoutlabel}{the sentence to write when the choice has been made}

\item{decal}{Default year will be current year, use -1 to set the default value in the interface to the year before}
}
\description{
Multiple Choice method for ref_horodate referential objects, to put together with notebook widgets
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}
