% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_par.R
\docType{methods}
\name{choicemult,ref_par-method}
\alias{choicemult,ref_par-method}
\title{Multiple Choice method for ref_par referential objects, internal use}
\usage{
\S4method{choicemult}{ref_par}(object, objectreport = NULL,
  label = gettext("Sample characteristic", domain = "R-stacomiR"),
  nomassign = "ref_par")
}
\arguments{
\item{object}{An object of class \link{ref_par-class}}

\item{objectreport}{An object report which includes the \link{ref_par-class}, default NULL}

\item{label}{The name for the frame}

\item{nomassign}{The name by which the widget will be assigned in envir_stacomi}
}
\description{
Multiple Choice method for ref_par referential objects, internal use
}
\examples{
\dontrun{
 object=new("ref_par")
win=gwindow()
group=ggroup(container=win,horizontal=FALSE)
object<-charge(object)
report_mig_char=new("report_mig_char")
objectreport=report_taille # for other test
choicemult(object,objectreport=report_mig_char)	
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

