% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_parqual.R
\docType{methods}
\name{choicemult,ref_parqual-method}
\alias{choicemult,ref_parqual-method}
\title{Multiple Choice method for ref_parqual referential objects internal use}
\usage{
\S4method{choicemult}{ref_parqual}(object, objectreport = NULL,
  nomassign = "ref_parqual", label = gettext("Qualitative", domain =
  "R-stacomiR"))
}
\arguments{
\item{object}{An object of class \link{ref_parqual-class}}

\item{objectreport}{An object report which includes the \link{ref_parqual-class}, default NULL}

\item{nomassign}{The name used when assigning the object ref_parqual to the \code{envir_stacomi} environment}

\item{label}{The name of the frame}
}
\description{
Multiple Choice method for ref_parqual referential objects internal use
}
\note{
this methods rewrites that of the ref_par as it integrates a call to chargecomplement to load
the list of possible values for a qualitative parameter
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}
