% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_taxa.R
\docType{methods}
\name{choicemult,ref_taxa-method}
\alias{choicemult,ref_taxa-method}
\title{Multiple Choice method for reftaxa referential objects, the graphical interface is built to allow
for multiple choices. See load for method in the command line.}
\usage{
\S4method{choicemult}{ref_taxa}(object, objectreport = NULL,
  is.enabled = TRUE)
}
\arguments{
\item{object}{An object of class \link{ref_taxa-class}}

\item{objectreport}{An object report which includes the \link{ref_taxa-class}, default NULL}

\item{is.enabled}{Sets if the frame is enabled at launch, defaut TRUE}
}
\description{
Multiple Choice method for reftaxa referential objects, the graphical interface is built to allow
for multiple choices. See load for method in the command line.
}
\examples{
\dontrun{ 
 object=new("ref_taxa")
win=gwindow()
group=ggroup(container=win,horizontal=FALSE)
object<-charge(object)
report_mig=new(report_mig)
choicemult(object,objectreport=report_mig)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
