% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_timestep.R
\docType{methods}
\name{choicemult,ref_timestep-method}
\alias{choicemult,ref_timestep-method}
\title{Graphical interface for multiple choice method for PasdeTemps (used in report_mig_mult)}
\usage{
\S4method{choicemult}{ref_timestep}(object)
}
\arguments{
\item{object}{An object of class \link{ref_timestep-class}}
}
\description{
Graphical interface for multiple choice method for PasdeTemps (used in report_mig_mult)
}
\note{
this method differs from choice as it is called within a notebook,
it does not allow for multiple choice to be made
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

