% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefChoix.r
\docType{methods}
\name{choicemult,RefChoix-method}
\alias{choicemult,RefChoix-method}
\title{Multiple Choice method for RefChoix referential objects, to put together with notebook widgets}
\usage{
\S4method{choicemult}{RefChoix}(object, selected_value)
}
\arguments{
\item{object}{An object of class \link{RefChoix-class}}

\item{selected_value}{the value selected in the combo}
}
\description{
Multiple Choice method for RefChoix referential objects, to put together with notebook widgets
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

