% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefDC.r
\docType{methods}
\name{choicemult,RefDC-method}
\alias{choicemult,RefDC-method}
\title{choicemult, selection method for refDC allowing to select several DC}
\usage{
\S4method{choicemult}{RefDC}(object, objectBilan = NULL, is.enabled = TRUE)
}
\arguments{
\item{object}{An object of class RefDC}

\item{objectBilan}{A bilan object}

\item{is.enabled}{A boolean indicating if the widget can be seleted at launch}
}
\description{
choicemult, selection method for refDC allowing to select several DC
}
\note{
The choice method has for arguments a report (bilan) object
 (e.g) is called from a report Bilan(e.g Bilan_carlot).
  By default,  the value of the objectbilan is null.
  When it is not   the method calls daughter widgets (e.g. the dc widget will call species) 
and fills it with the method \link{charge_avec_filtre,RefTaxon-method}
}
\examples{
\dontrun{ 
win=gwindow()
group=ggroup(container=win,horizontal=FALSE)
object=new("RefDC")
object<-charge(object)
objectBilan=new("BilanMigrationMult")
choicemult(object=object,objectBilan=objectBilan)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

