% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refpar.r
\docType{methods}
\name{choicemult,Refpar-method}
\alias{choicemult,Refpar-method}
\title{Multiple Choice method for Refpar referential objects, internal use}
\usage{
\S4method{choicemult}{Refpar}(object, objectBilan = NULL,
  label = gettext("Sample characteristic", domain = "R-stacomiR"),
  nomassign = "refpar")
}
\arguments{
\item{object}{An object of class \link{Refpar-class}}

\item{objectBilan}{An object Bilan which includes the \link{Refpar-class}, default NULL}

\item{label}{The name for the frame}

\item{nomassign}{The name by which the widget will be assigned in envir_stacomi}
}
\description{
Multiple Choice method for Refpar referential objects, internal use
}
\examples{
\dontrun{
 object=new("Refpar")
win=gwindow()
group=ggroup(container=win,horizontal=FALSE)
object<-charge(object)
bilanMigrationCar=new("BilanMigrationCar")
objectBilan=bilan_taille # for other test
choicemult(object,objectBilan=bilanMigrationCar)	
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

