% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refparqual.r
\docType{methods}
\name{choicemult,Refparqual-method}
\alias{choicemult,Refparqual-method}
\title{Multiple Choice method for Refparqual referential objects internal use}
\usage{
\S4method{choicemult}{Refparqual}(object, objectBilan = NULL,
  nomassign = "refparqual", label = gettext("Qualitative", domain =
  "R-stacomiR"))
}
\arguments{
\item{object}{An object of class \link{Refparqual-class}}

\item{objectBilan}{An object Bilan which includes the \link{Refparqual-class}, default NULL}

\item{nomassign}{The name used when assigning the object Refparqual to the \code{envir_stacomi} environment}

\item{label}{The name of the frame}
}
\description{
Multiple Choice method for Refparqual referential objects internal use
}
\note{
this methods rewrites that of the Refpar as it integrates a call to chargecomplement to load
the list of possible values for a qualitative parameter
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

