% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefTaxon.r
\docType{methods}
\name{choicemult,RefTaxon-method}
\alias{choicemult,RefTaxon-method}
\title{Multiple Choice method for Reftaxon referential objects, the graphical interface is built to allow
for multiple choices. See load for method in the command line.}
\usage{
\S4method{choicemult}{RefTaxon}(object, objectBilan = NULL,
  is.enabled = TRUE)
}
\arguments{
\item{object}{An object of class \link{RefTaxon-class}}

\item{objectBilan}{An object Bilan which includes the \link{RefTaxon-class}, default NULL}

\item{is.enabled}{Sets if the frame is enabled at launch, defaut TRUE}
}
\description{
Multiple Choice method for Reftaxon referential objects, the graphical interface is built to allow
for multiple choices. See load for method in the command line.
}
\examples{
\dontrun{ 
 object=new("RefTaxon")
win=gwindow()
group=ggroup(container=win,horizontal=FALSE)
object<-charge(object)
bilanMigration=new(BilanMigration)
choicemult(object,objectBilan=bilanMigration)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

