\name{choix-methods}
\docType{methods}
\alias{choix-methods}
\alias{choix,PasDeTemps-method}
\alias{choix,PasDeTempsJournalier-method}
\alias{choix,RefAnnee-method}
\alias{choix,RefCheckBox-method}
\alias{choix,RefChoix-method}
\alias{choix,RefDC-method}
\alias{choix,RefDF-method}
\alias{choix,RefHorodate-method}
\alias{choix,RefListe-method}
\alias{choix,Refpar-method}
\alias{choix,Refparqual-method}
\alias{choix,RefStades-method}
\alias{choix,RefStationMesure-method}
\alias{choix,RefTaxon-method}
\alias{choix,RefTextBox-method}
\title{Methods for Function choix}
\description{
These methods are used within the graphical interface to load the referential elements of each
class (ex Reftaxon) and copy the user choice in the environment "envir_stacomi"
}
\section{Methods}{
\describe{

\item{\code{signature(objet = "PasDeTemps")}}{ ~~describe this method here }

\item{\code{signature(objet = "PasDeTempsJournalier")}}{ ~~describe this method here }

\item{\code{signature(objet = "RefAnnee")}}{ ~~describe this method here }

\item{\code{signature(objet = "RefCheckBox")}}{ ~~describe this method here }

\item{\code{signature(objet = "RefChoix")}}{ ~~describe this method here }

\item{\code{signature(objet = "RefDC")}}{ ~~describe this method here }

\item{\code{signature(objet = "RefDF")}}{ ~~describe this method here }

\item{\code{signature(objet = "RefHorodate")}}{ ~~describe this method here }

\item{\code{signature(objet = "RefListe")}}{ ~~describe this method here }

\item{\code{signature(objet = "Refpar")}}{ ~~describe this method here }

\item{\code{signature(objet = "Refparqual")}}{ ~~describe this method here }

\item{\code{signature(objet = "RefStades")}}{ ~~describe this method here }

\item{\code{signature(objet = "RefStationMesure")}}{ ~~describe this method here }

\item{\code{signature(objet = "RefTaxon")}}{ ~~describe this method here }

\item{\code{signature(objet = "RefTextBox")}}{ ~~describe this method here }
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s)}
