% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefAnnee.r
\docType{methods}
\name{choix,RefAnnee-method}
\alias{choix,RefAnnee-method}
\title{choice method for RefAnnee referential}
\usage{
\S4method{choix}{RefAnnee}(object, nomassign = "refAnnee",
  funoutlabel = get("msg", envir = envir_stacomi)$RefAnnee.2,
  titleFrame = get("msg", envir = envir_stacomi)$RefAnnee.1, preselect = 1)
}
\arguments{
\item{nomassign}{The name to be asssigned in envir_stacomi}

\item{funoutlabel}{The label that appears in funout}

\item{titleFrame}{Title for the frame}

\item{preselect}{The number of the year selected in the gdroplist (integer)}
}
\description{
Allows the selection of year and the assignment in environment envir_stacomi
}
\examples{
 object=new("RefAnnee")
object<-charge(object)
win=gwindow(title="test refAnnee")
group=ggroup(container=win,horizontal=FALSE)
choix(object,nomassign="refAnnee",funoutlabel="essai",titleFrame="essai RefAnnee",preselect=1)
dispose(win)
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

