% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefCheckBox.r
\docType{methods}
\name{choix,RefCheckBox-method}
\alias{choix,RefCheckBox-method}
\title{Choice method for ReCheckBox referential objects}
\usage{
\S4method{choix}{RefCheckBox}(object)
}
\description{
Choice method for ReCheckBox referential objects
}
\examples{
object=new("RefCheckBox")
object<- charge(object,title="essai",labels=c("par1","par2","par3"),checked=c(TRUE,TRUE,TRUE))
win=gwindow(title="test RefCheckBox")
group=ggroup(container=win,horizontal=FALSE)
choix(object) 
dispose(win)
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

