% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefChoix.r
\docType{methods}
\name{choix,RefChoix-method}
\alias{choix,RefChoix-method}
\title{Choice method for Rechoix referential objects}
\usage{
\S4method{choix}{RefChoix}(object)
}
\description{
Used by the graphical interface.
}
\examples{
 object=new("RefChoix")
 object<-charge(object,vecteur=c("oui","non"),label="essai",selected=as.integer(1))
win=gwindow(title="test refChoix")
group=ggroup(container=win,horizontal=FALSE)
choix(object) 
dispose(win)
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

