% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefDC.r
\docType{methods}
\name{choix,RefDC-method}
\alias{choix,RefDC-method}
\title{choice method for RefDC}
\usage{
\S4method{choix}{RefDC}(object, objectBilan = NULL, is.enabled = TRUE)
}
\arguments{
\item{objectBilan}{the objectBilan from which this frame was called}

\item{is.enabled}{a boolean indicating whether the current frame will be enabled. Selecting a "parent"
frame may cause some frame to be disabled. When created the frame is enabled.}
}
\description{
choice method for RefDC
}
\note{
The choix method has for arguments a report (bilan) object
 (e.g) is called from a report Bilan(e.g Bilan_lot).
  By default,  the value of the objectbilan is null.
  When it is not   the method calls daughter widgets (e.g. the dc widget will call species) 
and fills it with the method \link {charge_avec_filtre}
}
\examples{
\dontrun{ win=gwindow()
group=ggroup(container=win,horizontal=FALSE)
object=new("RefDC")
object<-charge(object)
objectBilan=new("BilanMigration")
choix(object=object,objectBilan=objectBilan)}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

