% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefDF.r
\docType{methods}
\name{choix,RefDF-method}
\alias{choix,RefDF-method}
\title{Choice method for DF referential objects}
\usage{
\S4method{choix}{RefDF}(object)
}
\description{
Choice method for DF referential objects
}
\note{
the choice method assigns an object of class refDF in the environment envir_stacomi
}
\examples{
 
 \dontrun{ 
win=gwindow()
group=ggroup(container=win,horizontal=FALSE)
object=new("RefDF")
object<-charge(object)
choix(object)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

