% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefHorodate.r
\docType{methods}
\name{choix,RefHorodate-method}
\alias{choix,RefHorodate-method}
\title{choix method for RefHorodate}
\usage{
\S4method{choix}{RefHorodate}(object, label = "date",
  nomassign = "horodate",
  funoutlabel = "nous avons le choix dans la date\\n", decal = 0,
  affichecal = TRUE)
}
\value{
Selects the date in the graphical interface
}
\description{
choix method for RefHorodate
}
\author{
cedric.briand
}

