% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coef_durif}
\alias{coef_durif}
\title{Silvering index coefficients from Caroline Durif (2009) to predict silvering stage from morphological parameters}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 5 rows and 6 columns.
}
\usage{
coef_durif
}
\description{
Classification scores are calculated by multiplying the metrics 
BL = body length, W = weight, MD = mean eye diameter (Dv+Dh)/2, and FL length of the pectoral fin,
with each parameter p as S=Constant+BL*p(bl)+W*p(W)... The stage chosen is the one achieving the 
highest score
}
\references{
Durif, C.M., Guibert, A., and Elie, P. 2009.
Morphological discrimination of the silvering stages of the European eel. 
In American Fisheries Society Symposium. pp. 103-111.
 \url{https://fishlarvae.org/common/SiteMedia/durif\%20et\%20al\%202009b.pdf}
}
\keyword{datasets}
