\name{coerce-methods}
\docType{methods}
\alias{coerce-methods}
\alias{coerce,BilanMigration,BilanMigrationInterAnnuelle-method}
\alias{coerce,PasDeTempsChar,PasDeTemps-method}
\alias{coerce,RequeteODBCwhere,RequeteODBC-method}
\alias{coerce,RequeteODBCwheredate,RequeteODBCwhere-method}
\title{Methods for Function coerce}
\description{
Methods for function \code{coerce}  ie inherited from a mother class
}
\section{Methods}{
\describe{

\item{\code{signature(from = "BilanMigration", to = "BilanMigrationInterAnnuelle")}}{ ~~describe this method here }

\item{\code{signature(from = "PasDeTempsChar", to = "PasDeTemps")}}{ ~~describe this method here }

\item{\code{signature(from = "RequeteODBCwhere", to = "RequeteODBC")}}{ ~~describe this method here }

\item{\code{signature(from = "RequeteODBCwheredate", to = "RequeteODBCwhere")}}{ ~~describe this method here }
}}
\keyword{methods}

