\name{coerce<--methods}
\docType{methods}
\alias{coerce<--methods}
\alias{coerce<-,RequeteODBCwhere,RequeteODBC-method}
\alias{coerce<-,RequeteODBCwheredate,RequeteODBCwhere-method}
\title{Methods for Function coerce<-}
\description{
used in setAs functions to change from one class to the next
}
\section{Methods}{
\describe{

\item{\code{signature(from = "RequeteODBCwhere", to = "RequeteODBC")}}{ ~~describe this method here }

\item{\code{signature(from = "RequeteODBCwheredate", to = "RequeteODBCwhere")}}{ ~~describe this method here }
}}
\keyword{methods}
