% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_carlot.r
\docType{methods}
\name{connect,Bilan_carlot-method}
\alias{connect,Bilan_carlot-method}
\title{connect method for Bilan_carlot}
\usage{
\S4method{connect}{Bilan_carlot}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{Bilan_carlot-class}}

\item{silent}{Boolean if TRUE messages are not displayed}
}
\value{
An object of class \link{Bilan_carlot-class}
}
\description{
connect method for Bilan_carlot
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
