% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_poids_moyen.r
\docType{methods}
\name{connect,Bilan_poids_moyen-method}
\alias{connect,Bilan_poids_moyen-method}
\title{connect method for Bilan_Poids_moyen}
\usage{
\S4method{connect}{Bilan_poids_moyen}(object)
}
\arguments{
\item{object}{An object of class \link{Bilan_poids_moyen-class}}
}
\value{
Bilan_Poids_Moyen request corresponding to user choices, mean weight
 w is calculated as car_valeur_quantitatif/lot_effectif. These coefficients are stored in the database,
and the connect method loads them from the table using the \link{RefCoe-class}
}
\description{
connect method for Bilan_Poids_moyen
}
\note{
dates for the request are from august to august (a glass eel season)
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

