% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_stades_pigm.r
\docType{methods}
\name{connect,Bilan_stades_pigm-method}
\alias{connect,Bilan_stades_pigm-method}
\title{connect method for Bilan_stades_pigm}
\usage{
\S4method{connect}{Bilan_stades_pigm}(object)
}
\arguments{
\item{object}{An object of class \link{Bilan_stades_pigm-class}}
}
\value{
An object of class Bilan_stades_pigm
}
\description{
connect method for Bilan_stades_pigm
}
\note{
will try to get data for the temperature (refstation) only if retrocalcul is checked
 by default it is not when lanching
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

