% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_taille.r
\docType{methods}
\name{connect,Bilan_taille-method}
\alias{connect,Bilan_taille-method}
\title{connect method for class Bilan_taille}
\usage{
\S4method{connect}{Bilan_taille}(object)
}
\arguments{
\item{object}{An object of class \code{\link{Bilan_taille-class}}}
}
\value{
An object of class \code{\link{Bilan_taille-class}}  with requete field filled
}
\description{
connect method for class Bilan_taille
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

