% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanAgedemer.r
\docType{methods}
\name{connect,BilanAgedemer-method}
\alias{connect,BilanAgedemer-method}
\title{connect method for BilanAgedemer}
\usage{
\S4method{connect}{BilanAgedemer}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{BilanAgedemer-class}}

\item{silent}{Default FALSE, if TRUE the program should no display messages}
}
\value{
An object of class \link{BilanAgedemer-class}
}
\description{
connect method for BilanAgedemer
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
