% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanAnnuels.r
\docType{methods}
\name{connect,BilanAnnuels-method}
\alias{connect,BilanAnnuels-method}
\title{connect method for BilanAnnuels class
this method performs the sum over the year attention this function does
not count subsamples.}
\usage{
\S4method{connect}{BilanAnnuels}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{BilanAnnuels-class}}

\item{silent}{Stops messages from being displayed if silent=TRUE, default FALSE}
}
\value{
An instantianted object with values filled with user choice

A dataframe with column effectif, comprising the sum of bilanMigration counts
}
\description{
connect method for BilanAnnuels class
this method performs the sum over the year attention this function does
not count subsamples.
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
