% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanArgentee.r
\docType{methods}
\name{connect,BilanArgentee-method}
\alias{connect,BilanArgentee-method}
\title{connect method for BilanArgentee}
\usage{
\S4method{connect}{BilanArgentee}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{BilanArgentee-class}}

\item{silent}{Boolean if TRUE messages are not displayed}
}
\value{
An object of class \link{BilanArgentee-class}
}
\description{
connect method for BilanArgentee
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
