% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanEspeces.r
\docType{methods}
\name{connect,BilanEspeces-method}
\alias{connect,BilanEspeces-method}
\title{connect method for BilanEspeces}
\usage{
\S4method{connect}{BilanEspeces}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class BilanEspeces}

\item{silent}{Boolean, if TRUE, information messages are not displayed}
}
\value{
An object of class BilanEspeces with data slot filled
}
\description{
connect method for BilanEspeces
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

