% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanFonctionnementDC.r
\docType{methods}
\name{connect,BilanFonctionnementDC-method}
\alias{connect,BilanFonctionnementDC-method}
\title{connect method for BilanFonctionnementDC}
\usage{
\S4method{connect}{BilanFonctionnementDC}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{BilanFonctionnementDC-class}}

\item{silent}{boolean, default FALSE, if TRUE messages are not displayed}
}
\value{
An object of class \link{BilanFonctionnementDC-class}
}
\description{
loads the working periods and type of arrest or disfunction of the DC
}
\author{
cedric.briand
}

