% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigration.r
\docType{methods}
\name{connect,BilanMigration-method}
\alias{connect,BilanMigration-method}
\alias{connect.bilanmigration}
\title{connect method for BilanMigration}
\usage{
\S4method{connect}{BilanMigration}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{BilanMigration-class}}

\item{silent}{Boolean default FALSE, if TRUE information messages not displayed}
}
\value{
BilanMigration with slot @data filled from the database
}
\description{
uses the BilanMigrationMult method
}

