% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationCar.r
\docType{methods}
\name{connect,BilanMigrationCar-method}
\alias{connect,BilanMigrationCar-method}
\title{connect method for BilanMigrationCar}
\usage{
\S4method{connect}{BilanMigrationCar}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{BilanMigrationCar-class}}

\item{silent}{Boolean default FALSE, if TRUE information messages not displayed}
}
\value{
BilanMigrationCar with slot @data filled from the database
}
\description{
uses the BilanMigrationMult method
}

