% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMult.r
\docType{methods}
\name{connect,BilanMigrationMult-method}
\alias{connect,BilanMigrationMult-method}
\title{connect method for BilanMigrationMult}
\usage{
\S4method{connect}{BilanMigrationMult}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{An object of class BilanMigrationMult}

\item{silent}{Boolean, if TRUE messages are not displayed}
}
\value{
An object of class \link{BilanMigrationMult-class} with slot @data filled from the database
}
\description{
this method loads data from the database for BilanMigration but also fills the table of conversion coefficient, if 
the taxa is eel. It also calls connect method for \link{BilanFonctionnementDF-class}, 
\link{BilanFonctionnementDC-class} and \link{BilanOperation-class} associated with the Bilan
and used by the \link{fungraph} and \link{fungraph_civelle} functions.
}

