% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMultConditionEnv.r
\docType{methods}
\name{connect,BilanMigrationMultConditionEnv-method}
\alias{connect,BilanMigrationMultConditionEnv-method}
\title{connect method for BilanMigrationMultConditionEnv class}
\usage{
\S4method{connect}{BilanMigrationMultConditionEnv}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{BilanMigrationMultConditionEnv-class}}

\item{silent}{Default FALSE, if TRUE the program should no display messages}
}
\value{
an object of BilanMigrationMultConditionEnv class
}
\description{
connect method for BilanMigrationMultConditionEnv class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
