% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanOperation.r
\docType{methods}
\name{connect,BilanOperation-method}
\alias{connect,BilanOperation-method}
\title{connect method for BilanOperation}
\usage{
\S4method{connect}{BilanOperation}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{BilanOperation-class}
load data from the operation table, one dataset per DC}

\item{silent}{Boolean, TRUE removes messages.}
}
\value{
An object of class \link{BilanOperation-class}
}
\description{
connect method for BilanOperation
}
\author{
cedric.briand
}

