% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_dc.R
\name{connect,report_dc-method}
\alias{connect,report_dc-method}
\alias{connect.report_dc}
\title{connect method for report_dc}
\usage{
\S4method{connect}{report_dc}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_dc-class}}

\item{silent}{boolean, default FALSE, if TRUE messages are not displayed}
}
\value{
An object of class \link{report_dc-class} with slot data filled from the database
}
\description{
loads the working periods and type of arrest or disfunction of the DC
}
\author{
cedric.briand
}
