% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_df.R
\name{connect,report_df-method}
\alias{connect,report_df-method}
\alias{connect.report_df}
\title{connect method for report_df}
\usage{
\S4method{connect}{report_df}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_df-class}
loads the working periods and type of arrest or disfunction of the DF}

\item{silent}{Boolean, TRUE removes messages.}
}
\value{
An object of class \code{report_df} with slot data filled from the database
}
\description{
connect method for report_df
}
\author{
cedric.briand
}
