% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_env.R
\docType{methods}
\name{connect,report_env-method}
\alias{connect,report_env-method}
\alias{connect.report_env}
\title{connect method for report_env class}
\usage{
\S4method{connect}{report_env}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_env-class}}

\item{silent}{Default FALSE, if TRUE the program should no display messages}
}
\value{
an object of report_env class
}
\description{
connect method for report_env class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

