% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig.R
\name{connect,report_mig-method}
\alias{connect,report_mig-method}
\alias{connect.report_mig}
\title{connect method for report_mig}
\usage{
\S4method{connect}{report_mig}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_mig-class}}

\item{silent}{Boolean default FALSE, if TRUE information messages not displayed}
}
\value{
report_mig with slot \code{@data} filled from the database
}
\description{
uses the report_mig_mult method
}
