% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_char.R
\docType{methods}
\name{connect,report_mig_char-method}
\alias{connect,report_mig_char-method}
\alias{connect.report_mig_char}
\title{connect method for report_mig_char}
\usage{
\S4method{connect}{report_mig_char}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_mig_char-class}}

\item{silent}{Boolean default FALSE, if TRUE information messages not displayed}
}
\value{
report_mig_char with slot @data filled from the database
}
\description{
uses the report_mig_mult method
}

