% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_env.R
\name{connect,report_mig_env-method}
\alias{connect,report_mig_env-method}
\alias{summary.report_mig_env}
\title{connect method for report_mig_env class}
\usage{
\S4method{connect}{report_mig_env}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_mig_env-class}}

\item{silent}{Default FALSE, if TRUE the program should no display messages}
}
\value{
an object of report_mig_env class
}
\description{
connect method for report_mig_env class
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
