% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_interannual.R
\name{connect,report_mig_interannual-method}
\alias{connect,report_mig_interannual-method}
\alias{connect.report_mig_interannual}
\title{connect method for report_mig_interannual}
\usage{
\S4method{connect}{report_mig_interannual}(object, silent = FALSE, check = TRUE)
}
\arguments{
\item{object}{An object of class \link{report_mig_interannual-class}}

\item{silent}{Stops messages from being displayed if silent=TRUE, default FALSE}

\item{check}{Checks that data are corresponding between report_annual and report_mig}
}
\value{
report_mig_interannual an instantiated object with values filled with user choice
}
\description{
This method will check if the data in the t_reportjournalier_bjo table has no missing data,
if missing the program will load missing data. As a second step,
the program will check if the numbers in the table  t_reportjournalier_bjo differ from those in the database,
and propose to re-run the report_mig (which has a write_database methode to write daily reports) for those years.
}
\note{
We expect different results between daily reports from the t_reportjournalier_bjo table and the annual sums
from report_annual for glass eels as those may have been weighted and not only counted. The t_reportjournalier_bjo table used by report_mig_interannual
contains the sum of glass eel numbers converted from weights and those directly counted. The report_annual does not.
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
