% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_mult.R
\name{connect,report_mig_mult-method}
\alias{connect,report_mig_mult-method}
\alias{connect.report_mig_mult}
\alias{'}
\title{connect method for report_mig_mult}
\usage{
\S4method{connect}{report_mig_mult}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_mig_mult-class}}

\item{silent}{Boolean, if TRUE messages are not displayed}
}
\value{
An object of class \link{report_mig_mult-class} with slot @data filled from the database
}
\description{
this method loads data from the database for report_mig but also fills the table of conversion coefficient, if 
the taxa is eel. It also calls connect method for \link{report_df-class}, 
\link{report_dc-class} and \link{report_ope-class} associated with the report
and used by the \link{fungraph} and \link{fungraph_glasseel} functions. As a side effect it assigns
objects 	\link{report_dc-class}, \link{report_df-class} and \link{report_ope-class} in environment \code{envir_stacomi}
}
