% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_ope.R
\name{connect,report_ope-method}
\alias{connect,report_ope-method}
\alias{connect.report_ope}
\title{connect method for report_ope}
\usage{
\S4method{connect}{report_ope}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_ope-class}
load data from the operation table, one dataset per DC}

\item{silent}{Boolean, TRUE removes messages.}
}
\value{
An object of class \link{report_ope-class} with slot data \code{@data} filled
}
\description{
connect method for report_ope
}
\author{
cedric.briand
}
