% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_sample_char.R
\name{connect,report_sample_char-method}
\alias{connect,report_sample_char-method}
\alias{connect.report_sample_char}
\title{connect method for report_sample_char}
\usage{
\S4method{connect}{report_sample_char}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_sample_char-class}}

\item{silent}{Boolean if TRUE messages are not displayed}
}
\value{
An object of class \link{report_sample_char-class} with slot data \code{@data} filled
}
\description{
connect method for report_sample_char
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
