% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_sea_age.R
\name{connect,report_sea_age-method}
\alias{connect,report_sea_age-method}
\alias{connect.report_sea_age}
\title{connect method for report_sea_age}
\usage{
\S4method{connect}{report_sea_age}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_sea_age-class}}

\item{silent}{Default FALSE, if TRUE the program should no display messages}
}
\value{
An object of class \link{report_sea_age-class} with slot data \code{@data} filled
}
\description{
connect method for report_sea_age
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
