% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_silver_eel.R
\name{connect,report_silver_eel-method}
\alias{connect,report_silver_eel-method}
\alias{connect.report_silver_eel}
\title{connect method for report_silver_eel}
\usage{
\S4method{connect}{report_silver_eel}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_silver_eel-class}}

\item{silent}{Boolean if TRUE messages are not displayed}
}
\value{
An object of class \link{report_silver_eel-class} with slot data \code{@data} filled
}
\description{
connect method for report_silver_eel
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
