% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_species.R
\name{connect,report_species-method}
\alias{connect,report_species-method}
\alias{connect.report_species}
\title{connect method for report_species}
\usage{
\S4method{connect}{report_species}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class report_species}

\item{silent}{Boolean, if TRUE, information messages are not displayed}
}
\value{
An object of class \link{report_species-class} with data slot filled with slot data \code{@data} filled
}
\description{
connect method for report_species
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
