% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefMsg.r
\docType{methods}
\name{createmessage,RefMsg-method}
\alias{createmessage,RefMsg-method}
\title{createmessage method for RefMsg referential objects}
\usage{
\S4method{createmessage}{RefMsg}(object)
}
\value{
An S4 object of class RefMsg
}
\description{
createmessage method for RefMsg referential objects
}
\note{
When coming from the database, " are now /", those at the beginning and end are turned into ", the others are single quote when they are to be pasted within the text as code example. The remainder "c("a","b","c") are rebuilt into vectors by the function
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

