% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generic.r
\name{createmessage}
\alias{createmessage}
\title{Generic method to create messages}
\usage{
createmessage(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Additional parms}
}
\description{
Generic method to create messages
}
\author{
cedric.briand
}

