% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMult.r
\docType{methods}
\name{cumplot,BilanMigrationMult-method}
\alias{cumplot,BilanMigrationMult-method}
\title{Method to create Cumulated graphs for BilanMigrationMult.}
\usage{
\S4method{cumplot}{BilanMigrationMult}(object, ...)
}
\arguments{
\item{...}{}
}
\description{
Data are summed per day for different dc taxa and stages
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

