% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_timestep.R
\docType{methods}
\name{current_end_date,ref_timestep-method}
\alias{current_end_date,ref_timestep-method}
\title{Gets the ending date of a time step for an object of class \link{ref_timestep-class}}
\usage{
\S4method{current_end_date}{ref_timestep}(object)
}
\arguments{
\item{object}{An object of class \link{ref_timestep-class}}
}
\value{
Currentend_date, The ending date for the current timestep
}
\description{
Gets the ending date of a time step for an object of class \link{ref_timestep-class}
}
