% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_timestep.R
\name{current_end_date}
\alias{current_end_date}
\title{Generic method for getting the ending date for current time step}
\usage{
current_end_date(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional parameters passed to the method}
}
\description{
Generic method for getting the ending date for current time step
}
\keyword{internal}

