\name{currentDateDebut-methods}
\docType{methods}
\alias{currentDateDebut-methods}
\alias{currentDateDebut,PasDeTemps-method}
\title{ Methods for Function currentDateDebut}
\description{
 used by class PasDetemps
}
\section{Methods}{
\describe{

\item{\code{signature(objet = "PasDeTemps")}}{ ~~describe this method here }
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s)}
