% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasdeTemps.r
\docType{methods}
\name{currentDateDebut,PasDeTemps-method}
\alias{currentDateDebut,PasDeTemps-method}
\title{Gets the starting date of a time step for an object of class \link{PasDeTemps-class}}
\usage{
\S4method{currentDateDebut}{PasDeTemps}(object)
}
\arguments{
\item{object}{An object of class \link{PasDeTemps-class}}
}
\value{
CurrentDateDebut, The starting date for the current timestep
}
\description{
Gets the starting date of a time step for an object of class \link{PasDeTemps-class}
}

