% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasdeTemps.r
\name{currentDateDebut}
\alias{currentDateDebut}
\title{Generic method for getting the beginning date for current time step}
\usage{
currentDateDebut(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional parameters passed to the method}
}
\description{
Generic method for getting the beginning date for current time step
}

