\name{currentDateFin-methods}
\docType{methods}
\alias{currentDateFin-methods}
\alias{currentDateFin,PasDeTemps-method}
\title{ Methods for Function currentDateFin }
\description{
 used by class PasDetemps
}
\section{Methods}{
\describe{

\item{\code{signature(objet = "PasDeTemps")}}{ finishing time of the current time step}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s)}
