% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasdeTemps.r
\docType{methods}
\name{currentDateFin,PasDeTemps-method}
\alias{currentDateFin,PasDeTemps-method}
\title{Gets the ending date of a time step for an object of class \link{PasDeTemps-class}}
\usage{
\S4method{currentDateFin}{PasDeTemps}(object)
}
\arguments{
\item{object}{An object of class \link{PasDeTemps-class}}
}
\value{
CurrentDateFin, The ending date for the current timestep
}
\description{
Gets the ending date of a time step for an object of class \link{PasDeTemps-class}
}

