% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasdeTemps.r
\name{currentDateFin}
\alias{currentDateFin}
\title{Generic method for getting the ending date for current time step}
\usage{
currentDateFin(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional parameters passed to the method}
}
\description{
Generic method for getting the ending date for current time step
}
\keyword{internal}

