% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_timestep.R
\docType{methods}
\name{currentstart_date,ref_timestep-method}
\alias{currentstart_date,ref_timestep-method}
\title{Gets the starting date of a time step for an object of class \link{ref_timestep-class}}
\usage{
\S4method{currentstart_date}{ref_timestep}(object)
}
\arguments{
\item{object}{An object of class \link{ref_timestep-class}}
}
\value{
current_start_date, The starting date for the current timestep
}
\description{
Gets the starting date of a time step for an object of class \link{ref_timestep-class}
}
\keyword{internal}
