% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_timestep.R
\name{currentstart_date}
\alias{currentstart_date}
\title{Generic method for getting the beginning date for current time step}
\usage{
currentstart_date(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional parameters passed to the method}
}
\description{
Generic method for getting the beginning date for current time step
}
\keyword{internal}

