\name{DateFin-methods}
\docType{methods}
\alias{DateFin-methods}
\alias{DateFin,PasDeTemps-method}
\title{Methods for Function DateFin}
\description{
 used by class PasDetemps
}
\section{Methods}{
\describe{

\item{\code{signature(objet = "PasDeTemps")}}{ ~~describe this method here }
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s)}
