% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasdeTemps.r
\docType{methods}
\name{DateFin,PasDeTemps-method}
\alias{DateFin,PasDeTemps-method}
\title{Gets the final horodate for an object of class \link{PasDeTemps-class}}
\usage{
\S4method{DateFin}{PasDeTemps}(object)
}
\arguments{
\item{object}{An object of class \link{PasDeTemps-class}}
}
\value{
DateFin, The final date corresponding to nbStep*time duration + initial date
}
\description{
Gets the final horodate for an object of class \link{PasDeTemps-class}
}
\keyword{internal}

