% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasdeTemps.r
\name{DateFin}
\alias{DateFin}
\title{Generic method for getting the final date}
\usage{
DateFin(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional parameters passed to the method}
}
\description{
Generic method for getting the final date
}
\keyword{internal}

