% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_timestep.R
\name{end_date,ref_timestep-method}
\alias{end_date,ref_timestep-method}
\title{Gets the final horodate for an object of class \link{ref_timestep-class}}
\usage{
\S4method{end_date}{ref_timestep}(object)
}
\arguments{
\item{object}{An object of class \link{ref_timestep-class}}
}
\value{
end_date, The final date corresponding to nb_step*time duration + initial date
}
\description{
Gets the final horodate for an object of class \link{ref_timestep-class}
}
\keyword{internal}
