% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generic.R
\name{end_date}
\alias{end_date}
\title{Generic method for getting the final date}
\usage{
end_date(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional parameters passed to the method}
}
\description{
Generic method for getting the final date
}
\keyword{internal}
