% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generic.r
\docType{data}
\name{envir_stacomi}
\alias{envir_stacomi}
\title{Environment where most objects from the package are stored and then loaded
by the charge method}
\format{An object of class \code{environment} of length 0.}
\usage{
envir_stacomi
}
\description{
envir_stacomi \code{envir_stacomi <- new.env(parent = baseenv())} is the
environment where most object created by the interface are stored
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{datasets}

