% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_interannual.R
\name{fn_connect_report_mig_interannual}
\alias{fn_connect_report_mig_interannual}
\title{Get table content for table t_bilanmigrationjournalier_bjo in report_mig_interannual}
\usage{
fn_connect_report_mig_interannual(years, taxa, stage, dc)
}
\arguments{
\item{years}{A vector of years}

\item{taxa}{One taxa}

\item{stage}{One stage}

\item{dc}{A vector of counting devices}
}
\value{
a data frame with the content of table t_bilanmigrationjournalier_bjo in the database
}
\description{
Each time a report mig runs, it can write its content in the t_bilanmigrationjournalier_bjo
table which stores the results of the report_mig with one value per day
}
