% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_EcritBilanJournalier.r
\name{fn_EcritBilanJournalier}
\alias{fn_EcritBilanJournalier}
\title{fn_EcritBilanJournier writes the daily migration in the
t_bilanmigrationjournalier_bjo table}
\usage{
fn_EcritBilanJournalier(bilanMigration)
}
\arguments{
\item{bilanMigration}{an object of class \code{\linkS4class{BilanMigration}}}
}
\description{
Daily values are needed to compare migrations from year to year, by the class \link{BilanMigrationInterAnnuelle-class}. They are added by
the function fn_EcritBilanJournalier
}
\note{
the user is asked whether or not he wants to overwrite data, if no
data are present in the database, the import is done anyway
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

