\name{fn_table_per_dis}
\alias{fn_table_per_dis}
\title{functions called in DF and DC...}
\usage{fn_table_per_dis(typeperiode, tempsdebut, tempsfin, libelle, date=TRUE)}
\description{functions called in DF and DC}
\value{listeg}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{typeperiode}{}
\item{tempsdebut}{}
\item{tempsfin}{}
\item{libelle}{}
\item{date}{}}
