% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_stades_pigm.r
\name{fntablestade}
\alias{fntablestade}
\title{function calculating a table with pigment stages VB to VIA3 from lines retreived from the database
containing individual characteristic of glass eel#' 
this function is called from within the charge method it was separated from the charge method
as it it convenient to use elsewhere}
\usage{
fntablestade(stades,choicepere="lotpere")
}
\arguments{
\item{stades}{a data frame containing stage values}

\item{choicepere}{either "date" or "lot_pere" the first will group pigment stage by date, 
the second will allow to keep separate lines when several samples have been collected a given day}
}
\value{
a list with tablestades atable with numbers per stage for a given date or lotpere (sample), and date
}
\description{
function calculating a table with pigment stages VB to VIA3 from lines retreived from the database
containing individual characteristic of glass eel#' 
this function is called from within the charge method it was separated from the charge method
as it it convenient to use elsewhere
}
\author{
Cedric Briand \\email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
\code{\linkS4class{Bilan_stades_pigm}}
}

