\name{fun_char_spe}
\alias{fun_char_spe}
\title{function used to remove special non utf8 character which cause the gtk interface to crash...}
\usage{fun_char_spe(text)}
\description{function used to remove special non utf8 character which cause the gtk interface to crash}
\value{text}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{text}{}}
