% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_dailystats.R
\name{fun_dailystats}
\alias{fun_dailystats}
\title{function to create daily statistics}
\usage{
fun_dailystats(tableau, time.sequence, taxa, stage, DC)
}
\arguments{
\item{tableau}{A table}

\item{time.sequence}{Time sequence from report_mig and report_mig_mult}

\item{taxa}{A taxa}

\item{stage}{A stage}

\item{DC}{A counting device}
}
\description{
function to create daily statistics
}
\note{
to be used in shiny
}
\author{
Cedric Briand \email{cedric.briand'at'eptb-vilaine.fr}
}
