% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_interannual.R
\name{fun_report_mig_interannual}
\alias{fun_report_mig_interannual}
\title{statistics per time period}
\usage{
fun_report_mig_interannual(dat, year = NULL, timesplit = NULL)
}
\arguments{
\item{dat}{a data frame with columns ("bjo_annee","bjo_jour","bjo_labelquantite","bjo_valeur")}

\item{year}{The year to exclude from the historical series (it will be plotted against the historical series)}

\item{timesplit}{"week" "2 weeks" "month" as provided to seq.POSIXt, default NULL}
}
\value{
a data frame with mean, max, and min calculated for each timesplit
}
\description{
statistics per time period
}
