% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funschema.R
\name{fun_schema}
\alias{fun_schema}
\title{Creates a list of available schemas in the db}
\usage{
fun_schema()
}
\value{
A table with of data providers with org_code, the user of each schema, and  org_description the description of the schema
}
\description{
Creates a list of available schemas in the db
}
