% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_table_per_dis.R
\name{fun_table_per_dis}
\alias{fun_table_per_dis}
\title{functions called in DF and DC}
\usage{
fun_table_per_dis(
  typeperiode,
  tempsdebut,
  tempsfin,
  libelle,
  color,
  date = TRUE
)
}
\arguments{
\item{typeperiode}{ref.tr_typearretdisp_tar(per_tar_code) the code of the
period (see table ref.tr_typearretdisp_tar)}

\item{tempsdebut}{ref.tr_typearretdisp_tar(per_date_debut) starting
timestamp of the period}

\item{tempsfin}{The postgres column ref.tr_typearretdisp_tar(per_date_fin) ending timestamp of
the period}

\item{libelle}{The postgres column ref.tr_typearretdisp_tar(libelle )description of the period
type}

\item{color}{A named vector of color matching libelle.}

\item{date}{Boolean, should the function return a POSIXt or date value}
}
\value{
A list
}
\description{
functions called in DF and DC
}
\note{
returns either POSIXt or date if date=TRUE
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
