% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMult.r
\name{fun_weight_conversion}
\alias{fun_weight_conversion}
\title{returns a table where weights and number are calculated from number and weights respectively
performs a query to collect the conversion coefficients}
\usage{
fun_weight_conversion(tableau, duree)
}
\arguments{
\item{tableau}{}

\item{duree}{}
}
\value{
tableau, the data frame
}
\description{
returns a table where weights and number are calculated from number and weights respectively
performs a query to collect the conversion coefficients
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

