% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_write_monthly.R
\name{fun_write_monthly}
\alias{fun_write_monthly}
\title{This writes monthly data in t_reportmensuel_mens table}
\usage{
fun_write_monthly(report_mig, resum, silent)
}
\arguments{
\item{report_mig}{an object of class \code{\linkS4class{report_mig}}}

\item{resum}{data frame with summary per month}

\item{silent}{Suppresses messages}
}
\value{
No return value, called for side effects
}
\description{
This writes monthly data in t_reportmensuel_mens table
}
\note{
This function is launched by fun_write_daily, the resum 
dataset is created by the \link{funstat} function
}
