% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_dc.R
\name{funbarchart1DC}
\alias{funbarchart1DC}
\title{Handler for barchart for report_df class from the graphical interface}
\usage{
funbarchart1DC(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
Handler for barchart for report_df class from the graphical interface
}
\note{
The program cuts periods which overlap between two month
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

