% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_dc.R
\name{funbarchartDC}
\alias{funbarchartDC}
\title{Function to create a barchart (lattice) corresponding to the periods}
\usage{
funbarchartDC(h, ...)
}
\arguments{
\item{h}{a handler}

\item{...}{Additional parameters}
}
\value{
assigns the data frame \code{periodeDC} allowing to build the lattice graph in the environment envir_stacomi
}
\description{
Function to create a barchart (lattice) corresponding to the periods
}
\author{
cedric.briand
}
\keyword{internal}

