\name{funbarchartDF}
\alias{funbarchartDF}
\title{funbarchartDF creates a barchart for BilanFonctionnementDF class}
\usage{
  funbarchartDF(h, ...)
}
\arguments{
  \item{h}{a handler}
}
\description{
  funbarchartDF creates a barchart for
  BilanFonctionnementDF class
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

