\name{funBilanMigrationAnnuel}
\alias{funBilanMigrationAnnuel}
\title{this functions performs the sum over the year
attention this function does not count subsamples.}
\usage{
  funBilanMigrationAnnuel(bilanMigration)
}
\description{
  this functions performs the sum over the year attention
  this function does not count subsamples.
}

