% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanFonctionnementDC.r
\name{funboxDC}
\alias{funboxDC}
\title{function used for some lattice graphes with dates}
\usage{
funboxDC(h, ...)
}
\arguments{
\item{vectordate}{date or POSIXt}
}
\value{
vectordate (without class)
}
\description{
function used for some lattice graphes with dates
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

