% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanFonctionnementDF.r
\name{funboxDF}
\alias{funboxDF}
\title{FunboxDF draws rectangles to describe the DF work for BilanFonctionnementDF class}
\usage{
funboxDF(h, ...)
}
\arguments{
\item{h}{a handler}
}
\description{
FunboxDF draws rectangles to describe the DF work for BilanFonctionnementDF class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

