% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanArgentee.r
\name{funboxplotBilanArgentee}
\alias{funboxplotBilanArgentee}
\title{Boxplots for ggplot2}
\usage{
funboxplotBilanArgentee(h, ...)
}
\arguments{
\item{h}{A handler passed by the graphical interface}

\item{...}{Additional parameters}
}
\description{
assigns an object g in envir_stacomi for eventual modification of the plot
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

