% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_sample_char.R
\name{funboxplotreport_sample_char}
\alias{funboxplotreport_sample_char}
\title{Boxplots for ggplot2}
\usage{
funboxplotreport_sample_char(...)
}
\arguments{
\item{...}{Additional parameters}
}
\value{
Nothing, calls plot method \code{plot.type = "2"} on data assigned in envir_stacomi
}
\description{
assigns an object g in envir_stacomi for eventual modification of the plot
}
\author{
Cedric Briand \email{cedric.briand'at'eptb-vilaine.fr}
}
\keyword{internal}
