% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_stades_pigm.r
\name{funcalcbilan_stades_pigm}
\alias{funcalcbilan_stades_pigm}
\title{main launching function for class Bilan_stades_pigm}
\usage{
funcalcbilan_stades_pigm(h, ...)
}
\arguments{
\item{h}{A handler}

\item{additional}{arguments}

\item{list()}{additional arguments}
}
\description{
Function with handler which calls charge (and thus connect) and calculates
the title
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

