\name{funCalculeBilanTaille}
\alias{funCalculeBilanTaille}
\title{funCalculeBilanTaille...}
\usage{funCalculeBilanTaille(h, ...)}
\description{funCalculeBilanTaille calls charge method and within charge the connect method for Bilan_taille class}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{h}{}
\item{...}{}}
