% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_dc.R
\name{funchartDC}
\alias{funchartDC}
\title{Handler function to plot calendar like graph, internal use}
\usage{
funchartDC(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
Handler function to plot calendar like graph, internal use
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

