% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanFonctionnementDF.r
\name{funchartDF}
\alias{funchartDF}
\title{Handler fonction to plot calendar like graph, internal use}
\usage{
funchartDF(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
Handler fonction to plot calendar like graph, internal use
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

