% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanArgentee.r
\name{fundensityBilanArgentee}
\alias{fundensityBilanArgentee}
\title{fundensityBilanArgentee uses ggplot2 to draw density plots}
\usage{
fundensityBilanArgentee(h, ...)
}
\arguments{
\item{h}{A handler}

\item{...}{Additional parameters}
}
\description{
assigns an object g in envir_stacomi for eventual modification of the plot
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

