\name{fungraph}
\alias{fungraph}
\title{Function for BilanMigration graphes including numbers DF DC operations}
\usage{
  fungraph(bilanMigration, tableau, duree, taxon, stade)
}
\arguments{
  \item{bilanMigration}{an objet of class \code{\linkS4class{BilanMigration}}}
  \item{tableau}{a data frame with the results}
  \item{duree}{a vector POSIXt}
  \item{taxon}{the species}
  \item{stade}{the stage}
}
\description{
  Function for BilanMigration graphes including numbers DF
  DC operations
}
\author{
  Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

