\name{fungraph_civelle}
\alias{fungraph_civelle}
\title{Graph function for glass eel migration.}
\usage{fungraph_civelle(bilanMigration, table, duree, taxon, stade)}
\description{Graph function for glass eel migration. Differs from fungraph as it does not draw the ggplot graph for month}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{bilanMigration}{}
\item{table=tableau}{}
\item{duree}{}
\item{taxon}{}
\item{stade}{}}
