\name{fungraph_env}
\alias{fungraph_env}
\title{graph function for BilanMigrationEnv...}
\usage{fungraph_env(tableau, duree, taxon, stade, stations)}
\description{graph function for BilanMigrationEnv}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{tableau}{}
\item{duree}{}
\item{taxon}{}
\item{stade}{}
\item{stations}{}}
