% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_stades_pigm.r
\name{fungraphgg}
\alias{fungraphgg}
\title{gglot function to draw graps of pigment stages}
\usage{
fungraphgg(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
gglot function to draw graps of pigment stages
}

