\name{fungraphInteract_lot}
\alias{fungraphInteract_lot}
\title{fungraphInteract_lot calls the ggplotusr interface to draw plots...}
\usage{fungraphInteract_lot(h, ...)}
\description{fungraphInteract_lot calls the ggplotusr interface to draw plots}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{h}{}
\item{...}{}}
