\name{fungraphInteract_tail}
\alias{fungraphInteract_tail}
\title{fungraphInteract_tail uses the ggplot2usr interface to build the graphes...}
\usage{fungraphInteract_tail(h, ...)}
\description{fungraphInteract_tail uses the ggplot2usr interface to build the graphes}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{h}{a handler}
\item{...}{}}
