\name{fungraphstades}
\alias{fungraphstades}
\title{Fonction principale de calcul et de tracage des graphiques...}
\usage{fungraphstades(tablestades, retrocalcul=TRUE, datatempsal, points=TRUE,
    nb=TRUE, graphstades=TRUE, lmax=1, labelretro, labelgraphstades,
    phi, maxVIA3=10, dates, Vparm, effectifs)}
\description{Fonction principale de calcul et de tracage des graphiques}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{tablestades}{}
\item{retrocalcul}{}
\item{datatempsal}{}
\item{points}{}
\item{nb}{}
\item{graphstades}{}
\item{lmax}{}
\item{labelretro}{}
\item{labelgraphstades}{}
\item{phi}{}
\item{maxVIA3}{}
\item{dates}{}
\item{Vparm}{}}
