\name{funhtml}
\alias{funhtml}
\title{function used to print the html tables of output (see xtable documentation)...}
\usage{funhtml(data, caption, top=TRUE, outfile, clipboard=FALSE, append=TRUE,
    digits, ...)}
\description{function used to print the html tables of output (see xtable documentation)}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{data}{a data frame}
\item{caption}{the caption}
\item{top}{of top the caption is placed on top}
\item{outfile}{outfile is the path to the file}
\item{clipboard}{if clipboard TRUE, a copy to the clipboard is made}
\item{append}{is the file appended to the previous one ?}
\item{digits}{}
\item{...}{}}
