\name{funout}
\alias{funout}
\title{this function uses gfile, edits a text with info and changing colors...}
\usage{funout(text, arret=FALSE, wash=FALSE)}
\description{this function uses gfile, edits a text with info and changing colors}
\value{nblignes assigned in .Global}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{text}{}
\item{arret}{}
\item{wash}{}}
