% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_stades_pigm.r
\name{funphi}
\alias{funphi}
\title{Function to calculate pigmentation times.}
\usage{
funphi(parm, datatempsal)
}
\arguments{
\item{parm}{parameters of the model}

\item{datatempsal}{data.frame containing temperatures and salinities}
}
\value{
list("dates"=time.sequence,"phi_jour"=phi_jour)
}
\description{
Function to calculate pigmentation times.
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

