% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanAgedemer.r
\name{funplotBilanAgedemer}
\alias{funplotBilanAgedemer}
\title{funplotBilanAgedemer}
\usage{
funplotBilanAgedemer(h, ...)
}
\arguments{
\item{h}{A handler, with action 1,2,3 or 4}

\item{...}{Additional parameters}
}
\description{
assigns an object g in envir_stacomi for eventual modification of the plot
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

