% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanArgentee.r
\name{funpointBilanArgentee}
\alias{funpointBilanArgentee}
\title{Point graph from ggplot}
\usage{
funpointBilanArgentee(h, ...)
}
\arguments{
\item{h}{handler passed by the graphical interface}

\item{...}{Additional parameters}
}
\description{
assigns an object g in envir_stacomi for eventual modification of the plot
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

