% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_sample_char.R
\name{funpointreport_sample_char}
\alias{funpointreport_sample_char}
\title{Point graph from ggplot}
\usage{
funpointreport_sample_char(...)
}
\arguments{
\item{...}{Additional parameters}
}
\value{
Nothing, calls plot method \code{plot.type = "3"} on data assigned in envir_stacomi
}
\description{
assigns an object g in envir_stacomi for eventual modification of the plot
}
\author{
Cedric Briand \email{cedric.briand'at'eptb-vilaine.fr}
}
\keyword{internal}
