% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funSousListeBilanMigration.r
\name{funSousListeBilanMigration}
\alias{funSousListeBilanMigration}
\title{funSousListeBilanMigration}
\usage{
funSousListeBilanMigration(bilanMigration)
}
\arguments{
\item{bilanMigration}{an object of class \code{\linkS4class{BilanMigration}}}
}
\description{
workhorse function for bilanMigration. Calculates the number for a stage and
a taxa per day.  The operation for the fishway is never from 00:00 to 00:00
so the number per day is calculated according to the ration between the
duration of the operation and the duration of the day. This function will
allow daily reports to be saved into the database when graph is launched
}

