\name{funstat}
\alias{funstat}
\title{Function to calculate statistics per month...}
\usage{funstat(tableau, duree, taxon, stade, DC)}
\description{Function to calculate statistics per month}
\value{resum}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{tableau}{}
\item{duree}{}
\item{taxon}{}
\item{stade}{}
\item{DC}{}}
