% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funstat_daily.R
\name{funstat_daily}
\alias{funstat_daily}
\title{function to create daily statistics}
\usage{
funstat_daily(tableau, time.sequence, taxa, stage, DC)
}
\arguments{
\item{tableau}{A table}

\item{time.sequence}{Time sequence from report_mig and report_mig_mult}

\item{taxa}{A taxa}

\item{stage}{A stage}

\item{DC}{A counting device}
}
\description{
function to create daily statistics
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
