\name{funstatJournalier}
\alias{funstatJournalier}
\title{function to create daily statistics...}
\usage{funstatJournalier(tableau, duree, taxon, stade, DC)}
\description{function to create daily statistics}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{tableau}{}
\item{duree}{}
\item{taxon}{}
\item{stade}{}
\item{DC}{}}
