\name{funtable}
\alias{funtable}
\title{function to print and save statistics in...}
\usage{funtable(tableau, duree, taxon, stade, DC, resum)}
\description{function to print and save statistics in .csv and .html formats}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{tableau}{}
\item{duree}{}
\item{taxon}{}
\item{stade}{}
\item{DC}{}
\item{resum}{}}
