% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_carlot.r
\name{funtableBilan_carlot}
\alias{funtableBilan_carlot}
\title{table function}
\usage{
funtableBilan_carlot(h, ...)
}
\arguments{
\item{h}{hanlder passed by the graphical interface}

\item{...}{Additional parameters}
}
\description{
funtableBilan_carlot shows a table of results in gdf
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

