\name{funtableBilan_lot}
\alias{funtableBilan_lot}
\title{funtableBilan_lot shows a table of results in gdf...}
\usage{funtableBilan_lot(h, ...)}
\description{funtableBilan_lot shows a table of results in gdf}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{h}{}
\item{...}{}}
