% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanAgedemer.r
\name{funtableBilanAgedemer}
\alias{funtableBilanAgedemer}
\title{table function}
\usage{
funtableBilanAgedemer(h, ...)
}
\arguments{
\item{h}{hanlder passed by the graphical interface}

\item{...}{Additional parameters}
}
\description{
funtableBilanAgedemer shows a table of results in gdf
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

