% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanArgentee.r
\name{funtableBilanArgentee}
\alias{funtableBilanArgentee}
\title{table function}
\usage{
funtableBilanArgentee(h, ...)
}
\arguments{
\item{h}{hanlder passed by the graphical interface}

\item{...}{Additional parameters}
}
\description{
funtableBilanArgentee shows a table of results in gdf
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

