% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanFonctionnementDC.r
\name{funtableDC}
\alias{funtableDC}
\title{FuntableDC create a table output for BilanFonctionnementDC class}
\usage{
funtableDC(h, ...)
}
\arguments{
\item{h}{Handler}

\item{...}{Additional parameters}
}
\description{
FuntableDC create a table output for BilanFonctionnementDC class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

