% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanFonctionnementDF.r
\name{funtableDF}
\alias{funtableDF}
\title{Table output for BilanFonctionnementDF class}
\usage{
funtableDF(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
Table output for BilanFonctionnementDF class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

