% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_sample_char.R
\name{funtablereport_sample_char}
\alias{funtablereport_sample_char}
\title{table function}
\usage{
funtablereport_sample_char(h, ...)
}
\arguments{
\item{h}{handler passed by the graphical interface}

\item{...}{Additional parameters}
}
\description{
funtablereport_sample_char shows a table of results in gdf
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

