% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_sea_age.R
\name{funtablereport_sea_age}
\alias{funtablereport_sea_age}
\title{table function}
\usage{
funtablereport_sea_age(h, ...)
}
\arguments{
\item{h}{handler passed by the graphical interface}

\item{...}{Additional parameters}
}
\description{
funtablereport_sea_age shows a table of results in gdf
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

