% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_silver_eel.R
\name{funtablereport_silver_eel}
\alias{funtablereport_silver_eel}
\title{table function}
\usage{
funtablereport_silver_eel(h, ...)
}
\arguments{
\item{h}{handler passed by the graphical interface}

\item{...}{Additional parameters}
}
\description{
funtablereport_silver_eel shows a table of results in gdf
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}
