\name{funtitle_bilan_stades_pigm}
\alias{funtitle_bilan_stades_pigm}
\title{Fonction handler qui retourne le titre du graphique apres le choix dans la date...}
\usage{funtitle_bilan_stades_pigm(h, ...)}
\description{Fonction handler qui retourne le titre du graphique après le choix dans la date}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{h}{}}
