\name{funtraitement_poids}
\alias{funtraitement_poids}
\title{returns a table where weights and number are calculated from number and weights respectively...}
\usage{funtraitement_poids(tableau, duree)}
\description{returns a table where weights and number are calculated from number and weights respectively
performs a query to collect the conversion coefficients}
\value{tableau}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{tableau}{}
\item{duree}{}}
