% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_org}
\alias{get_org}
\title{this function gets the name of the stucture as it is set in the database}
\usage{
get_org()
}
\value{
The name of the structure (org_code)
}
\description{
this function gets the name of the stucture as it is set in the database
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
\keyword{internal}
