% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_schema}
\alias{get_schema}
\title{this function gets the schema from envir stacomi and throws warning}
\usage{
get_schema(default = NULL)
}
\arguments{
\item{default}{passed to rlang::get_env}
}
\value{
The schema in envir_stacomi
}
\description{
this function gets the schema from envir stacomi and throws warning
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
\keyword{internal}
