% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_timestep.R
\docType{methods}
\name{get_step_label,ref_timestep-method}
\alias{get_step_label,ref_timestep-method}
\title{Gets the string value of time step}
\usage{
\S4method{get_step_label}{ref_timestep}(object)
}
\arguments{
\item{object}{An object of class \link{ref_timestep-class}}
}
\value{
A string corresponding to the value of current time step
}
\description{
Gets the string value of time step
}
\keyword{internal}

