% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_timestep.R
\name{get_step_label}
\alias{get_step_label}
\title{Generic method to get the string value of time step}
\usage{
get_step_label(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional parameters passed to the method}
}
\description{
Generic method to get the string value of time step
}
\keyword{internal}
