% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_timestep.R
\docType{methods}
\name{get_year,ref_timestep-method}
\alias{get_year,ref_timestep-method}
\title{Gets the year or a vector of years corresponding to the timestep ("ref_timestep") object}
\usage{
\S4method{get_year}{ref_timestep}(object)
}
\arguments{
\item{object}{An object of class \link{ref_timestep-class}}
}
\description{
Gets the year or a vector of years corresponding to the timestep ("ref_timestep") object
}
\keyword{internal}
