% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generic.R
\name{get_year}
\alias{get_year}
\title{Generic method to get the years}
\usage{
get_year(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional parameters passed to the method}
}
\description{
Generic method to get the years
}
\keyword{internal}
