\name{getanneeprec-methods}
\docType{methods}
\alias{getanneeprec-methods}
\alias{getanneeprec,RefHorodate-method}
\title{ Methods for Function getanneeprec }
\description{
Used by class RefAnnee
}
\section{Methods}{
\describe{

\item{\code{signature(objet = "RefHorodate")}}{ ~~describe this method here }
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s)}
