\name{getAnnees-methods}
\docType{methods}
\alias{getAnnees-methods}
\alias{getAnnees,PasDeTemps-method}
\title{ Methods for Function getAnnees}
\description{
 Used by class PasDeTemps
}
\section{Methods}{
\describe{

\item{\code{signature(objet = "PasDeTemps")}}{ ~~describe this method here }
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s)}
