% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasdeTemps.r
\docType{methods}
\name{getAnnees,PasDeTemps-method}
\alias{getAnnees,PasDeTemps-method}
\title{Gets the year or a vector of years corresponding to the timestep ("PasDeTemps") object}
\usage{
\S4method{getAnnees}{PasDeTemps}(object)
}
\arguments{
\item{object}{An object of class \link{PasDeTemps-class}}
}
\description{
Gets the year or a vector of years corresponding to the timestep ("PasDeTemps") object
}
