% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasdeTemps.r
\name{getAnnees}
\alias{getAnnees}
\title{Generic method to get the years}
\usage{
getAnnees(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional parameters passed to the method}
}
\description{
Generic method to get the years
}
\keyword{internal}

