% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasdeTemps.r
\docType{methods}
\name{getdateDebut,PasDeTemps-method}
\alias{getdateDebut,PasDeTemps-method}
\title{Returns the starting date as character}
\usage{
\S4method{getdateDebut}{PasDeTemps}(object)
}
\arguments{
\item{object}{An object of class \link{PasDeTemps-class}}

\item{...}{Additional parameters passed to the method}
}
\description{
Returns the starting date as character
}

