% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_timestep.R
\docType{methods}
\name{getdateDebut,ref_timestep-method}
\alias{getdateDebut,ref_timestep-method}
\title{Returns the starting date as character}
\usage{
\S4method{getdateDebut}{ref_timestep}(object)
}
\arguments{
\item{object}{An object of class \link{ref_timestep-class}}

\item{...}{Additional parameters passed to the method}
}
\description{
Returns the starting date as character
}
\keyword{internal}

