% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_timestep.R
\name{getdateDebut}
\alias{getdateDebut}
\title{Generic method the get starting date}
\usage{
getdateDebut(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional parameters passed to the method}
}
\description{
Generic method the get starting date
}
\keyword{internal}

