% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitaires.r
\name{getdbname}
\alias{getdbname}
\title{Retreives the dbnames from a connection using "baseOBC"}
\usage{
getdbname()
}
\value{
A string with the name of the database
}
\description{
Retreives the dbnames from a connection using "baseOBC"
}

