\name{getLibellesPas-methods}
\docType{methods}
\alias{getLibellesPas-methods}
\alias{getLibellesPas,PasDeTemps-method}
\title{ Methods for Function getLibellesPas}
\description{
Get the text value of a pas de temps (ex 1 semaine)
}
\section{Methods}{
\describe{

\item{\code{signature(objet = "PasDeTemps")}}{ ~~describe this method here }
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s)}
