% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasdeTemps.r
\docType{methods}
\name{getLibellesPas,PasDeTemps-method}
\alias{getLibellesPas,PasDeTemps-method}
\title{Gets the string value of time step}
\usage{
\S4method{getLibellesPas}{PasDeTemps}(object)
}
\arguments{
\item{object}{An object of class \link{PasDeTemps-class}}
}
\value{
A string corresponding to the value of current time step
}
\description{
Gets the string value of time step
}

