% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasdeTemps.r
\name{getLibellesPas}
\alias{getLibellesPas}
\title{Generic method to get the string value of time step}
\usage{
getLibellesPas(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional parameters passed to the method}
}
\description{
Generic method to get the string value of time step
}
