% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_timestep.R
\docType{methods}
\name{getnocurrent_step,ref_timestep-method}
\alias{getnocurrent_step,ref_timestep-method}
\title{Gets the current time step of an object of class \link{ref_timestep-class}}
\usage{
\S4method{getnocurrent_step}{ref_timestep}(object)
}
\arguments{
\item{object}{An object of class \link{ref_timestep-class}}
}
\value{
the current time step of the object
}
\description{
Gets the current time step of an object of class \link{ref_timestep-class}
}
\keyword{internal}

