% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_timestep.R
\name{getnocurrent_step}
\alias{getnocurrent_step}
\title{Generic method to get current time step}
\usage{
getnocurrent_step(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional parameters passed to the method}
}
\description{
Generic method to get current time step
}
