% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasdeTemps.r
\docType{methods}
\name{getnoPasCourant,PasDeTemps-method}
\alias{getnoPasCourant,PasDeTemps-method}
\title{Gets the current time step of an object of class \link{PasDeTemps-class}}
\usage{
\S4method{getnoPasCourant}{PasDeTemps}(object)
}
\arguments{
\item{object}{An object of class \link{PasDeTemps-class}}
}
\value{
the current time step of the object
}
\description{
Gets the current time step of an object of class \link{PasDeTemps-class}
}
