% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasdeTemps.r
\name{getnoPasCourant}
\alias{getnoPasCourant}
\title{Generic method to get current time step}
\usage{
getnoPasCourant(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional parameters passed to the method}
}
\description{
Generic method to get current time step
}

