\name{getRefHorodate-methods}
\docType{methods}
\alias{getRefHorodate-methods}
\alias{getRefHorodate,RefHorodate-method}
\title{Methods for Function getRefHorodate}
\description{
 see class RefHorodate for use
}
\section{Methods}{
\describe{

\item{\code{signature(objet = "RefHorodate")}}{ ~~describe this method here }
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s)}
