\name{getvalue-methods}
\docType{methods}
\alias{getvalue-methods}
\alias{getvalue,Refperiode-method}
\title{ Methods for Function getvalue }
\description{
 Methods for function \code{getvalue}
}
\section{Methods}{
\describe{

\item{\code{signature(objet = "Refperiode")}}{ ~~describe this method here }
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s)}
