% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_period.R
\docType{methods}
\name{getvalue,ref_period-method}
\alias{getvalue,ref_period-method}
\title{Returns the POSIXt value of a given name}
\usage{
\S4method{getvalue}{ref_period}(object, id)
}
\arguments{
\item{object}{An object of class \link{ref_period-class}}

\item{id}{one of "jour", "semaine", "quinzaine", "mois"}
}
\value{
"a character to be used in seq.POSIXt
}
\description{
Returns the POSIXt value of a given name
}
\examples{
\dontrun{
 getvalue(new("ref_period"),"quinzaine")
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
