% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refperiode.r
\docType{methods}
\name{getvalue,Refperiode-method}
\alias{getvalue,Refperiode-method}
\title{Returns the POSIXt value of a given name}
\usage{
\S4method{getvalue}{Refperiode}(object, id, ...)
}
\value{
"a character to be used in seq.POSIXt
}
\description{
Returns the POSIXt value of a given name
}
\examples{
 getvalue(new("Refperiode"),"quinzaine")
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

