% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generic.r
\name{getvalue}
\alias{getvalue}
\title{Generic method getvalue}
\usage{
getvalue(object, ...)

getvalue(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Additional parms}

\item{object}{Object}

\item{...}{Additional parms}
}
\description{
Generic method getvalue

Generic method getvalue
}
\author{
cedric.briand

cedric.briand
}

