\name{graphdate}
\alias{graphdate}
\title{function used for some lattice graphes with dates...}
\usage{graphdate(vectordate)}
\description{function used for some lattice graphes with dates}
\value{vectordate (without class)}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{vectordate}{date or POSIXt}}
