\name{graphe-methods}
\docType{methods}
\alias{graphe-methods}
\alias{graphe}
\alias{graphe,BilanMigrationPar-method}
\title{ Methods for Function graphe }
\description{
method with a parameter choice to join several graphes called by several handlers (buttons)
}
\section{Methods}{
\describe{
\item{\code{signature(objet = "BilanMigrationPar")}}{
method with a parameter choice to join several graphes called by several handlers (buttons) }
}}
\keyword{methods}