% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacomi.R
\name{h_report_dc}
\alias{h_report_dc}
\title{handler function used by the main interface}
\usage{
h_report_dc(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
handler function used by the main interface
}
\keyword{internal}

