% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacomi.R
\name{h_report_df}
\alias{h_report_df}
\title{handler function used by the main interface}
\usage{
h_report_df(h, ...)
}
\arguments{
\item{h}{A handler}

\item{...}{Other parameters}
}
\description{
handler function used by the main interface
}
\keyword{internal}
