% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_env.R
\name{h_report_envgraph}
\alias{h_report_envgraph}
\title{h_report_envgraph Internal method}
\usage{
h_report_envgraph(h, ...)
}
\arguments{
\item{h}{a handler}

\item{...}{Additional parameters}
}
\description{
h_report_envgraph Internal method
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}
