% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_mult.R
\name{h_report_mig_mult_graph}
\alias{h_report_mig_mult_graph}
\title{handler for plot method in report_mig_mult class}
\usage{
h_report_mig_mult_graph(h = null, ...)
}
\arguments{
\item{h}{a handler}

\item{...}{Additional parameters}
}
\description{
internal use
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}
