% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_mult.R
\name{h_report_mig_multcalc}
\alias{h_report_mig_multcalc}
\title{handler for calculations}
\usage{
h_report_mig_multcalc(h = null, ...)
}
\arguments{
\item{h}{a handler}

\item{...}{Additional parameters}
}
\description{
internal use for graphical interface
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}
