% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_mult.R
\name{h_report_mig_multgraph2}
\alias{h_report_mig_multgraph2}
\title{handler du calcul h_report_miggraph2}
\usage{
h_report_mig_multgraph2(h = null, ...)
}
\arguments{
\item{h}{a handler}

\item{...}{Additional parameters}
}
\description{
Cumulated migration over time
}
\keyword{internal}
