% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_mult.R
\name{h_report_mig_multgraph3}
\alias{h_report_mig_multgraph3}
\title{handler}
\usage{
h_report_mig_multgraph3(h = null, ...)
}
\arguments{
\item{h}{a handler}

\item{...}{Additional parameters}
}
\description{
internal use
}
\keyword{internal}

