% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig.R
\name{h_report_migcalc}
\alias{h_report_migcalc}
\title{handler for calculations report_mig}
\usage{
h_report_migcalc(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
internal use
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

