% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig.R
\name{h_report_miggraph}
\alias{h_report_miggraph}
\title{handler for  h_report_miggraph}
\usage{
h_report_miggraph(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
Standard report_mig graph over time
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

