% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig.R
\name{h_report_miggraph2}
\alias{h_report_miggraph2}
\title{handler for h_report_miggraph2}
\usage{
h_report_miggraph2(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
Step plot over time
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

