% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig.R
\name{h_report_migwrite}
\alias{h_report_migwrite}
\title{handler h_report_migwrite
Allows the saving of daily and monthly counts in the database}
\usage{
h_report_migwrite(h, ...)
}
\arguments{
\item{h}{a handler}

\item{...}{Additional parameters}
}
\description{
handler h_report_migwrite
Allows the saving of daily and monthly counts in the database
}
\note{
these entries are necessary to run the Interannual Migration class. 
then no entry will be written to the database
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}
