\name{handler_interfacegraphique}
\alias{hbilanMigration}
\alias{hBilanEspeces}
\alias{h0}
\alias{hBilanConditionEnv}
\alias{hBilanLots}
\alias{hBilanMigrationPar}
\alias{hDC}
\alias{hOPE}
\alias{hDFDC}
\alias{hBilanMigration}
\alias{hBilanMigrationInterannuelle}
\alias{hBilanMigrationConditionEnv}
\alias{hDF}
\alias{hhelp}
\alias{hpds}
\alias{hSt}
\alias{hTail}
\alias{htodo}
\alias{hx11}
\title{handler function used by the main interface to call bilan classes}
\usage{
 hBilanLots(h, ...)
 hBilanMigration(h, ...)
 hBilanEspeces(h,...)
}
\arguments{
  \item{h}{the handler}
  \item{...}{additional arguments passed to the function}
}
\description{
  Handler functions used by interface_graphique the main interface
}

\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

