% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanAnnuels.r
\name{hbarplotBilanAnnuels}
\alias{hbarplotBilanAnnuels}
\title{Barplot handler}
\usage{
hbarplotBilanAnnuels(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
Barplot handler
}
