% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_annual.R
\name{hbarplotreport_annual}
\alias{hbarplotreport_annual}
\title{Barplot handler}
\usage{
hbarplotreport_annual(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
Barplot handler
}
\keyword{internal}
