% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacomi.r
\name{hBilan_carlot}
\alias{hBilan_carlot}
\title{handler function used by the main interface}
\usage{
hBilan_carlot(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
handler function used by the main interface
}
\keyword{internal}

