% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_graphique.r
\name{hBilanEspeces}
\alias{hBilanEspeces}
\title{handler function used by the main interface}
\usage{
hBilanEspeces(h, ...)
}
\description{
handler function used by the main interface
}

