\name{hBilanEspecescalc}
\alias{hBilanEspecescalc}
\title{handler du calcul du BilanEspeces...}
\usage{hBilanEspecescalc(h, ...)}
\description{handler du calcul du BilanEspeces
realise le calcul du bilan especes, l'ecrit dans l'environnement envir_stacomi
traite eventuellement les quantites de lots (si c'est des civelles)}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{h}{}
\item{...}{}}
