\name{hbilanMigrationcalc}
\alias{hbilanMigrationcalc}
\title{handler du graphique BilanMigration...}
\usage{hbilanMigrationcalc(h, ...)}
\description{handler du graphique BilanMigration
realise le calcul du bilan migration, l'ecrit dans l'environnement envir_stacomi
traite eventuellement les quantites de lots (si c'est des civelles)}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{h}{a handler}
\item{...}{}}
