% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacomi.r
\name{hBilanMigrationConditionEnv}
\alias{hBilanMigrationConditionEnv}
\title{handler function used by the main interface}
\usage{
hBilanMigrationConditionEnv(h, ...)
}
\description{
handler function used by the main interface
}

