\name{hbilanMigrationConditionEnvcalc}
\alias{hbilanMigrationConditionEnvcalc}
\title{handler du graphique BilanMigrationConditionEnv...}
\usage{hbilanMigrationConditionEnvcalc(h, ...)}
\description{handler du graphique BilanMigrationConditionEnv
realise le calcul du bilan migration avec CE, l'ecrit dans l'environnement envir_stacomi
traite eventuellement les quantites de lots (si c'est des civelles)}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{h}{}
\item{...}{}}
