\name{hbilanMigrationgraph2}
\alias{hbilanMigrationgraph2}
\title{handler du calcul hBilanMigrationgraph2...}
\usage{hbilanMigrationgraph2(h, ...)}
\description{handler du calcul hBilanMigrationgraph2
appelle les fonctions fungraph pour faire un graphe annuel des 
cumuls de migration au cours du temps}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{h}{}
\item{...}{}}
