% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacomi.r
\name{hBilanMigrationInterAnnuelle}
\alias{hBilanMigrationInterAnnuelle}
\title{handler function used by the main interface}
\usage{
hBilanMigrationInterAnnuelle(h, ...)
}
\description{
handler function used by the main interface
}

