% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_graphique.r
\name{hBilanMigrationMult}
\alias{hBilanMigrationMult}
\title{handler function used by the main interface}
\usage{
hBilanMigrationMult(h, ...)
}
\description{
handler function used by the main interface
}

