% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMult.r
\name{hbilanMigrationMult_graph}
\alias{hbilanMigrationMult_graph}
\title{handler for graphe method in BilanMigrationMult class}
\usage{
hbilanMigrationMult_graph(h = null, ...)
}
\arguments{
\item{h}{a handler}

\item{...}{Additional parameters}
}
\description{
internal use
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
