% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMult.r
\name{hbilanMigrationMultcalc}
\alias{hbilanMigrationMultcalc}
\title{handler for calculations}
\usage{
hbilanMigrationMultcalc(h = null, ...)
}
\arguments{
\item{h}{a handler}

\item{...}{Additional parameters}
}
\description{
internal use for graphical interface
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
BilanMigrationMult
}

