% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMult.r
\name{hbilanMigrationMultgraph2}
\alias{hbilanMigrationMultgraph2}
\title{handler du calcul hBilanMigrationgraph2}
\usage{
hbilanMigrationMultgraph2(h = null, ...)
}
\arguments{
\item{h}{a handler}

\item{...}{Additional parameters}
}
\description{
cumuls de migration au cours du temps
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
