% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationPar.r
\name{hbilanMigrationParcalc}
\alias{hbilanMigrationParcalc}
\title{handler for bilanmigrationpar}
\usage{
hbilanMigrationParcalc(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{Additional parameters}
}
\description{
handler for bilanmigrationpar
}

