% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationPar.r
\name{hbilanMigrationPargraph2}
\alias{hbilanMigrationPargraph2}
\title{le handler appelle la methode generique graphe sur l'object plot.type=2}
\usage{
hbilanMigrationPargraph2(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{Additional parameters}
}
\description{
le handler appelle la methode generique graphe sur l'object plot.type=2
}

