\name{hbilanMigrationParstat}
\alias{hbilanMigrationParstat}
\title{le handler appelle la methode generique graphe sur l'objet choix 3}
\usage{
  hbilanMigrationParstat(h, ...)
}
\arguments{
  \item{h,}{passé par le handler}
}
\description{
  le handler appelle la methode generique graphe sur
  l'objet choix 3
}

