% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigration.r
\name{hbilanMigrationwrite}
\alias{hbilanMigrationwrite}
\title{handler hBilanMigrationwrite
Allows the saving of daily and monthly counts in the database, this method is also called from hbilanMigrationgraph}
\usage{
hbilanMigrationwrite(h, ...)
}
\arguments{
\item{h}{a handler}

\item{...}{Additional parameters}
}
\description{
handler hBilanMigrationwrite
Allows the saving of daily and monthly counts in the database, this method is also called from hbilanMigrationgraph
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

