% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_species.R
\name{hbilespcalc}
\alias{hbilespcalc}
\title{handler for calculation}
\usage{
hbilespcalc(h, ...)
}
\arguments{
\item{h}{a handler}

\item{...}{Additional parameters}
}
\description{
internal use
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

