% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationCar.r
\name{hbmCcalc}
\alias{hbmCcalc}
\title{handler for bilanmigrationpar}
\usage{
hbmCcalc(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{Additional parameters}
}
\description{
handler for bilanmigrationpar
}
