% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationCar.r
\name{hbmCgraph}
\alias{hbmCgraph}
\title{le handler appelle la methode generique graphe sur l'object plot.type=1}
\usage{
hbmCgraph(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{Additional parameters}
}
\description{
le handler appelle la methode generique graphe sur l'object plot.type=1
}
