% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationCar.r
\name{hbmCgraph2}
\alias{hbmCgraph2}
\title{le handler appelle la methode generique graphe sur l'object plot.type=2}
\usage{
hbmCgraph2(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{Additional parameters}
}
\description{
le handler appelle la methode generique graphe sur l'object plot.type=2
}
