% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_char.R
\name{hbmCplotcrossed}
\alias{hbmCplotcrossed}
\title{handler for plot}
\usage{
hbmCplotcrossed(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{Additional parameters}
}
\description{
handler for plot
}
\keyword{internal}

