% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_char.R
\name{hbmCplotqual}
\alias{hbmCplotqual}
\title{handler for plot}
\usage{
hbmCplotqual(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{Additional parameters}
}
\description{
handler for plot
}
\keyword{internal}
