% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_char.R
\name{hbmCstat}
\alias{hbmCstat}
\title{This handler calls the generic method graphe on object plot.type 3}
\usage{
hbmCstat(h)
}
\arguments{
\item{h}{handler}

\item{...}{Additional parameters}
}
\description{
This handler calls the generic method graphe on object plot.type 3
}
\keyword{internal}

