% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMultConditionEnv.r
\name{hbmmCEcalc}
\alias{hbmmCEcalc}
\title{Internal handler function}
\usage{
hbmmCEcalc(h = null, ...)
}
\arguments{
\item{h}{a handler}

\item{...}{Additional parameters}
}
\description{
Internal handler function
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
