% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMultConditionEnv.r
\name{hbmmCEgraph}
\alias{hbmmCEgraph}
\title{internal method for graphical interface}
\usage{
hbmmCEgraph(h = null, ...)
}
\arguments{
\item{h}{A handler}
}
\description{
internal method for graphical interface
}
