% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_poids_moyen.r
\name{hcalc}
\alias{hcalc}
\title{Internal handler for calcule, class \code{\link{Bilan_poids_moyen-class}}.}
\usage{
hcalc(h, ...)
}
\arguments{
\item{h}{handler}

\item{\dots}{additional arguments passed to the function}
}
\description{
Internal handler for calcule, class \code{\link{Bilan_poids_moyen-class}}.
}

