% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacomi.r
\name{hDC}
\alias{hDC}
\title{handler function used by the main interface}
\usage{
hDC(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
handler function used by the main interface
}
\keyword{internal}

