\name{hDF}
\title{handler function used by the main interface...}
\usage{hDF(h, ...)}
\description{handler function used by the main interface}
\alias{hDC,hOPE,hDFDC,hBilanMigration,hBilanMigrationInterannuelle,hBilanMigrationConditionEnv,}
\alias{hBilanMigrationPar,}
\alias{hBilanConditionEnv,}
\alias{hBilanLots,}
\alias{hTail,}
\alias{hpds,}
\alias{hSt,}
\alias{htodo,}
\alias{hhelp,}
\alias{h0,}
\alias{hx11}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{h}{}
\item{...}{}}
