% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_stades_pigm.r
\name{hfungraphstades}
\alias{hfungraphstades}
\title{handler function for fungraphstades}
\usage{
hfungraphstades(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
handler function for fungraphstades
}

