% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_interannual.R
\name{hgraph_report_mig_interannual}
\alias{hgraph_report_mig_interannual}
\title{Plot of all interannual from top to bottom}
\usage{
hgraph_report_mig_interannual(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
Plot of all interannual from top to bottom
}
\keyword{internal}

