% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_interannual.R
\name{hgraph_report_mig_interannual2}
\alias{hgraph_report_mig_interannual2}
\title{Plot of daily migrations}
\usage{
hgraph_report_mig_interannual2(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
Plot of daily migrations
}
\keyword{internal}

