% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_interannual.R
\name{hgraph_report_mig_interannual3}
\alias{hgraph_report_mig_interannual3}
\title{Step plot with different years displayed on the same graph. One year
can be highlighted against the others}
\usage{
hgraph_report_mig_interannual3(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
Step plot with different years displayed on the same graph. One year
can be highlighted against the others
}
\keyword{internal}

