% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_interannual.R
\name{hgraph_report_mig_interannual4}
\alias{hgraph_report_mig_interannual4}
\title{Plot comparing the migration  to the migration  
computed for all years available in the daily migration table.
This function plots comparisons for periods of 1 week, 2 weeks, month}
\usage{
hgraph_report_mig_interannual4(h, ...)
}
\arguments{
\item{h}{A handler}

\item{...}{Additional parameters}
}
\description{
Plot comparing the migration  to the migration  
computed for all years available in the daily migration table.
This function plots comparisons for periods of 1 week, 2 weeks, month
}
\keyword{internal}
