% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_interannual.R
\name{hgraph_report_mig_interannual5}
\alias{hgraph_report_mig_interannual5}
\title{Function displaying comparison similar to \link{hgraph_report_mig_interannual4} but using pointrange and geom_bar
This function plots comparisons for periods of 1 week, 2 weeks, month}
\usage{
hgraph_report_mig_interannual5(h, ...)
}
\arguments{
\item{h}{A handler}

\item{...}{Additional parameters}
}
\description{
Function displaying comparison similar to \link{hgraph_report_mig_interannual4} but using pointrange and geom_bar
This function plots comparisons for periods of 1 week, 2 weeks, month
}
\keyword{internal}
