% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationInterAnnuelle.r
\name{hgraphBilanMigrationInterAnnuelle2}
\alias{hgraphBilanMigrationInterAnnuelle2}
\title{Plot of daily migrations}
\usage{
hgraphBilanMigrationInterAnnuelle2(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
Plot of daily migrations
}
