% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationInterAnnuelle.r
\name{hgraphBilanMigrationInterAnnuelle3}
\alias{hgraphBilanMigrationInterAnnuelle3}
\title{Step plot with different years displayed on the same graph. One year
can be highlighted against the others}
\usage{
hgraphBilanMigrationInterAnnuelle3(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
Step plot with different years displayed on the same graph. One year
can be highlighted against the others
}

