% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationInterAnnuelle.r
\name{hgraphBilanMigrationInterAnnuelle4}
\alias{hgraphBilanMigrationInterAnnuelle4}
\title{Plot comparing the migration  to the migration  
computed for all years available in the daily migration table.
This function plots comparisions for periods of 1 week, 2 weeks, month
 @param h A handler
 @param ... Additional parameters}
\usage{
hgraphBilanMigrationInterAnnuelle4(h, ...)
}
\description{
Plot comparing the migration  to the migration  
computed for all years available in the daily migration table.
This function plots comparisions for periods of 1 week, 2 weeks, month
 @param h A handler
 @param ... Additional parameters
}

