% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationInterAnnuelle.r
\name{hgraphBilanMigrationInterAnnuelle5}
\alias{hgraphBilanMigrationInterAnnuelle5}
\title{Function displaying comparaison similar to \link{hgraphBilanMigrationInterAnnuelle4} but using pointrange and geom_bar
This function plots comparisions for periods of 1 week, 2 weeks, month}
\usage{
hgraphBilanMigrationInterAnnuelle5(h, ...)
}
\arguments{
\item{h}{A handler}

\item{...}{Additional parameters}
}
\description{
Function displaying comparaison similar to \link{hgraphBilanMigrationInterAnnuelle4} but using pointrange and geom_bar
This function plots comparisions for periods of 1 week, 2 weeks, month
}
