\name{hHistogramme}
\alias{hHistogramme}
\title{handler du calcul hHistogramme...}
\usage{hHistogramme(h, ...)}
\description{handler du calcul hHistogramme
trace un histogramme des espèces fréquentant la passe...}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{h}{}
\item{...}{}}
