% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigration.r
\name{houtBilanMigration}
\alias{houtBilanMigration}
\title{handler to print the command line}
\usage{
houtBilanMigration(h = null, ...)
}
\arguments{
\item{h}{a handler}

\item{...}{Additional parameters}
}
\description{
handler to print the command line
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
