% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_dc.R
\name{houtDC}
\alias{houtDC}
\title{handler to print the command line}
\usage{
houtDC(h, ...)
}
\arguments{
\item{h}{a handler}

\item{...}{Additional parameters}
}
\description{
handler to print the command line
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

