% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_mult.R
\name{houtreport_mig_mult}
\alias{houtreport_mig_mult}
\title{handler to print the command line}
\usage{
houtreport_mig_mult(h = null, ...)
}
\arguments{
\item{h}{a handler}

\item{...}{Additional parameters}
}
\description{
handler to print the command line
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}
