% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanAnnuels.r
\name{hplotBilanAnnuels}
\alias{hplotBilanAnnuels}
\title{plot handler}
\usage{
hplotBilanAnnuels(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
plot handler
}
