% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_species.R
\name{hplotbilesp}
\alias{hplotbilesp}
\title{handler for plot internal use}
\usage{
hplotbilesp(h)
}
\arguments{
\item{h}{Handler}
}
\description{
handler for plot internal use
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

