% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_annual.R
\name{hplotreport_annual}
\alias{hplotreport_annual}
\title{plot handler}
\usage{
hplotreport_annual(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
plot handler
}
\keyword{internal}

