% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_poids_moyen.r
\name{hreg}
\alias{hreg}
\title{Internal handler for reg, class \code{\link{Bilan_poids_moyen-class}}.}
\usage{
hreg(h, ...)
}
\arguments{
\item{h}{handler}

\item{\dots}{additional arguments passed to the function}
}
\description{
Internal handler for reg, class \code{\link{Bilan_poids_moyen-class}}.
}

