% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacomi.r
\name{hsilver}
\alias{hsilver}
\title{handler function used by the main interface}
\usage{
hsilver(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
handler function used by the main interface
}
