% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationInterAnnuelle.r
\name{hsummaryBilanMigrationInterannuelle}
\alias{hsummaryBilanMigrationInterannuelle}
\title{Summary handler internal method}
\usage{
hsummaryBilanMigrationInterannuelle(h, ...)
}
\arguments{
\item{h}{A handler}

\item{...}{Additional parameters}
}
\description{
Summary handler internal method
}
