% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanEspeces.r
\name{hsummarybilesp}
\alias{hsummarybilesp}
\title{handler for summary BilanEspeces, internal use}
\usage{
hsummarybilesp(h)
}
\arguments{
\item{h}{a handler}
}
\description{
handler for summary BilanEspeces, internal use
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

