% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_interannual.R
\name{hsummaryreport_migInterannuelle}
\alias{hsummaryreport_migInterannuelle}
\title{Summary handler internal method}
\usage{
hsummaryreport_migInterannuelle(h, ...)
}
\arguments{
\item{h}{A handler}

\item{...}{Additional parameters}
}
\description{
Summary handler internal method
}
\keyword{internal}

