\name{hTableBilanEspeces}
\alias{hTableBilanEspeces}
\title{handler du calcul BilanEspeces : traitements
appelle les fonctions funstat et funtable pour faire le bilan des migrations
dans des fichiers csv}
\usage{
  hTableBilanEspeces(h, ...)
}
\arguments{
  \item{h}{}

  \item{...}{}
}
\description{
  handler du calcul BilanEspeces : traitements appelle les
  fonctions funstat et funtable pour faire le bilan des
  migrations dans des fichiers csv
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

