\name{hTableBilanMigration}
\alias{hTableBilanMigration}
\title{handler du calcul BilanMigrationstat : traitements...}
\usage{hTableBilanMigration(h, ...)}
\description{handler du calcul BilanMigrationstat : traitements 
appelle les fonctions funstat et funtable pour faire le bilan des migrations
dans des fichiers csv}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{h}{}
\item{...}{}}
