% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationInterAnnuelle.r
\name{htableBilanMigrationInterAnnuelle}
\alias{htableBilanMigrationInterAnnuelle}
\title{Writing of results in datawd}
\usage{
htableBilanMigrationInterAnnuelle(h, ...)
}
\arguments{
\item{h}{A handler}

\item{...}{Additional parameters}
}
\description{
Writing of results in datawd
}

