% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMult.r
\name{hTableBilanMigrationMult}
\alias{hTableBilanMigrationMult}
\title{handler function table}
\usage{
hTableBilanMigrationMult(h = null, ...)
}
\arguments{
\item{h}{a handler}

\item{...}{Additional parameters}
}
\description{
internal use
}
\keyword{internal}

