% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig.R
\name{hTablereport_mig}
\alias{hTablereport_mig}
\title{handler for summary function, internal use
calls functions funstat and funtable to build summary tables in html and
csv files}
\usage{
hTablereport_mig(h, ...)
}
\arguments{
\item{h}{Handler}

\item{...}{Additional parameters}
}
\description{
handler for summary function, internal use
calls functions funstat and funtable to build summary tables in html and
csv files
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}
