% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_mult.R
\name{hTablereport_mig_mult}
\alias{hTablereport_mig_mult}
\title{handler function table}
\usage{
hTablereport_mig_mult(h = null, ...)
}
\arguments{
\item{h}{a handler}

\item{...}{Additional parameters}
}
\description{
internal use
}
\keyword{internal}

