\name{husr}
\alias{husr}
\title{this handler test the connexion and if it works loads the stacomi interface...}
\usage{husr(h, ...)}
\description{this handler test the connexion and if it works loads the stacomi interface}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{h}{}
\item{...}{}}
