% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanAnnuels.r
\name{hxtableBilanAnnuels}
\alias{hxtableBilanAnnuels}
\title{xtable handler}
\usage{
hxtableBilanAnnuels(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
xtable handler
}

