% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_annual.R
\name{hxtablereport_annual}
\alias{hxtablereport_annual}
\title{xtable handler}
\usage{
hxtablereport_annual(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
xtable handler
}
\keyword{internal}
