\name{ind}
\alias{ind}
\title{fonction pour renvoyer les index dans b  des valeurs du vecteur a...}
\usage{ind(a, b)}
\description{fonction pour renvoyer les index dans b  des valeurs du vecteur a
b peut apparaitre plusieurs fois dans a}
\note{attention le vecteur de resultat est dans le desordre}
\value{index of b in a}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{a}{}
\item{b}{}}
