\name{indrepeated}
\alias{indrepeated}
\title{fonction qui retourne l'index des valeurs repetees d'un vecteur...}
\usage{indrepeated(a)}
\description{fonction qui retourne l'index des valeurs repetees d'un vecteur}
\value{the index of repeated values within a vector}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{a}{}}
