\name{induk}
\alias{induk}
\title{fonction qui renvoit l'index des valeurs apparaissant une seule fois...}
\usage{induk(a)}
\description{fonction qui renvoit l'index des valeurs apparaissant une seule fois}
\value{the index unique  values within a vector}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{a}{}}
