% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMult.r
\docType{methods}
\name{initialize,BilanMigrationMult-method}
\alias{initialize,BilanMigrationMult-method}
\title{initialize method for BilanMigrationMult}
\usage{
\S4method{initialize}{BilanMigrationMult}(.Object, ...)
}
\description{
allows a more elaborate constuctor than new through use of charge methods
of Referential objects in the class
}

