% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_Bilan_carlot.r
\name{interface_Bilan_carlot}
\alias{interface_Bilan_carlot}
\title{An interface that calls the object to build the user interface}
\usage{
interface_Bilan_carlot()
}
\description{
An interface that calls the object to build the user interface
}
\note{
always has to be called within a group constructed and deleted using quitte()
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

