% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_BilanAgedemer.r
\name{interface_BilanAgedemer}
\alias{interface_BilanAgedemer}
\title{An interface that calls the object to build the user interface}
\usage{
interface_BilanAgedemer()
}
\description{
An interface that calls the object to build the user interface
}
\note{
always has to be called within a group constructed and deleted using quitte()
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

