% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_BilanAnnuels.r
\name{interface_BilanAnnuels}
\alias{interface_BilanAnnuels}
\title{interface for BilanAnnuels class}
\usage{
interface_BilanAnnuels()
}
\description{
interface for BilanAnnuels class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
