% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_BilanEspeces.r
\name{interface_BilanEspeces}
\alias{interface_BilanEspeces}
\title{Interface for BilanEspece class}
\usage{
interface_BilanEspeces()
}
\description{
Interface for BilanEspece class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

