% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_BilanFonctionnementDC.r
\name{interface_BilanFonctionnementDC}
\alias{interface_BilanFonctionnementDC}
\title{interface for BilanFonctionnementDC class}
\usage{
interface_BilanFonctionnementDC()
}
\description{
interface for BilanFonctionnementDC class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
