\name{interface_BilanFonctionnementDF}
\alias{interface_BilanFonctionnementDF}
\title{interface for BilanFonctionnementDF class}
\usage{
  interface_BilanFonctionnementDF()
}
\description{
  interface for BilanFonctionnementDF class
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

