% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_BilanMigration.r
\name{interface_BilanMigration}
\alias{interface_BilanMigration}
\title{Interface for BilanMigration class}
\usage{
interface_BilanMigration()
}
\description{
Interface for BilanMigration class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

