\name{interface_BilanMigrationInterAnnuelle}
\alias{interface_BilanMigrationInterAnnuelle}
\title{interface for BilanMigrationInterannuelle class}
\usage{
  interface_BilanMigrationInterAnnuelle()
}
\description{
  interface for BilanMigrationInterannuelle class
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

