% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_BilanMigrationMult.r
\name{interface_BilanMigrationMult}
\alias{interface_BilanMigrationMult}
\title{graphical interface for BilanMigrationMult class}
\usage{
interface_BilanMigrationMult()
}
\description{
launches the graphical interface after the main gwidget dialog has been launched
This function is called from a handler in the main graphical interface
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

