% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_BilanMigrationMultConditionEnv.r
\name{interface_BilanMigrationMultConditionEnv}
\alias{interface_BilanMigrationMultConditionEnv}
\title{interface for BilanMigrationConditionEnv class}
\usage{
interface_BilanMigrationMultConditionEnv()
}
\description{
interface for BilanMigrationConditionEnv class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

