% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_BilanMigrationCar.r
\name{interface_BilanMigrationPar}
\alias{interface_BilanMigrationPar}
\title{interface for BilanMigrationPar class}
\usage{
interface_BilanMigrationPar()
}
\description{
interface for BilanMigrationPar class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

