% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_bilan_poids_moyen.r
\name{interface_BilanPoidsMoyen}
\alias{interface_BilanPoidsMoyen}
\title{interface for Bilan_poids_moyen class}
\usage{
interface_BilanPoidsMoyen()
}
\description{
interface for Bilan_poids_moyen class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

