% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_BilanConditionEnv.r
\name{interface_ConditionEnv}
\alias{interface_ConditionEnv}
\title{Interface for class conditionEnv}
\usage{
interface_ConditionEnv()
}
\description{
Interface for class conditionEnv
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

