\name{interface_functions}
\alias{interface_Bilan_stades_pigm}
\alias{interface_BilanEspeces}
\alias{interface_BilanFonctionnementDC}
\alias{interface_BilanFonctionnementDF}
\alias{interface_BilanLot}
\alias{interface_BilanMigration}
\alias{interface_BilanMigrationConditionEnv}
\alias{interface_BilanMigrationInterAnnuelle}
\alias{interface_BilanMigrationPar}
\alias{interface_BilanPoidsMoyen}
\alias{interface_BilanTaille}
\alias{interface_chooselang}
\alias{interface_ConditionEnv}
\title{An interface that calls the object to build the user interface, one for each class}
\usage{
  interface_Bilan_stades_pigm()

}
\description{
  An interface that calls the object to build the user
  interface
}
\note{
  always has to be called within a group constructed and
  deleted using quitte()
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

