\name{interface_graphique}
\alias{interface_graphique}
\title{lancement du programme, cette fonction recupere d'abord les chemins ODBC et le repertoire de travail à partir du fichier XML}
\usage{
  interface_graphique()
}
\description{
  lancement du programme, cette fonction recupere d'abord
  les chemins ODBC et le repertoire de travail à partir du
  fichier XML
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

