% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_report_annual.R
\name{interface_report_annual}
\alias{interface_report_annual}
\title{interface for report_annual class}
\usage{
interface_report_annual()
}
\description{
interface for report_annual class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

