% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_report_dc.R
\name{interface_report_dc}
\alias{interface_report_dc}
\title{interface for report_dc class}
\usage{
interface_report_dc()
}
\description{
interface for report_dc class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

