% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_report_df.R
\name{interface_report_df}
\alias{interface_report_df}
\title{interface for report_df class}
\usage{
interface_report_df()
}
\description{
interface for report_df class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

