% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_report_env.R
\name{interface_report_env}
\alias{interface_report_env}
\title{Interface for class conditionEnv}
\usage{
interface_report_env()
}
\description{
Interface for class conditionEnv
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

