% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_report_ge_weight.R
\name{interface_report_ge_weight}
\alias{interface_report_ge_weight}
\title{interface for report_ge_weight class}
\usage{
interface_report_ge_weight()
}
\description{
interface for report_ge_weight class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

