% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_report_mig.R
\name{interface_report_mig}
\alias{interface_report_mig}
\title{Interface for report_mig class, internal use, this function is called
by a handler in the main graphical interface}
\usage{
interface_report_mig()
}
\description{
Interface for report_mig class, internal use, this function is called
by a handler in the main graphical interface
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}
