% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_report_mig_char.R
\name{interface_report_mig_char}
\alias{interface_report_mig_char}
\title{interface for report_migPar class}
\usage{
interface_report_mig_char()
}
\description{
interface for report_migPar class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}
