% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_report_mig_interannual.R
\name{interface_report_mig_interannual}
\alias{interface_report_mig_interannual}
\title{interface for report_migInterannuelle class}
\usage{
interface_report_mig_interannual()
}
\description{
interface for report_migInterannuelle class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

