% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_report_mig_mult.R
\name{interface_report_mig_mult}
\alias{interface_report_mig_mult}
\title{graphical interface for report_mig_mult class}
\usage{
interface_report_mig_mult()
}
\description{
launches the graphical interface after the main gwidget dialog has been launched
This function is called from a handler in the main graphical interface
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

