% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_report_sea_age.R
\name{interface_report_sea_age}
\alias{interface_report_sea_age}
\title{An interface that calls the object to build the user interface}
\usage{
interface_report_sea_age()
}
\description{
An interface that calls the object to build the user interface
}
\note{
always has to be called within a group constructed and deleted using quitte()
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}
