% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_report_species.R
\name{interface_report_species}
\alias{interface_report_species}
\title{Interface for report_species class}
\usage{
interface_report_species()
}
\description{
Interface for report_species class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}
