\name{killfactor}
\alias{killfactor}
\title{very usefull function used to "kill" these bloody factors that appears, noticeably after loading with odbc...}
\usage{killfactor(df)}
\description{very usefull function used to "kill" these bloody factors that appears, noticeably after loading with odbc}
\value{df}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{df}{a data.frame}}
