% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraries.r
\name{libraries}
\alias{libraries}
\title{function to call and load the libraries used in stacomi}
\usage{
libraries()
}
\description{
function to call and load the libraries used in stacomi
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
function to call and load the libraries used in stacomi

function to call and load the libraries used in stacomi

Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

