% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefDC.r
\docType{methods}
\name{load,RefDC-method}
\alias{load,RefDC-method}
\title{load method for RefDC}
\usage{
\S4method{load}{RefDC}(object, dc)
}
\arguments{
\item{object}{an object of class RefDC}

\item{dc}{a character vector of dc chosen}
}
\description{
the load method is intented to have the same behaviour as choix (which creates a
widget in the graphical interface) but from the command line.  The parameters for dc are transformed to integer as the RefDC only 
takes integer in the dc slots. The method also loads the stations and ouvrages (dams) associated with the counting device (dc).
The values passed to the load function are then checked with the setValidty method.
Finally, if an objectBilan is passed as a parameter, the method will do a charge_avec_filtre to select only the taxa present in the counting devices
}
\examples{
\dontrun{
win=gwindow()
group=ggroup(container=win,horizontal=FALSE)
object=new("RefDC")
object<-charge(object)
objectBilan=new("BilanMigrationMult")
load(object=object,objectBilan=objectBilan,dc=1)}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other load functions: \code{\link{load,RefStades-method}}
}

