% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefStades.r
\docType{methods}
\name{load,RefStades-method}
\alias{load,RefStades-method}
\title{load method for RefStades}
\usage{
\S4method{load}{RefStades}(object, stades)
}
\arguments{
\item{stades}{the vector of stages chosen}
}
\description{
the load method is intented to have the same behaviour as choix (which creates a
widget in the graphical interface) but from the command line. The values passed to the load function
for stades is the code.  Any numeric value will be discarded
}
\examples{
object=new("RefTaxon")
object<-charge(object)
objectBilan=new("BilanMigrationMult")
load(object=object,objectBilan=objectBilan,"Anguilla anguilla")
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other load functions: \code{\link{load,RefDC-method}}
}

