% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefTaxon.r
\docType{methods}
\name{load,RefTaxon-method}
\alias{load,RefTaxon-method}
\title{load method for RefTaxon}
\usage{
\S4method{load}{RefTaxon}(object, taxons)
}
\arguments{
\item{object}{An object from the class RefTaxon}

\item{taxons}{The vector of taxon, can be either code (numeric) or latin name}
}
\description{
the load method is intented to have the same behaviour as choix (which creates a
widget in the graphical interface) but from the command line. The values passed to the load function
for taxon can be either numeric (2038 = Anguilla anguilla) or character.
}
\examples{
object=new("RefTaxon")
object<-charge(object)
objectBilan=new("BilanMigrationMult")
load(object=object,objectBilan=objectBilan,"Anguilla anguilla")
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

