% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacomi.R
\name{load_stacomi}
\alias{load_stacomi}
\title{Internal function, tests the connection and if it works loads the stacomi interface}
\usage{
load_stacomi(...)
}
\arguments{
\item{...}{Other arguments}
}
\value{
Nothing
}
\description{
Internal function, tests the connection and if it works loads the stacomi interface
}
\note{
\code{base} is copied by stacomi into envir_stacomi. Same for \code{database_expected}
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
\keyword{internal}
