\name{messages}
\alias{messages}
\title{Function for creating a list gathering all msgs for internationalization.}
\usage{quitte(...)}
\description{
The message list is saved in the envir_stacomi environment.
Attention all msgs must be starting with a letter "l" as numbers are wrong when used as msgs within a list
}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
