% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_poids_moyen.r
\docType{methods}
\name{model,Bilan_poids_moyen-method}
\alias{model,Bilan_poids_moyen-method}
\alias{model.Bilan_poids_moyen}
\alias{model.bilPM}
\title{model method for Bilan_poids_moyen' 
this method uses samples collected over the season to model the variation in weight of
glass eel or yellow eels.}
\usage{
model(object,model.type=c("seasonal","seasonal1","seasonal2","manual"),silent=FALSE)
}
\arguments{
\item{object}{An object of class \link{Bilan_pois_moyen-class}}
}
\description{
model method for Bilan_poids_moyen' 
this method uses samples collected over the season to model the variation in weight of
glass eel or yellow eels.
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

