\name{monthDays}
\alias{monthDays}
\title{monthDays function imported from depuis Hmisc that cannot be loaded for reasons
of compatibility with xtable}
\usage{
  monthDays(time)
}
\arguments{
  \item{time}{A POSIXt or Date object describing the month or year in question}
}  
\description{
  monthDays function imported from depuis Hmisc 
}
\author{
  Charles Dupont 
}

