% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitaires.r
\name{mygtkProgressBar}
\alias{mygtkProgressBar}
\title{Progress bar using a gtkdialog, the progress bar is assigned in .GlobalEnv
This progress bar has a button to close. Use}
\usage{
mygtkProgressBar(title, progress_text, width = 400, height = 50,
  pulse = TRUE)
}
\arguments{
\item{title}{THe title of the bar}

\item{progress_text}{The text to display for progression}

\item{width}{Width of the progress bar}

\item{height}{Height of the progress bar}

\item{pulse}{Do you want the widget to pulse}
}
\value{
nothing but progress_bar and dialog gwidgets are assigned to .globalEnv
}
\description{
Progress bar using a gtkdialog, the progress bar is assigned in .GlobalEnv
This progress bar has a button to close. Use
}
\examples{
\dontrun{
mygtkProgressBar("Trial","progress text")
fraction_progressed=seq(0,1,length.out=50)
for(i in fraction_progressed) {
     Sys.sleep(0.1)
    progress_bar$setFraction(i)
}
dispose(dialog)
}
}
\author{
cedric.briand
}

