% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generic.r
\name{out}
\alias{out}
\title{Generic method getvalue}
\usage{
out(object, ...)

out(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Additional parms}

\item{object}{Object}

\item{...}{Additional parms}
}
\description{
Generic method getvalue

Generic method getvalue
}
\author{
cedric.briand

cedric.briand
}

