% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasdeTemps.r
\docType{class}
\name{PasDeTemps-class}
\alias{DateFin}
\alias{DateFin,PasDeTemps-method}
\alias{LesPasDeTemps}
\alias{PasDeTemps}
\alias{PasDeTemps-class}
\alias{ValeurPasDeTemps}
\alias{currentDateDebut}
\alias{currentDateDebut,PasDeTemps-method}
\alias{currentDateFin}
\alias{currentDateFin,PasDeTemps-method}
\alias{dureePas}
\alias{getAnnees}
\alias{getAnnees,PasDeTemps-method}
\alias{getLibellesPas}
\alias{getLibellesPas,PasDeTemps-method}
\alias{getdateDebut}
\alias{getdateDebut,PasDeTemps-method}
\alias{getnoPasCourant}
\alias{getnoPasCourant,PasDeTemps-method}
\alias{setdateDebut}
\alias{setdateDebut,PasDeTemps-method}
\alias{suivant}
\alias{suivant,PasDeTemps-method}
\alias{validite_PasDeTemps}
\title{Class "PasDeTemps"}
\description{
Describes a time step
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("PasDeTemps",
dateDebut="POSIXt",dureePas=numeric(),nbPas=numeric(),noPasCourant=integer())}.
\describe{ 
\item{list("dateDebut")}{Object of class \code{"POSIXt"} Starting
date }
\item{:}{Object of class \code{"POSIXt"} Starting date }
\item{list("dureePas")}{Object of class \code{"numeric"} Step length
}\item{:}{Object of class \code{"numeric"} Step length }
\item{list("nbPas")}{Object of class \code{"numeric"} Number of steps
}\item{:}{Object of class \code{"numeric"} Number of steps }
\item{list("noPasCourant")}{Object of class \code{"integer"} Number of the
current step }\item{:}{Object of class \code{"integer"} Number of the
current step } }
}
\examples{

showClass("PasDeTemps")

}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\seealso{
\code{\linkS4class{PasDeTempsJournalier}}
}
\keyword{classes}

