% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_carlot.r
\docType{methods}
\name{plot,Bilan_carlot,missing-method}
\alias{plot,Bilan_carlot,missing-method}
\alias{plot.Bilan_carlot}
\alias{plot.b_carlot}
\alias{plot.bilan_carlot}
\title{Plots of various type for BilanMigration, and performs writing to the database of daily values.}
\usage{
\S4method{plot}{Bilan_carlot,missing}(x, plot.type = "1", silent = FALSE)
}
\arguments{
\item{x}{An object of class Bilan_carlot}

\item{plot.type}{One of "1","violin plot". Defaut to \code{1} , can also be \code{2} boxplot or 
\code{3} points.}

\item{silent}{Stops displaying the messages.}

\item{...}{Additional arguments, see \code{plot}, \code{plot.default} and \code{par}}
}
\description{
Plots of various type for BilanMigration, and performs writing to the database of daily values.
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

