% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_poids_moyen.r
\docType{methods}
\name{plot,Bilan_poids_moyen,missing-method}
\alias{plot,Bilan_poids_moyen,missing-method}
\alias{plot.Bilan_poids_moyen}
\alias{plot.bilPM}
\title{Plot method for Bilan_poids_moyen'}
\usage{
\S4method{plot}{Bilan_poids_moyen,missing}(x, plot.type = "point",
  silent = FALSE)
}
\arguments{
\item{x}{An object of class \link{Bilan_poids_moyen-class}}

\item{plot.type}{Default "1". "1" plot of mean weight of glass eel against the mean date of operation (halfway between start,
and end of operation). The ggplot "p" can be accessed from envir_stacomi using \code{get("p",envir_stacomi)}. "2" standard plot of current coefficent.
"3" same as "1" but with size according to number.}

\item{silent}{Stops displaying the messages.}
}
\description{
Plot method for Bilan_poids_moyen'
}
\note{
the model method provides plots for the fitted models
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

