% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanAgedemer.r
\docType{methods}
\name{plot,BilanAgedemer,missing-method}
\alias{plot,BilanAgedemer,missing-method}
\alias{plot.BilanAgedemer}
\alias{plot.bilan_adm}
\title{Plots of various type for BilanAgedemer}
\usage{
\S4method{plot}{BilanAgedemer,missing}(x, plot.type = "1", silent = FALSE)
}
\arguments{
\item{x}{An object of class \link{BilanAgedemer-class}}

\item{plot.type}{Default "1"
 \itemize{
        \item{plot.type="1"}{density plot by sea age}
        \item{plot.type="2"}{Density plot by sea age and dc}
}}

\item{silent}{Default FALSE, if TRUE the program should no display messages.}
}
\description{
Plots of various type for BilanAgedemer
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

