% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanAnnuels.r
\docType{methods}
\name{plot,BilanAnnuels,missing-method}
\alias{plot,BilanAnnuels,missing-method}
\alias{plot.BilanAnnuels}
\alias{plot.bilA}
\title{Plot method for BilanAnnuels}
\usage{
\S4method{plot}{BilanAnnuels,missing}(x, plot.type = "point",
  silent = FALSE)
}
\arguments{
\item{x}{An object of class \link{BilanAnnuels-class}}

\item{plot.type}{Default point}

\item{silent}{Stops displaying the messages.
\itemize{
        \item{plot.type="point": ggplot+geom_point}'         
}}
}
\description{
Plot method for BilanAnnuels
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
\link{BilanMigrationInterAnnuelle-class} for examples
}

