% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanEspeces.r
\docType{methods}
\name{plot,BilanEspeces,missing-method}
\alias{plot,BilanEspeces,missing-method}
\alias{plot.BilanBilanEspeces}
\alias{plot.bilesp}
\title{Plot method for BilanEspeces}
\usage{
\S4method{plot}{BilanEspeces,missing}(x, plot.type = "pie", color = NULL,
  silent = FALSE)
}
\arguments{
\item{x}{An object of class \link{BilanEspeces-class}}

\item{plot.type}{Default pie
#' \itemize{
        \item{plot.type="pie": A pie}'     
        \item{plot.type="barchart" : A barchart}
}}

\item{color}{Default NULL, a vector of colors of length corresponding to the number of taxa-stage
different values, use unique(bilesp@calcdata$taxon_stade) to get that number. The color applies to both
pie and barchart plots}

\item{silent}{Stops displaying the messages.}
}
\description{
Plot method for BilanEspeces
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

