% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanFonctionnementDC.r
\docType{methods}
\name{plot,BilanFonctionnementDC,ANY-method}
\alias{plot,BilanFonctionnementDC,ANY-method}
\title{Different plots for BilanFonctionnementDC}
\usage{
\S4method{plot}{BilanFonctionnementDC,ANY}(x, y, plot.type = 1,
  silent = FALSE, title = NULL)
}
\arguments{
\item{x}{An object of class \link{BilanFonctionnementDC-class}}

\item{y}{From the formals but missing}

\item{plot.type}{One of \code{barchart},\code{box}. Defaut to \code{barchart} showing
 a summary of the df operation per month, can also be \code{box}, 
a plot with adjacent rectangles.}

\item{silent}{Stops displaying the messages.}

\item{title}{The title of the graph, if NULL a default title will be plotted 
with the number of the DF}
}
\value{
Nothing but prints the different plots
}
\description{
\itemize{
     \item{plot.type=1}{A barplot of the operation time per month}
		\item{plot.type=2}{Barchat giving the time per type of operation }
		\item{plot.type=2}{Rectangle plots drawn along a line}
     \item{plot.type=4}{Plots per day drawn over the period to show the operation of a df, days in x, hours in y} 
	}
}
\note{
The program cuts periods which overlap between two month. 
The splitting of different periods into month is 
assigned to the \code{envir_stacomi} environment
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

