% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigration.r
\docType{methods}
\name{plot,BilanMigration,ANY-method}
\alias{plot,BilanMigration,ANY-method}
\title{Plots of various type for BilanMigration.}
\usage{
\S4method{plot}{BilanMigration,ANY}(x, y, plot.type = "standard",
  silent = FALSE, ...)
}
\arguments{
\item{x}{An object of class BilanMigration}

\item{y}{From the formals but missing}

\item{plot.type}{One of "standard","step". Defaut to \code{standard} the standard BilanMigration with dc and operation displayed, can also be \code{step} or 
\code{multiple}}

\item{silent}{Stops displaying the messages.}

\item{...}{Additional arguments, see \code{plot}, \code{plot.default} and \code{par}}
}
\description{
\itemize{
		\item{plot.type="standard"}{calls \code{\link{fungraph}} and \code{\link{fungraph_civelle}} functions to plot as many "bilanmigration"
			as needed, the function will test for the existence of data for one dc, one taxa, and one stage}
		\item{plot.type="step"}{creates Cumulated graphs for BilanMigrationMult.  Data are summed per day for different dc taxa and stages}
		\item{plot.type="multiple"}{Method to overlay graphs for BilanMigrationMult (multiple dc/taxa/stage in the same plot)}
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

