% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationCar.r
\docType{methods}
\name{plot,BilanMigrationCar,missing-method}
\alias{plot,BilanMigrationCar,missing-method}
\title{plot method for BilanMigrationCar}
\usage{
\S4method{plot}{BilanMigrationCar,missing}(x, color_parm = NULL,
  plot.type = "barplot", ...)
}
\arguments{
\item{x}{An object of class BilanMigrationCar}

\item{plot.type}{One of "qual", "quant" "crossed"}

\item{...}{Additional parameters}

\item{y}{not used there}

\item{color}{A named vector for the colors of either parameters (if plot.type=quant) or levels for
parameters (if plot.type=qual).}
}
\description{
plot method for BilanMigrationCar
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
