% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMultConditionEnv.r
\docType{methods}
\name{plot,BilanMigrationMultConditionEnv,missing-method}
\alias{plot,BilanMigrationMultConditionEnv,missing-method}
\alias{plot.BilanMigrationMultConditionEnv}
\alias{plot.bmmCE}
\title{Plot method for BilanMigrationMultConditionEnv}
\usage{
\S4method{plot}{BilanMigrationMultConditionEnv,missing}(x,
  color_station = NULL, color_dc = NULL, silent = FALSE)
}
\arguments{
\item{x}{An object of class \link{BilanMigrationMultConditionEnv}}

\item{color_station}{A named vector of station color (e.g. c("temp_gabion"="red","coef_maree"="blue","phases_lune"="green")) default null}

\item{color_dc}{A named vector giving the color for each dc default null (e.g. c("5"="#4D4D4D","6"="#E6E6E6","12"="#AEAEAE"))}

\item{silent}{Stops displaying the messages.}
}
\description{
Plot method for BilanMigrationMultConditionEnv
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

