% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationPar.r
\docType{methods}
\name{plot,BilanMigrationPar,missing-method}
\alias{plot,BilanMigrationPar,missing-method}
\title{plot method for BilanMigrationPar}
\usage{
plot(x,plot.type=c("barplot","xyplot","summary"))
}
\arguments{
\item{x}{An object of class BilanMigrationPar}

\item{plot.type}{=barplot, 2=xyplot, 3=summary table}
}
\description{
plot method for BilanMigrationPar
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

