% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_annual.R
\name{plot,report_annual,missing-method}
\alias{plot,report_annual,missing-method}
\alias{plot.report_annual}
\title{Plot method for report_annual}
\usage{
\S4method{plot}{report_annual,missing}(x, plot.type = "point", silent = FALSE)
}
\arguments{
\item{x}{An object of class \link{report_annual-class}}

\item{plot.type}{Default point}

\item{silent}{Stops displaying the messages.
\itemize{
        \item{plot.type="point": ggplot+geom_point}'
}}
}
\value{
No return value, called for side effects
}
\description{
Plot method for report_annual
}
\seealso{
\link{report_mig_interannual-class} for examples
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
