% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_df.R
\docType{methods}
\name{plot,report_df,ANY-method}
\alias{plot,report_df,ANY-method}
\alias{plot.report_df}
\title{Different plots for report_df}
\usage{
\S4method{plot}{report_df,ANY}(x, y, plot.type = 1, silent = FALSE,
  main = NULL)
}
\arguments{
\item{x}{An object of class \link{report_df-class}}

\item{y}{From the formals but missing}

\item{plot.type}{One of \code{barchart},\code{box}. Defaut to \code{barchart} showing a summary of the df operation per month, can also be \code{box}, 
a plot with adjacent rectangles.}

\item{silent}{Stops displaying the messages.}

\item{main}{The title of the graph, if NULL a default title will be plotted with the number of the DF}
}
\value{
Nothing but prints the different plots
}
\description{
\itemize{
     \item{plot.type=1}{A barplot of the operation time per month}
		\item{plot.type=2}{Barchat giving the time per type of operation }
		\item{plot.type=2}{Rectangle plots drawn along a line}
     \item{plot.type=4}{Plots per day drawn over the period to show the operation of a df, days in x, hours in y} 
	}
}
\note{
The program cuts periods which overlap between two month. The splitting of different periods into month is 
assigned to the \code{envir_stacomi} environment
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

