% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_ge_weight.R
\name{plot,report_ge_weight,missing-method}
\alias{plot,report_ge_weight,missing-method}
\alias{plot.report_ge_weight}
\title{Internal handler for reg, class \code{\link{report_ge_weight-class}}.}
\usage{
\S4method{plot}{report_ge_weight,missing}(x, plot.type = 1, silent = FALSE)
}
\arguments{
\item{x}{An object of class \link{report_ge_weight-class}}

\item{plot.type}{Default '1'. '1' plot of mean weight of glass eel against the mean date of operation (halfway between start,
and end of operation). The ggplot 'p' can be accessed from envir_stacomi using \code{get('p',envir_stacomi)}. '2' standard plot of current coefficent.
'3' same as '1' but with size according to number.}

\item{silent}{Stops displaying the messages.}

\item{h}{handler}

\item{\dots}{additional arguments passed to the function}
}
\description{
Internal handler for reg, class \code{\link{report_ge_weight-class}}.
}
\note{
the model method provides plots for the fitted models
}
\author{
Cedric Briand \email{cedric.briand'at'eptb-vilaine.fr}
}
\keyword{Plot}
\keyword{for}
\keyword{internal}
\keyword{method}
\keyword{report_ge_weight'}
