% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_env.R
\name{plot,report_mig_env,missing-method}
\alias{plot,report_mig_env,missing-method}
\alias{plot.report_mig_env}
\title{Plot method for report_mig_env}
\usage{
\S4method{plot}{report_mig_env,missing}(x, color_station = NULL, color_dc = NULL, silent = FALSE)
}
\arguments{
\item{x}{An object of class \link{report_mig_env}}

\item{color_station}{A named vector of station color (e.g. c("temp_gabion"="red","coef_maree"="blue","phases_lune"="green")) default null}

\item{color_dc}{A named vector giving the color for each dc default null (e.g. c("5"="#4D4D4D","6"="#E6E6E6","12"="#AEAEAE"))}

\item{silent}{Stops displaying the messages.}
}
\value{
Nothing, called for its side effect of plotting
}
\description{
Plot method for report_mig_env
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
