% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_mult.R
\name{plot,report_mig_mult,missing-method}
\alias{plot,report_mig_mult,missing-method}
\alias{plot.report_mig_mult}
\title{Plots of various type for report_mig_mult}
\usage{
\S4method{plot}{report_mig_mult,missing}(
  x,
  plot.type = "standard",
  color = NULL,
  color_ope = NULL,
  silent = FALSE,
  ...
)
}
\arguments{
\item{x}{An object of class report_mig_mult}

\item{plot.type}{One of 'standard','step','multiple'. Defaut to \code{standard} the standard report_mig with dc and operation displayed, can also be \code{step} or 
\code{multiple}}

\item{color}{Default NULL, argument passed for the plot.type='standard' method. A vector of color in the following order : (1) working, (2) stopped, (3:7) 1...5 types of operation,
(8:11) numbers, weight, NULL, NULL (if glass eel), (8:11)  measured, calculated, expert, direct observation for other taxa. If null will be set to brewer.pal(12,'Paired')[c(8,10,4,6,1,2,3,5,7)]}

\item{color_ope}{Default NULL, argument passed for the plot.type='standard' method. A vector of color for the operations. Default to brewer.pal(4,'Paired')}

\item{silent}{Stops most messages from being displayed}

\item{...}{Additional arguments passed to matplot or plot if plot.type='standard', see ... in \link{fungraph_glasseel} and \link{fungraph}}
}
\value{
Nothing, called for its side effect of plotting
}
\description{
\describe{
  \item{plot.type='standard'}{calls \code{\link{fungraph}} and \code{\link{fungraph_glasseel}} functions to plot as many 'report_mig'
  as needed, the function will test for the existence of data for one dc, one taxa, and one stage}
  \item{plot.type='step'}{creates Cumulated graphs for report_mig_mult.  Data are summed per day for different dc taxa and stages}
  \item{plot.type='multiple'}{Method to overlay graphs for report_mig_mult (multiple dc/taxa/stage in the same plot)}
}
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
