% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_sample_char.R
\name{plot,report_sample_char,missing-method}
\alias{plot,report_sample_char,missing-method}
\alias{plot.report_sample_char}
\title{Plots of various type for reportcarlot}
\usage{
\S4method{plot}{report_sample_char,missing}(x, plot.type = "1", silent = FALSE)
}
\arguments{
\item{x}{An object of class report_sample_char}

\item{plot.type}{One of '1','violin plot'. Defaut to \code{1} , can also be \code{2} boxplot or 
\code{3} points.}

\item{silent}{Stops displaying the messages}
}
\value{
Nothing, called for its side effect, plotting
}
\description{
Plots of various type for reportcarlot
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
