% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_sea_age.R
\name{plot,report_sea_age,missing-method}
\alias{plot,report_sea_age,missing-method}
\alias{plot.report_sea_age}
\title{Plots of various type for report_sea_age}
\usage{
\S4method{plot}{report_sea_age,missing}(x, plot.type = "1", silent = FALSE)
}
\arguments{
\item{x}{An object of class \link{report_sea_age-class}}

\item{plot.type}{Default "1"
 \describe{
        \item{plot.type="1"}{density plot by sea age}
        \item{plot.type="2"}{Density plot by sea age and dc}
}}

\item{silent}{Default FALSE, if TRUE the program should no display messages.}
}
\value{
Nothing, called for its side effect of plotting
}
\description{
Plots of various type for report_sea_age
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
