% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_silver_eel.R
\name{plot,report_silver_eel,missing-method}
\alias{plot,report_silver_eel,missing-method}
\alias{plot.report_silver_eel}
\title{Plots of various type for report_silver_eel}
\usage{
\S4method{plot}{report_silver_eel,missing}(x, plot.type = c("1", "2", "3", "4"), silent = FALSE)
}
\arguments{
\item{x}{An object of class \link{report_silver_eel-class}}

\item{plot.type}{Default "1"
 \describe{
        \item{plot.type="1"}{Lattice plot of Durif's stages according to Body Length and Eye Index (average of vertical and horizontal diameters).
If several DC are provided then a comparison of data per dc is provided}
        \item{plot.type="2"}{Lattice plot giving a comparison of Durif's stage proportion over time, if several DC are provided an annual comparison
is proposed, if only one DC is provided then the migration is split into month.}
        \item{plot.type="3"}{ Series of graphs showing  mean Fulton's coefficient, Pankhurst eye index,    along
with a size weight analysis and regression using robust regression (rlm more robust to the presence of outliers)}
            \item{plot.type="4"}{ Lattice cloud plot of Pankurst~ Body Length ~ weight)}
}}

\item{silent}{Stops displaying the messages}
}
\value{
A lattice xy.plot if \code{plot.type =1}, a lattice barchart if \code{plot.type=2}, nothing but plots a series of graphs in 
a single plot if \code{plot.type=3}, a lattice cloud object if \code{plot.type=4}
}
\description{
Plots of various type for report_silver_eel
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
