% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_species.R
\name{plot,report_species,missing-method}
\alias{plot,report_species,missing-method}
\alias{plot.reportreport_species}
\title{Plot method for report_species}
\usage{
\S4method{plot}{report_species,missing}(x, plot.type = "pie", color = NULL, silent = FALSE)
}
\arguments{
\item{x}{An object of class \link{report_species-class}}

\item{plot.type}{Default pie
#' \describe{
  \item{plot.type='pie'}{A pie}' 
  \item{plot.type='barchart'}{A barchart}
}}

\item{color}{Default NULL, a vector of colors of length corresponding to the number of taxa-stage
different values, use unique(bilesp@calcdata$taxa_stage) to get that number. The color applies to both
pie and barchart plots}

\item{silent}{Stops displaying the messages}
}
\value{
Nothing, called for producing plots
}
\description{
Plot method for report_species
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
