% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMult.r
\docType{methods}
\name{plot1,BilanMigrationMult-method}
\alias{plot1,BilanMigrationMult-method}
\title{Method to overlay graphs for BilanMigrationMult (multiple dc/taxa/stage in the same plot)}
\usage{
\S4method{plot1}{BilanMigrationMult}(object, ...)
}
\arguments{
\item{...}{}
}
\description{
Tests are done to get the structure of
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

