% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_carlot.r
\docType{methods}
\name{print,Bilan_carlot-method}
\alias{print,Bilan_carlot-method}
\title{Method to print the command line of the object}
\usage{
\S4method{print}{Bilan_carlot}(x, ...)
}
\arguments{
\item{x}{An object of class Bilan_carlot}

\item{...}{Additional parameters passed to print}
}
\description{
Method to print the command line of the object
}
\author{
cedric.briand
}

