% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanAgedemer.r
\docType{methods}
\name{print,BilanAgedemer-method}
\alias{print,BilanAgedemer-method}
\title{Method to print the command line of the object}
\usage{
\S4method{print}{BilanAgedemer}(x, ...)
}
\arguments{
\item{x}{An object of class BilanAgedemer}

\item{...}{Additional parameters passed to print}
}
\description{
Method to print the command line of the object
}
\author{
cedric.briand
}
