% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanArgentee.r
\docType{methods}
\name{print,BilanArgentee-method}
\alias{print,BilanArgentee-method}
\title{Method to print the command line of the object}
\usage{
\S4method{print}{BilanArgentee}(x, ...)
}
\arguments{
\item{x}{An object of class BilanArgentee}

\item{...}{Additional parameters passed to print}
}
\description{
Method to print the command line of the object
}
\author{
cedric.briand
}
