% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_df.R
\name{print,report_df-method}
\alias{print,report_df-method}
\alias{print.report_df}
\title{handler to print the command line}
\usage{
\S4method{print}{report_df}(x, ...)
}
\arguments{
\item{x}{An object of class report_df}

\item{...}{Additional parameters passed to print}

\item{h}{a handler}
}
\description{
handler to print the command line
}
\author{
cedric.briand
}
\keyword{Method}
\keyword{command}
\keyword{internal}
\keyword{line}
\keyword{object}
\keyword{of}
\keyword{print}
\keyword{the}
\keyword{to}
