% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_mult.R
\name{print,report_mig_mult-method}
\alias{print,report_mig_mult-method}
\alias{print.report_mig_mult}
\title{Method to print the command line of the object}
\usage{
\S4method{print}{report_mig_mult}(x, ...)
}
\arguments{
\item{x}{An object of class report_mig_mult}

\item{...}{Additional parameters passed to print}
}
\description{
Method to print the command line of the object
}
\author{
cedric.briand
}
