% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_sample_char.R
\name{print,report_sample_char-method}
\alias{print,report_sample_char-method}
\alias{print.report_sample_char}
\title{Method to print the command line of the object}
\usage{
\S4method{print}{report_sample_char}(x, ...)
}
\arguments{
\item{x}{An object of class report_sample_char}

\item{...}{Additional parameters passed to print}
}
\description{
Method to print the command line of the object
}
\author{
cedric.briand
}
