% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_sea_age.R
\name{print,report_sea_age-method}
\alias{print,report_sea_age-method}
\alias{print.report_sea_age}
\title{Method to print the command line of the object}
\usage{
\S4method{print}{report_sea_age}(x, ...)
}
\arguments{
\item{x}{An object of class report_sea_age}

\item{...}{Additional parameters passed to print}
}
\description{
Method to print the command line of the object
}
\author{
cedric.briand
}
