% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_silver_eel.R
\name{print,report_silver_eel-method}
\alias{print,report_silver_eel-method}
\alias{print.report_silver_eel}
\title{Method to print the command line of the object}
\usage{
\S4method{print}{report_silver_eel}(x, ...)
}
\arguments{
\item{x}{An object of class report_silver_eel}

\item{...}{Additional parameters passed to print}
}
\value{
NULL, prints data in the console
}
\description{
Method to print the command line of the object
}
\author{
cedric.briand
}
