% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{progress_bar}
\alias{progress_bar}
\title{Progress bar using a gtkdialog, the progress bar is assigned in envir_stacomi
This progress bar has a button to close.}
\usage{
progress_bar(title, progress_text, width = 400, height = 50, pulse = TRUE)
}
\arguments{
\item{title}{The title of the bar}

\item{progress_text}{The text to display for progression}

\item{width}{Width of the progress bar}

\item{height}{Height of the progress bar}

\item{pulse}{Do you want the widget to pulse}
}
\value{
nothing
}
\description{
Progress bar using a gtkdialog, the progress bar is assigned in envir_stacomi
This progress bar has a button to close.
}
\note{
The name of the progress bar is \code{progres}, it will be assigned in envir_stacomi,
it contains a progress bar widget named progress bar, also assigned in envir_stacomi See example for use.
}
\examples{
\dontrun{
progress_bar(title="Trial",progress_text="progress text")
fraction_progressed=seq(0,1,length.out=50)
 progress_bar<-get("progress_bar",envir_stacomi)
for(i in fraction_progressed){ 
     Sys.sleep(0.1)
   progress_bar$setText(sprintf("\%d\%\% progression",round(100*i)))
    progress_bar$setFraction(i)
}
dispose(progres)
}
}
\author{
cedric.briand
}

