\name{quitte}
\alias{quitte}
\title{function used to clean the objects whithin the group and the graphes...}
\usage{quitte(...)}
\description{function used to clean the objects whithin the group and the graphes
and also elements remaining in the envir_stacomi environment}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{...}{}}
