% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_ann}
\alias{r_ann}
\title{Annual migration of yellow and silver eel for three fishways / counting devices at the
Arzal dam (data from 1995 to 2016)}
\format{
An object of class \link{report_annual-class} with data slot loaded.
}
\usage{
r_ann
}
\description{
The dataset corresponds to the three fishways located on the Arzal dam, filled with annual data
}
\keyword{data}
