% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_ann_adour}
\alias{r_ann_adour}
\title{Annual migration of salmon in the Adour and tributaries}
\format{
An object of class \link{report_annual-class} with data slot loaded.
}
\usage{
r_ann_adour
}
\description{
The dataset corresponds to the fishways DC=33:40 of the Adour for adult migrant salmons
from 1996 to 2005 (annual counts). It has been kindly provided as an example set by the Migradour
association.
}
\keyword{data}
