% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_dc}
\alias{r_dc}
\title{Counting Device (DC) operation from 2000 to 2015 at the Arzal dam (Vilaine, France)}
\format{
An object of class report_dc with 4 slots:
\describe{
  \item{data}{ A dataframe with 544 obs. of  7 variables
  \describe{
    \item{per_dis_identifiant}{The number of the DC}
    \item{per_date_debut}{Starting time a POSIXct}
    \item{per_date_fin }{Ending time a POSIXct}
    \item{ope_dic_identifiant}{DC id}
    \item{per_commentaires }{A comment}
    \item{per_etat_fonctionnement}{Integer 1= working, 0 not working}
    \item{per_tar_code}{The type of operation ('1'=normal operation,
             '2'=Device stopped in normal operation (e.g. the trap is disactivated for the duration of the
    fish sorting and counting by operators),
   '3'='Stopped for maintenance or other problem',
             '4'='Works but not fully operational, i.e. the camera is not working properly because of high turbidity...',
             '5'='Not known')}
    \item{libelle}{label corresponding to per_tar_code}
           }
       }
  \item{df}{the \code{ref_dc} object with 3 slots filled with data corresponding to the iav postgres schema}
  \item{horodatedebut}{the \code{ref_horodate} with horodate set for starting date}
  \item{horodatefin}{the \code{ref_horodate} with horodate set for ending date}   
}
}
\usage{
r_dc
}
\description{
This data corresponds to the data collected at the vertical slot fishway camera
from 2000 to 2015. It represents an object of class \link{report_dc-class} 
with data loaded
}
\keyword{data}
