% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_df}
\alias{r_df}
\title{Overview of the fishway operation at Arzal in (Vilaine France).}
\format{
An object of class report_df with 4 slots:
\describe{
  \item{data}{ A dataframe with 4261 obs. of  7 variables
  \describe{
    \item{per_dis_identifiant}{The number of the DF}
    \item{per_date_debut}{Starting time a POSIXct}
    \item{per_date_fin }{Ending time a POSIXct}
    \item{ope_dic_identifiant}{DF id}
    \item{per_commentaires }{A comment}
    \item{per_etat_fonctionnement}{Integer 1= working, 0 not working}
    \item{per_tar_code}{The type of operation ('1'=normal operation,
             '2'=Device stopped in normal operation (ie lift ascending, high tide...),
   '3'='Stopped for maintenance or other problem',
             '4'='Works but not fully operational,i.e.flow problem, flood, clogged with debris...',
             '5'='Not known')}
    \item{libelle}{label corresponding to per_tar_code}
           }
       }
  \item{df}{the \code{ref_df} object with 3 slots filled with data corresponding to the iav postgres schema}
  \item{horodatedebut}{the \code{ref_horodate} with horodate set for starting date}
  \item{horodatefin}{the \code{ref_horodate} with horodate set for ending date}'   
}
}
\usage{
r_df
}
\description{
This dataset corresponds to the data collected at the vertical slot fishway
in 2015, the fishway is working daily with a cycle depending on tide. This dataset
is used to show an example of detailed output for an object of class \link{report_df-class} with data loaded
}
\keyword{data}
