% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_gew}
\alias{r_gew}
\title{Wet weight of glass eel from the trapping ladder (Arzal, Vilaine France)}
\format{
An object of class \code{report_ge_weight} of length 1.
}
\usage{
r_gew
}
\description{
Data correspond to glass eel collected in the Vilaine at the trapping ladder (Arzal, France). The years selected are 2009 to 2012,
the query used in the \link{report_ge_weight-class} loads from 2008-08-01 to 2012-08-01
Glass eel are too numerous to be counted. They are weighted and in the stacomi database,
a table with daily coefficients (in  N glass eel/g) to transform weight into number.
The weight is called a 'wet weight' as we don't wan't to drain any of the mucus in glass eel
when weighting them. Samples of 50 to 200 glass eel are weighted and then counted to provide an idea of
the seasonal evolution of wet weight.
}
\keyword{datasets}
