% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_mig}
\alias{r_mig}
\title{Video counting of Marine lamprey (Petromyzon marinus) in 2012 in the Vilaine (France)}
\format{
An object of class report_mig with 8 slots:
\describe{
  \item{dc}{the \code{ref_dc} object with 4 slots filled with data corresponding to the iav postgres schema}
  \item{taxa}{the \code{ref_taxa} the taxa selected}
  \item{stage}{the \code{ref_stage} the stage selected}
  \item{timestep}{the \code{ref_timestep_daily} calculated for all 2015}
  \item{data}{ A dataframe with 10304 rows and 11 variables
  \describe{
    \item{ope_identifiant}{operation id}
    \item{lot_identifiant}{sample id}
    \item{lot_identifiant}{sample id}
    \item{ope_dic_identifiant}{dc id}
    \item{lot_tax_code}{species id}
    \item{lot_std_code}{stage id}
    \item{value}{the value}
    \item{type_de_quantite}{either effectif (number) or poids (weights)}
    \item{lot_dev_code}{destination of the fishes}
    \item{lot_methode_obtention}{method of data collection, measured, calculated...} 
    }
  }
  \item{coef_conversion}{A data frame with 0 observations : no quantity are reported for video recording of mullets, only numbers}
  \item{time.sequence}{A time sequence generated for the report, used internally}
}
}
\usage{
r_mig
}
\description{
This dataset corresponds to the data collected at the vertical slot fishway
in 2012, video recording marine lamprey migration
}
\keyword{data}
