% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_mig_char}
\alias{r_mig_char}
\title{Qualitative and quantitative parameters describing Salmon migration at Decize (Loire)}
\format{
An object of class \link{report_mig_char-class} with data slot loaded:
\describe{
  \item{calcdata}{slot to be filled with the calcule method}
  \item{data}{ A list of 2 elements
  \describe{
    \item{parqual}{values of all the qualitative parameters}
    \item{parquan}{values of all the quantitative parameters}
    }
  }
 \item{dc}{the \code{ref_dc} : the control devices selected}
 \item{taxa}{the \code{ref_taxa} : Salmo salar selected}
 \item{stage}{the \code{ref_stage} : the stages selected}
 \item{par}{an object of class \link{ref_par-class}: the parameters used}
 \item{horodatedebut}{an object of class \code{ref_horodate-class} : the start date selected}
 \item{horodatefin}{an object of class \code{ref_horodate-class} : the end date selected}
}
}
\usage{
r_mig_char
}
\description{
The dataset corresponds to the characteristics (qualitative and quantitative) of salmo salar migrating
at Decize (Loire river) and Vichy (Allier river) counting device in 2012. It has been loaded as 
an example for the report_mig_char-class and kindly provided by Loire Grands Migrateurs (LOGRAMI).
}
\keyword{data}
