% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_mig_dc}
\alias{r_mig_dc}
\title{Counting device operation for the video recording (Arzal dam, Vilaine, France).}
\format{
An object of class \link{report_dc-class}
}
\usage{
r_mig_dc
}
\description{
This dataset corresponds to the data collected in the vertical slot fishway for the video 
 recording operation. It is loaded along with \link{r_mig} to
demonstrate the use of the \link{report_mig-class} when the database is not loaded
}
\keyword{data}
