% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_mig_df}
\alias{r_mig_df}
\title{Fishway operation for the vertical slot fishway (Arzal dam, Vilaine, France).}
\format{
An object of class \link{report_df-class}
}
\usage{
r_mig_df
}
\description{
This dataset corresponds to the data collected at in the vertical slot fishway
it is loaded along with \link{r_mig} and used to demonstrate the \link{report_mig-class}
when the database is not installed.
}
\keyword{data}
