% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_mig_env}
\alias{r_mig_env}
\title{An object of class report_mig_env with data loaded}
\format{
An object of class \link{report_env-class} with data slot loaded:
\describe{
  \item{report_mig_mult}{An object of class \link{report_mig_mult-class}}
  \item{report_env}{ An object of class \link{report_env-class}}#' 
  }
}
\usage{
r_mig_env
}
\description{
The dataset correspond to data loaded for the Arzal dam (Vilaine) in 2008, two quantitative
parameters (temperature and tide coefficient) and a qualitative parameter (moon phase) are loaded.
}
\keyword{data}
