% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_mig_interannual}
\alias{r_mig_interannual}
\title{Daily glass eel and elver migration from 1984 to 2016 in the Sevre Niortaise}
\format{
An object of class \link{report_mig_interannual-class} with data loaded.
}
\usage{
r_mig_interannual
}
\description{
The first eel trapping ladder in France was built by Antoine Legault and the team from Rennes
in the Sevre Niortaise, Marais Poitevin. Also refurbished several times since 1984 it has been 
operational at the same location and provides one of the longest series of eel migration. 
For this reason,
the dataset has been loaded as an example for the report_mig_interannual-class. It has been
kindly provided by the parc du Marais Poitevin. The stage corresponds to small eels (elvers)
less than 150 mmm stage name 'PANG'
}
\keyword{data}
