% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_mig_mult}
\alias{r_mig_mult}
\title{Anguilla migration at the Arzal station (report_mig_mult-class)}
\format{
An object of class report_mig_mult with slots:
\describe{
  \item{dc}{the \code{ref_dc} object filled with data}
  \item{taxa}{the \code{ref_taxa} object filled in with data corresponding to dc}
  \item{stage}{the \code{ref_stage} object filled in with data corresponding to dc, and taxa}
  \item{timestep}{the \code{ref_timestep_daily} calculated for all 2011}
  \item{data}{ A dataframe with 400 rows and 11 variables
  \describe{
    \item{ope_identifiant}{operation id}
    \item{lot_identifiant}{sample id}
    \item{lot_identifiant}{sample id}
    \item{ope_dic_identifiant}{dc id}
    \item{lot_tax_code}{species id}
    \item{lot_std_code}{stage id}
    \item{value}{the value}
    \item{type_de_quantite}{either effectif (number) or poids (weights)}
    \item{lot_dev_code}{destination of the fishes}
    \item{lot_methode_obtention}{method of data collection, measured, calculated...} 
    }
  }
  \item{calcdata}{slot to be filled with the calcule method} 
  \item{coef_conversion}{A data frame with 364 observations with daily coefficients to convert from weight to numbers}
  \item{time.sequence}{A time sequence generated for the report, used internally by the object}
}
}
\usage{
r_mig_mult
}
\description{
This data corresponds to data collected from three fishways
and correspond to the migration station at Arzal (Vilaine estuary, France) in 2011 for 
three continental stages of eel (Anguilla anguilla) : glass eel, yellow eel
and silver eel.
}
\keyword{data}
