% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_mig_mult_df}
\alias{r_mig_mult_df}
\title{Fishway operation at the Arzal Dam (Vilaine France) (3 Fishways in 2011)}
\format{
An object of class report_df  \link{report_df-class}
}
\usage{
r_mig_mult_df
}
\description{
This dataset corresponds to the data collected at three different fishways
it is loaded along with \link{r_mig_mult} and used in demonstration for the
\link{report_mig_mult-class}
}
\keyword{data}
