% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_mig_mult_ope}
\alias{r_mig_mult_ope}
\title{Counting operations for three different counting device in Arzal (Vilaine, France)}
\format{
An object of class report_ope
}
\usage{
r_mig_mult_ope
}
\description{
This dataset corresponds to the data collected at three different control devices
It is an object of class \link{report_ope-class} with data loaded.
it is loaded along with \link{r_mig_mult}
}
\keyword{data}
