% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_mig_ope}
\alias{r_mig_ope}
\title{An object of class \link{report_ope-class} with data loaded}
\format{
An object of class report_ope
}
\usage{
r_mig_ope
}
\description{
This dataset corresponds to the data collected at the vertical slot fishway in Arzal (Vilaine river
estuary, France). The operation of the fishway is dependent on tide and is recorded every 10 minutes. This dataset has
to be loaded along with \link{r_mig} to demonstrate the use of the \link{report_mig-class}
}
\keyword{data}
