% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_sample_char}
\alias{r_sample_char}
\title{Size of yellow and glass eel at the Arzal dam (Vilaine, France) in the fishway and main eel trapping ladder.}
\format{
An object of class \link{report_sample_char-class}
}
\usage{
r_sample_char
}
\description{
This dataset corresponds to the data collected at two different control devices
at the Arzal control station (see example in \link{report_sample_char-class}), all body size 
parameters (total size, size converted from pixel in video control) are used in example
}
\keyword{data}
