% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_seaa}
\alias{r_seaa}
\title{An object of class report_sea_age with data loaded}
\format{
An object of class \link{report_sea_age-class} with 8 slots:
\describe{
  \item{dc}{the \code{ref_dc} : the control devices selected}
  \item{taxa}{the \code{ref_taxa} : Salmo salar selected}
  \item{stage}{the \code{ref_stage} : the stages selected}
  \item{par}{Object of class \link{ref_par-class}: the parameters used}
\item{horodatedebut}{object of class \code{ref_horodate-class} : the start date selected}
\item{horodatefin}{object of class \code{ref_horodate-class} : the end date selected} 
  \item{limit1hm}{The size limit, in mm between 1 sea winter fishes and 2 sea winter fishes}
  \item{limit2hm}{The size limit, in mm between 2 sea winter fishes and 3 sea winter fishes}
  \item{data}{ A dataframe with 898  rows and 20 variables
  \describe{
    \item{ope_identifiant}{operation id}
    \item{lot_identifiant}{sample id}
    \item{ope_dic_identifiant}{dc id}
    \item{ope_date_debut}{start date}
    \item{ope_date_fin}{end date}
    \item{lot_effectif}{number of fishes}
    \item{lot_tax_code}{species id}
    \item{lot_std_code}{stages id}
    \item{tax_nom_latin}{species latin names}
    \item{std_libelle}{stages names} 
    \item{dev_code}{destination of the fishes id}
    \item{dev_libelle}{destination of the fishes names}
    \item{par_nom}{parameter name}
    \item{car_par_code}{parameter id}
    \item{car_methode_obtention}{method of data collection, measured, calculated...}
    \item{car_valeur_quantitatif}{the value of the parameter}
    }
  }
 
}
}
\usage{
r_seaa
}
\description{
This dataset corresponds to the data collected at Vichy (left and right bank fishways) and Decize-Saint 
Leger des Vignes fishways (respectively on the Allier and Loire river, France) in 2012 on the size structure of Salmo salar.
It has been kindly provided by the Loire Grands Migrateurs (LOGRAMI) association.
}
\keyword{data}
