% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_silver}
\alias{r_silver}
\title{Silver eel migration in the Somme}
\format{
An object of class \link{report_silver_eel-class} with data slot loaded.
}
\usage{
r_silver
}
\description{
The dataset corresponds to the silver eel traps ('anguilleres) for 2015-2016.
This dataset has been kindly provided by the Federation de Peche de la Somme,
given the upstream location of the trap, most individuals are female
}
\keyword{data}
