% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_checkbox.R
\docType{class}
\name{ref_checkbox-class}
\alias{ref_checkbox-class}
\title{ref_checkbox referencial class}
\description{
referential class allows to choose for several parms with checkbox
}
\section{Slots}{

\describe{
\item{\code{title}}{A "character", the title of the box giving the possible choices}

\item{\code{labels}}{The logical parameters choice}

\item{\code{checked}}{A boolean vector}
}}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("ref_checkbox", ...)}.
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other referential.objects: \code{\link{charge,ref_choice-method}},
  \code{\link{ref_choice-class}},
  \code{\link{ref_coe-class}}, \code{\link{ref_dc-class}},
  \code{\link{ref_df-class}},
  \code{\link{ref_horodate-class}},
  \code{\link{ref_list-class}},
  \code{\link{ref_par-class}},
  \code{\link{ref_parqual-class}},
  \code{\link{ref_parquan-class}},
  \code{\link{ref_period-class}},
  \code{\link{ref_stage-class}},
  \code{\link{ref_taxa-class}},
  \code{\link{ref_year-class}}
}

