% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_choice.R
\docType{class}
\name{ref_choice-class}
\alias{ref_choice-class}
\title{Class 'ref_choice'}
\description{
ref_choice referential class allows to choose within several values with
radiobuttons interface
}
\section{Slots}{

\describe{
\item{\code{listechoice}}{A character vector giving possible choices}

\item{\code{label}}{A character, title of the box giving the possible choices}

\item{\code{selected}}{An \code{Integer}  the initial selected value (as an index), first=1 used in gradio}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new('ref_choice', listechoice=character() ,label=character()
,selected=integer())}.
}

\seealso{
Other referential objects: 
\code{\link{charge,ref_choice-method}},
\code{\link{ref_coe-class}},
\code{\link{ref_dc-class}},
\code{\link{ref_df-class}},
\code{\link{ref_horodate-class}},
\code{\link{ref_list-class}},
\code{\link{ref_par-class}},
\code{\link{ref_parqual-class}},
\code{\link{ref_parquan-class}},
\code{\link{ref_stage-class}},
\code{\link{ref_taxa-class}},
\code{\link{ref_year-class}}
}
\author{
cedric.briand@eptb-vilaine.fr
}
\concept{referential objects}
