% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_dc.R
\docType{class}
\name{ref_dc-class}
\alias{ref_dc-class}
\title{Class 'ref_dc'}
\description{
Description of a control device.
}
\section{Slots}{

\describe{
\item{\code{dc_selected}}{Object of class \code{'integer'}, The selected device}

\item{\code{ouvrage}}{Object of class \code{'integer'}, the attached dam}

\item{\code{station}}{Object of class \code{'character'}, the attached migration monitoring station, this is necessary to join the
table of escapements calculated at the station level.}

\item{\code{data}}{Object of class \code{'data.frame'} data pertaining to the control device}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new('ref_dc', dc_selected=integer(), ouvrage=integer(),
data=data.frame())}.
}

\seealso{
Other referential objects: 
\code{\link{charge,ref_choice-method}},
\code{\link{ref_choice-class}},
\code{\link{ref_coe-class}},
\code{\link{ref_df-class}},
\code{\link{ref_horodate-class}},
\code{\link{ref_list-class}},
\code{\link{ref_par-class}},
\code{\link{ref_parqual-class}},
\code{\link{ref_parquan-class}},
\code{\link{ref_stage-class}},
\code{\link{ref_taxa-class}},
\code{\link{ref_year-class}}
}
\author{
cedric.briand@eptb-vilaine.fr
}
\concept{referential objects}
\keyword{classes}
