% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_df.R
\docType{class}
\name{ref_df-class}
\alias{ref_df-class}
\title{Class 'ref_df'}
\arguments{
\item{df_selected}{Object of class \code{'integer'} The identifier of the fishway}

\item{ouvrage}{Object of class \code{'integer'} The attached dam}

\item{data}{Object of class \code{'data.frame'} Data concerning the fishway}
}
\description{
Representation of a fishway, contains description data of all fishways from
the database along with the selected fishways (df) (integer)
Objects from the Class: Objects can be created by calls of the form
\code{new('ref_df', df_selected=integer(), ouvrage=integer(),
data=data.frame())}.
}
\seealso{
Other referential objects: 
\code{\link{charge,ref_choice-method}},
\code{\link{ref_choice-class}},
\code{\link{ref_coe-class}},
\code{\link{ref_dc-class}},
\code{\link{ref_horodate-class}},
\code{\link{ref_list-class}},
\code{\link{ref_par-class}},
\code{\link{ref_parqual-class}},
\code{\link{ref_parquan-class}},
\code{\link{ref_stage-class}},
\code{\link{ref_taxa-class}},
\code{\link{ref_year-class}}
}
\author{
cedric.briand@eptb-vilaine.fr
}
\concept{referential objects}
