% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_env.R
\docType{class}
\name{ref_env-class}
\alias{ref_env-class}
\title{Class 'ref_env'}
\description{
Enables to load measure stations and to select one of them
}
\section{Slots}{

\describe{
\item{\code{dataframe}}{Data concerning the
measure station}

\item{\code{env_selected}}{The selected measure station}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new('ref_env', ...)}.
}

\author{
cedric.briand@eptb-vilaine.fr
}
\keyword{classes}
