% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_horodate.R
\docType{class}
\name{ref_horodate-class}
\alias{ref_horodate-class}
\title{Class ref_horodate}
\description{
choice of date with method to show current and previous year
}
\section{Slots}{

\describe{
\item{\code{horodate}}{a "POSIXt"}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("ref_horodate", \dots{})}.
}

\seealso{
Other referential objects: 
\code{\link{charge,ref_choice-method}},
\code{\link{ref_choice-class}},
\code{\link{ref_coe-class}},
\code{\link{ref_dc-class}},
\code{\link{ref_df-class}},
\code{\link{ref_list-class}},
\code{\link{ref_par-class}},
\code{\link{ref_parqual-class}},
\code{\link{ref_parquan-class}},
\code{\link{ref_stage-class}},
\code{\link{ref_taxa-class}},
\code{\link{ref_year-class}}
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
\concept{referential objects}
