% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_list.R
\docType{class}
\name{ref_list-class}
\alias{ref_list-class}
\title{Class 'ref_list'}
\arguments{
\item{liste}{choice='character' A vector of character to choose within a droplist}

\item{label='character'}{the title of the box}
}
\description{
Enables to load a 'ref_list' object from a list given by a 'report' object
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new('ref_list', listechoice, label)}.  \describe{
\item{list('listechoice')}{Object of class \code{'character'}}\item{:}{Object
of class \code{'character'}} \item{list('label')}{Object of class
\code{'character'}}\item{:}{Object of class \code{'character'}} }
}

\seealso{
Other referential objects: 
\code{\link{charge,ref_choice-method}},
\code{\link{ref_choice-class}},
\code{\link{ref_coe-class}},
\code{\link{ref_dc-class}},
\code{\link{ref_df-class}},
\code{\link{ref_horodate-class}},
\code{\link{ref_par-class}},
\code{\link{ref_parqual-class}},
\code{\link{ref_parquan-class}},
\code{\link{ref_stage-class}},
\code{\link{ref_taxa-class}},
\code{\link{ref_year-class}}
}
\author{
cedric.briand@eptb-vilaine.fr
}
\concept{referential objects}
\keyword{internal}
