% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_par.R
\docType{class}
\name{ref_par-class}
\alias{ref_par-class}
\title{Class 'ref_par'}
\description{
Class enabling to load the list of parameters and select one of them
}
\section{Slots}{

\describe{
\item{\code{data}}{A data.frame}

\item{\code{par_selected}}{A character vector corresponding to par_code}

\item{\code{data='data.frame'}}{the list of parameters}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
}

\seealso{
Other referential objects: 
\code{\link{charge,ref_choice-method}},
\code{\link{ref_choice-class}},
\code{\link{ref_coe-class}},
\code{\link{ref_dc-class}},
\code{\link{ref_df-class}},
\code{\link{ref_horodate-class}},
\code{\link{ref_list-class}},
\code{\link{ref_parqual-class}},
\code{\link{ref_parquan-class}},
\code{\link{ref_stage-class}},
\code{\link{ref_taxa-class}},
\code{\link{ref_year-class}}
}
\author{
cedric.briand@eptb-vilaine.fr
}
\concept{referential objects}
\keyword{classes}
