% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_parqual.R
\docType{class}
\name{ref_parqual-class}
\alias{ref_parqual-class}
\title{Class "ref_parqual"}
\description{
Class enabling to load the list of qualitative parameters and to select one
of them. It inherits from 'ref_par', uses its 'choice' method
}
\section{Slots}{

\describe{
\item{\code{valqual="data.frame"}}{the list of qualitative parameters}
}}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("ref_parqual", ...)}.  \describe{ \item{list("valqual")}{Object of
class \code{"data.frame"} Values of qualitatives parameters}\item{:}{Object
of class \code{"data.frame"} Values of qualitatives parameters}
\item{list("data")}{Object of class \code{"data.frame"} Qualitatives
parameters }\item{:}{Object of class \code{"data.frame"} Qualitatives
parameters } }
}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\seealso{
Other referential.objects: \code{\link{charge,ref_choice-method}},
  \code{\link{ref_checkbox-class}},
  \code{\link{ref_choice-class}},
  \code{\link{ref_coe-class}}, \code{\link{ref_dc-class}},
  \code{\link{ref_df-class}},
  \code{\link{ref_horodate-class}},
  \code{\link{ref_list-class}},
  \code{\link{ref_par-class}},
  \code{\link{ref_parquan-class}},
  \code{\link{ref_period-class}},
  \code{\link{ref_stage-class}},
  \code{\link{ref_taxa-class}},
  \code{\link{ref_year-class}}
}

