% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_stage.R
\docType{class}
\name{ref_stage-class}
\alias{ref_stage-class}
\title{Class 'ref_stage'}
\description{
Representation of a fish phase
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame containing data loaded from the database by either charge or charge_with_filter methods}

\item{\code{stage_selected}}{Contains the code \code{'tax_code'} of the stage selected by choice_c() method}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new('ref_stage', data='data.frame')}.  \describe{
\item{list('data')}{Object of class \code{'data.frame'} ~ The phases
available in the database}\item{:}{Object of class \code{'data.frame'} ~ The
phases available in the database} }
}

\seealso{
Other referential objects: 
\code{\link{charge,ref_choice-method}},
\code{\link{ref_choice-class}},
\code{\link{ref_coe-class}},
\code{\link{ref_dc-class}},
\code{\link{ref_df-class}},
\code{\link{ref_horodate-class}},
\code{\link{ref_list-class}},
\code{\link{ref_par-class}},
\code{\link{ref_parqual-class}},
\code{\link{ref_parquan-class}},
\code{\link{ref_taxa-class}},
\code{\link{ref_year-class}}
}
\author{
cedric.briand@eptb-vilaine.fr
}
\concept{referential objects}
\keyword{classes}
