% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_taxa.R
\docType{class}
\name{ref_taxa-class}
\alias{ref_taxa-class}
\title{Class "ref_taxa"}
\description{
Loading and selection of fish species. This class is a referential class, and it is 
integrated into refreport objects.
}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{"data.frame"} of species available in the database}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("ref_taxa", ...)}.
}

\seealso{
Other referential objects: \code{\link{charge,ref_choice-method}},
  \code{\link{ref_checkbox-class}},
  \code{\link{ref_choice-class}},
  \code{\link{ref_coe-class}}, \code{\link{ref_dc-class}},
  \code{\link{ref_df-class}},
  \code{\link{ref_horodate-class}},
  \code{\link{ref_list-class}},
  \code{\link{ref_par-class}},
  \code{\link{ref_parqual-class}},
  \code{\link{ref_parquan-class}},
  \code{\link{ref_period-class}},
  \code{\link{ref_stage-class}},
  \code{\link{ref_year-class}}
}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\concept{referential objects}
