% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_textbox.R
\docType{class}
\name{ref_textbox-class}
\alias{ref_textbox-class}
\title{ref_textbox referencial class}
\description{
allows to a put a value within a glabel
}
\section{Slots}{

\describe{
\item{\code{title='character'}}{the title of the box giving the possible choices}

\item{\code{labels}}{the logical parameters choice}

\item{\code{checked}}{a vector}
}}

\author{
cedric.briand@eptb-vilaine.fr
}
