% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_timestep.R
\docType{class}
\name{ref_timestep-class}
\alias{ref_timestep-class}
\title{Class "ref_timestep"}
\description{
Describes a time step
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("ref_timestep",
dateDebut="POSIXt",step_duration=numeric(),nb_step=numeric(),nocurrent_step=integer())}.
\describe{
\item{list("dateDebut")}{Object of class \code{"POSIXt"} Starting
date }
\item{:}{Object of class \code{"POSIXt"} Starting date }
\item{list("step_duration")}{Object of class \code{"numeric"} Step length
}\item{:}{Object of class \code{"numeric"} Step length }
\item{list("nb_step")}{Object of class \code{"numeric"} Number of steps
}\item{:}{Object of class \code{"numeric"} Number of steps }
\item{list("nocurrent_step")}{Object of class \code{"integer"} Number of the
current step }\item{:}{Object of class \code{"integer"} Number of the
current step } }
}

\seealso{
\code{\linkS4class{ref_timestep_daily}}
}
\author{
cedric.briand@eptb-vilaine.fr
}
\concept{report Object}
