% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_timestep_daily.R
\docType{class}
\name{ref_timestep_daily-class}
\alias{ref_timestep_daily-class}
\title{Class 'ref_timestep_daily'}
\description{
Representation of a ref_timestep object with a step length equal to one day.
It receives an inheritance from ref_timestep
}
\details{
validity_ref_timestep_daily
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new('ref_timestep_daily',
dateDebut='POSIXt',step_duration=numeric(),nb_step=numeric(),nocurrent_step=integer())}.
\describe{ \item{list('dateDebut')}{Object of class \code{'POSIXt'} Starting
date }\item{:}{Object of class \code{'POSIXt'} Starting date }
\item{list('step_duration')}{Object of class \code{'numeric'} Step length
}\item{:}{Object of class \code{'numeric'} Step length }
\item{list('nb_step')}{Object of class \code{'numeric'} Number of steps
}\item{:}{Object of class \code{'numeric'} Number of steps }
\item{list('nocurrent_step')}{Object of class \code{'integer'} Number of the
current step }\item{:}{Object of class \code{'integer'} Number of the
current step } }
}

\seealso{
\code{\linkS4class{ref_timestep}}
}
\author{
cedric.briand@eptb-vilaine.fr
}
\keyword{classes}
