% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_timestep.R
\docType{class}
\name{ref_timestepChar-class}
\alias{ref_timestepChar-class}
\title{Class "ref_timestepChar"}
\description{
Character to represent a ref_timestep
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("ref_timestepChar", \dots{})}
}

\examples{

showClass("ref_timestepChar")

}
\seealso{
\code{\linkS4class{ref_timestep}}
}
\author{
cedric.briand@eptb-vilaine.fr
}
\keyword{classes}
