% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_year.R
\docType{class}
\name{ref_year-class}
\alias{ref_year-class}
\title{Year reference class}
\description{
Class used to select one or several years
}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{data.frame} with the list of possible years selected as numerics}

\item{\code{year_selected}}{A numeric vector}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("ref_year", data=data.frame(), year_selected=numeric())}.
}

\seealso{
Other referential objects: 
\code{\link{charge,ref_choice-method}},
\code{\link{ref_choice-class}},
\code{\link{ref_coe-class}},
\code{\link{ref_dc-class}},
\code{\link{ref_df-class}},
\code{\link{ref_horodate-class}},
\code{\link{ref_list-class}},
\code{\link{ref_par-class}},
\code{\link{ref_parqual-class}},
\code{\link{ref_parquan-class}},
\code{\link{ref_stage-class}},
\code{\link{ref_taxa-class}}
}
\author{
cedric.briand@eptb-vilaine.fr
}
\concept{referential objects}
\keyword{classes}
