% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefAnnee.r
\docType{class}
\name{RefAnnee-class}
\alias{RefAnnee-class}
\title{Year reference class}
\description{
Class used to select one or several years
}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{data.frame} with the list of possible years selected as numerics}

\item{\code{annee_selectionnee}}{A numeric vector}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("RefAnnee", data=data.frame(), annee_selectionnee=numeric())}.
}

\seealso{
Other Referential objects: \code{\link{RefCheckBox-class}},
  \code{\link{RefChoix-class}}, \code{\link{RefCoe-class}},
  \code{\link{RefDC-class}}, \code{\link{RefDF-class}},
  \code{\link{RefHorodate-class}},
  \code{\link{RefListe-class}},
  \code{\link{RefStades-class}},
  \code{\link{RefTaxon-class}}, \code{\link{Refpar-class}},
  \code{\link{Refparqual-class}},
  \code{\link{Refparquan-class}},
  \code{\link{Refperiode-class}},
  \code{\link{charge,RefChoix-method}},
  \code{\link{charge,RefMsg-method}}
}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\keyword{classes}
