\name{RefChoix}
\alias{RefChoix}
\title{RefChoix referencial class allows to choose within several values with radiobuttons}
\description{
  RefChoix referencial class allows to choose within
  several values with radiobuttons
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

