\name{RefCoe}
\alias{RefCoe}
\title{RefCoe referential class}
\description{
  RefCoe referential class
}
\note{
  Class loading coefficient of conversion between quantity
  (weights or volumes of glass eel) and numbers between a
  starting and finishing date
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

