% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefDF.r
\docType{class}
\name{RefDF-class}
\alias{RefDF-class}
\title{Class "RefDF"}
\arguments{
\item{df_selectionne}{Object of class \code{"integer"} The identifier of the fishway}

\item{ouvrage}{Object of class \code{"integer"} The attached dam}

\item{data}{Object of class \code{"data.frame"} Data concerning the fishway}
}
\description{
Representation of a fishway, contains description data of all fishways from
the database along with the selected fishways (df) (integer)
Objects from the Class: Objects can be created by calls of the form
\code{new("RefDF", df_selectionne=integer(), ouvrage=integer(),
data=data.frame())}.
}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\seealso{
Other Referential.objects: \code{\link{RefAnnee-class}},
  \code{\link{RefCheckBox-class}},
  \code{\link{RefChoix-class}}, \code{\link{RefCoe-class}},
  \code{\link{RefDC-class}},
  \code{\link{RefHorodate-class}},
  \code{\link{RefListe-class}},
  \code{\link{RefStades-class}},
  \code{\link{RefTaxon-class}}, \code{\link{Refpar-class}},
  \code{\link{Refparqual-class}},
  \code{\link{Refparquan-class}},
  \code{\link{Refperiode-class}},
  \code{\link{charge,RefChoix-method}}
}

