\name{RefDF}
\alias{RefDF}
\title{RefDF referential class to load the DF dataframe}
\description{
  RefDF referential class to load the DF dataframe
}
\note{
  Contains description data of all DF from the database
  along with the selected df (integer)
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

