\name{RefHorodate}
\alias{RefHorodate}
\title{Classe objet utilisee pour sa methode de chargement du choix dans la date
Utilise dans fonctionnement DC, avec methode de affichage annee en cours et annee precedente
Possibilite de modifier facilement l'annee affichee par defaut en changeant decal}
\description{
  Classe objet utilisee pour sa methode de chargement du
  choix dans la date Utilise dans fonctionnement DC, avec
  methode de affichage annee en cours et annee precedente
  Possibilite de modifier facilement l'annee affichee par
  defaut en changeant decal
}

