% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefListe.r
\docType{class}
\name{RefListe-class}
\alias{RefListe-class}
\title{Class "RefListe"}
\description{
Enables to load a "RefChoix" object fom a list given by a "Bilan" object
}
\section{Slots}{

\describe{
\item{\code{liste}}{choice="character" A vector of character to choose within a droplist}

\item{\code{label="character"}}{the title of the box}
}}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("RefListe", listechoice, label)}.  \describe{
\item{list("listechoice")}{Object of class \code{"character"}}\item{:}{Object
of class \code{"character"}} \item{list("label")}{Object of class
\code{"character"}}\item{:}{Object of class \code{"character"}} }
}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\seealso{
Other referential classes \code{\linkS4class{RefAnnee}}
\code{\linkS4class{RefCheckBox}} \code{\linkS4class{RefChoix}}
\code{\linkS4class{RefCoe}} \code{\linkS4class{RefDC}}
\code{\linkS4class{RefDF}} \code{\linkS4class{RefListe}}
\code{\linkS4class{Refpar}} \code{\linkS4class{Refparqual}}
\code{\linkS4class{Refparquan}} \code{\linkS4class{RefPoidsMoyenPeche}}
\code{\linkS4class{RefStades}} \code{\linkS4class{RefStationMesure}}
\code{\linkS4class{RefTaxon}}

Other Referential.objects: \code{\link{RefAnnee-class}},
  \code{\link{RefCheckBox-class}},
  \code{\link{RefChoix-class}}, \code{\link{RefCoe-class}},
  \code{\link{RefDC-class}}, \code{\link{RefDF-class}},
  \code{\link{RefHorodate-class}},
  \code{\link{RefPoidsMoyenPeche-class}},
  \code{\link{RefStades-class}},
  \code{\link{Refpar-class}},
  \code{\link{Refparqual-class}},
  \code{\link{Refparquan-class}},
  \code{\link{Refperiode-class}},
  \code{\link{charge,RefChoix-method}},
  \code{\link{charge,RefMsg-method}}
}
\keyword{classes}

