% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefMsg.r
\docType{class}
\name{RefMsg-class}
\alias{RefMsg-class}
\title{RefMsg referential class to load message according to the language chosen}
\description{
RefMsg referential class to load message according to the language chosen
}
\section{Slots}{

\describe{
\item{\code{data}}{A data.frame}
}}
\section{Methods}{
 \describe{ \item{createmessage}{\code{signature(object =
"RefMsg")}: creates a message } }
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

