% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refpar.r
\docType{class}
\name{Refpar-class}
\alias{Refpar-class}
\title{Class "Refpar"}
\description{
Class enabling to load the list of parameters and select one of them
}
\section{Slots}{

\describe{
\item{\code{data="data.frame"}}{the list of parameters}
}}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("Refpar", data)}.  \describe{ \item{list("data")}{Object of class
\code{"data.frame"} ~ All the parameters stored in the
database}\item{:}{Object of class \code{"data.frame"} ~ All the parameters
stored in the database} }
}
\examples{

showClass("Refpar")

}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\seealso{
Other referential classes \code{\linkS4class{RefAnnee}}
\code{\linkS4class{RefCheckBox}} \code{\linkS4class{RefChoix}}
\code{\linkS4class{RefCoe}} \code{\linkS4class{RefDC}}
\code{\linkS4class{RefDF}} \code{\linkS4class{RefListe}}
\code{\linkS4class{Refpar}} \code{\linkS4class{Refparqual}}
\code{\linkS4class{Refparquan}} \code{\linkS4class{RefPoidsMoyenPeche}}
\code{\linkS4class{RefStades}} \code{\linkS4class{RefStationMesure}}
\code{\linkS4class{RefTaxon}}

Other Referential.objects: \code{\link{RefAnnee-class}},
  \code{\link{RefCheckBox-class}},
  \code{\link{RefChoix-class}}, \code{\link{RefCoe-class}},
  \code{\link{RefDC-class}}, \code{\link{RefDF-class}},
  \code{\link{RefHorodate-class}},
  \code{\link{RefListe-class}},
  \code{\link{RefPoidsMoyenPeche-class}},
  \code{\link{RefStades-class}},
  \code{\link{Refparqual-class}},
  \code{\link{Refparquan-class}},
  \code{\link{Refperiode-class}},
  \code{\link{charge,RefChoix-method}},
  \code{\link{charge,RefMsg-method}}
}
\keyword{classes}

