% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refparqual.r
\docType{class}
\name{Refparqual-class}
\alias{Refparqual-class}
\title{Class "Refparqual"}
\description{
Class enabling to load the list of qualitative parameters and to select one
of them. It inherits from 'Refpar', uses its 'choice' method
}
\section{Slots}{

\describe{
\item{\code{valqual="data.frame"}}{the list of qualitative parameters}
}}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("Refparqual", ...)}.  \describe{ \item{list("valqual")}{Object of
class \code{"data.frame"} Values of qualitatives parameters}\item{:}{Object
of class \code{"data.frame"} Values of qualitatives parameters}
\item{list("data")}{Object of class \code{"data.frame"} Qualitatives
parameters }\item{:}{Object of class \code{"data.frame"} Qualitatives
parameters } }
}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\seealso{
Other Referential objects: \code{\link{RefAnnee-class}},
  \code{\link{RefCheckBox-class}},
  \code{\link{RefChoix-class}}, \code{\link{RefCoe-class}},
  \code{\link{RefDC-class}}, \code{\link{RefDF-class}},
  \code{\link{RefHorodate-class}},
  \code{\link{RefListe-class}},
  \code{\link{RefStades-class}},
  \code{\link{RefTaxon-class}}, \code{\link{Refpar-class}},
  \code{\link{Refparquan-class}},
  \code{\link{Refperiode-class}},
  \code{\link{charge,RefChoix-method}},
  \code{\link{charge,RefMsg-method}}
}

