\name{Refparqual}
\alias{Refparqual}
\title{Refparqual referential class choose a parameter}
\description{
  Refparqual referential class choose a parameter
}
\note{
  Classe permettant de charger la liste de parametres
  qualitatifs et de les selectionner herite de Refpar dont
  elle utilie la methode choix par rapport à parquan, cette
  classe possede en plus lun vecteur des valeurs possibles
  des parametres qualitatifs
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

