\name{Refparquan}
\alias{Refparquan}
\title{Refparquan referential class choose a quantitative parameter}
\description{
  Refparquan referential class choose a quantitative
  parameter
}
\note{
  The choice method of this class is inherited from the
  parent class
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

