% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefStades.r
\docType{class}
\name{RefStades-class}
\alias{RefStades-class}
\title{Class "RefStades"}
\description{
Representation of a fish phase
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("RefStades", data="data.frame")}.  \describe{
\item{list("data")}{Object of class \code{"data.frame"} ~ The phases
available in the database}\item{:}{Object of class \code{"data.frame"} ~ The
phases available in the database} }
}

\seealso{
Other Referential objects: \code{\link{RefAnnee-class}},
  \code{\link{RefCheckBox-class}},
  \code{\link{RefChoix-class}}, \code{\link{RefCoe-class}},
  \code{\link{RefDC-class}}, \code{\link{RefDF-class}},
  \code{\link{RefHorodate-class}},
  \code{\link{RefListe-class}},
  \code{\link{RefTaxon-class}}, \code{\link{Refpar-class}},
  \code{\link{Refparqual-class}},
  \code{\link{Refparquan-class}},
  \code{\link{Refperiode-class}},
  \code{\link{charge,RefChoix-method}},
  \code{\link{charge,RefMsg-method}}
}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\keyword{classes}
