% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefStationMesure.r
\docType{class}
\name{RefStationMesure-class}
\alias{RefStationMesure-class}
\title{Class "RefStationMesure"}
\description{
Enables to load measure stations and to select one of them
}
\section{Slots}{

\describe{
\item{\code{dataframe}}{Data concerning the
measure station}
}}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("RefStationMesure", ...)}.
}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\seealso{
Other referential classes 
\code{\linkS4class{RefAnnee}}
\code{\linkS4class{RefCheckBox}} 
\code{\linkS4class{RefChoix}}
\code{\linkS4class{RefCoe}} 
\code{\linkS4class{RefDC}}
\code{\linkS4class{RefDF}} 
\code{\linkS4class{RefListe}}
\code{\linkS4class{Refpar}} 
\code{\linkS4class{Refparqual}}
\code{\linkS4class{Refparquan}} 
\code{\linkS4class{RefPoidsMoyenPeche}}
\code{\linkS4class{RefStades}} 
\code{\linkS4class{RefStationMesure}}
\code{\linkS4class{RefTaxon}}
}
\keyword{classes}

