% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefTaxon.r
\docType{class}
\name{RefTaxon-class}
\alias{RefTaxon-class}
\title{Class "RefTaxon"}
\description{
Loading and selection of fish species. This class is a referential class, and it is 
integrated into refBilan objects.
}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{"data.frame"} of species available in the database}
}}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("RefTaxon", ...)}.
}
\examples{

showClass("RefTaxon")
}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\seealso{
Other referential classes 
\code{\linkS4class{RefAnnee}}
\code{\linkS4class{RefCheckBox}} 
\code{\linkS4class{RefChoix}}
\code{\linkS4class{RefCoe}} 
\code{\linkS4class{RefDC}}
\code{\linkS4class{RefDF}} 
\code{\linkS4class{RefListe}}
\code{\linkS4class{Refpar}} 
\code{\linkS4class{Refparqual}}
\code{\linkS4class{Refparquan}} 
\code{\linkS4class{RefPoidsMoyenPeche}}
\code{\linkS4class{RefStades}} 
\code{\linkS4class{RefStationMesure}}
\code{\linkS4class{RefTaxon}}
}
\keyword{classes}

