% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReftextBox.r
\docType{class}
\name{RefTextBox-class}
\alias{RefTextBox-class}
\title{RefTextBox referencial class}
\description{
allows to a put a value within a glabel
}
\section{Slots}{

\describe{
\item{\code{title="character"}}{the title of the box giving the possible choices}

\item{\code{labels}}{the logical parameters choice}

\item{\code{checked}}{a vector}
}}
\author{
cedric.briand"at"eptb-vilaine.fr
}

