% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_ope.R
\docType{class}
\name{report_ope-class}
\alias{report_ope-class}
\alias{report_ope}
\title{Report on operations}
\description{
Operations are monitoring operations generated between two dates. In the case of video monitoring
or similar, they can be instantaneous
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("report_ope")}.
}

\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
\concept{report Object}
\keyword{classes}
