% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitaires.r
\name{round.POSIXt}
\alias{round.POSIXt}
\title{round.POSIXT function imported from  Hmisc that cannot be loaded}
\usage{
\method{round}{POSIXt}(x, digits = c("secs", "mins", "hours", "days",
  "months", "years"), ...)
}
\arguments{
\item{x}{}

\item{digits}{}
}
\value{
a rounded time value
}
\description{
round.POSIXT function imported from  Hmisc that cannot be loaded
}
\author{
Charles Dupont
}

