% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_timestep.R
\docType{methods}
\name{set_starting_date,ref_timestep-method}
\alias{set_starting_date,ref_timestep-method}
\title{Sets starting date from a character}
\usage{
\S4method{set_starting_date}{ref_timestep}(object, string)
}
\arguments{
\item{object}{An object of class \link{ref_timestep-class}}

\item{string}{Character string of type"\%Y-\%m-\%d \%H:\%M:\%S" or "\%Y-\%m-\%d".
this allows to use either horodate or date}
}
\value{
An object of class \link{ref_timestep-class}
}
\description{
Sets starting date from a character
}
\keyword{internal}
