% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_timestep.R
\name{set_starting_date}
\alias{set_starting_date}
\title{Generic method to set the starting date}
\usage{
set_starting_date(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional parameters passed to the method}
}
\description{
Generic method to set the starting date
}
\keyword{internal}

