% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationCar.r
\docType{methods}
\name{setasqualitative,BilanMigrationCar-method}
\alias{setasqualitative,BilanMigrationCar-method}
\title{Turns a quantitative parameter into qualitative}
\usage{
\S4method{setasqualitative}{BilanMigrationCar}(object, par, silent = FALSE,
  ...)
}
\arguments{
\item{object}{An object of class \link{Refparquan-class}}

\item{par}{The code of a quantitative parameter}

\item{silent}{Default FALSE, if TRUE the program should not display messages}

\item{...}{Additional parms to the cut method \link[base]{cut}}
}
\description{
Turns a quantitative parameter into qualitative
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
