% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_char.R
\name{setasqualitative,report_mig_char-method}
\alias{setasqualitative,report_mig_char-method}
\title{handler for report_mig_char}
\usage{
\S4method{setasqualitative}{report_mig_char}(object, par, silent = FALSE, ...)
}
\arguments{
\item{object}{An object of class \link{ref_parquan-class}}

\item{par}{The code of a quantitative parameter}

\item{silent}{Default FALSE, if TRUE the program should not display messages}

\item{...}{Additional parms to the cut method \link[base]{cut}}

\item{h}{handler}
}
\value{
An object of class \link{ref_parquan-class} with lines removed from \code{r@data[["parquan"]]}
and added (after transformation to qualitative values) in \code{r@data[["parqal"]]}
}
\description{
internal use
}
\author{
Cedric Briand \email{cedric.briand'at'eptb-vilaine.fr}
}
\keyword{Turns}
\keyword{a}
\keyword{internal}
\keyword{into}
\keyword{parameter}
\keyword{qualitative}
\keyword{quantitative}
