% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generic.R
\name{setasqualitative}
\alias{setasqualitative}
\title{Generic method to transform quantitative par into a qualitative one}
\usage{
setasqualitative(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Additional parms}
}
\description{
Generic method to transform quantitative par into a qualitative one
}
\author{
cedric.briand
}
