\name{setdateDebut-methods}
\docType{methods}
\alias{setdateDebut-methods}
\alias{setdateDebut,PasDeTemps-method}
\title{Methods for Function setdateDebut}
\description{
see PasDeTemps.r for description of the class
}
\section{Methods}{
\describe{

\item{\code{signature(objet = "PasDeTemps")}}{ ~~describe this method here }
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s)}
