% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasdeTemps.r
\docType{methods}
\name{setdateDebut,PasDeTemps-method}
\alias{setdateDebut,PasDeTemps-method}
\title{Sets starting date from a character}
\usage{
\S4method{setdateDebut}{PasDeTemps}(object, string)
}
\arguments{
\item{object}{An object of class \link{PasDeTemps-class}}

\item{string}{Character string of type"\%Y-\%m-\%d \%H:\%M:\%S" or "\%Y-\%m-\%d".
this allows to use either horodate or date}
}
\value{
An object of class \link{PasDeTemps-class}
}
\description{
Sets starting date from a character
}

