% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasdeTemps.r
\name{setdateDebut}
\alias{setdateDebut}
\title{Generic method to set the starting date}
\usage{
setdateDebut(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional parameters passed to the method}
}
\description{
Generic method to set the starting date
}
\keyword{internal}

