\name{setHorodate-methods}
\docType{methods}
\alias{setHorodate-methods}
\alias{setHorodate,Horodate-method}
\title{Function setHorodate}
\description{
 Set the horodate
}
\section{Methods}{
\describe{

\item{objet = "Horodate"}{ The new horodate. The horodate is set to objet }
}}
\keyword{methods}
