% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_horodate.R
\docType{methods}
\name{setref_horodate,ref_horodate-method}
\alias{setref_horodate,ref_horodate-method}
\title{Method to set the Horodate}
\usage{
\S4method{setref_horodate}{ref_horodate}(object, string)
}
\arguments{
\item{object}{An object of class \link{ref_horodate-class}}

\item{string}{A string representing an horodate in the format "\%Y-\%m-\%d \%H:\%M:\%S"}
}
\value{
An Object of class "ref_horodate"
}
\description{
Method to set the Horodate
}
\author{
cedric.briand
}
\keyword{internal}
