\name{setRefHorodate-methods}
\docType{methods}
\alias{setRefHorodate-methods}
\alias{setRefHorodate,RefHorodate-method}
\title{Methods for Function setRefHorodate}
\description{
Methods for function \code{setRefHorodate}
}
\section{Methods}{
\describe{

\item{\code{signature(objet = "RefHorodate")}}{ ~~describe this method here }
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s)}
