% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefHorodate.r
\docType{methods}
\name{setRefHorodate,RefHorodate-method}
\alias{setRefHorodate,RefHorodate-method}
\title{Method to set the Horodate}
\usage{
\S4method{setRefHorodate}{RefHorodate}(object, string)
}
\value{
An Object of class "RefHorodate"
}
\description{
Method to set the Horodate
}
\author{
cedric.briand
}

