% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{stacomi_installed}
\alias{stacomi_installed}
\title{Test that the program is installed.}
\usage{
stacomi_installed()
}
\value{
A logical checking if there is calcmig.csv folder
}
\description{
Test that the program is installed.
}
