\name{suivant-methods}
\docType{methods}
\alias{suivant-methods}
\alias{suivant,PasDeTemps-method}
\title{Methods for Function suivant}
\description{
 Methods for function \code{suivant} get the next PasDeTemps
}
\section{Methods}{
\describe{

\item{\code{signature(objet = "PasDeTemps")}}{ ~~describe this method here }
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s)}
