% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasdeTemps.r
\docType{methods}
\name{suivant,PasDeTemps-method}
\alias{suivant,PasDeTemps-method}
\title{Gets the next time step}
\usage{
\S4method{suivant}{PasDeTemps}(object)
}
\arguments{
\item{object}{An object of class \link{PasDeTemps-class}}
}
\value{
An object of class \link{PasDeTemps-class} with current time step set
}
\description{
Gets the next time step
}
\keyword{internal}

