% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasdeTemps.r
\name{suivant}
\alias{suivant}
\title{Generic method next}
\usage{
suivant(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional parameters passed to the method}
}
\description{
Generic method next
}
\keyword{internal}

