% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_carlot.r
\docType{methods}
\name{summary,Bilan_carlot-method}
\alias{summary,Bilan_carlot-method}
\title{summary for Bilan_carlot}
\usage{
\S4method{summary}{Bilan_carlot}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{Bilan_carlot-class}}}

\item{silent}{Should the program stay silent or display messages, default FALSE}

\item{...}{Additional parameters}
}
\description{
summary for Bilan_carlot
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
