% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanArgentee.r
\docType{methods}
\name{summary,BilanArgentee-method}
\alias{summary,BilanArgentee-method}
\title{summary for BilanArgentee}
\usage{
\S4method{summary}{BilanArgentee}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{BilanArgentee-class}}}

\item{silent}{Should the program stay silent or display messages, default FALSE}

\item{...}{Additional parameters}
}
\description{
summary for BilanArgentee
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

