% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanFonctionnementDC.r
\docType{methods}
\name{summary,BilanFonctionnementDC-method}
\alias{summary,BilanFonctionnementDC-method}
\title{summary for BilanFonctionnementDC, write csv and html output, and prints summary statistics}
\usage{
\S4method{summary}{BilanFonctionnementDC}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{BilanFonctionnementDC-class}}}

\item{silent}{Should the program stay silent or display messages, default FALSE}

\item{...}{Additional parameters (not used there)}
}
\description{
summary for BilanFonctionnementDC, write csv and html output, and prints summary statistics
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

