% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanFonctionnementDF.r
\docType{methods}
\name{summary,BilanFonctionnementDF-method}
\alias{summary,BilanFonctionnementDF-method}
\title{summary for BilanFonctionnementDF, write csv and html output, and prints summary statistics}
\usage{
\S4method{summary}{BilanFonctionnementDF}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{BilanFonctionnementDF-class}}}

\item{silent}{Should the program stay silent or display messages, default FALSE}

\item{...}{Additional parameters (not used there)}
}
\description{
summary for BilanFonctionnementDF, write csv and html output, and prints summary statistics
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

