% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigration.r
\docType{methods}
\name{summary,BilanMigration-method}
\alias{summary,BilanMigration-method}
\alias{summary.bilanmigration}
\title{summary for bilanMigration 
calls functions funstat and funtable to create migration overviews
and generate csv and html output in the user data directory}
\usage{
\S4method{summary}{BilanMigration}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{BilanMigration-class}}}

\item{silent}{Should the program stay silent or display messages, default FALSE}

\item{...}{Additional parameters (not used there)}
}
\description{
summary for bilanMigration 
calls functions funstat and funtable to create migration overviews
and generate csv and html output in the user data directory
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

