% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationCar.r
\docType{methods}
\name{summary,BilanMigrationCar-method}
\alias{summary,BilanMigrationCar-method}
\title{summary for BilanMigrationCar}
\usage{
\S4method{summary}{BilanMigrationCar}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{BilanMigrationCar-class}}}

\item{silent}{Should the program stay silent or display messages, default FALSE}

\item{...}{Additional parameters}
}
\description{
summary for BilanMigrationCar
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
