% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationInterAnnuelle.r
\docType{methods}
\name{summary,BilanMigrationInterAnnuelle-method}
\alias{summary,BilanMigrationInterAnnuelle-method}
\alias{summary.BilanMigrationInterannuelle}
\title{summary for BilanMigrationInterAnnuelle
provides summary statistics for the latest year (if silent=TRUE), or the year selected in the interface,
if silent=FALSE. Mean, min and max are historical statistics with the selected year excluded from the 
historical dataset.}
\usage{
\S4method{summary}{BilanMigrationInterAnnuelle}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{BilanMigrationInterAnnuelle-class}}}

\item{silent}{Should the program stay silent or display messages, default FALSE}

\item{...}{Additional parameters (not used there)}
}
\description{
summary for BilanMigrationInterAnnuelle
provides summary statistics for the latest year (if silent=TRUE), or the year selected in the interface,
if silent=FALSE. Mean, min and max are historical statistics with the selected year excluded from the 
historical dataset.
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
