% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMult.r
\docType{methods}
\name{summary,BilanMigrationMult-method}
\alias{summary,BilanMigrationMult-method}
\title{summary for BilanMigrationMult 
calls functions funstat and funtable to create migration overviews
and generate csv and html output in the user data directory}
\usage{
\S4method{summary}{BilanMigrationMult}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{BilanMigrationMult-class}}}

\item{silent}{Should the program stay silent or display messages, default FALSE}

\item{...}{Additional parameters (not used there)}
}
\description{
summary for BilanMigrationMult 
calls functions funstat and funtable to create migration overviews
and generate csv and html output in the user data directory
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

