% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_dc.R
\name{summary,report_dc-method}
\alias{summary,report_dc-method}
\alias{summary.report_dc}
\title{summary for report_dc, write csv and html output, and prints summary statistics}
\usage{
\S4method{summary}{report_dc}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{report_dc-class}}}

\item{silent}{Should the program stay silent or display messages, default FALSE}

\item{...}{Additional parameters (not used there)}
}
\value{
Nothing, called for its side effect of writing html, csv files and printing summary
}
\description{
summary for report_dc, write csv and html output, and prints summary statistics
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
