% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig.R
\name{summary,report_mig-method}
\alias{summary,report_mig-method}
\alias{summary.report_mig}
\title{summary for report_mig
calls functions funstat and funtable to create migration overviews
and generate csv and html output in the user data directory}
\usage{
\S4method{summary}{report_mig}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{report_mig-class}}}

\item{silent}{Should the program stay silent or display messages, default FALSE}

\item{...}{Additional parameters (not used there)}
}
\value{
Nothing, calls the \link{summary.report_mig_mult} method
}
\description{
summary for report_mig
calls functions funstat and funtable to create migration overviews
and generate csv and html output in the user data directory
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
