% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_sea_age.R
\name{summary,report_sea_age-method}
\alias{summary,report_sea_age-method}
\alias{summary.report_sea_age}
\title{summary for report_sea_age}
\usage{
\S4method{summary}{report_sea_age}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{An object of class \link{report_sea_age-class}}

\item{silent}{Default FALSE, if TRUE the program should no display messages.}

\item{...}{Additional parameters}
}
\value{
The summary
}
\description{
summary for report_sea_age
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
