% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_species.R
\name{summary,report_species-method}
\alias{summary,report_species-method}
\alias{summary.report_species}
\title{handler for summary report_species, internal use}
\usage{
\S4method{summary}{report_species}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{report_species-class}}}

\item{silent}{Should the program stay silent or display messages, default FALSE}

\item{h}{a handler}
}
\description{
handler for summary report_species, internal use
}
\author{
Cedric Briand \email{cedric.briand'at'eptb-vilaine.fr}

Cedric Briand \email{cedric.briand'at'eptb-vilaine.fr}
}
\keyword{and}
\keyword{csv}
\keyword{data}
\keyword{directory}
\keyword{for}
\keyword{generate}
\keyword{html}
\keyword{in}
\keyword{internal}
\keyword{output}
\keyword{report_species}
\keyword{summary}
\keyword{the}
\keyword{user}
