% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanAgedemer.r
\docType{methods}
\name{supprime,BilanAgedemer-method}
\alias{supprime,BilanAgedemer-method}
\title{supprime method for BilanMigrationInterannuelle class}
\usage{
\S4method{supprime}{BilanAgedemer}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{BilanAgedemer-class}}
}
\value{
nothing
}
\description{
supprime method for BilanMigrationInterannuelle class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
