% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationInterAnnuelle.r
\docType{methods}
\name{supprime,BilanMigrationInterAnnuelle-method}
\alias{supprime,BilanMigrationInterAnnuelle-method}
\title{supprime method for BilanMigrationInterannuelle class}
\usage{
\S4method{supprime}{BilanMigrationInterAnnuelle}(object)
}
\arguments{
\item{object}{An object of class \link{BilanMigrationInterAnnuelle-class}}
}
\value{
nothing
}
\description{
supprime method for BilanMigrationInterannuelle class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
