% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefCoe.r
\docType{methods}
\name{supprime,RefCoe-method}
\alias{supprime,RefCoe-method}
\title{supprime method for "RefCoe" class}
\usage{
\S4method{supprime}{RefCoe}(object, tax, std, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{RefCoe-class}}

\item{tax}{'2038=Anguilla anguilla'}

\item{std}{'CIV=civelle'}

\item{silent}{Default FALSE, if TRUE the program should no display messages}
}
\value{
nothing
}
\description{
supprime method for "RefCoe" class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

