% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_sea_age.R
\name{supprime,report_sea_age-method}
\alias{supprime,report_sea_age-method}
\alias{supprime.report_sea_age}
\title{supprime method for report_mig_interannual class}
\usage{
\S4method{supprime}{report_sea_age}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_sea_age-class}}

\item{silent}{Default FALSE, if TRUE the program should no display messages}
}
\value{
Nothing, called for its side effect of deleting data in the database
}
\description{
supprime method for report_mig_interannual class
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
