% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generic.R
\name{supprime}
\alias{supprime}
\title{Generic method to delete entires from the database}
\usage{
supprime(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Additional parms}
}
\description{
Generic method to delete entires from the database
}
\author{
cedric.briand
}
