\name{surface}
\alias{surface}
\title{plots polygons}
\usage{
  surface(xmat, ymat, ordre = 1:dim(ymat)[2],
    couleur = 1:dim(ymat)[2], \dots)
}
\arguments{
  \item{xmat}{}
  \item{ymat}{}
  \item{ordre}{the order in which the polygons will be drawn}
  \item{couleur}{a color vector}
  \item{\dots}{additional arguments passed to the function}
}
\value{
  Bilan_stades_pigm with slots filled with user choice
}
\description{
  plots polygons
}
\author{
  Laurent Beaulaton \email{laurent.beaulaton@onema.fr}
}

