\name{trunc.POSIXt}
\alias{trunc.POSIXt}
\title{trunc...}
\usage{trunc.POSIXt(x, units=c("secs", "mins", "hours", "days", "months",
    "years"), ...)}
\description{trunc.POSIXT function imported from depuis Hmisc that cannot be loaded for reasons
of compatibility with xtable}
\value{a truncated time value}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{x}{}
\item{digits}{}}
