\name{validite_Annee}
\alias{validite_Annee,validite_RefHorodate,validite_ODBC,validite_PasDeTemps,validite_PasDeTempsChar,validite_RefHorodate}
\title{validite_Annee tests for validity within the class...}
\usage{validite_Annee(object)}
\description{validite_Annee tests for validity within the class}
\value{the test for the object Refannee}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{object}{}}
