% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefHorodate.r
\name{validite_RefHorodate}
\alias{validite_RefHorodate}
\title{Validity check for RefHorodate}
\usage{
validite_RefHorodate(object)
}
\arguments{
\item{object}{A refHorodate object}
}
\description{
Validity check for RefHorodate
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
