% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_horodate.R
\name{validity_ref_horodate}
\alias{validity_ref_horodate}
\title{Validity check for ref_horodate}
\usage{
validity_ref_horodate(object)
}
\arguments{
\item{object}{A ref_horodate object}
}
\description{
Validity check for ref_horodate
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
\keyword{internal}
