% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_year.R
\name{validity_year}
\alias{validity_year}
\title{Validity check for ref_year}
\usage{
validity_year(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{ref_year}}}
}
\value{
boolean The test for the object refannee
}
\description{
validity_year tests for validity within the class
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
\keyword{internal}
