\name{vector_to_listsql}
\alias{vector_to_listsql}
\title{Transforms a vector into a string called within an sql command  e...}
\usage{vector_to_listsql(vect)}
\description{Transforms a vector into a string called within an sql command  e.g. c(A,B,C) => in ('A','B','C')}
\value{listsql a list of value}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{vect}{}}
