% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_poids_moyen.r
\docType{methods}
\name{write_database,Bilan_poids_moyen-method}
\alias{write_database,Bilan_poids_moyen-method}
\title{Function to write data to the stacomi database for \link{Bilan_poids_moyen-class}}
\usage{
\S4method{write_database}{Bilan_poids_moyen}(object, silent,
  dbname = "bd_contmig_nat", host = "localhost", port = 5432)
}
\arguments{
\item{object}{An object of class \link{Bilan_poids_moyen-class}}

\item{silent}{Boolean, if TRUE, information messages are not displayed}
}
\value{
An object of class \link{Bilan_poids_moyen-class}
}
\description{
Function to write data to the stacomi database for \link{Bilan_poids_moyen-class}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

