% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanAgedemer.r
\docType{methods}
\name{write_database,BilanAgedemer-method}
\alias{write_database,BilanAgedemer-method}
\title{Command line method to write the caracteristic "sea age" (car_par_code='A124')
into the tj_caracteristiquelot_car table in the user's scheme}
\usage{
\S4method{write_database}{BilanAgedemer}(object, silent = TRUE,
  dbname = "bd_contmig_nat")
}
\arguments{
\item{object}{an object of class \link{BilanAgedemer-class}}

\item{silent}{: Default FALSE, if TRUE the program should no display messages.}

\item{dbname}{: the name of the database, defaults to "bd_contmig_nat"}
}
\description{
The sea age caracteristic is calculated from the mesured or calculated size of salmon and with a size/age rule
defined by the user
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
