% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigration.r
\docType{methods}
\name{write_database,BilanMigration-method}
\alias{write_database,BilanMigration-method}
\title{Command line method to write the daily and monthly counts to the 
t_bilanmigrationjournalier_bjo table}
\usage{
\S4method{write_database}{BilanMigration}(object, silent = TRUE,
  dbname = "bd_contmig_nat", host = "localhost", port = 5432)
}
\arguments{
\item{silent}{: TRUE to avoid messages}

\item{dbname}{: the name of the database, defaults to "bd_contmig_nat"}

\item{host}{: the host for sqldf, defaults to "localhost"}

\item{port}{: the port, defaults to 5432}

\item{bilanMigration}{an object of class \code{\linkS4class{BilanMigration}}}
}
\description{
Daily values are needed to compare migrations from year to year, by the class \link{BilanMigrationInterAnnuelle-class}. They are added by
by this function.
}
\note{
the user is asked whether or not he wants to overwrite data, if no
data are present in the database, the import is done anyway. The name of the database
is not passed in odbc link, here defaults to "bd_contmig_nat"
}
\examples{
\dontrun{
stacomi(gr_interface=FALSE,login_window=FALSE,database_expected=FALSE) 
data("bM_Arzal")
bM_Arzal<-calcule(bM_Arzal)
write_database(bilanMigration=bM_Arzal,silent=FALSE)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

