% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_ge_weight.R
\name{write_database,report_ge_weight-method}
\alias{write_database,report_ge_weight-method}
\alias{write_database.report_ge_weight}
\title{Method to write data to the stacomi database for \link{report_ge_weight-class}}
\usage{
\S4method{write_database}{report_ge_weight}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_ge_weight-class}}

\item{silent}{Boolean, if TRUE, information messages are not displayed}
}
\value{
Nothing, called for its side effect of writing to the database
}
\description{
Data will be written in tj_coefficientconversion_coe table, if the class retrieves some data
from the database, those will be deleted first.
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
