% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig.R
\name{write_database,report_mig-method}
\alias{write_database,report_mig-method}
\alias{write_database.report_mig}
\title{handler h_report_migwrite
Allows the saving of daily and monthly counts in the database}
\usage{
\S4method{write_database}{report_mig}(object, silent = TRUE, check_for_bjo = TRUE)
}
\arguments{
\item{object}{an object of class \linkS4class{report_mig}}

\item{silent}{: TRUE to avoid messages}

\item{check_for_bjo}{: do you want to check if data are already present in the bjo table, and delete them,
this param was added otherwise connect method when called from report_mig_interannual runs in loops}

\item{h}{a handler}

\item{...}{Additional parameters}
}
\description{
handler h_report_migwrite
Allows the saving of daily and monthly counts in the database
}
\note{
these entries are necessary to run the Interannual Migration class.
then no entry will be written to the database

the user is asked whether or not he wants to overwrite data only when silent is FALSE, if no
data are present in the database, the import is done anyway.
}
\examples{
\dontrun{
stacomi(database_expected=FALSE)
data("r_mig")
r_mig<-calcule(r_mig)
write_database(report_mig=r_mig,silent=FALSE)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}

Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{Command}
\keyword{Daily}
\keyword{They}
\keyword{\link{report_mig_interannual-class}.}
\keyword{added}
\keyword{and}
\keyword{are}
\keyword{by}
\keyword{class}
\keyword{compare}
\keyword{counts}
\keyword{daily}
\keyword{from}
\keyword{function.}
\keyword{internal}
\keyword{line}
\keyword{method}
\keyword{migrations}
\keyword{monthly}
\keyword{needed}
\keyword{t_bilanmigrationjournalier_bjo}
\keyword{table}
\keyword{the}
\keyword{this}
\keyword{to}
\keyword{values}
\keyword{write}
\keyword{year}
\keyword{year,}
