% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_sea_age.R
\name{write_database,report_sea_age-method}
\alias{write_database,report_sea_age-method}
\alias{write_database.report_sea_age}
\title{Command line method to write the characteristic "sea age" (car_par_code='A124')
into the tj_caracteristiquelot_car table in the user's scheme}
\usage{
\S4method{write_database}{report_sea_age}(object, silent = TRUE)
}
\arguments{
\item{object}{an object of class \link{report_sea_age-class}}

\item{silent}{: Default FALSE, if TRUE the program should no display messages.}
}
\value{
Nothing, called for its side effect of writing data to the database
}
\description{
The sea age characteristic is calculated from the measured or calculated size of salmon and with a size/age rule
defined by the user.
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
