% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_annual.R
\name{xtable,report_annual-method}
\alias{xtable,report_annual-method}
\alias{xtable.report_annual}
\title{xtable function for \link{report_annual-class}
create an xtable objet but also assigns an add.to.column argument in envir_stacomi,
for later use by the print.xtable method.}
\usage{
\S4method{xtable}{report_annual}(
  x,
  caption = NULL,
  label = NULL,
  align = NULL,
  digits = 0,
  display = NULL,
  auto = FALSE,
  dc_name = NULL,
  tax_name = NULL,
  std_name = NULL
)
}
\arguments{
\item{x, }{an object of class "report_annual"}

\item{caption, }{see xtable}

\item{label, }{see xtable}

\item{align, }{see xtable, overidden if NULL}

\item{digits}{default 0}

\item{display}{see xtable}

\item{auto}{see xtable}

\item{dc_name}{A string indicating the names of the DC, in the order of  \code{x@dc@dc_selected}
if not provided DC codes are used.}

\item{tax_name}{A string indicating the names of the taxa, if not provided latin names are used}

\item{std_name}{A string indicating the stages names, if not provided then std_libelle are used}
}
\value{
A xtable for annual report
}
\description{
xtable function for \link{report_annual-class}
create an xtable objet but also assigns an add.to.column argument in envir_stacomi,
for later use by the print.xtable method.
}
