% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_char.R
\name{xtable,report_mig_char-method}
\alias{xtable,report_mig_char-method}
\alias{xtable.report_mig_char}
\title{xtable function for \link{report_mig_char-class}
create an xtable objet to be later used by the print.xtable method.}
\usage{
\S4method{xtable}{report_mig_char}(
  x,
  caption = NULL,
  label = NULL,
  align = NULL,
  digits = NULL,
  display = NULL,
  auto = FALSE,
  ...
)
}
\arguments{
\item{x, }{an object of class 'report_mig_char'}

\item{caption, }{see xtable}

\item{label, }{see xtable}

\item{align, }{see xtable, overidden if NULL}

\item{digits, }{see xtable}

\item{display}{see xtable}

\item{auto}{see xtable}

\item{...}{Additional parameters}
}
\value{
A xtable
}
\description{
xtable function for \link{report_mig_char-class}
create an xtable objet to be later used by the print.xtable method.
}
