% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{b_carlot}
\alias{b_carlot}
\title{Size of yellow and glass eel at the Arzal dam (Vilaine, France) in the fishway and main eel trapping ladder.}
\format{An object of class \link{Bilan_carlot-class}}
\usage{
b_carlot
}
\description{
This dataset corresponds to the data collected at two different control devices
at the Arzal control station (see example in \link{Bilan_carlot-class}), all body size 
parameters (total size, size converted from pixel in video control) are used in example
}
\keyword{data}
