% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bil_CE}
\alias{bil_CE}
\title{Dayly temperature at Gabion station during the year 2008}
\format{An object of class \link{BilanConditionEnv-class} with data slot loaded.}
\usage{
bil_CE
}
\description{
The dataset corresponds to the dayly temperature at Gabion station during the year 2008. This environmental station is used to
analyse conditions in which fish migrated at Arzal dam
}
\keyword{data}

