% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bilA}
\alias{bilA}
\title{Annual migration of yellow and silver eel for three fishways / counting devices at the
Arzal dam (data from 1995 to 2016)}
\format{An object of class \link{BilanAnnuels-class} with data slot loaded.}
\usage{
bilA
}
\description{
The dataset corresponds to the three fishways located on the Arzal dam, filled with annual data
}
\keyword{data}
