% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bilAM}
\alias{bilAM}
\title{Annual migration of salmon by migradour}
\format{An object of class \link{BilanAnnuels-class} with data slot loaded.}
\usage{
bilAM
}
\description{
The dataset corresponds to the fishways DC=33:40 of the Adour for adult migrant salmons
from 1996 to 2005 (annual counts). It has been kinly provided as an example set by the Migradour
association.
}
\keyword{data}
