% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_poids_moyen.r
\docType{class}
\name{Bilan_poids_moyen-class}
\alias{Bilan_poids_moyen-class}
\title{Class "Bilan_poids_moyen"}
\description{
Bilan_poids_moyen class The objective is to calculate mean weight of glass
eel which are counted from weight measurements and to reintegrate weight to
number coefficients
#'
}
\section{Slots}{

\describe{
\item{\code{data}}{Object of class \code{"data.frame"} data for bilan lot}

\item{\code{dc}}{Object of class \code{\link{RefDC-class}}, the counting device}

\item{\code{anneedebut}}{Object of class \code{\link{RefAnnee-class}}. refAnnee allows to choose year of beginning}

\item{\code{anneefin}}{Object of class \code{\link{RefAnnee-class}}
refAnnee allows to choose last year of the Bilan}

\item{\code{coe}}{Object of class \code{\link{RefCoe-class}} class loading coefficient
of conversion between quantity (weights or volumes of glass eel) and numbers}

\item{\code{liste}}{Object of class \code{\link{RefListe-class}} RefListe referential
class choose within a list, here the choice is wether subsamples or not. Subsamples
in the stacomi database are samples with a non null value for parent sample. Migration
counts are never made on subsamples but those can be integrated to calculate mean weights.}
}}
\note{
We have also tools available to import glass eel measurement from
experimental fishing in the estuary For the charge method dates for the
request are from august to august (a glass eel season)
}
\examples{
 \dontrun{
showClass("Bilan_poids_moyen")
object=new("Bilan_poids_moyen")
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other Bilan Classes \code{\linkS4class{Bilan_carlot}}
\code{\linkS4class{Bilan_poids_moyen}}
\code{\linkS4class{Bilan_stades_pigm}} 
\code{\linkS4class{Bilan_taille}}
\code{\linkS4class{BilanConditionEnv}} 
\code{\linkS4class{BilanEspeces}}
\code{\linkS4class{BilanFonctionnementDC}}
\code{\linkS4class{BilanFonctionnementDF}}
\code{\linkS4class{BilanMigration}}
\code{\linkS4class{BilanMigrationConditionEnv}}
\code{\linkS4class{BilanMigrationInterAnnuelle}}
\code{\linkS4class{BilanMigrationPar}}
}
\concept{
Bilan Object 
 @keywords classes
}

