% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_poids_moyen.r
\docType{class}
\name{Bilan_poids_moyen-class}
\alias{Bilan_poids_moyen-class}
\title{Class "Bilan_poids_moyen"}
\description{
Bilan_poids_moyen class The objective is to calculate mean weight of glass
eel which are counted from weight measurements and to reintegrate weight to
number coefficients
}
\note{
We have also tools available to import glass eel measurement from
experimental fishing in the estuary For the charge method dates for the
request are from august to august (a glass eel season)
}
\section{Slots}{
 \describe{ \item{list("data")}{Object of class
\code{"data.frame"} data for bilan lot }\item{:}{Object of class
\code{"data.frame"} data for bilan lot } \item{list("dc")}{Object of class
\code{"RefDC"} refDC an instantiation of the counting device
class}\item{:}{Object of class \code{"RefDC"} refDC an instantiation of the
counting device class} \item{list("anneedebut")}{Object of class
\code{"RefAnnee"} refAnnee allows to choose year of beginning
}\item{:}{Object of class \code{"RefAnnee"} refAnnee allows to choose year
of beginning } \item{list("anneefin")}{Object of class \code{"RefAnnee"}
refAnnee allows to choose year of ending }\item{:}{Object of class
\code{"RefAnnee"} refAnnee allows to choose year of ending }
\item{list("coe")}{Object of class \code{"RefCoe"} class loading coefficient
of conversion between quantity (weights or volumes of glass eel) and numbers
}\item{:}{Object of class \code{"RefCoe"} class loading coefficient of
conversion between quantity (weights or volumes of glass eel) and numbers }
\item{list("liste")}{Object of class \code{"RefListe"} RefListe referential
class choose within a list, here do you want subsamples or not
}\item{:}{Object of class \code{"RefListe"} RefListe referential class
choose within a list, here do you want subsamples or not } }
}
\examples{

showClass("Bilan_poids_moyen")
object=new("Bilan_poids_moyen")

}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other Bilan Classes \code{\linkS4class{Bilan_lot}}
\code{\linkS4class{Bilan_poids_moyen}}
\code{\linkS4class{Bilan_stades_pigm}} 
\code{\linkS4class{Bilan_taille}}
\code{\linkS4class{BilanConditionEnv}} 
\code{\linkS4class{BilanEspeces}}
\code{\linkS4class{BilanFonctionnementDC}}
\code{\linkS4class{BilanFonctionnementDF}}
\code{\linkS4class{BilanMigration}}
\code{\linkS4class{BilanMigrationConditionEnv}}
\code{\linkS4class{BilanMigrationInterAnnuelle}}
\code{\linkS4class{BilanMigrationPar}}
}
\concept{
Bilan Object
}

