% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_poids_moyen.r
\docType{class}
\name{Bilan_poids_moyen-class}
\alias{Bilan_poids_moyen-class}
\title{Class "Bilan_poids_moyen"}
\description{
Bilan_poids_moyen class The objective is to calculate mean weight of glass
eel which are counted from weight measurements and to reintegrate weight to
number coefficients
}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{"data.frame"} data for bilan lot}

\item{\code{calcdata}}{A list containing two processed data frames, data and coe}

\item{\code{dc}}{Object of class \code{\link{RefDC-class}}, the counting device}

\item{\code{anneedebut}}{Object of class \code{\link{RefAnnee-class}}. RefAnnee allows to choose the year of beginning}

\item{\code{anneefin}}{Object of class \code{\link{RefAnnee-class}}
refAnnee allows to choose last year of the Bilan}

\item{\code{coe}}{Object of class \code{\link{RefCoe-class}} class loading coefficient
of conversion between quantity (weights or volumes of glass eel) and numbers}

\item{\code{liste}}{Object of class \code{\link{RefListe-class}} RefListe referential
class choose within a list, here the choice is wether subsamples or not. Subsamples
in the stacomi database are samples with a non null value for parent sample. Migration
counts are never made on subsamples but those can be integrated to calculate mean weights.}
}}
\note{
We have also tools available to import glass eel measurement from
experimental fishing in the estuary For the charge method dates for the
request are from august to august (a glass eel season)
}
\examples{
require(stacomiR)
# launching stacomi without selecting the scheme or interface
stacomi(gr_interface=FALSE,
		login_window=FALSE,
		database_expected=FALSE)
# the following script will load data from the two Anguillere monitored in the Somme

\dontrun{
	#create an instance of the class
	bilPM<-new("Bilan_poids_moyen")
	baseODBC<-get("baseODBC",envir=envir_stacomi)
	baseODBC[c(2,3)]<-rep("iav",2)
	assign("baseODBC",baseODBC,envir_stacomi)
	sch<-get("sch",envir=envir_stacomi)
	assign("sch","iav.",envir_stacomi)
	bilPM@liste<-charge(object=bilPM@liste,listechoice=c("=1",">1","tous"),label="")
	# here I'm using weights when number are larger than 1 ie wet weight
	bilPM<-choice_c(bilPM,
			dc=c(6),			
			anneedebut="2009",
			anneefin="2015",
			selectedvalue=">1",
			silent=FALSE)
	bilPM<-connect(bilPM)	
}	
# load the dataset generated by previous lines
data("bilPM")
# the calculation will fill the slot calcdata
bilPM<-calcule(bilPM)

# A ggplot showing the trend in weight
plot(bilPM, plot.type=1)
# A plot showing both the data and the trend as recorded in the database
plot(bilPM, plot.type=2)
# Same as plot.type=1 but with size according to size of the sample,
# usefull for wet weights where weight are recorded on a number of glass eel
plot(bilPM, plot.type=3)

# First model with nls, see Guerault and Desaunay (1993) 
model(bilPM,model.type="seasonal")
model(bilPM,model.type="seasonal1")
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other Bilan Objects: \code{\link{BilanAgedemer-class}},
  \code{\link{BilanAnnuels-class}},
  \code{\link{BilanArgentee-class}},
  \code{\link{BilanEspeces-class}},
  \code{\link{BilanFonctionnementDC-class}},
  \code{\link{BilanFonctionnementDF-class}},
  \code{\link{BilanMigration-class}},
  \code{\link{BilanMigrationConditionEnv-class}},
  \code{\link{BilanMigrationInterAnnuelle-class}},
  \code{\link{BilanMigrationMult-class}},
  \code{\link{Bilan_carlot-class}},
  \code{\link{Bilan_stades_pigm-class}},
  \code{\link{Bilan_taille-class}}
}
\keyword{classes}

