% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_stades_pigm.r
\docType{class}
\name{Bilan_stades_pigm-class}
\alias{Bilan_stades_pigm-class}
\title{Bilan class for pigment stage structure analysis (glass eel)}
\description{
The pigment stages analysis has been developed to allow to analyze the
change in pigment stage structure for glass eel (Anguilla anguilla).  The
class uses the parameters calibrated by Briand et al. , 2005 to
backcalculate the probable date when the glass eels arrived in the estuary
(i.e. at a fully transparent stage VB. The evolution of pigment stages is
modeled with gamma functions which use a pigment time calculated from daily
temperatures and salinities.  The temperatures has a major influence on the
glass eel pigment stage evolution.
}
\note{
This class is displayed by interface_bilan_stades_pigm, The class uses
temperature (from an abiotic measure station) and mean salinity to calculate
the change towards one stage
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("Bilan_stades_pigm", ...)}.
}
\examples{
#' showClass("Bilan_stades_pigm")
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\references{
BRIAND C., FATIN D., CICCOTTI E. and LAMBERT P., 2005. A
stage-structured model to predict the effect of temperature and salinity on
glass eel Anguilla anguilla pigmentation development. J Fish Biol, 67,
995-1009.
\url{http://www3.interscience.wiley.com/journal/118686679/abstract}
\url{http://www.eptb-vilaine.fr/site/index.php/publications-scientifiques/46-publications-migrateurs/60-dynamique-de-population-et-de-migration-des-civelles-en-estuaire-de-vilaine.}
}
\seealso{
Other Bilan Classes \code{\linkS4class{Bilan_carlot}}
\code{\linkS4class{Bilan_poids_moyen}}
\code{\linkS4class{Bilan_stades_pigm}} 
\code{\linkS4class{Bilan_taille}}
\code{\linkS4class{BilanConditionEnv}} 
\code{\linkS4class{BilanEspeces}}
\code{\linkS4class{BilanFonctionnementDC}}
\code{\linkS4class{BilanFonctionnementDF}}
\code{\linkS4class{BilanMigration}}
\code{\linkS4class{BilanMigrationConditionEnv}}
\code{\linkS4class{BilanMigrationInterAnnuelle}}
\code{\linkS4class{BilanMigrationPar}} 
\code{\link{fnstade}}
}

