% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_taille.r
\docType{class}
\name{Bilan_taille-class}
\alias{Bilan_taille-class}
\title{Class "Bilan_taille" Bilan class}
\description{
Bilan_taille class This class allows to load the following quantitative
sample characteristics : full size, mostly used for yellow eels.
}
\details{
Eels trapped are usually screened and separated into groups, this is not
good practise but still in use The groups used in stacomi are based on the
selectivity curve according to the mesh, roughly the separation takes place
at the L50 for Arzal So individuals might refer to one size class (eg the
size group of large eels) This affiliation is sought at the subsample level
by a query to display on a single line two different characteristics size
and size group Once the data are loaded, the program checks if the size
class group is indicated at the sample or susample level
}
\section{Slots}{

\describe{
\item{\code{dc}}{an object of class \link{RefDC-class}}

\item{\code{taxons}}{An object of class \link{RefTaxon-class}}

\item{\code{stades}}{An object of class \link{RefStades-class} inherited from \link{BilanMigration-class}}

\item{\code{parquan}}{An object of class \link{Refparquan-class}, quantitative parameter}

\item{\code{parqual}}{An object of class \link{Refparqual-class}, quanlitative parameter}

\item{\code{horodate}}{An object of class link{RefHorodate-class}}

\item{\code{requete}}{An Object of class \link[stacomirtools]{RequeteODBC-class}}

\item{\code{datedebut}}{A POSIXlt value, starting horodate for the Bilan}

\item{\code{datefin}}{A POSIXlt value, ending horodate for the Bilan}
}}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("Bilan_taille", ...)}
}
\examples{

showClass("Bilan_taille")
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other Bilan Classes \code{\linkS4class{Bilan_carlot}},
\code{\linkS4class{Bilan_poids_moyen}},
\code{\linkS4class{Bilan_stades_pigm}}, \code{\linkS4class{Bilan_taille}},
\code{\linkS4class{BilanConditionEnv}}, \code{\linkS4class{BilanEspeces}},
\code{\linkS4class{BilanFonctionnementDC}},
\code{\linkS4class{BilanFonctionnementDF}},
\code{\linkS4class{BilanMigration}},
\code{\linkS4class{BilanMigrationConditionEnv}},
\code{\linkS4class{BilanMigrationInterAnnuelle}},
\code{\linkS4class{BilanMigrationPar}}
}
\concept{
Bilan Object
}

