% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanAgedemer.r
\docType{class}
\name{BilanAgedemer-class}
\alias{BilanAgedemer-class}
\title{Class "BilanAgedemer"}
\description{
the BilanAgedemer class is used to dispatch adult salmons to age class according
their size and to basin dependent limits set by the user. Once checked with graphs and summary
statistics, the results are to be written to the database.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame with data generated from the database}

\item{\code{calcdata}}{A list of dc with processed data. This lists consists of two elements
\itemize{
\item (1) data A dataset with age set to be used by the plot and summary methods
\item (2) tj_caracteristitiquelot_car A dataset to import into the database
}}

\item{\code{dc}}{Object of class \link{RefDC-class}: the control devices}

\item{\code{taxons}}{Object of class \link{RefTaxon-class}: the speciess}

\item{\code{stades}}{Object of class \link{RefStades-class} : the stages of the fish}

\item{\code{par}}{Object of class \link{Refpar-class}: the parameters used}

\item{\code{horodatedebut}}{An object of class \code{RefHorodate-class}}

\item{\code{horodatefin}}{An object of class \code{RefHorodate-class}}

\item{\code{limit1hm}}{The size limit, in mm between 1 sea winter fishes and 2 sea winter fishes}

\item{\code{limit2hm}}{The size limit, in mm between 3 sea winter fishes and 3 sea winter fishes}
}}
\note{
This class is displayed by interface_bilan_agedemer
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("BilanAgedemer", ...)}
}
\examples{
require(stacomiR)
# launching stacomi without selecting the scheme or interface
stacomi(gr_interface=FALSE,
		login_window=FALSE,
		database_expected=TRUE)
# the following script will load data from the two Anguillere monitored in the Somme

\dontrun{
	#create an instance of the class
	bilan_adm<-new("BilanAgedemer")
	baseODBC<-get("baseODBC",envir=envir_stacomi)
	baseODBC[c(2,3)]<-rep("logrami",2)
	assign("baseODBC",baseODBC,envir_stacomi)
	sch<-get("sch",envir=envir_stacomi)
	assign("sch","logrami.",envir_stacomi)
	bilan_adm<-choice_c(bilan_adm,
			dc=c(107,108,101),			
			horodatedebut="2012-01-01",
			horodatefin="2012-12-31",
			limit1hm=675,
			limit2hm=875,
			silent=FALSE
			)
	bilan_adm<-connect(bilan_adm)
	
}	
# load the dataset generated by previous lines
data("bilan_adm")
# the calculation will fill the slot calcdata
bilan_adm<-calcule(bilan_adm)
# stages are in bilan_adm@calcdata[["6"]][,"stage"] 
#look at data structure using str(bilan_adm@calcdata[["6"]])

# standard plot as drawn by Laurent Beaulaton (Analyse des données d'argenture acquises en France)
# showing Durif's stage according to size and eye diameter
plot(bilan_adm, plot.type=1)


# print a summary statistic, and save the output in a list for later use
stats<-summary(bilan_adm)

object=new("RefMsg")
msgfr<-charge_avec_filtre(object,lang='French')
str(msgfr)
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other Bilan Objects: \code{\link{BilanAnnuels-class}},
  \code{\link{BilanArgentee-class}},
  \code{\link{BilanEspeces-class}},
  \code{\link{BilanFonctionnementDC-class}},
  \code{\link{BilanFonctionnementDF-class}},
  \code{\link{BilanMigration-class}},
  \code{\link{BilanMigrationConditionEnv-class}},
  \code{\link{BilanMigrationInterAnnuelle-class}},
  \code{\link{BilanMigrationMult-class}},
  \code{\link{Bilan_carlot-class}},
  \code{\link{Bilan_poids_moyen-class}},
  \code{\link{Bilan_stades_pigm-class}},
  \code{\link{Bilan_taille-class}}
}
\keyword{classes}

