% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanConditionEnv.r
\docType{class}
\name{BilanConditionEnv-class}
\alias{BilanConditionEnv}
\alias{BilanConditionEnv-class}
\title{class BilanConditionEnv simple output of one or several environmental
conditions...}
\description{
Annual overview of environmental conditions. Enables to draw charts and
write files.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("BilanConditionEnv", horodate=new("Horodate"),
stationMesure=new("RefStationMesure"), data=data.frame(),
requete=new("RequeteODBCwheredate"))}.
}
\examples{

showClass("BilanConditionEnv")

}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\seealso{
Other Bilan Classes \code{\linkS4class{Bilan_lot}}
\code{\linkS4class{Bilan_poids_moyen}}
\code{\linkS4class{Bilan_stades_pigm}} \code{\linkS4class{Bilan_taille}}
\code{\linkS4class{BilanConditionEnv}} \code{\linkS4class{BilanEspeces}}
\code{\linkS4class{BilanFonctionnementDC}}
\code{\linkS4class{BilanFonctionnementDF}}
\code{\linkS4class{BilanMigration}}
\code{\linkS4class{BilanMigrationConditionEnv}}
\code{\linkS4class{BilanMigrationInterAnnuelle}}
\code{\linkS4class{BilanMigrationPar}}
}
\keyword{classes}
\keyword{dynamic}

