% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanConditionEnv.r
\docType{class}
\name{BilanConditionEnv-class}
\alias{BilanConditionEnv-class}
\title{class BilanConditionEnv simple output of one or several environmental
conditions...}
\description{
Annual overview of environmental conditions. Enables to draw charts and
write files.
}
\section{Slots}{

\describe{
\item{\code{horodate}}{\link{RefHorodate-class}}

\item{\code{stationMesure}}{\link{RefStationMesure-class}}

\item{\code{data}}{\code{data.frame}}

\item{\code{datedebut}}{A \link[base]{-.POSIXt} value}

\item{\code{datefin}}{A \link[base]{-.POSIXt} value}
}}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("BilanConditionEnv", horodate=new("Horodate"),
stationMesure=new("RefStationMesure"), data=data.frame(),
requete=new("RequeteODBCwheredate"))}.
}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\seealso{
Other Bilan Classes \code{\linkS4class{Bilan_carlot}}
\code{\linkS4class{Bilan_poids_moyen}}
\code{\linkS4class{Bilan_stades_pigm}} \code{\linkS4class{Bilan_taille}}
\code{\linkS4class{BilanConditionEnv}} \code{\linkS4class{BilanEspeces}}
\code{\linkS4class{BilanFonctionnementDC}}
\code{\linkS4class{BilanFonctionnementDF}}
\code{\linkS4class{BilanMigration}}
\code{\linkS4class{BilanMigrationConditionEnv}}
\code{\linkS4class{BilanMigrationInterAnnuelle}}
\code{\linkS4class{BilanMigrationPar}}
}
\concept{
Bilan Object
}
\keyword{classes}

