% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanConditionEnv.r
\docType{class}
\name{BilanConditionEnv-class}
\alias{BilanConditionEnv}
\alias{BilanConditionEnv-class}
\alias{bilanConditionEnv}
\title{class BilanConditionEnv simple output of one or several environmental
conditions...}
\description{
Annual overview of environmental conditions. This class enables to draw some plot, but will mostly used to build
joined graphs crossing the information from \link{BilanMigrationMult-class} and \link{BilanMigrationMultConditionEnv-class}
}
\section{Slots}{

\describe{
\item{\code{horodatedebut}}{\link{RefHorodate-class}}

\item{\code{horodatefin}}{\link{RefHorodate-class}}

\item{\code{stationMesure}}{\link{RefStationMesure-class}}

\item{\code{data}}{\code{data.frame}}
}}
\examples{
require(stacomiR)
stacomi(gr_interface=FALSE,
		login_window=FALSE,
		database_expected=FALSE)
\dontrun{
	bil_CE<-new("BilanConditionEnv")
	bil_CE<-choice_c(bil_CE,
			stationMesure=c("temp_gabion","coef_maree"),
			datedebut="2008-01-01",
			datefin="2008-12-31",
			silent=FALSE)	
	bil_CE<-connect(bil_CE)
	
}	

data("bil_CE")
plot(bil_CE,silent=TRUE)
}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\seealso{
Other Bilan.Objects: \code{\link{BilanAgedemer-class}},
  \code{\link{BilanAnnuels-class}},
  \code{\link{BilanArgentee-class}},
  \code{\link{BilanEspeces-class}},
  \code{\link{BilanFonctionnementDC-class}},
  \code{\link{BilanFonctionnementDF-class}},
  \code{\link{BilanMigration-class}},
  \code{\link{BilanMigrationCar-class}},
  \code{\link{BilanMigrationInterAnnuelle-class}},
  \code{\link{BilanMigrationMult-class}},
  \code{\link{BilanMigrationMultConditionEnv-class}},
  \code{\link{Bilan_carlot-class}},
  \code{\link{Bilan_poids_moyen-class}}
}
\keyword{classes}

