% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanEspeces.r
\docType{class}
\name{BilanEspeces-class}
\alias{BilanEspeces-class}
\title{Class "BilanEspeces" Report of the species present at a counting device for
a given period}
\description{
this class is used to make the assessment of all species, and their number,
per month it writes either an histogram or a pie chart of number per
year/week/month
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("BilanEspeces", ...)}.
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other Bilan Classes \code{\linkS4class{Bilan_carlot}}
\code{\linkS4class{Bilan_poids_moyen}}
\code{\linkS4class{Bilan_stades_pigm}} \code{\linkS4class{Bilan_taille}}
\code{\linkS4class{BilanConditionEnv}} \code{\linkS4class{BilanEspeces}}
\code{\linkS4class{BilanFonctionnementDC}}
\code{\linkS4class{BilanFonctionnementDF}}
\code{\linkS4class{BilanMigration}}
\code{\linkS4class{BilanMigrationConditionEnv}}
\code{\linkS4class{BilanMigrationInterAnnuelle}}
\code{\linkS4class{BilanMigrationPar}}
}
\concept{
Bilan Object
}

