% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanFonctionnementDC.r
\docType{class}
\name{BilanFonctionnementDC-class}
\alias{BilanFonctionnementDC-class}
\title{Class "BilanFonctionnementDC" Bilan du fonctionnement du dispositif de
comptage}
\description{
The counting device is not always working. It may me stopped either
following a monitoring protocol, or due to misfunction of the device, this
class allows to draw graphics allowing an overview of the device operation
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("BilanFonctionnementDC", ...)}.
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other Bilan Class \code{\linkS4class{Bilan_carlot}}
\code{\linkS4class{Bilan_poids_moyen}}
\code{\linkS4class{Bilan_stades_pigm}} \code{\linkS4class{Bilan_taille}}
\code{\linkS4class{BilanConditionEnv}} \code{\linkS4class{BilanEspeces}}
\code{\linkS4class{BilanFonctionnementDC}}
\code{\linkS4class{BilanFonctionnementDF}}
\code{\linkS4class{BilanMigration}}
\code{\linkS4class{BilanMigrationConditionEnv}}
\code{\linkS4class{BilanMigrationInterAnnuelle}}
\code{\linkS4class{BilanMigrationPar}}
}
\concept{
Bilan Object
}

