% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bilanFonctionnementDC}
\alias{bilanFonctionnementDC}
\title{Counting device operation for three different counting device in Arzal (Vilaine, France)}
\format{An object of class BilanFonctionnementDC with 4 slots
\describe{
  \item{data}{ A dataframe with 25 rows and 7 variables
			\describe{
  			\item{per_dis_identifiant}{the df or dc unique id}
  			\item{per_date_debut}{the starting date of the counting device operation \code{POSIXct}}
  			\item{per_date_fin}{the ending date of the counting device operation \code{POSIXct}}
  			\item{per_commentaires}{comments on the counting device operation}
  			\item{per_etat_fonctionnement}{Boolean, is the counting device working ?}
  			\item{lot_std_code}{stage id}
  			\item{per_tar_code}{The type of operation for the DC, 1 normal operation, 2 device stopped in normal
				operation (the stop is considered as normal, e.g. you don't monitor video if a cage has been placed to trap fishes), 
				3 stopped for maintenance or other problem, 4 the DC is working but not well (escapement in a tank, high turbidity preventing
video counting...), 5 unknown operation.}
  			\item{libelle}{The label for the type or operation}
  			}
	}
  \item{dc}{the \code{RefDC} the DC with 4 slots
			\describe{
 	 			\item{dc_selectionne}{the selected device}
 				\item{ouvrage}{the dam}
 	 			\item{station}{the monitoring station, a section of river}
 	 			\item{data}{A dataset of all dc present in the database with 10 observations}
			}
	}
  \item{horodatedebut}{the beginning date, a \link{RefHorodate-class}}
  \item{horodatefin}{the ending date, a \link{RefHorodate-class}}
}}
\usage{
bilanFonctionnementDC
}
\description{
This dataset corresponds to data collected at three different control devices.
This object is of class \link{BilanFonctionnementDC-class} with data loaded
it is loaded along with \link{bMM_Arzal} and used in demonstration for the
\link{BilanMigrationMult-class}
}
\keyword{data}

