% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bilanFonctionnementDC_bM}
\alias{bilanFonctionnementDC_bM}
\title{Counting device operation for the video recording (Arzal dam, Vilaine, France).}
\format{An object of class \link{BilanFonctionnementDC-class}}
\usage{
bilanFonctionnementDC_bM
}
\description{
This dataset corresponds to the data collected in the vertical slot fishway for the video 
 recording operation. It is loaded along with \link{bM_Arzal} to
demonstrate the use of the \link{BilanMigration-class} when the database is not loaded
}
\keyword{data}

