% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanFonctionnementDF.r
\docType{class}
\name{BilanFonctionnementDF-class}
\alias{BilanFonctionnementDF-class}
\title{Report on fishway operation}
\description{
Fishways (DF) are of various nature, from very simple eel ladders fed by water discharged from the river,
to more complex fishways with levels adjusted by the opening of various gates and regulators. 
The objective of this class is to provide an assessment of the working status of a fishway throughout the year.
A number of fishes ascending a fishway has meaning only if we know that the fishway is operational, and that the counting 
orerated on the fishway has remained operational.
In the database the operation of the fishway (DF) and counting device (DC) is agregated in one table (t_periodefonctdispositif_per).
The column  per_etat_fonctionnement indicates whether the fishway is operational (with a boolean) and the column per_tar_code indicates
the status of either the fishway or DC. In the database four types of operation are set,  "1"=normal operation,
"2"=Device stopped in nomral operation (ie lift ascending, high tide...),
"3"="Stopped for maintenance or other problem",
"4"="Works but not fully operational, ie flow problem, flood, clogged with debris...",
"5"="Not known")
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame}

\item{\code{df}}{An object of class \code{RefDF-class}}

\item{\code{horodatedebut}}{An object of class \code{RefHorodate-class}}

\item{\code{horodatefin}}{An object of class \code{RefHorodate-class}}
}}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("BilanFonctionnementDF")}.
}
\examples{
require(stacomiR)
stacomi(gr_interface=FALSE,
		login_window=FALSE,
		database_expected=FALSE)
## An example that will work with the database installed only
\dontrun{
	bfDF=new("BilanFonctionnementDF")
	bfDF<-choice_c(bfDF,
			1,
			horodatedebut="2015-01-01",
			horodatefin="2015-12-31",
			silent=TRUE)
	Sys.setenv(TZ='GMT')
	# the times at Arzal are recorded continuously
	# they are converted to date when a time appears while the hour is changing
	# hence the following
	bfDF<-connect(bfDF)
}

	data("bfDF")
	plot(bfDF,plot.type="4")
	# the following examples work but take a while to compute
	\dontrun{
	plot(bfDF,plot.type="1")
	plot(bfDF,plot.type="2",main="A nice title")
	plot(bfDF,plot.type="3",main="A nice title")	
	}


  
	

}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other Bilan.Objects: \code{\link{BilanAgedemer-class}},
  \code{\link{BilanAnnuels-class}},
  \code{\link{BilanArgentee-class}},
  \code{\link{BilanConditionEnv-class}},
  \code{\link{BilanEspeces-class}},
  \code{\link{BilanFonctionnementDC-class}},
  \code{\link{BilanMigration-class}},
  \code{\link{BilanMigrationCar-class}},
  \code{\link{BilanMigrationInterAnnuelle-class}},
  \code{\link{BilanMigrationMult-class}},
  \code{\link{BilanMigrationMultConditionEnv-class}},
  \code{\link{Bilan_carlot-class}},
  \code{\link{Bilan_poids_moyen-class}}
}
\keyword{classes}

