% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanFonctionnementDF.r
\docType{class}
\name{BilanFonctionnementDF-class}
\alias{BilanFonctionnementDF-class}
\title{Class "BilanFonctionnementDF" Report fishway work}
\description{
The DF (Dispositif de Franchissement) is a fishway. It may be automated and
work only at certain times This report allows to see the detail of its work.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("BilanFonctionnementDF")}.
}
\examples{

showClass("BilanFonctionnementDF")

}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\references{
~put references to the literature/web site here ~
}
\seealso{
Other Bilan Class \code{\linkS4class{Bilan_lot}}
\code{\linkS4class{Bilan_poids_moyen}}
\code{\linkS4class{Bilan_stades_pigm}} 
\code{\linkS4class{Bilan_taille}}
\code{\linkS4class{BilanConditionEnv}} 
\code{\linkS4class{BilanEspeces}}
\code{\linkS4class{BilanFonctionnementDC}}
\code{\linkS4class{BilanFonctionnementDF}}
\code{\linkS4class{BilanMigration}}
\code{\linkS4class{BilanMigrationConditionEnv}}
\code{\linkS4class{BilanMigrationInterAnnuelle}}
\code{\linkS4class{BilanMigrationPar}}
}
\keyword{classes}
\keyword{dynamic}

