% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bilanFonctionnementDF}
\alias{bilanFonctionnementDF}
\title{Fishway operation at the Arzal Dam (Vilaine France) (3 Fishways in 2011)}
\format{An object of class BilanFonctionnementDF  \link{BilanFonctionnementDF-class}}
\usage{
bilanFonctionnementDF
}
\description{
This dataset corresponds to the data collected at three different fishways
it is loaded along with \link{bMM_Arzal}
}
\keyword{data}
