% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bilanFonctionnementDF_bM}
\alias{bilanFonctionnementDF_bM}
\title{An object of class \link{BilanFonctionnementDF-class} with data loaded}
\format{An object of class BilanFonctionnementDF}
\usage{
bilanFonctionnementDF_bM
}
\description{
This dataset corresponds to the data collected at three different fishways
it is loaded along with \link{bM_Arzal}
}
\keyword{data}
