% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigration.r
\docType{class}
\name{BilanMigration-class}
\alias{BilanMigration-class}
\title{Migration report for one DC, one species and one stage}
\description{
Migration report for one DC, one species and one stage
}
\section{Slots}{

\describe{
\item{\code{dc}}{Object of class \link{RefDC-class}: the control device}

\item{\code{taxons}}{Object of class \link{RefTaxon-class}: the species}

\item{\code{stades}}{Object of class \link{RefStades-class} : the stage of the fish}

\item{\code{pasDeTemps}}{Object of class \link{PasDeTempsJournalier-class} : the time step 
constrained to daily value and 365 days}

\item{\code{data}}{Object of class \code{data.frame} with data filled in from the connect method}

\item{\code{calcdata}}{A "list" of calculated daily data, one per dc, filled in by the calcule method}

\item{\code{coef_conversion}}{A data.frame of daily weight to number conversion coefficients, filled in by the connect 
method if any weight are found in the data slot.}

\item{\code{time.sequence}}{Object of class \code{POSIXct} : a time sequence of days generated by the calcule method}
}}
\note{
Method \code{plot(...,type="standard")} also calls a function that will write to the database if
a connection to the database is expected.
}
\examples{
library(stacomiR)

stacomi(gr_interface=FALSE,login_window=FALSE,database_expected=FALSE) 
## launches the application in the command line
## here an example of loading
## not run as the program is possibly not installed
## this example generates the bM_Arzal dataset
\dontrun{
	stacomi(gr_interface=FALSE,
			login_window=FALSE,
			database_expected=FALSE)	
	bM_Arzal=new("BilanMigration")
	bM_Arzal=choice_c(bM_Arzal,
			dc=5,
			taxons=c("Liza ramada"),
			stades=c("IND"),
			datedebut="2015-01-01",
			datefin="2015-12-31")
	bM_Arzal<-charge(bM_Arzal)
	bM_Arzal<-connect(bM_Arzal)
}
########################
# loading data
## use the following to get the raw data loaded by the connect method
# not shown there as the database and program might not be installed
# All three classes bilan... were created by the charge and connect method 
# of BilanMigrationMult
# in the previous example
################################
data("bM_Arzal")
data("bilanOperation_bM")
assign("bilanOperation",bilanOperation_bM,envir=envir_stacomi)
data("bilanFonctionnementDF_bM")
assign("bilanFonctionnementDF",bilanFonctionnementDF_bM,envir=envir_stacomi)
data("bilanFonctionnementDC_bM")
assign("bilanFonctionnementDC",bilanFonctionnementDC_bM,envir=envir_stacomi)

########################
# calculations
########################
bM_Arzal<-calcule(bM_Arzal,silent=TRUE)
#Individual plot for all DC (standard), taxon and stage where data present
#silent argument to stop all messages
plot(bM_Arzal,plot.type="standard",silent=TRUE)

#cumulated migration at the station (all stages and DC grouped)
plot(bM_Arzal,plot.type="step")

# data will be written in the data directory specified in the stacomi/calcmig.csv
#file

\dontrun{
	summary(bM_Arzal,silent=TRUE)
}
# this will write the daily bilan for later in in the bilannMigrationInterannuelle-class
\dontrun{
	fn_EcritBilanJournalier(bM_Arzal,silent=TRUE)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other Bilan Objects: \code{\link{BilanAnnuels-class}},
  \code{\link{BilanArgentee-class}},
  \code{\link{BilanEspeces-class}},
  \code{\link{BilanFonctionnementDC-class}},
  \code{\link{BilanFonctionnementDF-class}},
  \code{\link{BilanMigrationConditionEnv-class}},
  \code{\link{BilanMigrationInterAnnuelle-class}},
  \code{\link{BilanMigrationMult-class}},
  \code{\link{Bilan_carlot-class}},
  \code{\link{Bilan_poids_moyen-class}},
  \code{\link{Bilan_stades_pigm-class}},
  \code{\link{Bilan_taille-class}}
}
\keyword{classes}

