% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigration.r
\docType{class}
\name{BilanMigration-class}
\alias{BilanMigration-class}
\title{Migration report for one DC, one species and one stage}
\description{
Migration report for one DC, one species and one stage
}
\section{Slots}{

\describe{
\item{\code{dc}}{Object of class \link{RefDC-class}: the control device}

\item{\code{taxons}}{Object of class \link{RefTaxon-class}: the species}

\item{\code{stades}}{Object of class \link{RefStades-class} : the stage of the fish}

\item{\code{pasDeTemps}}{Object of class \link{PasDeTempsJournalier-class} : the time step 
constrained to daily value and 365 days}

\item{\code{data}}{Object of class \code{data.frame} with data filled in from the connect method}

\item{\code{calcdata}}{A "list" of calculated daily data, one per dc, filled in by the calcule method}

\item{\code{coef_conversion}}{A data.frame of daily weight to number conversion coefficients, filled in by the connect 
method if any weight are found in the data slot.}

\item{\code{time.sequence}}{Object of class \code{POSIXct} : a time sequence of days generated by the calcule method}
}}
\note{
TODO discuss  and how it is used to "write" in the database
}
\examples{
library(stacomiR)

stacomi(gr_interface=FALSE,login_window=FALSE,database_expected=FALSE) 
## launches the application in the command line
## here an example of loading
## not run as the program is possibly not installed
## this example generates the bM_Arzal dataset
\dontrun{
	stacomi(gr_interface=FALSE,
			login_window=FALSE,
			database_expected=FALSE)	
	bM_Arzal=new("BilanMigration")
	bM_Arzal=choice_c(bM_Arzal,
			dc=5,
			taxons=c("Liza ramada"),
			stades=c("IND"),
			datedebut="2015-01-01",
			datefin="2015-12-31")
	bM_Arzal<-connect(bM_Arzal)
}


data("bM_Arzal")
# use the following to get the raw data loaded by the connect method
# not shown there as the database and program might not be installed

#bM_Arzal<-connect(bM_Arzal)
#str(bM_Arzal@data)
# calculations
bM_Arzal<-calcule(bM_Arzal,silent=TRUE)
#Individual plot for all DC (standard), taxon and stage where data present
#silent argument to stop all messages
# this will also write daily and monthly migration balance to the database,
# but only if stacomi has been run with argument : database_expected=FALSE
plot(bM_Arzal,plot.type="standard",silent=TRUE)

#cumulated migration at the station (all stages and DC grouped)
plot(bM_Arzal,plot.type="step")

# data will be written in the data directory specified in the stacomi/calcmig.csv file
summary(bM_Arzal,silent=TRUE)
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other Bilan Class \code{\linkS4class{Bilan_carlot}}, 
\code{\linkS4class{Bilan_poids_moyen}}, 
\code{\linkS4class{Bilan_stades_pigm}}, \code{\linkS4class{Bilan_taille}}, 
\code{\linkS4class{BilanConditionEnv}}, \code{\linkS4class{BilanEspeces}}, 
\code{\linkS4class{BilanFonctionnementDC}}, 
\code{\linkS4class{BilanFonctionnementDF}}, 
\code{\linkS4class{BilanMigration}}, 
\code{\linkS4class{BilanMigrationConditionEnv}}, 
\code{\linkS4class{BilanMigrationInterAnnuelle}}, 
\code{\linkS4class{BilanMigrationPar}}
}
\concept{
Bilan Object
}

