% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigration.r
\docType{class}
\name{BilanMigration-class}
\alias{BilanMigration-class}
\title{Migration report for one DC, one species and one stage}
\description{
Migration report for one DC, one species and one stage
}
\section{Slots}{

\describe{
\item{\code{dc}}{Object of class \link{RefDC-class}: the control device}

\item{\code{taxons}}{Object of class \link{RefTaxon-class}: the species}

\item{\code{stades}}{Object of class \link{RefStades-class} : the stage of the fish}

\item{\code{pasDeTemps}}{Object of class \link{PasDeTempsJournalier-class} : the time step 
constrained to daily value and 365 days}

\item{\code{data}}{Object of class \code{data.frame}}
}}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other Bilan Class \code{\linkS4class{Bilan_carlot}}, 
\code{\linkS4class{Bilan_poids_moyen}}, 
\code{\linkS4class{Bilan_stades_pigm}}, \code{\linkS4class{Bilan_taille}}, 
\code{\linkS4class{BilanConditionEnv}}, \code{\linkS4class{BilanEspeces}}, 
\code{\linkS4class{BilanFonctionnementDC}}, 
\code{\linkS4class{BilanFonctionnementDF}}, 
\code{\linkS4class{BilanMigration}}, 
\code{\linkS4class{BilanMigrationConditionEnv}}, 
\code{\linkS4class{BilanMigrationInterAnnuelle}}, 
\code{\linkS4class{BilanMigrationPar}}
}
\concept{
Bilan Object
}

