% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationConditionEnv.r
\docType{class}
\name{BilanMigrationConditionEnv-class}
\alias{BilanMigrationConditionEnv-class}
\title{Class "BilanMigrationConditionEnv"}
\description{
Enables to compute an annual overview of fish migration and environmental
conditions in the same chart
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("BilanMigrationConditionEnv",
bilanMigration=new("BilanMigration"),
bilanConditionEnv=new("BilanConditionEnv"))}.  \describe{
\item{list("bilanMigration")}{Object of class \code{"BilanMigration"} The
migration overview }\item{:}{Object of class \code{"BilanMigration"} The
migration overview } \item{list("bilanConditionEnv")}{Object of class
\code{"BilanConditionEnv"} The environmental overview}\item{:}{Object of
class \code{"BilanConditionEnv"} The environmental overview} }
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other Bilan Objects: \code{\link{BilanAgedemer-class}},
  \code{\link{BilanAnnuels-class}},
  \code{\link{BilanArgentee-class}},
  \code{\link{BilanEspeces-class}},
  \code{\link{BilanFonctionnementDC-class}},
  \code{\link{BilanFonctionnementDF-class}},
  \code{\link{BilanMigration-class}},
  \code{\link{BilanMigrationInterAnnuelle-class}},
  \code{\link{BilanMigrationMult-class}},
  \code{\link{Bilan_carlot-class}},
  \code{\link{Bilan_poids_moyen-class}},
  \code{\link{Bilan_stades_pigm-class}},
  \code{\link{Bilan_taille-class}}
}
\keyword{classes}

