% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationConditionEnv.r, R/BilanMigrationInterAnnuelle.r
\docType{class}
\name{BilanMigrationConditionEnv-class}
\alias{BilanMigrationConditionEnv}
\alias{BilanMigrationConditionEnv-class}
\title{Class "BilanMigrationConditionEnv"}
\description{
Enables to compute an annual overview of fish migration and environmental
conditions in the same chart

Enables to compute an annual overview of fish migration and environmental
conditions in the same chart
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("BilanMigrationConditionEnv",
bilanMigration=new("BilanMigration"),
bilanConditionEnv=new("BilanConditionEnv"))}.  \describe{
\item{list("bilanMigration")}{Object of class \code{"BilanMigration"} The
migration overview }\item{:}{Object of class \code{"BilanMigration"} The
migration overview } \item{list("bilanConditionEnv")}{Object of class
\code{"BilanConditionEnv"} The environmental overview}\item{:}{Object of
class \code{"BilanConditionEnv"} The environmental overview} }

 Objects can be created by calls of the form
\code{new("BilanMigrationConditionEnv",
bilanMigration=new("BilanMigration"),
bilanConditionEnv=new("BilanConditionEnv"))}.  \describe{
\item{list("bilanMigration")}{Object of class \code{"BilanMigration"} The
migration overview }\item{:}{Object of class \code{"BilanMigration"} The
migration overview } \item{list("bilanConditionEnv")}{Object of class
\code{"BilanConditionEnv"} The environmental overview}\item{:}{Object of
class \code{"BilanConditionEnv"} The environmental overview} }
}
\examples{

showClass("BilanMigrationConditionEnv")


showClass("BilanMigrationConditionEnv")

}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}

Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\references{
\url{http://w3.eptb-vilaine.fr:8080/tracstacomi}

\url{http://w3.eptb-vilaine.fr:8080/tracstacomi}
}
\seealso{
Other Bilan Class \code{\linkS4class{Bilan_lot}}
\code{\linkS4class{Bilan_poids_moyen}}
\code{\linkS4class{Bilan_stades_pigm}} \code{\linkS4class{Bilan_taille}}
\code{\linkS4class{BilanConditionEnv}} \code{\linkS4class{BilanEspeces}}
\code{\linkS4class{BilanFonctionnementDC}}
\code{\linkS4class{BilanFonctionnementDF}}
\code{\linkS4class{BilanMigration}}
\code{\linkS4class{BilanMigrationConditionEnv}}
\code{\linkS4class{BilanMigrationInterAnnuelle}}
\code{\linkS4class{BilanMigrationPar}}

Other Bilan Class \code{\linkS4class{Bilan_lot}}
\code{\linkS4class{Bilan_poids_moyen}}
\code{\linkS4class{Bilan_stades_pigm}} \code{\linkS4class{Bilan_taille}}
\code{\linkS4class{BilanConditionEnv}} \code{\linkS4class{BilanEspeces}}
\code{\linkS4class{BilanFonctionnementDC}}
\code{\linkS4class{BilanFonctionnementDF}}
\code{\linkS4class{BilanMigration}}
\code{\linkS4class{BilanMigrationConditionEnv}}
\code{\linkS4class{BilanMigrationInterAnnuelle}}
\code{\linkS4class{BilanMigrationPar}}
}
\keyword{classes}
\keyword{dynamic}

