% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMult.r
\docType{class}
\name{BilanMigrationMult-class}
\alias{BilanMigrationMult-class}
\title{Migration reports for multiple DC / species / stages}
\description{
Migration counts for several Fish counting devices (DC), several taxa and several stages.
This migration count can be built either by the graphical interface or from the command line 
(see examples).
}
\section{Slots}{

\describe{
\item{\code{dc}}{An object of class \code{RefDC-class}}

\item{\code{taxons}}{An object of class \code{\link{RefTaxon-class}}}

\item{\code{stades}}{An object of class \code{\link{RefStades-class}}}

\item{\code{pasDeTemps}}{An object of class \code{\link{PasDeTempsJournalier-class}}}

\item{\code{data}}{A data.frame containing raw data filled by the connect method}

\item{\code{calcdata}}{A "list" of calculated daily data, one per dc, filled in by the calcule method}

\item{\code{coef_conversion}}{A data.frame of daily weight to number conversion coefficients, filled in by the connect
method if any weight are found in the data slot.}

\item{\code{time.sequence}}{A POSIXt time sequence}
}}
\note{
A Migration Bilan comes from a migration monitoring : the fishes are monitored in a section of river, this section is
called a control station (station). Most often, there is a dam, one or several fishways (DF) which comprise one or several counting devices (DC).
On each counting device, the migration is recorded. It can be either an instant recording (video control) or the use of traps,
Operations are monitoring operations during a period. For each operation, several species of fishes can be recorded (samples). In the case of migratory 
fishes the stage of development is important as it may indicate generic migrations, to and fro, between the river and the sea.

Hence a Multiple Migration Bilan is built from several one or several counting devices (DC), one or several Taxa (Taxon), one or several stages
(stage). The migration can be also recorded not as numbers, but in the case of glass eels, as weight, which will be later transformed to number, 
from daily conversion coefficients. The methods in this class test whether the counts are numbers or another type of quantity.
This class makes different calculations than BilanMigration, it does not handle escapement coefficients,
it uses quantities other than numbers if necessary (only used for glass eel in the connect method).
}
\examples{
library(stacomiR)

stacomi(gr_interface=FALSE,login_window=FALSE,database_expected=FALSE) 
## launches the application in the command line
## here an example of loading
## not run as the program is possibly not installed
## this example generates the bMM_Arzal dataset
\dontrun{
	stacomi(gr_interface=FALSE,
			login_window=FALSE,
			database_expected=FALSE)	
	bMM_Arzal=new("BilanMigrationMult")
	bMM_Arzal=choice_c(bMM_Arzal,
			dc=c(5,6,12),
			taxons=c("Anguilla anguilla"),
			stades=c("AGG","AGJ","CIV"),datedebut="2011-01-01",datefin="2011-12-31")
	bMM_Arzal<-connect(bMM_Arzal)
}


data("bMM_Arzal")
# use the following to get the raw data loaded by the connect method
# not shown there as the database and program might not be installed

#bMM_Arzal<-connect(bMM_Arzal)
#str(bMM_Arzal@data)
# calculations
bMM_Arzal<-calcule(bMM_Arzal,silent=TRUE)
#Individual plot for all DC, taxon and stage where data present
#Silent=TRUE to turn off messages
plot(bMM_Arzal,plot.type="standard",silent=TRUE)

#cumulated migration at the station (all stages and DC grouped)
plot(bMM_Arzal,plot.type="step")
#some plots are a mixture of french and english
#changing the reference table names to english should shift all to english.
#same for spanish.
# combined plot for ggplot2
plot(bMM_Arzal,plot.type="multiple")
# data will be written in the data directory specified in the stacomi/calcmig.csv file
summary(bMM_Arzal,silent=TRUE)
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

