% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMultConditionEnv.r
\docType{class}
\name{BilanMigrationMultConditionEnv-class}
\alias{BilanMigrationMultConditionEnv-class}
\alias{BilanMigrationMultConditionEnv}
\alias{bilanmigrationmultconditionenv}
\alias{bmmCE}
\title{Class "BilanMigrationMultConditionEnv"}
\description{
Enables to compute an annual overview of fish migration and environmental
conditions in the same chart
}
\section{Slots}{

\describe{
\item{\code{bilanMigrationMult}}{\link{BilanMigrationMult-class}}

\item{\code{bilanConditionEnv}}{\link{BilanConditionEnv-class}}
}}

\examples{
require(stacomiR)
stacomi(gr_interface=FALSE,
		login_window=FALSE,
		database_expected=FALSE)
\dontrun{
	bmmCE<-new("BilanMigrationMultConditionEnv")
	bmmCE<-choice_c(bmmCE,
			dc=c(5,6,12),
			taxon=c("Anguilla anguilla"),
			stade=c("AGJ","AGG","CIV"),
			stationMesure=c("temp_gabion","coef_maree","phases_lune"),
			datedebut="2008-01-01",
			datefin="2008-12-31",
			silent=FALSE)	
	bmmCE<-charge(bmmCE)
	bmmCE<-connect(bmmCE)
	bmmCE<-calcule(bmmCE,silent=TRUE)
	
}	

data("bmmCE")

plot(bmmCE,
		color_station=c("temp_gabion"="red","coef_maree"="blue","phases_lune"="pink"),
 		color_dc=c("5"="yellow","6"="orange","12"="purple")
)
}
\seealso{
Other Bilan Objects: \code{\link{BilanAgedemer-class}},
  \code{\link{BilanAnnuels-class}},
  \code{\link{BilanArgentee-class}},
  \code{\link{BilanConditionEnv-class}},
  \code{\link{BilanEspeces-class}},
  \code{\link{BilanFonctionnementDC-class}},
  \code{\link{BilanFonctionnementDF-class}},
  \code{\link{BilanMigration-class}},
  \code{\link{BilanMigrationCar-class}},
  \code{\link{BilanMigrationInterAnnuelle-class}},
  \code{\link{BilanMigrationMult-class}},
  \code{\link{Bilan_carlot-class}},
  \code{\link{Bilan_poids_moyen-class}}

Other Bilan Objects: \code{\link{BilanAgedemer-class}},
  \code{\link{BilanAnnuels-class}},
  \code{\link{BilanArgentee-class}},
  \code{\link{BilanConditionEnv-class}},
  \code{\link{BilanEspeces-class}},
  \code{\link{BilanFonctionnementDC-class}},
  \code{\link{BilanFonctionnementDF-class}},
  \code{\link{BilanMigration-class}},
  \code{\link{BilanMigrationCar-class}},
  \code{\link{BilanMigrationInterAnnuelle-class}},
  \code{\link{BilanMigrationMult-class}},
  \code{\link{Bilan_carlot-class}},
  \code{\link{Bilan_poids_moyen-class}}
}
\author{
cedric.briand"at"eptb-vilaine.fr marion.legrand"at"logrami.fr

Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{classes}
