% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationPar.r
\docType{class}
\name{BilanMigrationPar-class}
\alias{BilanMigrationPar-class}
\title{Class BilanMigrationPar, Migration report along with quantitative and
qualitative characteristics}
\description{
Migration along with qualitative or quantitative characteristics or both
(e.g.) weight of eels according to the size class per period of time, weight
of fish according to gender. This class does not split migration evenly over 
time period. So, unlike calculations made in class BilanMigration and BilanMigrationMult
the whole time span of the migration operation is not considered, only  the date of beginning of 
the operation is used to perform calculation.
}
\section{Slots}{

\describe{
\item{\code{parquan}}{An object of class \link{Refparquan-class}, quantitative parameter}

\item{\code{parqual}}{An object of class \link{Refparqual-class}, quanlitative parameter}

\item{\code{echantillon}}{An object of class \link{RefChoix-class}", vector of choice}

\item{\code{valeurs_possibles}}{A \code{data.frame} choice among possible choice of a qualitative parameter (discrete)}

\item{\code{dc}}{an object of class \link{RefDC-class} inherited from \link{BilanMigration-class}}

\item{\code{taxons}}{An object of class \link{RefTaxon-class} inherited from \link{BilanMigration-class}}

\item{\code{stades}}{An object of class \link{RefStades-class} inherited from \link{BilanMigration-class}}

\item{\code{pasDeTemps}}{An object of class \link{PasDeTempsJournalier-class} inherited from \link{BilanMigration-class}}

\item{\code{data}}{A\code{data.frame} inherited from \link{BilanMigration-class}, stores the results}

\item{\code{time.sequence}}{An object of class "POSIXct" inherited from \link{BilanMigration-class}}
}}
\note{
the program by default uses two parameter choice, checking box "aucun" will
allow the program to ignore the parameter

program : default two parameter choice, checking box "aucun" will allow the program to ignore the parameter
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("BilanMigrationPar", ...)}.  they are loaded by the interface
using interface_BilanMigrationPar function.
}
\examples{

showClass("BilanMigrationPar")

}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}

Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other Bilan Class 
\code{\linkS4class{Bilan_lot}}
\code{\linkS4class{Bilan_poids_moyen}}
\code{\linkS4class{Bilan_stades_pigm}} 
\code{\linkS4class{Bilan_taille}}
\code{\linkS4class{BilanConditionEnv}} 
\code{\linkS4class{BilanEspeces}}
\code{\linkS4class{BilanFonctionnementDC}}
\code{\linkS4class{BilanFonctionnementDF}}
\code{\linkS4class{BilanMigration}}
\code{\linkS4class{BilanMigrationConditionEnv}}
\code{\linkS4class{BilanMigrationInterAnnuelle}}
\code{\linkS4class{BilanMigrationPar}}
}

