% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bilanOperation}
\alias{bilanOperation}
\title{Counting operations for three different counting device in Arzal (Vilaine, France)}
\format{An object of class BilanOperation}
\usage{
bilanOperation
}
\description{
This dataset corresponds to the data collected at three different control devices
It is an object of class \link{BilanOperation-class} with data loaded.
it is loaded along with \link{bMM_Arzal}
}
\keyword{data}
