% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bmi}
\alias{bmi}
\title{Daily glass eel and elver migration from 1984 to 2016 in the Sèvre Niortaise}
\format{An object of class \link{BilanMigrationInterAnnuelle-class} with data loaded.}
\usage{
bmi
}
\description{
The first eel trapping ladder in France was built by Antoine Legault and the team from Rennes
in the Sevre Niortaise, Marais Poitevin. Also refurbished several times since 1984 it has been 
operational at the same location and provides one of the longest series of eel migration. 
For this reason,
the dataset has been loaded as an example for the BilanMigrationInterAnnuelle-class. It has been
kindly given by the parc du Marais Poitevin.
}
\keyword{data}
