% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calcmig}
\alias{calcmig}
\title{A data frame containing the default connection arguments.
The program will use a file installed in c:/program files/stacomi but
if not found will switch to this default value}
\format{An object of class \code{data.frame} with 1 rows and 11 columns.}
\usage{
calcmig
}
\description{
A data frame containing the default connection arguments.
The program will use a file installed in c:/program files/stacomi but
if not found will switch to this default value
}
\keyword{datasets}

