% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanAgedemer.r
\docType{methods}
\name{calcule,BilanAgedemer-method}
\alias{calcule,BilanAgedemer-method}
\title{Calcule method for BilanAgedemer, this method will split the data according to cut}
\usage{
\S4method{calcule}{BilanAgedemer}(object, silent)
}
\arguments{
\item{object}{An object of class \code{\link{BilanAgedemer-class}}}

\item{silent}{Default FALSE, if TRUE the program should no display messages}
}
\value{
An object of class \link{BilanAgedemer-class}
}
\description{
Calcule method for BilanAgedemer, this method will split the data according to cut
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
