% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanAgedemer.r
\docType{methods}
\name{calcule,BilanAgedemer-method}
\alias{calcule,BilanAgedemer-method}
\title{Calcule method for BilanAgedemer, this method will pass the data from long to wide format 
( one line per individual) and calculate Durif silvering index and Pankhurst and Fulton's K.}
\usage{
\S4method{calcule}{BilanAgedemer}(object, silent)
}
\arguments{
\item{object}{An object of class \code{\link{BilanAgedemer-class}}}

\item{silent}{Boolean, if TRUE, information messages are not displayed, only warnings and errors}
}
\description{
Calcule method for BilanAgedemer, this method will pass the data from long to wide format 
( one line per individual) and calculate Durif silvering index and Pankhurst and Fulton's K.
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

