% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigration.r
\docType{methods}
\name{calcule,BilanMigration-method}
\alias{calcule,BilanMigration-method}
\title{calcule method for BilanMigration}
\usage{
\S4method{calcule}{BilanMigration}(object)
}
\arguments{
\item{object}{An object of class \code{\link{BilanMigrationMult-class}}}

\item{negative}{a boolean indicating if a separate sum must be done for positive and negative values, if true, positive and negative counts return 
different rows}
}
\value{
BilanMigrationMult with slots filled by user choice
}
\description{
does the calculation once data are filled 
in with the connect method
}
\note{
The class BilanMigrationMult does not handle  escapement rates.
'devenir' i.e. the destination of the fishes.
}

