% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationInterAnnuelle.r
\docType{methods}
\name{calcule,BilanMigrationInterAnnuelle-method}
\alias{calcule,BilanMigrationInterAnnuelle-method}
\title{calcule method for BilanMigrationInterannuelle}
\usage{
\S4method{calcule}{BilanMigrationInterAnnuelle}(object, silent = FALSE,
  timesplit = "mois")
}
\arguments{
\item{object}{An object of class \code{\link{BilanMigrationInterannuelle-class}}}

\item{silent}{Boolean, if TRUE, information messages are not displayed, only warnings and errors}

\item{timesplit}{One of "day","week","month","2 weeks", "jour","semaine","quinzaine","mois"}
}
\value{
BilanMigration with calcdata slot filled.
}
\description{
calcule method for BilanMigrationInterannuelle
}
\note{
The class BilanMigration does not handle escapement rates nor 
'devenir' i.e. the destination of the fishes.
}

