% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMult.r
\docType{methods}
\name{calcule,BilanMigrationMult-method}
\alias{calcule,BilanMigrationMult-method}
\alias{calcule.BilanMigrationMult}
\alias{calcule.bilanmigrationMult}
\title{#' Transforms migration per period to daily migrations, and performs the conversion from weights to number is data
are stored as weights (glass eel). This calculation is performed in a loop for all dc.}
\usage{
\S4method{calcule}{BilanMigrationMult}(object, negative = FALSE,
  silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{BilanMigrationMult-class}}

\item{negative}{a boolean indicating if a separate sum must be done for positive and negative values, if true, positive and negative counts return 
different rows}

\item{silent}{Default FALSE, should messages be stopped}
}
\value{
BilanMigrationMult with a list in slot calcdata. For each dc one will find a list with the following elements
\describe{
\item{method}{In the case of instantaneous periods (video counting) the sum of daily values is done by the \link{fun_bilanMigrationMult} method and the value indicated in method is "sum".
 If any migration monitoring period is longer than a day, then the migration is split using the \link{fun_bilanMigrationMult_Overlaps} function and the value indicated in the 
method is "overlaps" as the latter method uses the overlap package to split migration period.}
\item{data}{the calculated data.}
\item{contient_poids}{A boolean which indicates, in the case of glass eel, that the function \link{fun_weight_conversion} has been run to convert the weights to numbers using the weight
to number coefficients in the database (see link{Bilan_poids_moyen}).}
\item{negative}{A parameter indicating if negative migration (downstream in the case of upstream migration devices) have been converted to positive numbers,
not developped yet}}
}
\description{
The calculation must be launched once data are filled by the connect method. Currently the negative argument
has no effect.
}
\note{
The class does not handle escapement rates, though structurally those are present in the database. If you 
want to use those you will have to do the calculation manually from the data in \code{BilanMigrationMult@data}
}

