\name{calcule-methods}
\docType{methods}
\alias{calcule-methods}
\alias{calcule,BilanMigration-method}
\alias{calcule,BilanMigrationConditionEnv-method}
\alias{calcule,BilanMigrationPar-method}
\title{ Methods for Function calcule}
\description{
 calcule is called when a class needs computation and not only data collected from the database
}
\section{Methods}{
\describe{

\item{\code{signature(objet = "BilanMigration")}}{ ~~describe this method here }

\item{\code{signature(objet = "BilanMigrationConditionEnv")}}{ ~~describe this method here }

\item{\code{signature(objet = "BilanMigrationPar")}}{ ~~describe this method here }
}}
\keyword{methods}
\keyword{Bilan-methods}
