% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_ge_weight.R
\name{calcule,report_ge_weight-method}
\alias{calcule,report_ge_weight-method}
\alias{calcule.report_ge_weight}
\title{Calcule method for report_ge_weight}
\usage{
\S4method{calcule}{report_ge_weight}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_ge_weight-class}}

\item{silent}{Boolean, if TRUE, information messages are not displayed, only warnings and errors}
}
\value{
An object of class  \link{report_ge_weight-class} with  \code{@calcdata[["data"]]} (essentially a selection of 
columns and renaming from \code{@data}) and \code{coe} daily coefficients extracted from the database 
\code{@calcdata[["coe"]]} and prepared for graphs
}
\description{
Calcule method for report_ge_weight
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
