% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_ge_weight.R
\docType{methods}
\name{calcule,report_ge_weight-method}
\alias{calcule,report_ge_weight-method}
\alias{calcule.report_ge_weight}
\title{Calcule method for report_ge_weight}
\usage{
\S4method{calcule}{report_ge_weight}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{report_ge_weight-class}}}

\item{silent}{Boolean, if TRUE, information messages are not displayed, only warnings and errors}
}
\description{
Calcule method for report_ge_weight
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
