% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_env.R
\name{calcule,report_mig_env-method}
\alias{calcule,report_mig_env-method}
\alias{calcule.report_mig_env}
\title{Calculations for migration in the class \link{report_mig_env-class}}
\usage{
\S4method{calcule}{report_mig_env}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_mig_env-class}}

\item{silent}{Boolean default FALSE, if TRUE information messages not displayed}
}
\value{
\link{report_mig_env-class} with data in slot r_mig_env@report_mig_mult@calcdata
}
\description{
Runs the calcule method in \link{report_mig_mult-class}
}
