% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_species.R
\name{calcule,report_species-method}
\alias{calcule,report_species-method}
\alias{calcule.report_species}
\title{calcule method for report_species}
\usage{
\S4method{calcule}{report_species}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_species-class}}

\item{silent}{Boolean, if TRUE, information messages are not displayed, only warnings and errors}
}
\value{
An object of class \link{report_species-class} with calcdata slot filled.
}
\description{
The number will be split according to the split argument passed to the class, e.g.
per year or month or week. Data from different DC will be grouped. Counts are given per taxa,
unless there are several stages, in which case the counts correspond to taxa + stage.
}
