% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanEspeces.r
\docType{methods}
\name{charge,BilanEspeces-method}
\alias{charge,BilanEspeces-method}
\title{charge method for BilanEspeces
verifies the content of objects and calls the connect method}
\usage{
\S4method{charge}{BilanEspeces}(object, ...)
}
\value{
BilanEspeces with slots filled by user choice
}
\description{
charge method for BilanEspeces
verifies the content of objects and calls the connect method
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

