% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanFonctionnementDC.r
\docType{methods}
\name{charge,BilanFonctionnementDC-method}
\alias{charge,BilanFonctionnementDC-method}
\title{charge method for BilanFonctionnementDC}
\usage{
\S4method{charge}{BilanFonctionnementDC}(object, h)
}
\arguments{
\item{h}{A handler passed from the graphical interface}

\item{objet}{An object of class \link{BilanFonctionnementDC-class}}
}
\value{
An object of class \link{BilanFonctionnementDC-class}
}
\description{
used by the graphical interface to retreive the objects of Referential classes
assigned to envir_stacomi
}
\author{
cedric.briand
}

