% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanFonctionnementDF.r
\docType{methods}
\name{charge,BilanFonctionnementDF-method}
\alias{charge,BilanFonctionnementDF-method}
\title{charge method for BilanFonctionnementDF}
\usage{
\S4method{charge}{BilanFonctionnementDF}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{BilanFonctionnementDF-class}}

\item{silent}{Keeps program silent}
}
\value{
An object of class \link{BilanFonctionnementDF-class}
}
\description{
used by the graphical interface to retrieve the objects of Referential classes
assigned to envir_stacomi
}
\keyword{internal}

