% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigration.r
\docType{methods}
\name{charge,BilanMigration-method}
\alias{charge,BilanMigration-method}
\title{charge method for BilanMigration}
\usage{
\S4method{charge}{BilanMigration}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{BilanMigration-class}}}

\item{silent}{Should the program be returning messages}
}
\value{
An object of class \link{BilanMigration-class} with slots filled by user choice
}
\description{
this method creates additional classes in envir_stacomi for later use in plot (operations, 
DF operation, DC operation.
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

