% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMult.r
\docType{methods}
\name{charge,BilanMigrationMult-method}
\alias{charge,BilanMigrationMult-method}
\title{charge method for BilanMigrationMult}
\usage{
\S4method{charge}{BilanMigrationMult}(object, ...)
}
\value{
BilanMigrationMult with slots filled by user choice
}
\description{
Used by the graphical interface to collect and test objects in the environment envir_stacomi
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

