% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMult.r
\docType{methods}
\name{charge,BilanMigrationMult-method}
\alias{charge,BilanMigrationMult-method}
\title{charge method for BilanMigrationMult}
\usage{
\S4method{charge}{BilanMigrationMult}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{BilanMigrationMult-class}}

\item{silent}{Default FALSE, if TRUE the program should no display messages}
}
\value{
BilanMigrationMult with slots filled by user choice
}
\description{
Used by the graphical interface to collect and test objects in the environment envir_stacomi, 
fills also the data slot by the connect method
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

