% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMultConditionEnv.r
\docType{methods}
\name{charge,BilanMigrationMultConditionEnv-method}
\alias{charge,BilanMigrationMultConditionEnv-method}
\title{charge method for BilanMigrationMultConditionEnv class}
\usage{
\S4method{charge}{BilanMigrationMultConditionEnv}(object, silent)
}
\arguments{
\item{object}{An object of class \link{BilanMigrationMultConditionEnv-class}}

\item{silent}{Should the function remain silent (boolean)}
}
\description{
charge method for BilanMigrationMultConditionEnv class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
