% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_df.R
\docType{methods}
\name{charge,ref_df-method}
\alias{charge,ref_df-method}
\title{Loading method for DF referential objects}
\usage{
\S4method{charge}{ref_df}(object)
}
\arguments{
\item{object}{An object of class \link{ref_df-class}}
}
\value{
An object of class ref_df
}
\description{
Loading method for DF referential objects
}
\examples{
\dontrun{
 object=new("ref_df")
charge(object)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

