% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_env.R
\name{charge,ref_env-method}
\alias{charge,ref_env-method}
\title{Loading method for ref_env referential object}
\usage{
\S4method{charge}{ref_env}(object)
}
\arguments{
\item{object}{An object of class \link{ref_env-class}}
}
\value{
An S4 object of class ref_env with data loaded from the database
}
\description{
Loading method for ref_env referential object
}
\examples{
\dontrun{
 object=new('ref_env')
 charge(object)
}
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
