% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_par.R
\name{charge,ref_par-method}
\alias{charge,ref_par-method}
\alias{charge.ref_par}
\title{Loading method for ref_par referential objects}
\usage{
\S4method{charge}{ref_par}(object)
}
\arguments{
\item{object}{An object of class \link{ref_par-class}}
}
\value{
An S4 object of class ref_par

An S4 object of class \link{ref_par-class}
}
\description{
Loading method for ref_par referential objects
}
\examples{
\dontrun{
 object=new('ref_par')
charge(object)
}
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
