% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_stage.R
\docType{methods}
\name{charge,ref_stage-method}
\alias{charge,ref_stage-method}
\title{Loading method for ref_stage referential objects}
\usage{
\S4method{charge}{ref_stage}(object)
}
\arguments{
\item{object}{An object of class \link{ref_stage-class}}
}
\value{
An S4 object of class ref_stage
}
\description{
Loading method for ref_stage referential objects
}
\examples{
\dontrun{
 object=new("ref_stage")
 charge(object)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

