% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_taxa.R
\docType{methods}
\name{charge,ref_taxa-method}
\alias{charge,ref_taxa-method}
\title{Loading method for ref_taxa referential objects}
\usage{
\S4method{charge}{ref_taxa}(object)
}
\arguments{
\item{object}{An object of class \link{ref_taxa-class}}
}
\value{
An S4 object of class ref_taxa
}
\description{
Loading method for ref_taxa referential objects
}
\examples{
\dontrun{
 object=new("ref_taxa")
 charge(object)}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
