% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_year.R
\docType{methods}
\name{charge,ref_year-method}
\alias{charge,ref_year-method}
\title{Loading method for ref_year referential objects}
\usage{
\S4method{charge}{ref_year}(object, objectreport = "report_ge_weight")
}
\arguments{
\item{object}{An object of class ref_year}

\item{objectreport}{The object report, default "report_ge_weight" other possible value report_mig_interannual}
}
\value{
object An object of class ref_year with slot data filled with the selected value
}
\description{
Selects year available either in the bjo table (if report_object==report_migInterannelle) or in the t_operation_ope table
}
\examples{
  
\dontrun{
object=new("ref_year")
charge(object)
 validObject(annee)
showMethods("charge")
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

