% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefAnnee.r
\docType{methods}
\name{charge,RefAnnee-method}
\alias{charge,RefAnnee-method}
\title{Loading method for RefAnnee referential objects}
\usage{
\S4method{charge}{RefAnnee}(object)
}
\arguments{
\item{object}{An object of class RefAnnee}
}
\value{
object An object of class RefAnnee with slot data filled with the selected value
}
\description{
Selects year avalaible in the t_operation_ope table
}
\examples{
  object=new("RefAnnee")
charge(object)
 validObject( annee)
showMethods("charge")
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

