% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ref_year.r
\docType{methods}
\name{charge,Ref_year-method}
\alias{charge,Ref_year-method}
\title{Loading method for Ref_year referential objects}
\usage{
\S4method{charge}{Ref_year}(object, objectBilan = "Bilan_poids_moyen")
}
\arguments{
\item{object}{An object of class Ref_year}

\item{objectBilan}{The object Bilan, default "Bilan_poids_moyen" other possible value BilanMigrationInterAnnuelle}
}
\value{
object An object of class Ref_year with slot data filled with the selected value
}
\description{
Selects year avalaible either in the bjo table (if ObjetBilan==BilanMigrationInterannelle) or in the t_operation_ope table
}
\examples{
  
\dontrun{
object=new("Ref_year")
charge(object)
 validObject(annee)
showMethods("charge")
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

