% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefCheckBox.r
\docType{methods}
\name{charge,RefCheckBox-method}
\alias{charge,RefCheckBox-method}
\title{Loading method for ReCheckBox referential objects}
\usage{
\S4method{charge}{RefCheckBox}(object, title, labels, checked)
}
\arguments{
\item{object}{An object of class \link{ReCheckBox-class}}

\item{title}{Title of the frame}

\item{labels}{Labels for checked}

\item{checked}{Vector of boolean indicating if refCheckBox are checked}
}
\value{
An object of class \link{ReCheckBox-class}
}
\description{
Loading method for ReCheckBox referential objects
}
\examples{
\dontrun{
 object=new("RefCheckBox")
charge(object,title="essai",labels=c("par1","par2","par3"),checked=c(TRUE,TRUE,TRUE))
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

