% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refparqual.r
\docType{methods}
\name{charge,Refparqual-method}
\alias{charge,Refparqual-method}
\title{Loading method for Reparqual referential objects}
\usage{
\S4method{charge}{Refparqual}(object)
}
\value{
An S4 object of class Refparqual
}
\description{
Loading method for Reparqual referential objects
}
\examples{
 object=new("Refparqual")
 charge(object)
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

