% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReftextBox.r
\docType{methods}
\name{charge,RefTextBox-method}
\alias{charge,RefTextBox-method}
\title{Loading method for ReTextBox referential objects}
\usage{
\S4method{charge}{RefTextBox}(object, title, label)
}
\arguments{
\item{object}{An object of class \link{RefTextBox-class}}

\item{title}{A title for the frame}

\item{label}{A label for the TextBox}
}
\value{
An S4 object of class RefTextBox
}
\description{
Loading method for ReTextBox referential objects
}
\examples{
\dontrun{
 object=new("RefTextBox")
charge(object,title="un titre",label="20")
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
