% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_mult.R
\name{charge,report_mig_mult-method}
\alias{charge,report_mig_mult-method}
\alias{charge.report_mig_mult}
\title{charge method for report_mig_mult}
\usage{
\S4method{charge}{report_mig_mult}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_mig_mult-class}}

\item{silent}{Default FALSE, if TRUE the program should no display messages}
}
\value{
An object of class \link{report_mig_mult-class} with slots filled from values assigned in \code{envir_stacomi} environment
}
\description{
For the report_mig_mult class the charge method must be run 
to load data on migration control operations
fishway operations, and counting devices operations as data from those are displayed in the main plots.
For other classes the charge method is only used by the graphical interface (shiny)
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
