% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_sea_age.R
\name{charge,report_sea_age-method}
\alias{charge,report_sea_age-method}
\alias{charge.report_sea_age}
\title{Loads data and check that all choices in the graphical interface have been made.}
\usage{
\S4method{charge}{report_sea_age}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_sea_age-class}}

\item{silent}{Boolean, if TRUE, information messages are not displayed}
}
\value{
An object of class \link{report_sea_age-class} with slots filled with user choice

An object of class \link{report_sea_age-class} with slots filled from values assigned in \code{envir_stacomi} environment
}
\description{
It is not necessary to run this method if the choice_c method has been run.
This method verifies that boxes have been clicked in the user interface and gets the objects pasted in
envir_stacomi
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
\keyword{internal}
