% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_silver_eel.R
\docType{methods}
\name{charge,report_silver_eel-method}
\alias{charge,report_silver_eel-method}
\alias{charge.report_silver_eel}
\title{charge method for report_silver_eel class}
\usage{
\S4method{charge}{report_silver_eel}(object, h)
}
\arguments{
\item{object}{An object of class \link{report_silver_eel-class}}

\item{h}{a handler}
}
\value{
An object of class \link{report_silver_eel-class} with slots filled with user choice

An object of the class
}
\description{
this method verifies that boxes have been clicked in the user interface and gets the objects pasted in 
envir_stacomi. It is not necessary to run this method when loading from the command line using the
choice_c method
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}
