% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_species.R
\name{charge,report_species-method}
\alias{charge,report_species-method}
\alias{charge.report_species}
\title{charge method for report_species}
\usage{
\S4method{charge}{report_species}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_species-class}}

\item{silent}{Stops displaying the messages.}
}
\value{
report_species with slots filled by user choice

An object of class \link{report_species-class} with slots filled from values assigned in \code{envir_stacomi} environment
}
\description{
Verifies the content of objects when the graphical interface is used, it is not necessary
to call the charge method if the choice_c method has been used
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
\keyword{internal}
