% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generic.R
\name{charge}
\alias{charge}
\title{Generic method to load referentials}
\usage{
charge(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Additional parm}
}
\description{
Generic method to load referentials
}
\author{
cedric.briand
}
