% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefStades.r
\docType{methods}
\name{charge_avec_filtre,RefStades-method}
\alias{charge_avec_filtre,RefStades-method}
\title{Loading method for RefStades referential objects searching only those stages existing for a DC and a Taxon}
\usage{
\S4method{charge_avec_filtre}{RefStades}(object, dc_selectionne,
  taxon_selectionne)
}
\arguments{
\item{object}{An object of class \link{RefStades-class}}

\item{dc_selectionne}{The selected counting device}

\item{taxon_selectionne}{The selected species}
}
\value{
An S4 object of class \link{RefStades-class}
}
\description{
Loading method for RefStades referential objects searching only those stages existing for a DC and a Taxon
}
\examples{
\dontrun{
 dc_selectionne=6
taxon_selectionne=2038
 object=new("RefStades")
 charge_avec_filtre(object,dc_selectionne,taxon_selectionne)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

