% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_parqual.R
\docType{methods}
\name{charge_complement,ref_parqual-method}
\alias{charge_complement,ref_parqual-method}
\title{Loads an additional dataset
this method is loaded to obtain the possible values of a qualitative parameter. Here data only contains one line}
\usage{
\S4method{charge_complement}{ref_parqual}(object)
}
\arguments{
\item{object}{An object of class \link{ref_parqual-class}}
}
\value{
An S4 object of class ref_parqual with the valqual slot filled
}
\description{
Loads an additional dataset
this method is loaded to obtain the possible values of a qualitative parameter. Here data only contains one line
}
\examples{
\dontrun{
 dc_selectionne=6
taxa_selectionne=2038
 stage_selectionne="AGJ"
 object=new("ref_parqual")
 object<-charge(object)
 charge_complement(object)
}		
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
