% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refparqual.r
\docType{methods}
\name{charge_complement,Refparqual-method}
\alias{charge_complement,Refparqual-method}
\title{Loads an additional dataset
this method is loaded to obtain the possible values of a qualitative parameter. Here data only contains one line}
\usage{
\S4method{charge_complement}{Refparqual}(object)
}
\arguments{
\item{object}{An object of class \link{Refparqual-class}}
}
\value{
An S4 object of class Refparqual with the valqual slot filled
}
\description{
Loads an additional dataset
this method is loaded to obtain the possible values of a qualitative parameter. Here data only contains one line
}
\examples{
\dontrun{
 dc_selectionne=6
taxon_selectionne=2038
 stade_selectionne="AGJ"
 object=new("Refparqual")
 object<-charge(object)
 charge_complement(object)
}		
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

