% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generic.R
\name{charge_complement}
\alias{charge_complement}
\title{Generic method to load additional data}
\usage{
charge_complement(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Additional parms}
}
\description{
Generic method to load additional data
}
\author{
cedric.briand
}
