% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_taxa.R
\docType{methods}
\name{charge_with_filter,ref_taxa-method}
\alias{charge_with_filter,ref_taxa-method}
\title{Loading method for ref_taxa referential objects searching only taxa existing for a DC}
\usage{
\S4method{charge_with_filter}{ref_taxa}(object, dc_selectionne)
}
\arguments{
\item{object}{An object of class \link{ref_taxa-class}}

\item{dc_selectionne}{A counting device selected, only taxa attached to this dc are selected}
}
\value{
An S4 object of class ref_taxa
}
\description{
Loading method for ref_taxa referential objects searching only taxa existing for a DC
}
\examples{
\dontrun{
 dc_selectionne=6
 object=new("ref_taxa")
 charge_with_filter(object,dc_selectionne=dc_selectionne)}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
