% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitaires.r
\name{chargecsv}
\alias{chargecsv}
\title{chargecsv loads the informations stored in c:/program
files/stacomi/calcmig.csv file}
\usage{
chargecsv()
}
\value{
a list with the datawd place and the baseODBC vector
}
\description{
be sure to configure your odbc link to the
database, the name is the name of the first column of the calcmig.csv file. 
	\code{uid}, \code{pwd} are identifier and password to connect to the database, they should
correspond to your own schema in the database. \code{pgwd} is the path to the R
source if you plan not to use the compiler but run manually using inst/config/stacomi_manual_launch.R for development.\cr
	\code{datawd}, is the the
directory where you want to place the outputs, mostly tables, from the
program, default to ~//CalcmigData lang, is either one of French, English or
Spanish
	other fields correspond to sqldf options.
}
\note{
A version of the calcmig.csv is packaged in the config directory of the stacomiR library.
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

