\name{chargexml}
\alias{chargexml}
\title{chargexml loads the informations stored in c:/program files/stacomi/calcmig.xml file}
\usage{
  chargexml()
}
\value{
  a list with the datawd place and the baseODBC vector
}
\description{
  chargexml loads the informations stored in c:/program
  files/stacomi/calcmig.xml file
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

