% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefAnnee.r
\docType{methods}
\name{choice,RefAnnee-method}
\alias{choice,RefAnnee-method}
\title{choice method for RefAnnee referential}
\usage{
\S4method{choice}{RefAnnee}(object, nomassign = "refAnnee",
  funoutlabel = gettext("Year selected\\n", domain = "R-stacomiR"),
  titleFrame = gettext("Year choice", domain = "R-stacomiR"), preselect = 1)
}
\arguments{
\item{object}{An object of class \link{RefAnnee-class}}

\item{nomassign}{The name to be asssigned in envir_stacomi}

\item{funoutlabel}{The label that appears in funout}

\item{titleFrame}{Title for the frame}

\item{preselect}{The number of the year selected in the gdroplist (integer)}
}
\description{
Allows the selection of year and the assignment in environment envir_stacomi
}
\examples{
 
\dontrun{
object=new("RefAnnee")
object<-charge(object)
win=gwindow(title="test refAnnee")
group=ggroup(container=win,horizontal=FALSE)
choice(object,nomassign="refAnnee",funoutlabel="essai",titleFrame="essai RefAnnee",preselect=1)
dispose(win)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

