% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefHorodate.r
\docType{methods}
\name{choice,RefHorodate-method}
\alias{choice,RefHorodate-method}
\title{choice method for RefHorodate}
\usage{
\S4method{choice}{RefHorodate}(object, label = "date",
  nomassign = "horodate",
  funoutlabel = "nous avons le choice dans la date\\n", decal = 0,
  affichecal = TRUE)
}
\value{
Selects the date in the graphical interface
}
\description{
choice method for RefHorodate
}
\author{
cedric.briand
}

