% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefListe.r
\docType{methods}
\name{choice,RefListe-method}
\alias{choice,RefListe-method}
\title{Choice method for RefListe referential objects}
\usage{
\S4method{choice}{RefListe}(object, is.enabled = TRUE)
}
\arguments{
\item{object}{An object of class \link{RefListe-class}}

\item{is.enabled}{A boolean indicating whether the frame is enabled when first displayed, default to TRUE}
}
\description{
Choice method for RefListe referential objects
}
\note{
the choice method assigns an object of class refList named refListe in the environment envir_stacomi
}
\examples{
 
\dontrun{
 object=new("RefListe")
win=gwindow()
group=ggroup(container=win,horizontal=FALSE)
object<-charge(object,vecteur=c("choice1","choice2"),label="please choose")
choice(object)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

