% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refparqual.r
\docType{methods}
\name{choice,Refparqual-method}
\alias{choice,Refparqual-method}
\title{Choix=Choice method for Refparqual referential objects}
\usage{
\S4method{choice}{Refparqual}(object,
  label = "Choix d'une caracteristique qualitative de lot",
  nomassign = "refpar", frameassign = "frame_par", is.enabled = TRUE)
}
\description{
Choix=Choice method for Refparqual referential objects
}
\note{
the choice method assigns an object of class Refparqual named refparqual in the environment envir_stacomi
this method rewrites the method from Refpar, as it integrates a request of the possible values of qualitative parameters, hence the parameters,however it was redefined in refparqual
to load the possible values of qualitative parameters
}
\examples{
 
 object=new("Refparqual")
win=gwindow()
group=ggroup(container=win,horizontal=FALSE)
object<-charge(object)
choice(object)
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

