% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefStationMesure.r
\docType{methods}
\name{choice,RefStationMesure-method}
\alias{choice,RefStationMesure-method}
\title{Choice method for RefStationMesure referential object}
\usage{
\S4method{choice}{RefStationMesure}(object, is.enabled = TRUE,
  title = gettext("Monitoring stations selection", domain = "R-stacomiR"))
}
\arguments{
\item{object}{An object of class \link{RefStationMesure-class}}

\item{is.enabled}{A boolean parameter, if TRUE the frame is enabled when first displayed}

\item{title}{The title of the frame, defaut to "Monitoring stations selection"}
}
\description{
Choice method for RefStationMesure referential object
}
\examples{
 
\dontrun{
object=new("RefStationMesure")
win=gwindow()
group=ggroup(container=win,horizontal=FALSE)
object<-charge(object)
choice(object)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

