% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReftextBox.r
\docType{methods}
\name{choice,RefTextBox-method}
\alias{choice,RefTextBox-method}
\title{Choice method for ReTextBox referential objects}
\usage{
\S4method{choice}{RefTextBox}(object, nomassign = "refTextBox")
}
\arguments{
\item{object}{An object of class \link{RefTextBox-class}}

\item{nomassign}{The name with which the object will be assigned in envir_stacomi}
}
\description{
Choice method for ReTextBox referential objects
}
\examples{
\dontrun{
object=new("RefTextBox")
object<- charge(object,title="le titre",label="20")
win=gwindow(title="test RefTextBox")
group=ggroup(container=win,horizontal=FALSE)
choice(object) 
dispose(win)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

