% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generic.R
\name{choice}
\alias{choice}
\title{Generic method for choice (using the gwidget graphical interface)}
\usage{
choice(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Additional parms}
}
\description{
Generic method for choice (using the gwidget graphical interface)
}
\author{
cedric.briand
}
\keyword{internal}
