% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanAgedemer.r
\docType{methods}
\name{choice_c,BilanAgedemer-method}
\alias{choice_c,BilanAgedemer-method}
\title{command line interface for BilanAgedemer class}
\usage{
\S4method{choice_c}{BilanAgedemer}(object, dc, taxons = 2220,
  stades = c("5", "11", "BEC", "BER", "IND"), par = c("1786", "1785",
  "C001", "A124"), horodatedebut, horodatefin, limit1hm, limit2hm,
  silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{BilanAgedemer-class}}

\item{dc}{A numeric or integer, the code of the dc, coerced to integer,see \link{choice_c,RefDC-method}}

\item{taxons}{'2220=Salmo salar',
these should match the ref.tr_taxon_tax referential table in the stacomi database, see \link{choice_c,RefTaxon-method}}

\item{stades}{'5','11','BEC','BER','IND'}

\item{par}{Parameters chosen for the Bilan are mesured body size (1786), mesured fork length (1785),video size (C001) and number of year at sea (A124)}

\item{horodatedebut}{The starting date as a character, formats like \code{\%Y-\%m-\%d} or \code{\%d-\%m-\%Y} can be used as input}

\item{horodatefin}{The finishing date of the Bilan, for this class this will be used to calculate the number of daily steps.}

\item{silent}{Default FALSE, if TRUE the program should no display messages}
}
\value{
An object of class \link{BilanAgedemer-class}
The choice_c method fills in the data slot for classes \link{RefDC-class}, \link{RefTaxon-class}, \link{RefStades-class}, \link{Refpar-class} and two slots of \link{RefHorodate-class} and then 
uses the choice_c methods of these object to select the data.
}
\description{
command line interface for BilanAgedemer class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

