% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMult.r
\docType{methods}
\name{choice_c,BilanMigrationMult-method}
\alias{choice_c,BilanMigrationMult-method}
\title{command line interface for BilanMigrationMult class}
\usage{
\S4method{choice_c}{BilanMigrationMult}(object, dc, taxons, stades, datedebut,
  datefin, ...)
}
\description{
The choice_c method fills in the data slot for RefDC, RefTaxon, RefStades and then 
uses the choice_c methods of these object to "select" the data.
}
\examples{
 \dontrun{
bilanMigrationMult=choice_c(bilanMigrationMult,
 dc=c(5,6,12),
 taxons=c("Anguilla anguilla"),
 stades=c("AGG","AGJ","CIV"),datedebut="2011-01-01",datefin="2011-12-31")
bilanMigrationMult=calcule(bilanMigrationMult)
hbilanMigrationMultgraph()
hbilanMigrationMultgraph2()
hbilanMigrationMultgraph3()
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

