% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_choice.R
\docType{methods}
\name{choice_c,ref_choice-method}
\alias{choice_c,ref_choice-method}
\title{Choice_c method for refchoix referential objects}
\usage{
\S4method{choice_c}{ref_choice}(object, selectedvalue)
}
\arguments{
\item{object}{An object of class \link{ref_list-class}}

\item{selectedvalue}{the value selected in the combo}
}
\description{
Choice_c method for refchoix referential objects
}
\examples{
 
\dontrun{
object=new("ref_list")
object<-charge(object,vecteur=c("1","2"),label="please choose")
object<-choice_c(object)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

