% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_list.R
\name{choice_c,ref_list-method}
\alias{choice_c,ref_list-method}
\alias{choice_c.ref_list}
\title{Choice_c method for ref_list referential objects}
\usage{
\S4method{choice_c}{ref_list}(object, selectedvalue)
}
\arguments{
\item{object}{An object of class \link{ref_list-class}}

\item{selectedvalue}{the value selected in the combo}
}
\value{
An S4 object of class \link{ref_list-class}
}
\description{
Choice_c method for ref_list referential objects
}
\note{
the choice method assigns an object of class refList named ref_list in the environment envir_stacomi
}
\examples{
 
\dontrun{
object=new('ref_list')
object<-charge(object,vecteur=c('1','2'),label='please choose')
object<-choice_c(object)
}
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
