% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_par.R
\name{choice_c,ref_par-method}
\alias{choice_c,ref_par-method}
\alias{choice_c.ref_par}
\title{Command line interface to select a parameter}
\usage{
\S4method{choice_c}{ref_par}(object, par, silent = FALSE)
}
\arguments{
\item{object}{an object of class  \link{ref_par-class}}

\item{par}{A character vector of par}

\item{silent}{Default FALSE but not used there}
}
\value{
An object of class \link{ref_par-class}
}
\description{
Command line interface to select a parameter
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
