% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_par.R
\docType{methods}
\name{choice_c,ref_par-method}
\alias{choice_c,ref_par-method}
\title{Command line interface to select a parameter}
\usage{
\S4method{choice_c}{ref_par}(object, par, silent = FALSE)
}
\arguments{
\item{object}{an object of class  \link{ref_par-class}}

\item{par}{A character vector of par}

\item{silent}{Default FALSE but not used there}
}
\value{
An object of class \link{ref_par-class}
}
\description{
the choice_c method is intendedto have the same behaviour as choice (which creates a
widget in the graphical interface) but from the command line. 
If an objectreport is passed as a parameter, the method will do a charge_with_filter to select only the taxa present in the counting devices
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
