% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_textbox.R
\docType{methods}
\name{choice_c,ref_textbox-method}
\alias{choice_c,ref_textbox-method}
\title{Choice_c method for ReTextBox referential objects}
\usage{
\S4method{choice_c}{ref_textbox}(object, value, nomassign = "ref_textbox")
}
\arguments{
\item{object}{An object of class \link{ref_textbox-class}}

\item{value}{The value to set}

\item{nomassign}{The name with which the object will be assigned in envir_stacomi}
}
\description{
Choice_c method for ReTextBox referential objects
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

