% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefAnnee.r
\docType{methods}
\name{choice_c,RefAnnee-method}
\alias{choice_c,RefAnnee-method}
\title{choice_c method for RefAnnee referential from the command line}
\usage{
\S4method{choice_c}{RefAnnee}(object, annee, nomassign = "refAnnee",
  funoutlabel = gettext("Year selected\\n", domain = "R-stacomiR"),
  silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{RefAnnee-class}}

\item{annee}{The year to select, either as a character or as a numeric}

\item{nomassign}{The name to be asssigned in envir_stacomi}

\item{funoutlabel}{The label that appears in funout}

\item{silent}{Stops messages from being displayed if silent=TRUE, default FALSE}
}
\description{
The choice_c method will issue a warning if the year is not present in the database
Allows the selection of year and the assignment in environment envir_stacomi
}
\examples{
 
\dontrun{
object=new("RefAnnee")
object<-charge(object)
win=gwindow(title="test refAnnee")
group=ggroup(container=win,horizontal=FALSE)
choice(object,nomassign="refAnnee",funoutlabel="essai",titleFrame="essai RefAnnee",preselect=1)
dispose(win)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
