% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generic.R
\name{choice_c}
\alias{choice_c}
\title{Generic method for manual choice using the command line}
\usage{
choice_c(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Additional parms}
}
\description{
Generic method for manual choice using the command line
}
\author{
cedric.briand
}
\keyword{internal}
