% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasdeTemps.r
\docType{methods}
\name{choicemult,PasDeTemps-method}
\alias{choicemult,PasDeTemps-method}
\title{Graphical interface for multiple choice method for PasdeTemps (used in BilanMigrationMult)}
\usage{
\S4method{choicemult}{PasDeTemps}(object)
}
\arguments{
\item{object}{An object of class \link{PasDeTemps-class}}
}
\description{
Graphical interface for multiple choice method for PasdeTemps (used in BilanMigrationMult)
}
\note{
this method differs from choice as it is called within a notebook,
it does not allow for multiple choice to be made
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

