% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generic.R
\name{choicemult}
\alias{choicemult}
\title{Generic method for multiple choice (using the gwidget notebook for multiple choice)}
\usage{
choicemult(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Additional parms}
}
\description{
Generic method for multiple choice (using the gwidget notebook for multiple choice)
}
\author{
cedric.briand
}
\keyword{internal}

