% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_carlot.r
\docType{methods}
\name{connect,Bilan_carlot-method}
\alias{connect,Bilan_carlot-method}
\title{connect method for Bilan_carlot}
\usage{
\S4method{connect}{Bilan_carlot}(object, h, ...)
}
\arguments{
\item{object}{An object of class \link{Bilan_carlot-class}}

\item{h}{a handler}

\item{...}{additional parameters passed to the method}
}
\value{
An object of class \link{Bilan_carlot-class}
}
\description{
connect method for Bilan_carlot
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

