% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanConditionEnv.r
\docType{methods}
\name{connect,BilanConditionEnv-method}
\alias{connect,BilanConditionEnv-method}
\title{connect method for BilanConditionEnv class}
\usage{
\S4method{connect}{BilanConditionEnv}(object, h)
}
\arguments{
\item{object}{An objet of class \link{BilanConditionEnv-class}}

\item{h}{A handler}
}
\value{
an object of BilanConditionEnv class
}
\description{
connect method for BilanConditionEnv class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

