% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanEspeces.r
\docType{methods}
\name{connect,BilanEspeces-method}
\alias{connect,BilanEspeces-method}
\title{connect method for BilanEspeces}
\usage{
\S4method{connect}{BilanEspeces}(object)
}
\arguments{
\item{object}{An object of class \link{BilanEspeces-class}}
}
\value{
bilanEspeces instance with request corresponding to the user choices
}
\description{
connect method for BilanEspeces
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
