% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanFonctionnementDF.r
\docType{methods}
\name{connect,BilanFonctionnementDF-method}
\alias{connect,BilanFonctionnementDF-method}
\title{connect method for BilanFonctionnementDF}
\usage{
\S4method{connect}{BilanFonctionnementDF}(object)
}
\arguments{
\item{object}{An object of class \link{BilanFonctionnementDF-class}
loads the working periods and type of arrest or disfunction of the DF}
}
\value{
An object of class \code{BilanFonctionnementDF}
}
\description{
connect method for BilanFonctionnementDF
}
\author{
cedric.briand
}

