% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationInterAnnuelle.r
\docType{methods}
\name{connect,BilanMigrationInterAnnuelle-method}
\alias{connect,BilanMigrationInterAnnuelle-method}
\title{connect method for BilanMigrationInterannuelle class}
\usage{
\S4method{connect}{BilanMigrationInterAnnuelle}(object)
}
\arguments{
\item{object}{An object of class \link{BilanMigrationInterAnnuelle-class}}
}
\value{
bilanMigrationInterannuelle an instantianted object with values filled with user choice
}
\description{
connect method for BilanMigrationInterannuelle class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

