% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMult.r
\docType{methods}
\name{connect,BilanMigrationMult-method}
\alias{connect,BilanMigrationMult-method}
\title{connect method for BilanMigrationMult}
\usage{
\S4method{connect}{BilanMigrationMult}(object, ...)
}
\value{
BilanMigrationMult with slot @data filled from the database
}
\description{
a single query collects data from the database
}

