% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_annual.R
\docType{methods}
\name{connect,report_annual-method}
\alias{connect,report_annual-method}
\alias{connect.report_annual}
\title{connect method for report_annual class
this method performs the sum over the year attention this function does
not count subsamples.}
\usage{
\S4method{connect}{report_annual}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_annual-class}}

\item{silent}{Stops messages from being displayed if silent=TRUE, default FALSE}
}
\value{
An instantiated object with values filled with user choice

A dataframe with column effectif, comprising the sum of report_mig counts
}
\description{
connect method for report_annual class
this method performs the sum over the year attention this function does
not count subsamples.
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

