% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig.R
\name{connect,report_mig-method}
\alias{connect,report_mig-method}
\alias{connect.report_mig}
\title{handler for calculations report_mig}
\usage{
\S4method{connect}{report_mig}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{report_mig-class}}

\item{silent}{Boolean default FALSE, if TRUE information messages not displayed}

\item{h}{handler}

\item{...}{additional parameters}
}
\value{
report_mig with slot \code{@data} filled from the database
}
\description{
internal use
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{connect}
\keyword{for}
\keyword{internal}
\keyword{method}
\keyword{report_mig}
\keyword{report_mig_mult}
\keyword{the}
\keyword{uses}
