% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generic.R, R/data.R, R/stacomi.R
\docType{data}
\name{envir_stacomi}
\alias{envir_stacomi}
\title{Environment where most objects from the package are stored and then loaded
by the charge method}
\format{
An object of class \code{environment} of length 0.

An object of class \code{environment} of length 0.

An object of class \code{environment} of length 0.
}
\usage{
envir_stacomi

envir_stacomi

envir_stacomi
}
\description{
envir_stacomi \code{envir_stacomi <- new.env(parent = baseenv())} is the
environment where most object created by the interface are stored

This is where the graphical interface stores its objects
try \code{ls(envir=envir_stacomi)}

This is where the graphical interface stores its objects
try \code{ls(envir=envir_stacomi)}
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
\keyword{datasets}
\keyword{environment}
