\name{envir_stacomi}
\alias{envir_stacomi}
\title{Environment where most objects from the package are stored and then loaded by the charge method
}
\description{
    envir_stacomi \code{envir_stacomi <- new.env(parent = baseenv())} is the environment where most object created by the interface are stored
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

