% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanArgentee.r
\name{f_stade_Durif}
\alias{f_stade_Durif}
\title{Function to calculate the stages from Durif}
\usage{
f_stade_Durif(data)
}
\arguments{
\item{data}{A dataset with columns BL, W, Dv, Dh, FL corresponding to body length (mm),
Weight (g), vertical eye diameter (mm), vertical eye diameter (mm), and pectoral fin length (mm)}
}
\description{
Function to calculate the stages from Durif
}
\author{
Laurent Beaulaton \email{laurent.beaulaton"at"onema.fr}
}
