\name{fn_EcritBilanMensuel}
\alias{fn_EcritBilanMensuel}
\title{This writes monthly data in t_bilanmensuel_mens table}
\usage{
  fn_EcritBilanMensuel(bilanMigration, resum)
}
\arguments{
  \item{bilanMigration}{an object of class \code{\linkS4class{BilanMigration}}}
  \item{resum}{data frame with summary per month}  
}
\description{
  Launched by fun_EcritBilanJournalier
}

