% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_stades_pigm.r
\name{fnstade}
\alias{fnstade}
\title{function drawing polygon from gamma law describing pigmentation change in
glass eel}
\usage{
fnstade(par1, par2 = NULL, phicum, phidates, VB = FALSE, neg = TRUE,
  lmax = 1)
}
\arguments{
\item{par1}{Parameter describing the gamma law for the first stage}

\item{par2}{Parameter of the gamma law for the second stage}

\item{phicum}{cumulated pigmentation times for test : phicum=seq(0,20,
length.out=100)}

\item{phidates}{Dates}

\item{VB}{If TRUE, then calculation for first stage VB which differs from
the others}

\item{neg}{If FALSE then calculation of stages according to the pigmentation
time}

\item{lmax}{Scale parameter for the graphical function, lmax=0 allows to
draw the real values of abundances per stage along time, lmax=1 or 0.8 will
draw all stages at the same scale}
}
\value{
A list with x and y
}
\description{
function calculating from the gamma law the coordinates x and y allowing to
draw a polygon, the function fnstade may be used to draw a polygon(neg=TRUE)
or simply return the values from gamma function of each stage
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\references{
BRIAND C., FATIN D., CICCOTTI E. and LAMBERT P., 2005. A
stage-structured model to predict the effect of temperature and salinity on
glass eel Anguilla anguilla pigmentation development. J Fish Biol, 67,
995-1009.
\url{http://www3.interscience.wiley.com/journal/118686679/abstract}
\url{http://www.eptb-vilaine.fr/site/index.php/les-migrateurs/documents/publications-scientifiques/46-publications-migrateurs/60-dynamique-de-population-et-de-migration-des-civelles-en-estuaire-de-vilaine}
}
\seealso{
\code{\linkS4class{Bilan_stades_pigm}}
}

