\name{fun50}
\alias{fun50}
\title{fundist calcule la loi de repartition, fait la somme et calcule ou est le 50 pourcent de la  distribution...}
\usage{fun50(obj, objc)}
\description{fundist calcule la loi de repartition, fait la somme et calcule ou est le 50 pourcent de la  distribution}
\value{d50}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{obj}{}
\item{objc}{}}
