\name{fun50}
\alias{fun50}
\title{returns the value of obj where more than 50 percent of the distribution objc is reached}
\usage{
  fun50(obj, objc)
}
\arguments{
  \item{obj}{}

  \item{objc}{}
}
\value{
  d50
}
\seealso{
	\code{\linkS4class{Bilan_stades_pigm}}	
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

