% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_stades_pigm.r
\name{fun50}
\alias{fun50}
\title{used by \code{fundist} function, returns the value of obj where more than 50
percent of the distribution objc is reached}
\usage{
fun50(obj, objc)
}
\arguments{
\item{obj}{The pigment time scale}

\item{objc}{Cumulated gamma curve}
}
\value{
d50 The center of the distribution for the different stages
}
\description{
Allows to point the middle of the gamma distribution for each stage
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
\code{\linkS4class{Bilan_stades_pigm}}
}

