% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_mult.R
\name{fun_aggreg_for_plot}
\alias{fun_aggreg_for_plot}
\title{Calculates a data.frame where all components within the list calcdata are aggregated
and formatted for plot}
\usage{
fun_aggreg_for_plot(object)
}
\arguments{
\item{object}{An object of class \link{report_mig_mult-class}}
}
\value{
A data.frame
}
\description{
Calculates a data.frame where all components within the list calcdata are aggregated
and formatted for plot
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
