% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_mult.R
\name{fun_aggreg_for_plot}
\alias{fun_aggreg_for_plot}
\title{returns a table where all components within the list calcdata are aggregated
and formatted for plot}
\usage{
fun_aggreg_for_plot(object)
}
\arguments{
\item{object}{An object of class \link{report_mig_mult-class}}
}
\description{
returns a table where all components within the list calcdata are aggregated
and formatted for plot
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
