% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fun_char_spe}
\alias{fun_char_spe}
\title{function used to remove special non utf8 character which cause the gtk
interface to crash}
\usage{
fun_char_spe(text)
}
\arguments{
\item{text}{a text string which might contain no utf8 characters}
}
\value{
text
}
\description{
function used to remove special non utf8 character which cause the gtk
interface to crash
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
