% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_silver_eel.R
\name{fun_stage_durif}
\alias{fun_stage_durif}
\title{Function to calculate the stages from Durif}
\usage{
fun_stage_durif(data)
}
\arguments{
\item{data}{A dataset with columns BL, W, Dv, Dh, FL corresponding to body length (mm),
Weight (g), vertical eye diameter (mm), vertical eye diameter (mm), and pectoral fin length (mm)}
}
\value{
A data.frame with durif stages per individual
}
\description{
Function to calculate the stages from Durif
}
\author{
Laurent Beaulaton \email{laurent.beaulaton@ofb.fr}
}
