% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_mult.R
\name{fun_weight_conversion}
\alias{fun_weight_conversion}
\title{returns a table where weights and number are calculated from number and weights respectively
performs a query to collect the conversion coefficients}
\usage{
fun_weight_conversion(tableau, time.sequence, silent)
}
\arguments{
\item{tableau}{Table issued from report_mig}

\item{time.sequence}{Time sequence from report_mig}

\item{silent}{If silent=TRUE do not display messages}
}
\value{
tableau, the data frame with weight converted to numbers
}
\description{
returns a table where weights and number are calculated from number and weights respectively
performs a query to collect the conversion coefficients
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
