% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_df.R
\name{funbarchart1DF}
\alias{funbarchart1DF}
\title{Internal use barchart for report_df class from the graphical interface}
\usage{
funbarchart1DF(...)
}
\arguments{
\item{...}{additional parameters}
}
\value{
Nothing, called for its side effect of plotting data
}
\description{
Internal use barchart for report_df class from the graphical interface
}
\note{
The program cuts periods which overlap between two month
}
\keyword{internal}
