% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_df.R
\name{funbarchart1DF}
\alias{funbarchart1DF}
\title{Handler for barchart for report_df class from the graphical interface}
\usage{
funbarchart1DF(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
Handler for barchart for report_df class from the graphical interface
}
\note{
The program cuts periods which overlap between two month
}
\keyword{internal}

