% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_df.R
\name{funbarchartDF}
\alias{funbarchartDF}
\title{Internal use, function used in the graphical interface to create a barchart for report_df class}
\usage{
funbarchartDF(...)
}
\arguments{
\item{...}{additional parameters}
}
\value{
Nothing, called for its side effect of plotting data
}
\description{
Internal use, function used in the graphical interface to create a barchart for report_df class
}
\note{
The program cuts periods which overlap between two month
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
\keyword{internal}
