% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanFonctionnementDF.r
\name{funbarchartDF}
\alias{funbarchartDF}
\title{Handler for barchart for BilanFonctionnementDF class from the graphical interface}
\usage{
funbarchartDF(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
Handler for barchart for BilanFonctionnementDF class from the graphical interface
}
\note{
The program cuts periods which overlap between two month
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
