\name{funBilanMigrationAnnuel}
\alias{funBilanMigrationAnnuel}
\title{this functions performs the sum over the year
attention this function does not count subsamples.}
\usage{
  funBilanMigrationAnnuel(bilanMigration)
}
\arguments{
  \item{bilanMigration}{an objet of class \code{\linkS4class{BilanMigration}}}
}
\description{
  this functions performs the sum over the year attention
  this function does not count subsamples.
}

