% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funBilanMigrationAnnuel.r
\name{funBilanMigrationAnnuel}
\alias{funBilanMigrationAnnuel}
\title{this functions performs the sum over the year attention this function does
not count subsamples.}
\usage{
funBilanMigrationAnnuel(bilanMigration)
}
\arguments{
\item{bilanMigration}{an object of class \code{\linkS4class{BilanMigration}}}
}
\value{
A dataframe with column effectif, comprising the sum of bilanMigration counts
}
\description{
this functions performs the sum over the year attention this function does
not count subsamples.
}

