% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_dc.R
\name{funboxDC}
\alias{funboxDC}
\title{function used for some lattice graph}
\usage{
funboxDC(h, ...)
}
\arguments{
\item{h}{a handler}

\item{...}{Additional parameters}
}
\description{
function used for some lattice graph
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

