% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_df.R
\name{funboxDF}
\alias{funboxDF}
\title{Internal use, rectangles to describe the DF work for report_df class, 
graphical interface handler}
\usage{
funboxDF(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
Internal use, rectangles to describe the DF work for report_df class, 
graphical interface handler
}
\keyword{internal}

