% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanFonctionnementDF.r
\name{funboxDF}
\alias{funboxDF}
\title{Internal use, rectangles to describe the DF work for BilanFonctionnementDF class, 
graphical interface handler}
\usage{
funboxDF(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
Internal use, rectangles to describe the DF work for BilanFonctionnementDF class, 
graphical interface handler
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
