% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_sample_char.R
\name{funboxplotreport_sample_char}
\alias{funboxplotreport_sample_char}
\title{Boxplots for ggplot2}
\usage{
funboxplotreport_sample_char(h, ...)
}
\arguments{
\item{h}{A handler passed by the graphical interface}

\item{...}{Additional parameters}
}
\description{
assigns an object g in envir_stacomi for eventual modification of the plot
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

