\name{funcalcbilan_stades_pigm}
\alias{funcalcbilan_stades_pigm}
\title{Function with handler which calls charge (and thus connect) and calculates the title...}
\usage{funcalcbilan_stades_pigm(h, ...)}
\description{Function with handler which calls charge (and thus connect) and calculates the title}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{h}{}}
