\name{funcalcbilan_stades_pigm}
\alias{funcalcbilan_stades_pigm}
\title{main launching function for class Bilan_stades_pigm}
\usage{
  funcalcbilan_stades_pigm(h, ...)
}
\arguments{
  \item{h}{a handler}
  \item{...}{additional arguments}
}
\description{
  Function with handler which calls charge (and thus
  connect) and calculates the title
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

