% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_df.R
\name{funchartDF}
\alias{funchartDF}
\title{Function to plot calendar like graph, internal use}
\usage{
funchartDF(...)
}
\arguments{
\item{...}{additional parameters}
}
\value{
Nothing, called for its side effect of plotting data
}
\description{
Function to plot calendar like graph, internal use
}
\keyword{internal}
