% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_df.R
\name{funchartDF}
\alias{funchartDF}
\title{Handler function to plot calendar like graph, internal use}
\usage{
funchartDF(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
Handler function to plot calendar like graph, internal use
}
\keyword{internal}
