% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationInterAnnuelle.r
\name{fundat}
\alias{fundat}
\title{statistics per time period}
\usage{
fundat(dat, timesplit = NULL)
}
\arguments{
\item{dat}{a data frame}

\item{timesplit}{"week" "2 week" "month" as provided to seq.POSIXT, default NULL}
}
\value{
a data frame with mean, max, and min calculated for each timesplit
}
\description{
function called for bilamMigrationInterannelle objects renames columns
replaces nulls, and calculates reports with time period larger than day
}
\seealso{
\code{\linkS4class{Bilan_poids_moyen}}
}

