% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_carlot.r
\name{fundensityBilan_carlot}
\alias{fundensityBilan_carlot}
\title{fundensityBilan_carlot uses ggplot2 to draw density plots}
\usage{
fundensityBilan_carlot(h, ...)
}
\arguments{
\item{h}{}

\item{...}{}
}
\description{
assigns an object g in envir_stacomi for eventual modification of the plot
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

