\name{fundist}
\alias{fundist}
\title{fundist =fonction pour le calcul de la mediane de la distribution des donction pigmentaires...}
\usage{fundist(Vparm, phicum, graph=TRUE, lmax=1)}
\description{fundist =fonction pour le calcul de la mediane de la distribution des donction pigmentaires}
\note{les fonctions pigmentaires ne sont pas des fonctions statistiques standard,
le calcul du 50% de la distribution est fait à l'aide de fun50
cette fonction utilise fnstade pour calculer les valeurs des temps pigmentaires sur une echelle reguliere (phicum)}
\arguments{\item{lmax}{paramètre d'échelle de la fonction graphique, voir fnstades}
\item{graph}{logical, on peut voir les courbes en faisant graph = TRUE}}
