% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fungraph.R
\name{fungraph}
\alias{fungraph}
\title{Function for report_mig graphs including numbers DF DC operations}
\usage{
fungraph(
  report_mig,
  tableau,
  time.sequence,
  taxa,
  stage,
  dc = NULL,
  silent,
  color = NULL,
  color_ope = NULL,
  ...
)
}
\arguments{
\item{report_mig}{An object of class \code{\linkS4class{report_mig}}}

\item{tableau}{A data frame with the with the following columns : No.pas,debut_pas,fin_pas,
ope_dic_identifiant,lot_tax_code,lot_std_code,type_de_quantite,MESURE,CALCULE,
EXPERT,PONCTUEL,Effectif_total,taux_d_echappement,coe_valeur_coefficient}

\item{time.sequence}{A vector POSIXt}

\item{taxa}{The species}

\item{stage}{The stage}

\item{dc}{The DC}

\item{silent}{Message displayed or not}

\item{color}{Default NULL, a vector of color in the following order, working, stopped, 1...5 types of operation
for the fishway or DC, measured, calculated, expert, direct observation. If null will be set to brewer.pal(12,"Paired")[c(8,10,4,6,1,2,3,5,7)]}

\item{color_ope}{Default NULL, a vector of color for the operations. Default to brewer.pal(4,"Paired")}

\item{...}{additional parameters passed to matplot, main, ylab, ylim, lty, pch, bty, cex.main,
it is currenly not a good idea to change xlim (numbers are wrong, the month plot covers all month, and legend placement is wrong}
}
\value{
No return value, called for side effects
}
\description{
This graph is for species other than glass eel
}
\note{
this function is intended to be called from the plot method in report_mig_mult and report_mig
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
