% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fungraph.r
\name{fungraph}
\alias{fungraph}
\title{Function for BilanMigration graphes including numbers DF DC operations}
\usage{
fungraph(bilanMigration, tableau, time.sequence, taxon, stade,...)
}
\arguments{
\item{bilanMigration}{an object of class \code{\linkS4class{BilanMigration}}}

\item{time.sequence}{a vector POSIXt}

\item{taxon}{the species}

\item{stade}{the stage}

\item{...}{other parameters passed but not used}

\item{table=tableau}{a data frame with the results}
}
\description{
This graph is for species other than glass eel
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

