\name{fungraph}
\alias{fungraph}
\title{Function for BilanMigration graphes including numbers DF DC operations...}
\usage{fungraph(bilanMigration, tableau, duree, taxon, stade)}
\description{Function for BilanMigration graphes including numbers DF DC operations}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{bilanMigration}{}
\item{tableau}{}
\item{duree}{}
\item{taxon}{}
\item{stade}{}}
