% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fungraph.r
\name{fungraph}
\alias{fungraph}
\title{Function for BilanMigration graphes including numbers DF DC operations}
\usage{
fungraph(bilanMigration, tableau, time.sequence, taxon, stade, dc,...)
}
\arguments{
\item{bilanMigration}{An object of class \code{\linkS4class{BilanMigration}}}

\item{time.sequence}{A vector POSIXt}

\item{taxon}{The species}

\item{stade}{The stage}

\item{dc}{The DC}

\item{...}{other parameters passed but not used}

\item{table=tableau}{A data frame with the results}
}
\description{
This graph is for species other than glass eel
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

