% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fungraph_civelle.r
\name{fungraph_civelle}
\alias{fungraph_civelle}
\title{Graph function for glass eel migration. Differs from fungraph as it does not
draw the ggplot graph for month}
\usage{
fungraph_civelle(bilanMigration, table, duree, taxon, stade, dc = null)
}
\arguments{
\item{bilanMigration}{an object of class \code{\linkS4class{BilanMigration}} or an
#' object of class \code{\linkS4class{BilanMigrationMult}}}

\item{table}{a data frame with the results}

\item{duree}{a vector POSIXt}

\item{taxon}{the species}

\item{stade}{the stage}

\item{dc}{the dc, default to null, only necessary for \code{\linkS4class{BilanMigrationMult}}}
}
\description{
This graph will also plot numbers and bars according to whether the glass
eel have been counted through weight or numbers
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

