\name{fungraphBilan_poids_moyen}
\alias{fungraphBilan_poids_moyen}
\title{fungraphBilan_poids_moyen}
\description{Function for bilanPoidsMoyen also allowing to pass some objects from the peche iav database}
\usage{fungraphBilan_poids_moyen(h, ...)}
\arguments{
  \item{h}{handler}
  \item{...}{additional arguments passed to the function}
}

