% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_poids_moyen.r
\name{fungraphBilan_poids_moyen}
\alias{fungraphBilan_poids_moyen}
\title{fungraphBilan_poids_moyen}
\usage{
fungraphBilan_poids_moyen(h, ...)
}
\arguments{
\item{h}{handler}

\item{\dots}{additional arguments passed to the function}
}
\description{
Function for bilanPoidsMoyen also allowing to pass some objects from the
peche iav database
}

