% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_taille.r
\name{fungraphInteract_tail}
\alias{fungraphInteract_tail}
\title{fungraphInteract_tail uses the ggplot2usr interface to build the graphes}
\usage{
fungraphInteract_tail(h, ...)
}
\arguments{
\item{h}{a handler}

\item{...}{Additional parameters}
}
\description{
fungraphInteract_tail uses the ggplot2usr interface to build the graphes
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

