% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{funout}
\alias{funout}
\title{this function displays text and will be used to convey stacomiR message in shiny}
\usage{
funout(text, arret = FALSE, ...)
}
\arguments{
\item{text}{The text to displaying the R 
console and later in shiny}

\item{arret}{Should this cause the program to stop ?}

\item{...}{Additional parameters passed to print}
}
\value{
nblignes Assigned in envir_stacomi
}
\description{
this function displays text and will be used to convey stacomiR message in shiny
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
\keyword{internal}
