% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{funout}
\alias{funout}
\title{this function uses gfile, edits a text with info and changing colors}
\usage{
funout(text, arret = FALSE, wash = FALSE, ...)
}
\arguments{
\item{text}{The text to display both in the gtk interface and in the R
console}

\item{arret}{Should this cause the program to stop ?}

\item{wash}{Should the console be cleared after displaying the message}

\item{...}{Additional parameters passed to print}
}
\value{
nblignes Assigned in envir_stacomi
}
\description{
this function uses gfile, edits a text with info and changing colors
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}
