\name{funout}
\alias{funout}
\title{this function uses gfile, edits a text with info and changing colors}
\usage{
  funout(text, arret = FALSE, wash = FALSE)
}
\arguments{
  \item{text}{}

  \item{arret}{}

  \item{wash}{}
}
\value{
  nblignes assigned in .Global
}
\description{
  this function uses gfile, edits a text with info and
  changing colors
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

