\name{funout}
\alias{funout}
\title{this function uses gfile, edits a text with info and changing colors}
\usage{
  funout(text, arret = FALSE, wash = FALSE)
}
\arguments{
  \item{text}{the text to display both in the gtk interface and in the R console}

  \item{arret}{should this cause the program to stop ?}

  \item{wash}{Should the console be cleared after displaying the message}
}
\value{
  nblignes assigned in .Global
}
\description{
  this function uses gfile, edits a text with info and
  changing colors
}
\author{
  Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

