\name{funphi}
\alias{funphi}
\title{Function to calculate pigmentation times.}
\usage{funphi(parm, datatempsal)}
\description{Function to calculate pigmentation times.}
\value{list("dates"=duree,"phi_jour"=phi_jour)}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{parm}{parameters of the model}
\item{datatempsal}{data.frame containing temperatures and salinities}}
