\name{funphi}
\alias{funphi}
\title{Function to calculate pigmentation times.}
\usage{
  funphi(parm, datatempsal)
}
\arguments{
  \item{parm}{parameters of the model}
  \item{datatempsal}{data.frame containing temperatures and
  salinities}
}
\value{
  list("dates"=duree,"phi_jour"=phi_jour)
}
\description{
  Function to calculate pigmentation times.
}
\author{
  Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

