% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_silver_eel.R
\name{funplotreport_silver_eel}
\alias{funplotreport_silver_eel}
\title{funplotreport_silver_eel}
\usage{
funplotreport_silver_eel(h, ...)
}
\arguments{
\item{h}{A handler, with action 1,2,3 or 4}

\item{...}{Additional parameters}
}
\description{
assigns an object g in envir_stacomi for eventual modification of the plot
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

