\name{funSousListeBilanMigrationPar}
\alias{funSousListeBilanMigrationPar}
\title{funSousListeBilanMigrationPar}
\usage{funSousListeBilanMigrationPar(bilanMigrationPar)}
\description{
see \code{funSousListeBilanMigration} This function is similar to bilanMigration but allows
to count weights for a taxa and stage.
}