% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funSousListeBilanMigrationPar.r
\name{funSousListeBilanMigrationPar}
\alias{funSousListeBilanMigrationPar}
\title{funSousListeBilanMigrationPar}
\usage{
funSousListeBilanMigrationPar(bilanMigrationPar)
}
\arguments{
\item{bilanMigrationPar}{an object of class
\code{\linkS4class{BilanMigrationPar}}}
}
\description{
see \code{funSousListeBilanMigration} This function is similar to
bilanMigration but allows to count weights for a taxa and stage.
}

