% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funstat.r
\name{funstat}
\alias{funstat}
\title{Function to calculate statistics per month}
\usage{
funstat(tableau, time.sequence, taxon, stade, DC)
}
\arguments{
\item{tableau}{}

\item{time.sequence}{}

\item{taxon}{}

\item{stade}{}

\item{DC}{}
}
\value{
resum
}
\description{
Function to calculate statistics per month
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

