% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funstat.r
\name{funstat}
\alias{funstat}
\title{Function to calculate statistics per month}
\usage{
funstat(tableau, time.sequence, taxon, stade, DC, silent)
}
\arguments{
\item{tableau}{A table with the following columns : No.pas,debut_pas,fin_pas,
ope_dic_identifiant,lot_tax_code,lot_std_code,type_de_quantite,MESURE,CALCULE,               
EXPERT,PONCTUEL,Effectif_total,taux_d_echappement,coe_valeur_coefficient}

\item{time.sequence}{Passed from BilanMigration or BilanMigrationMult}

\item{taxon}{Taxa}

\item{stade}{The Stage}

\item{DC}{The counting device}

\item{silent}{Message displayed or not}
}
\description{
Function to calculate statistics per month
}
\note{
this function is intended to be called from within the summary method
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

