\name{funstat}
\alias{funstat}
\title{Function to calculate statistics per month}
\usage{
  funstat(tableau, duree, taxon, stade, DC)
}
\arguments{
  \item{tableau}{}

  \item{duree}{}

  \item{taxon}{}

  \item{stade}{}

  \item{DC}{}
}
\value{
  resum
}
\description{
  Function to calculate statistics per month
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

