% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funstatJournalier.r
\name{funstatJournalier}
\alias{funstatJournalier}
\title{function to create daily statistics}
\usage{
funstatJournalier(tableau, duree, taxon, stade, DC)
}
\arguments{
\item{tableau}{}

\item{duree}{}

\item{taxon}{}

\item{stade}{}

\item{DC}{}
}
\description{
function to create daily statistics
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

