\name{funstatJournalier}
\alias{funstatJournalier}
\title{function to create daily statistics}
\usage{
  funstatJournalier(tableau, duree, taxon, stade, DC)
}
\arguments{
  \item{tableau}{}

  \item{duree}{}

  \item{taxon}{}

  \item{stade}{}

  \item{DC}{}
}
\description{
  function to create daily statistics
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

