% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funstatJournalier.r
\name{funstatJournalier}
\alias{funstatJournalier}
\title{function to create daily statistics}
\usage{
funstatJournalier(tableau, time.sequence, taxon, stade, DC)
}
\arguments{
\item{tableau}{A table}

\item{time.sequence}{Time sequence from BilanMigration and BilanMigrationMult}

\item{taxon}{A taxa}

\item{stade}{A stage}

\item{DC}{A counting device}
}
\description{
function to create daily statistics
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
