% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funtable.R
\name{funtable}
\alias{funtable}
\title{function to print and save statistics in .csv and .html formats for report_mig and report_mig_mult class}
\usage{
funtable(tableau, time.sequence, taxa, stage, DC, resum, silent)
}
\arguments{
\item{tableau}{A table with the following columns : No.pas,debut_pas,fin_pas,
ope_dic_identifiant,lot_tax_code,lot_std_code,type_de_quantite,MESURE,CALCULE,               
EXPERT,PONCTUEL,Effectif_total,taux_d_echappement,coe_valeur_coefficient}

\item{time.sequence}{Passed from report_mig or report_mig_mult}

\item{taxa}{Taxa}

\item{stage}{The Stage}

\item{DC}{The counting device}

\item{resum}{A summary table generated by funstat}

\item{silent}{If TRUE, all messages turned off (except warnings and errors)}
}
\value{
No return value, called for side effects
}
\description{
function to print and save statistics in .csv and .html formats for report_mig and report_mig_mult class
}
\note{
this function is intended to be called from within the summary method
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
