\name{funtable}
\alias{funtable}
\title{function to print and save statistics in .csv and .html formats}
\usage{
  funtable(tableau, duree, taxon, stade, DC, resum)
}
\arguments{
  \item{tableau}{}

  \item{duree}{}

  \item{taxon}{}

  \item{stade}{}

  \item{DC}{}

  \item{resum}{}
}
\description{
  function to print and save statistics in .csv and .html
  formats
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

