% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funtable.r
\name{funtable}
\alias{funtable}
\title{function to print and save statistics in .csv and .html formats for BilanMigration and BilanMigrationMult class}
\usage{
funtable(tableau, time.sequence, taxon, stade, DC, resum)
}
\arguments{
\item{tableau}{A table with the following columns : No.pas,debut_pas,fin_pas,
ope_dic_identifiant,lot_tax_code,lot_std_code,type_de_quantite,MESURE,CALCULE,               
EXPERT,PONCTUEL,Effectif_total,taux_d_echappement,coe_valeur_coefficient}

\item{time.sequence}{Passed from BilanMigration or BilanMigrationMult}

\item{taxon}{Taxa}

\item{stade}{The Stage}

\item{DC}{The counting device}

\item{resum}{A summary table generated by funstat}
}
\description{
function to print and save statistics in .csv and .html formats for BilanMigration and BilanMigrationMult class
}
\note{
this function is intended to be called from within the summary method
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

