\name{funtableBilan_tail}
\alias{funtableBilan_tail}
\title{function used to display a table of the data}
\usage{
  funtableBilan_tail(h, ...)
}
\arguments{
  \item{h}{}

  \item{...}{}
}
\description{
  function used to display a table of the data
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

