% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_dc.R
\name{funtableDC}
\alias{funtableDC}
\title{funtableDC create a table output for report_dc class}
\usage{
funtableDC(...)
}
\arguments{
\item{...}{Additional parameters}
}
\value{
Nothing, calls the summary method on data assigned in envir_stacomi
}
\description{
funtableDC create a table output for report_dc class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}
