% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_df.R
\name{funtableDF}
\alias{funtableDF}
\title{Table output for report_df class}
\usage{
funtableDF(...)
}
\arguments{
\item{...}{additional parameters}
}
\value{
Nothing, called for its side effect of running summary
}
\description{
Table output for report_df class
}
\keyword{internal}
