\name{funtableDF}
\alias{funtableDF}
\title{FuntableDF create a table output for BilanFonctionnementDF class}
\usage{
  funtableDF(h, ...)
}
\arguments{
  \item{h}{a handler}
}
\description{
  FuntableDF create a table output for
  BilanFonctionnementDF class
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

