\name{funtitle_bilan_stades_pigm}
\alias{funtitle_bilan_stades_pigm}
\title{Fonction handler qui retourne le titre du graphique apres le choix dans la date}
\usage{
  funtitle_bilan_stades_pigm(h, ...)
}
\arguments{
  \item{h}{A handler}
  \item{...}{Other arguments passed to the function}
}
\description{
  Fonction handler qui retourne le titre du graphique apres
  le choix dans la date
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

