% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_stades_pigm.r
\name{funtitle_bilan_stades_pigm}
\alias{funtitle_bilan_stades_pigm}
\title{Handler function which retrun the graphics titles
date}
\usage{
funtitle_bilan_stades_pigm(h, ...)
}
\arguments{
\item{h}{A handler}

\item{...}{Other arguments passed to the function}
}
\description{
Handler function which retrun the graphics titles
date
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

