% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funtraitement_poids.r
\name{funtraitement_poids}
\alias{funtraitement_poids}
\title{returns a table where weights and number are calculated from number and weights respectively
performs a query to collect the conversion coefficients}
\usage{
funtraitement_poids(tableau, duree)
}
\arguments{
\item{tableau}{}

\item{duree}{}
}
\value{
tableau
}
\description{
returns a table where weights and number are calculated from number and weights respectively
performs a query to collect the conversion coefficients
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

