% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getdbname}
\alias{getdbname}
\title{Retrieves the dbname from a connection using "baseODBC"}
\usage{
getdbname()
}
\value{
A string with the name of the database
}
\description{
When running a connection using ODBC, the connection string does not contain
the name of the database. Pointing to the database is done while setting the ODBC
connection, but the program has no "way" to know what the name of the database is.
Since we are using sqldf to connect to the base, we need to know the database name, fortunately
it is stored in the ODBC connection string. This method gets this name, from
a "trial" connection to one of the table of the database.
}
