\name{getvalue-methods}
\docType{methods}
\alias{getvalue}
\alias{getvalue-methods}
\alias{getvalue,Refperiode-method}
\title{ Methods for Function getvalue }
\description{
 Methods for class Refperiod,  returns the POSIXt value of a given name
 \code{data=data.frame("id"=c("jour","semaine","quinzaine","mois"),
          "pgval"=c("day","week","2 week","month"))}
}
\section{Methods}{
\describe{
\item{\code{signature(object = "Refperiode")}}{ Returns the POSIXt value of a given name
}}
}
\keyword{methods}

