% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{graphdate}
\alias{graphdate}
\title{function used for some lattice graphs with dates}
\usage{
graphdate(vectordate)
}
\arguments{
\item{vectordate}{date or POSIXt}
}
\value{
vectordate (without class)
}
\description{
function used for some lattice graphs with dates
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
