\name{graphdate}
\alias{graphdate}
\title{function used for some lattice graphes with dates}
\usage{
  graphdate(vectordate)
}
\arguments{
  \item{vectordate}{date or POSIXt}
}
\value{
  vectordate (without class)
}
\description{
  function used for some lattice graphes with dates
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

