% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitaires.r
\name{graphdate}
\alias{graphdate}
\title{function used for some lattice graphes with dates}
\usage{
graphdate(vectordate)
}
\arguments{
\item{vectordate}{date or POSIXt}
}
\value{
vectordate (without class)
}
\description{
function used for some lattice graphes with dates
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

