\name{graphe-methods}
\docType{methods}
\alias{graphe-methods}
\alias{graphe}
\alias{graphe,report_migPar-method}
\title{ Methods for Function graphe }
\description{
method with a parameter choice to join several graphs called by several handlers (buttons)
}
\section{Methods}{
\describe{
\item{\code{signature(object = "report_migPar")}}{
method with a parameter choice to join several graphs called by several handlers (buttons) }
}}
\keyword{methods}