% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacomi.r
\name{hBilanConditionEnv}
\alias{hBilanConditionEnv}
\title{handler function used by the main interface}
\usage{
hBilanConditionEnv(h, ...)
}
\description{
handler function used by the main interface
}

