% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanEspeces.r
\name{hBilanEspecescalc}
\alias{hBilanEspecescalc}
\title{handler du calcul du BilanEspeces
realise le calcul du bilan especes, l'ecrit dans l'environnement envir_stacomi
traite eventuellement les quantites de lots (si c'est des civelles)}
\usage{
hBilanEspecescalc(h, ...)
}
\arguments{
\item{h}{}

\item{...}{}
}
\description{
handler du calcul du BilanEspeces
realise le calcul du bilan especes, l'ecrit dans l'environnement envir_stacomi
traite eventuellement les quantites de lots (si c'est des civelles)
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

