% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigration.r
\name{hbilanMigrationcalc}
\alias{hbilanMigrationcalc}
\title{handler for calculations BilanMigration}
\usage{
hbilanMigrationcalc(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
internal use
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

