% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigration.r
\name{hbilanMigrationcalc}
\alias{hbilanMigrationcalc}
\title{handler du graphique BilanMigration
realise le calcul du bilan migration, l'ecrit dans l'environnement envir_stacomi
traite eventuellement les quantites de lots (si c'est des civelles)}
\usage{
hbilanMigrationcalc(h, ...)
}
\arguments{
\item{h}{}

\item{...}{}
}
\description{
handler du graphique BilanMigration
realise le calcul du bilan migration, l'ecrit dans l'environnement envir_stacomi
traite eventuellement les quantites de lots (si c'est des civelles)
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

