% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationConditionEnv.r
\name{hbilanMigrationConditionEnvcalc}
\alias{hbilanMigrationConditionEnvcalc}
\title{handler du graphique BilanMigrationConditionEnv
realise le calcul du bilan migration avec CE, l'ecrit dans l'environnement envir_stacomi
traite eventuellement les quantites de lots (si c'est des civelles)}
\usage{
hbilanMigrationConditionEnvcalc(h, ...)
}
\arguments{
\item{h}{}

\item{...}{}
}
\description{
handler du graphique BilanMigrationConditionEnv
realise le calcul du bilan migration avec CE, l'ecrit dans l'environnement envir_stacomi
traite eventuellement les quantites de lots (si c'est des civelles)
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

