% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigration.r
\name{hbilanMigrationgraph}
\alias{hbilanMigrationgraph}
\title{handler for  hBilanMigrationgraph}
\usage{
hbilanMigrationgraph(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
Standard BilanMigration graph over time
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

