% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigration.r
\name{hbilanMigrationgraph}
\alias{hbilanMigrationgraph}
\title{handler hBilanMigrationgraph
calls the fungraph for BilanMigration and allows the saving of daily and monthly counts in the database}
\usage{
hbilanMigrationgraph(h, ...)
}
\arguments{
\item{h}{a handler}

\item{...}{Additional parameters}
}
\description{
handler hBilanMigrationgraph
calls the fungraph for BilanMigration and allows the saving of daily and monthly counts in the database
}
\note{
pb if other than daily value, the time steps have been constrained to daily values for this plot
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

