\name{hbilanMigrationgraph}
\alias{hbilanMigrationgraph}
\title{handler du calcul hBilanMigrationgraph...}
\usage{hbilanMigrationgraph(h, ...)}
\description{handler du calcul hBilanMigrationgraph
appelle les fonctions fungraph pour faire le bilan des migrations
et permet la sauvegarde des bilans journaliers dans la base}
\note{pb si autre chose que journalier les pas de temps ont été contraints à des pas de temps journaliers pour ce graphique}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{h}{}
\item{...}{}}
