% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigration.r
\name{hbilanMigrationgraph2}
\alias{hbilanMigrationgraph2}
\title{handler du calcul hBilanMigrationgraph2
appelle les fonctions fungraph pour faire un graphe annuel des 
cumuls de migration au cours du temps}
\usage{
hbilanMigrationgraph2(h, ...)
}
\arguments{
\item{h}{}

\item{...}{}
}
\description{
handler du calcul hBilanMigrationgraph2
appelle les fonctions fungraph pour faire un graphe annuel des 
cumuls de migration au cours du temps
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

