% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigration.r
\name{hbilanMigrationgraph2}
\alias{hbilanMigrationgraph2}
\title{handler for hBilanMigrationgraph2}
\usage{
hbilanMigrationgraph2(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{additional parameters}
}
\description{
Step plot over time
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

