% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacomi.r
\name{hBilanMigrationPar}
\alias{hBilanMigrationPar}
\title{handler function used by the main interface}
\usage{
hBilanMigrationPar(h, ...)
}
\description{
handler function used by the main interface
}

