% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationPar.r
\name{hbilanMigrationPargraph}
\alias{hbilanMigrationPargraph}
\title{le handler appelle la methode generique graphe sur l'object plot.type=1}
\usage{
hbilanMigrationPargraph(h, ...)
}
\arguments{
\item{h, }{passed by the handler}
}
\description{
le handler appelle la methode generique graphe sur l'object plot.type=1
}

