\name{hbilanMigrationPargraph}
\alias{hbilanMigrationPargraph}
\title{le handler appelle la methode generique graphe sur l'objet choix=1}
\usage{
  hbilanMigrationPargraph(h, ...)
}
\arguments{
  \item{h,}{passé par le handler}
}
\description{
  le handler appelle la methode generique graphe sur
  l'objet choix=1
}

