\name{hbilanMigrationPargraph2}
\alias{hbilanMigrationPargraph2}
\title{le handler appelle la methode generique graphe sur l'objet choix=2}
\usage{
  hbilanMigrationPargraph2(h, ...)
}
\arguments{
  \item{h,}{passé par le handler}
}
\description{
  le handler appelle la methode generique graphe sur
  l'objet choix=2
}

