% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationPar.r
\name{hbilanMigrationParstat}
\alias{hbilanMigrationParstat}
\title{This handler calls the generic method graphe on object choix 3}
\usage{
hbilanMigrationParstat(h, ...)
}
\arguments{
\item{h, }{passed by}
}
\description{
This handler calls the generic method graphe on object choix 3
}

