% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigration.r
\name{hbilanMigrationwrite}
\alias{hbilanMigrationwrite}
\title{handler hBilanMigrationwrite
Allows the saving of daily and monthly counts in the database}
\usage{
hbilanMigrationwrite(h, ...)
}
\arguments{
\item{h}{a handler}

\item{...}{Additional parameters}
}
\description{
handler hBilanMigrationwrite
Allows the saving of daily and monthly counts in the database
}
\note{
these entries are necessary to run the Interannual Migration class. 
then no entry will be written to the database
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

