% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_char.R
\name{hbmCcalc}
\alias{hbmCcalc}
\title{handler for report_mig_char}
\usage{
hbmCcalc(h, ...)
}
\arguments{
\item{h}{handler}

\item{...}{Additional parameters}
}
\description{
internal use
}
\keyword{internal}

