\name{hCamembert}
\alias{hCamembert}
\title{handler du calcul hCamembert...}
\usage{hCamembert(h, ...)}
\description{handler du calcul hCamembert
trace un camembert des especes ou un camembert par periode...}
\note{pas besoin de refaire tourner calcul si une autre liste à été chargée, les effectifs <0 sont transformés en positifs}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{h}{}
\item{...}{}}
