% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanEspeces.r
\name{hCamembert}
\alias{hCamembert}
\title{handler for pie chart
draws a pie chart of species or a pie chart per period}
\usage{
hCamembert(h, ...)
}
\arguments{
\item{h}{Handler}

\item{...}{Other parameters passed to the function}
}
\description{
handler for pie chart
draws a pie chart of species or a pie chart per period
}
\note{
no need to re-run calculation if another list has been loaded, negative numbers are converted to positive
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

