\name{hDF}
\alias{h0,}
\alias{hBilanConditionEnv,}
\alias{hBilanLots,}
\alias{hBilanMigrationPar,}
\alias{hDC,hOPE,hDFDC,hBilanMigration,hBilanMigrationInterannuelle,hBilanMigrationConditionEnv,}
\alias{hDF}
\alias{hhelp,}
\alias{hpds,}
\alias{hSt,}
\alias{hTail,}
\alias{htodo,}
\alias{hx11}
\title{handler function used by the main interface}
\usage{
  hDF(h, ...)
}
\arguments{
  \item{h}{}

  \item{...}{}
}
\description{
  handler function used by the main interface
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

