% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacomi.r
\name{hDF}
\alias{hDF}
\title{handler function used by the main interface}
\usage{
hDF(h, ...)
}
\arguments{
\item{h}{A handler}

\item{...}{Other parameters}
}
\description{
handler function used by the main interface
}
\keyword{internal}

