% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_graphique.r
\name{hDF}
\alias{h0,}
\alias{hBilanConditionEnv,}
\alias{hBilanLots,}
\alias{hBilanMigrationPar,}
\alias{hDC,hOPE,hDFDC,hBilanMigration,hBilanMigrationInterannuelle,hBilanMigrationConditionEnv,}
\alias{hDF}
\alias{hSt,}
\alias{hTail,}
\alias{hhelp,}
\alias{hpds,}
\alias{htodo,}
\alias{hx11}
\title{handler function used by the main interface}
\usage{
hDF(h, ...)
}
\arguments{
\item{h}{A handler}

\item{...}{Other parameters}
}
\description{
handler function used by the main interface
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

