\name{hHistogramme}
\alias{hHistogramme}
\title{handler du calcul hHistogramme
trace un histogramme des espèces fréquentant la passe...}
\usage{
  hHistogramme(h, ...)
}
\arguments{
  \item{h}{}

  \item{...}{}
}
\description{
  handler du calcul hHistogramme trace un histogramme des
  espèces fréquentant la passe...
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

