% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_df.R
\name{houtDF}
\alias{houtDF}
\title{handler to print the command line}
\usage{
houtDF(h, ...)
}
\arguments{
\item{h}{a handler}

\item{...}{Additional parameters}
}
\description{
handler to print the command line
}
\keyword{internal}

