% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_ge_weight.R
\name{hreg}
\alias{hreg}
\title{Internal handler for reg, class \code{\link{report_ge_weight-class}}.}
\usage{
hreg(h, ...)
}
\arguments{
\item{h}{handler}

\item{\dots}{additional arguments passed to the function}
}
\description{
Internal handler for reg, class \code{\link{report_ge_weight-class}}.
}
\keyword{internal}
