% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_species.R
\name{hsummarybilesp}
\alias{hsummarybilesp}
\title{handler for summary report_species, internal use}
\usage{
hsummarybilesp(h)
}
\arguments{
\item{h}{a handler}
}
\description{
handler for summary report_species, internal use
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}
