\name{hTableBilanEspeces}
\alias{hTableBilanEspeces}
\title{handler du calcul BilanEspeces : traitements...}
\usage{hTableBilanEspeces(h, ...)}
\description{handler du calcul BilanEspeces : traitements 
appelle les fonctions funstat et funtable pour faire le bilan des migrations
dans des fichiers csv}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{h}{}
\item{...}{}}
