% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanEspeces.r
\name{hTableBilanEspeces}
\alias{hTableBilanEspeces}
\title{handler du calcul BilanEspeces : traitements 
appelle les fonctions funstat et funtable pour faire le bilan des migrations
dans des fichiers csv}
\usage{
hTableBilanEspeces(h, ...)
}
\arguments{
\item{h}{a handler}

\item{...}{Additional parameters}
}
\description{
handler du calcul BilanEspeces : traitements 
appelle les fonctions funstat et funtable pour faire le bilan des migrations
dans des fichiers csv
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

