\name{hTableBilanMigration}
\alias{hTableBilanMigration}
\title{handler du calcul BilanMigrationstat : traitements
appelle les fonctions funstat et funtable pour faire le bilan des migrations
dans des fichiers csv}
\usage{
  hTableBilanMigration(h, ...)
}
\arguments{
  \item{h}{}

  \item{...}{}
}
\description{
  handler du calcul BilanMigrationstat : traitements
  appelle les fonctions funstat et funtable pour faire le
  bilan des migrations dans des fichiers csv
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

