% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigration.r
\name{hTableBilanMigration}
\alias{hTableBilanMigration}
\title{handler for summary function
calls functions funstat and funtable to build summary tables in html and
csv files}
\usage{
hTableBilanMigration(h, ...)
}
\arguments{
\item{h}{Handler}

\item{...}{Additional parameters}
}
\description{
handler for summary function
calls functions funstat and funtable to build summary tables in html and
csv files
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

