% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacomi.r
\name{husr}
\alias{husr}
\title{Internal function, tests the connection and if it works loads the stacomi interface}
\usage{
husr(h, ...)
}
\arguments{
\item{h}{A handler}

\item{...}{Other arguments}
}
\description{
Internal function, tests the connection and if it works loads the stacomi interface
}
\note{
\code{gr_interface} is copied by stacomi into envir_stacomi. Same for \code{database_expected}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}

