\name{husr}
\alias{husr}
\title{this handler test the connexion and if it works loads the stacomi interface}
\usage{
  husr(h, ...)
}
\arguments{
  \item{h}{}

  \item{...}{}
}
\description{
  this handler test the connexion and if it works loads the
  stacomi interface
}
\note{
  gr_interface is copied by stacomi into envir_stacomi.
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

