% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_graphique.r
\name{husr}
\alias{husr}
\title{this handler test the connection and if it works loads the stacomi interface}
\usage{
husr(h, ...)
}
\arguments{
\item{h}{}

\item{...}{}
}
\description{
this handler test the connection and if it works loads the stacomi interface
}
\note{
gr_interface is copied by stacomi into envir_stacomi.
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

