\name{interface_BilanFonctionnementDC}
\alias{interface_BilanFonctionnementDC}
\title{interface for BilanFonctionnementDC class}
\usage{
  interface_BilanFonctionnementDC()
}
\description{
  interface for BilanFonctionnementDC class
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

