% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_BilanFonctionnementDF.r
\name{interface_BilanFonctionnementDF}
\alias{interface_BilanFonctionnementDF}
\title{interface for BilanFonctionnementDF class}
\usage{
interface_BilanFonctionnementDF()
}
\description{
interface for BilanFonctionnementDF class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

