\name{interface_BilanLot}
\alias{interface_BilanLot}
\title{An interface that calls the object to build the user interface}
\usage{
  interface_BilanLot()
}
\description{
  An interface that calls the object to build the user
  interface
}
\note{
  always has to be called within a group constructed and
  deleted using quitte()
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

