% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_BilanMigrationInterannuelle.r
\name{interface_BilanMigrationInterAnnuelle}
\alias{interface_BilanMigrationInterAnnuelle}
\title{interface for BilanMigrationInterannuelle class}
\usage{
interface_BilanMigrationInterAnnuelle()
}
\description{
interface for BilanMigrationInterannuelle class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

