% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacomi.R
\name{interface_graphique}
\alias{interface_graphique}
\title{Program launch, this function launches the GwidgetRgtk graphical
interface to stacomi. To be able to run, some widgets (win, grouptotal, group...) 
are assigned in the user environment \code{.GlobalEnv}.}
\usage{
interface_graphique()
}
\description{
Program launch, this function launches the GwidgetRgtk graphical
interface to stacomi. To be able to run, some widgets (win, grouptotal, group...) 
are assigned in the user environment \code{.GlobalEnv}.
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\keyword{internal}
