\name{interface_graphique}
\alias{interface_graphique}
\title{Program launch, this function first gathers the ODBC path from the csv file}
\usage{
  interface_graphique()
}
\description{
  Program launch, this function fist gathers the ODBC path
  and working directory from the
  csv file
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

