\name{messages}
\alias{messages}
\title{Function used at the beginning of the program to display the first messages collected by method RODBC
Attention all msgs are starting with a letter "l" numbers are wrong when used as msgs within a list}
\usage{
  messages(lang = "French")
}
\arguments{
  \item{lang}{one of "French","Spanish","English"}
}
\description{
  Function used at the beginning of the program to display
  the first messages collected by method RODBC Attention
  all msgs are starting with a letter "l" numbers are wrong
  when used as msgs within a list
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

