% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generic.R
\name{model}
\alias{model}
\title{Generic for prediction}
\usage{
model(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Additional parms}
}
\description{
Generic for prediction
}
\author{
cedric.briand
}
