% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generic.r
\name{out}
\alias{out}
\title{Generic method to forward outputs}
\usage{
out(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Additional parms}
}
\description{
Generic method to forward outputs
}
\author{
cedric.briand
}
