% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasdeTemps.r
\docType{class}
\name{PasDeTempsChar-class}
\alias{PasDeTempsChar}
\alias{PasDeTempsChar-class}
\alias{validite_PasDeTempsChar}
\title{Class "PasDeTempsChar"}
\description{
Character to represent a PasDeTemps
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("PasDeTempsChar", \dots{})}
}
\examples{

showClass("PasDeTempsChar")

}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\seealso{
\code{\linkS4class{PasDeTemps}}
}
\keyword{classes}

