% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasDeTempsJournalier.r
\docType{class}
\name{PasDeTempsJournalier-class}
\alias{PasDeTempsJournalier-class}
\title{Class "PasDeTempsJournalier"}
\description{
Representation of a PasDeTemps object with a step length equal to one day.
It receives an heritance from PasDeTemps
}
\details{
validite_PasDeTempsJournalier
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("PasDeTempsJournalier",
dateDebut="POSIXt",dureePas=numeric(),nbPas=numeric(),noPasCourant=integer())}.
\describe{ \item{list("dateDebut")}{Object of class \code{"POSIXt"} Starting
date }\item{:}{Object of class \code{"POSIXt"} Starting date }
\item{list("dureePas")}{Object of class \code{"numeric"} Step length
}\item{:}{Object of class \code{"numeric"} Step length }
\item{list("nbPas")}{Object of class \code{"numeric"} Number of steps
}\item{:}{Object of class \code{"numeric"} Number of steps }
\item{list("noPasCourant")}{Object of class \code{"integer"} Number of the
current step }\item{:}{Object of class \code{"integer"} Number of the
current step } }
}
\examples{

showClass("PasDeTempsJournalier")

}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\seealso{
\code{\linkS4class{PasDeTemps}}
}
\keyword{classes}

