% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_carlot.r
\docType{methods}
\name{plot,Bilan_carlot,missing-method}
\alias{plot,Bilan_carlot,missing-method}
\alias{plot.Bilan_carlot}
\alias{plot.b_carlot}
\alias{plot.bilan_carlot}
\title{Plots of various type for BilanMigration, and performs writing to the database of daily values.}
\usage{
\S4method{plot}{Bilan_carlot,missing}(x, plot.type = "1", silent = FALSE)
}
\arguments{
\item{x}{An object of class BilanMigrationMult}

\item{plot.type}{One of "1","violin plot". Defaut to \code{1} , can also be \code{2} boxplot or 
\code{3} points.}

\item{silent}{Stops displaying the messages.}

\item{...}{Additional arguments, see \code{plot}, \code{plot.default} and \code{par}}
}
\description{
\itemize{
		\item{plot.type="standard"}{calls \code{\link{fungraph}} and \code{\link{fungraph_civelle}} functions to plot as many "bilanmigration"
			as needed, the function will test for the existence of data for one dc, one taxa, and one stage}
		\item{plot.type="step"}{creates Cumulated graphs for BilanMigrationMult.  Data are summed per day for different dc taxa and stages}
		\item{plot.type="multiple"}{Method to overlay graphs for BilanMigrationMult (multiple dc/taxa/stage in the same plot)}
}
}
\note{
When plotting the "standard" plot, the user will be prompted to "write" the daily migration and monthly migration in the database.
these entries are necessary to run the Interannual Migration class. If the stacomi has been launched with database_expected=FALSE,
then no entry will be written to the database
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

