% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMult.r
\docType{methods}
\name{plot,BilanMigrationMult,ANY-method}
\alias{plot,BilanMigrationMult,ANY-method}
\title{Main plot method}
\usage{
\S4method{plot}{BilanMigrationMult,ANY}(x, y = null, ...)
}
\arguments{
\item{x}{the bilanMigration}

\item{...}{}

\item{y=null}{to conform with plot method}
}
\description{
calls \link{fungraph} et \link{fungraph_civelle} functions to plot as many "bilanmigration"
as needed
}
\note{
the function will test for the existence of data for one dc, one taxa, and one stage
before running
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

