% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationPar.r
\docType{methods}
\name{plot,BilanMigrationPar,ANY-method}
\alias{plot,BilanMigrationPar,ANY-method}
\title{plot method for BilanMigrationPar}
\usage{
\S4method{plot}{BilanMigrationPar,ANY}(x, y, plot.type = "barplot", ...)
}
\arguments{
\item{x}{An object of class BilanMigrationPar}

\item{plot.type}{One of "barplot", "xyplot", "summary table"}
}
\description{
plot method for BilanMigrationPar
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

