% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_env.R
\name{plot,report_env,missing-method}
\alias{plot,report_env,missing-method}
\alias{plot.report_env}
\title{h_report_envgraph Internal method}
\usage{
\S4method{plot}{report_env,missing}(x, silent = FALSE)
}
\arguments{
\item{x}{An object of class \link{report_env-class}}

\item{silent}{Stops displaying the messages}

\item{h}{a handler}

\item{...}{Additional parameters}
}
\value{
Nothing, called for its side effect of plotting data
}
\description{
h_report_envgraph Internal method
}
\author{
Cedric Briand \email{cedric.briand'at'eptb-vilaine.fr}

Cedric Briand \email{cedric.briand'at'eptb-vilaine.fr}
}
\keyword{Plot}
\keyword{for}
\keyword{internal}
\keyword{method}
\keyword{report_env}
