% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_env.R
\docType{methods}
\name{plot,report_env,missing-method}
\alias{plot,report_env,missing-method}
\alias{plot.report_env}
\title{Plot method for report_env}
\usage{
\S4method{plot}{report_env,missing}(x, silent = FALSE)
}
\arguments{
\item{x}{An object of class \link{report_env-class}}

\item{silent}{Stops displaying the messages}
}
\description{
Plot method for report_env
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

