% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_ge_weight.R
\name{plot,report_ge_weight,missing-method}
\alias{plot,report_ge_weight,missing-method}
\alias{plot.report_ge_weight}
\title{Plot method for report_ge_weight'}
\usage{
\S4method{plot}{report_ge_weight,missing}(x, plot.type = 1, silent = FALSE)
}
\arguments{
\item{x}{An object of class \link{report_ge_weight-class}}

\item{plot.type}{Default '1'. '1' plot of mean weight of glass eel against the mean date of operation (halfway between start,
and end of operation). The ggplot 'p' can be accessed from envir_stacomi using \code{get('p',envir_stacomi)}. '2' standard plot of current coefficent.
'3' same as '1' but with size according to number.}

\item{silent}{Stops displaying the messages}
}
\value{
Nothing, called for its side effect of plotting data
}
\description{
Plot method for report_ge_weight'
}
\note{
the model method provides plots for the fitted models
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
