% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_char.R
\name{plot,report_mig_char,missing-method}
\alias{plot,report_mig_char,missing-method}
\alias{plot.report_mig_char}
\title{handler for plot}
\usage{
\S4method{plot}{report_mig_char,missing}(x, color_parm = NULL, plot.type = "qual", silent = FALSE, ...)
}
\arguments{
\item{x}{An object of class report_mig_char}

\item{color_parm}{A named vector for the colors of either parameters (if plot.type=quant) or levels for
parameters (if plot.type=qual).}

\item{plot.type}{One of 'qual', 'quant' 'crossed' default to qual}

\item{silent}{Boolean default FALSE, if TRUE information messages not displayed}

\item{...}{Additional parameters}

\item{h}{handler}
}
\value{
Nothing, called for its side effect of plotting data
}
\description{
handler for plot
}
\author{
Cedric Briand \email{cedric.briand'at'eptb-vilaine.fr}
}
\keyword{This}
\keyword{calls}
\keyword{for}
\keyword{generic}
\keyword{handler}
\keyword{internal}
\keyword{method}
\keyword{plot}
\keyword{report_mig_char}
\keyword{stat}
\keyword{the}
