% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_interannual.R
\docType{methods}
\name{plot,report_mig_interannual,missing-method}
\alias{plot,report_mig_interannual,missing-method}
\alias{plot.report_mig_interannual}
\title{Plot method for report_mig_interannual}
\usage{
\S4method{plot}{report_mig_interannual,missing}(x,
  plot.type = "standard", timesplit = "mois", silent = FALSE)
}
\arguments{
\item{x}{An object of class report_mig_interannual}

\item{plot.type}{Default standard}

\item{timesplit}{Used for plot.type barchart or dotplot, Default mois (month) other possible values are semaine (week), quinzaine (2 weeks),
English values within parenthesis are also accepted.}

\item{silent}{Stops displaying the messages.
\itemize{
        \item{plot.type="line": one line per daily report_mig}
        \item{plot.type="standard": the current year is displayed against a ribbon of historical values"}
        \item{plot.type="density": creates density plot to compare seasonality, data computed by 15 days period}
        \item{plot.type="step" : creates step plots to compare seasonality, the year chosen in the interface is the
latest if silent=TRUE, or it can be selected in the droplist. It is highlighted against the other with a dotted line}
        \item{plot.type="barchart": comparison of daily migration of one year against periodic migration for the other years available in the chronicle,
                                    different periods can be chosen with argument timesplit}
        \item{plot.type="pointrange": Pointrange graphs, different periods can be chosen with argument timesplit}
     \item{plot.type="seasonal": plot to display summary statistics about the migration period}
}}
}
\description{
Several of these plots are scaled against the same year, ie the comparison is based on
year 2000, meaning that day 1 would correspond to the first date of 2000,  which is also a
saturday, the last day of the week.
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
