% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generic.r
\name{plot}
\alias{plot}
\title{Generic method for plot}
\usage{
plot(x, y, ...)
}
\arguments{
\item{...}{Additional parms}

\item{object}{Object}
}
\description{
Generic method for plot
}
\author{
cedric.briand
}

