% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationMult.r
\docType{methods}
\name{print,BilanMigrationMult-method}
\alias{print,BilanMigrationMult-method}
\title{Method to print the command line of the object}
\usage{
\S4method{print}{BilanMigrationMult}(x, ...)
}
\arguments{
\item{x}{An object of class BilanMigrationMult}

\item{...}{Additional parameters passed to print}
}
\description{
Method to print the command line of the object
}
\author{
cedric.briand
}

