% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_dc.R
\name{print,report_dc-method}
\alias{print,report_dc-method}
\alias{print.report_dc}
\title{Method to print the command line of the object.}
\usage{
\S4method{print}{report_dc}(x, ...)
}
\arguments{
\item{x}{An object of class report_dc}

\item{...}{Additional parameters passed to print}
}
\value{
Nothing, called for its side effect
}
\description{
Method to print the command line of the object.
}
\author{
cedric.briand
}
