% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_df.R
\name{print,report_df-method}
\alias{print,report_df-method}
\alias{print.report_df}
\title{Method to print the command line of the object}
\usage{
\S4method{print}{report_df}(x, ...)
}
\arguments{
\item{x}{An object of class report_df}

\item{...}{Additional parameters passed to print}
}
\value{
Nothing, called for its side effect of printing data
}
\description{
Method to print the command line of the object
}
\author{
cedric.briand
}
