% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig.R
\name{print,report_mig-method}
\alias{print,report_mig-method}
\alias{print.report_mig}
\title{handler to print the command line}
\usage{
\S4method{print}{report_mig}(x, ...)
}
\arguments{
\item{x}{An object of class report_mig}

\item{...}{Additional parameters passed to print}

\item{h}{a handler}
}
\description{
handler to print the command line
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}

cedric.briand
}
\keyword{Method}
\keyword{command}
\keyword{internal}
\keyword{line}
\keyword{object}
\keyword{of}
\keyword{print}
\keyword{the}
\keyword{to}
