% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{quitte}
\alias{quitte}
\title{function used to clean the objects within the group and the graphs and
also elements remaining in the envir_stacomi environment}
\usage{
quitte(...)
}
\arguments{
\item{\dots}{additional arguments passed to the function}
}
\description{
function used to clean the objects within the group and the graphs and
also elements remaining in the envir_stacomi environment
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
