\name{quitte}
\alias{quitte}
\title{function used to clean the objects whithin the group and the graphes
and also elements remaining in the envir_stacomi environment}
\usage{
  quitte(...)
}
\arguments{
  \item{...}{}
}
\description{
  function used to clean the objects whithin the group and
  the graphes and also elements remaining in the
  envir_stacomi environment
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

