% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_parqual.R
\docType{class}
\name{ref_parqual-class}
\alias{ref_parqual-class}
\title{Class 'ref_parqual'}
\description{
Class enabling to load the list of qualitative parameters and to select one
of them. It inherits from 'ref_par', uses its 'choice' method
}
\section{Slots}{

\describe{
\item{\code{valqual='data.frame'}}{the list of qualitative parameters}
}}

\seealso{
Other referential objects: 
\code{\link{charge,ref_choice-method}},
\code{\link{ref_choice-class}},
\code{\link{ref_coe-class}},
\code{\link{ref_dc-class}},
\code{\link{ref_df-class}},
\code{\link{ref_horodate-class}},
\code{\link{ref_list-class}},
\code{\link{ref_par-class}},
\code{\link{ref_parquan-class}},
\code{\link{ref_stage-class}},
\code{\link{ref_taxa-class}},
\code{\link{ref_year-class}}
}
\author{
cedric.briand@eptb-vilaine.fr
}
\concept{referential objects}
