% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_stage.R
\docType{class}
\name{ref_stage-class}
\alias{ref_stage-class}
\title{Class "ref_stage"}
\description{
Representation of a fish phase
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("ref_stage", data="data.frame")}.  \describe{
\item{list("data")}{Object of class \code{"data.frame"} ~ The phases
available in the database}\item{:}{Object of class \code{"data.frame"} ~ The
phases available in the database} }
}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\seealso{
Other referential.objects: \code{\link{charge,ref_choice-method}},
  \code{\link{ref_checkbox-class}},
  \code{\link{ref_choice-class}},
  \code{\link{ref_coe-class}}, \code{\link{ref_dc-class}},
  \code{\link{ref_df-class}},
  \code{\link{ref_horodate-class}},
  \code{\link{ref_list-class}},
  \code{\link{ref_par-class}},
  \code{\link{ref_parqual-class}},
  \code{\link{ref_parquan-class}},
  \code{\link{ref_period-class}},
  \code{\link{ref_taxa-class}},
  \code{\link{ref_year-class}}
}
\keyword{classes}

