% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefCheckBox.r
\docType{class}
\name{RefCheckBox-class}
\alias{RefCheckBox-class}
\title{RefCheckBox referencial class}
\description{
referential class allows to choose for several parms with checkbox
}
\section{Slots}{

\describe{
\item{\code{title}}{A "character", the title of the box giving the possible choices}

\item{\code{labels}}{The logical parameters choice}

\item{\code{checked}}{A boolean vector}
}}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("RefCheckBox", ...)}.
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}

Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
\seealso{
Other referential classes 
\code{\linkS4class{RefAnnee}}
\code{\linkS4class{RefCheckBox}} 
\code{\linkS4class{RefChoix}}
\code{\linkS4class{RefCoe}} 
\code{\linkS4class{RefDC}}
\code{\linkS4class{RefDF}} 
\code{\linkS4class{RefListe}}
\code{\linkS4class{Refpar}} 
\code{\linkS4class{Refparqual}}
\code{\linkS4class{Refparquan}} 
\code{\linkS4class{RefPoidsMoyenPeche}}
\code{\linkS4class{RefStades}} 
\code{\linkS4class{RefStationMesure}}
\code{\linkS4class{RefTaxon}}

Other Referential.objects: \code{\link{RefAnnee-class}},
  \code{\link{RefChoix-class}}, \code{\link{RefCoe-class}},
  \code{\link{RefDC-class}}, \code{\link{RefDF-class}},
  \code{\link{RefHorodate-class}},
  \code{\link{RefListe-class}},
  \code{\link{RefStades-class}},
  \code{\link{Refpar-class}},
  \code{\link{Refparqual-class}},
  \code{\link{Refparquan-class}},
  \code{\link{Refperiode-class}},
  \code{\link{charge,RefChoix-method}},
  \code{\link{charge,RefMsg-method}}
}

