% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefChoix.r
\docType{class}
\name{RefChoix-class}
\alias{RefChoix-class}
\title{Class "RefChoix"}
\description{
RefChoix referential class allows to choose within several values with
radiobuttons
}
\section{Slots}{

\describe{
\item{\code{listechoix}}{A character vector giving possible choices}

\item{\code{label}}{A character, title of the box giving the possible choices}

\item{\code{selected}}{An \code{Integer}  the initial selected value (as an index), first=1 used in gradio}
}}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("RefChoix", listechoix=character() ,label=character()
,selected=integer())}.
}
\examples{

showClass("RefChoix")

}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\seealso{
Other referential classes 
\code{\linkS4class{RefAnnee}}
\code{\linkS4class{RefCheckBox}} 
\code{\linkS4class{RefChoix}}
\code{\linkS4class{RefCoe}} 
\code{\linkS4class{RefDC}}
\code{\linkS4class{RefDF}} 
\code{\linkS4class{RefListe}}
\code{\linkS4class{Refpar}} 
\code{\linkS4class{Refparqual}}
\code{\linkS4class{Refparquan}} 
\code{\linkS4class{RefPoidsMoyenPeche}}
\code{\linkS4class{RefStades}} 
\code{\linkS4class{RefStationMesure}}
\code{\linkS4class{RefTaxon}}

Other Referential objects: \code{\link{RefAnnee-class}},
  \code{\link{RefCheckBox-class}},
  \code{\link{RefCoe-class}}, \code{\link{RefDC-class}},
  \code{\link{RefDF-class}},
  \code{\link{RefHorodate-class}},
  \code{\link{RefListe-class}},
  \code{\link{RefPoidsMoyenPeche}},
  \code{\link{RefStades-class}},
  \code{\link{Refpar-class}},
  \code{\link{Refparqual-class}},
  \code{\link{Refparquan-class}},
  \code{\link{Refperiode-class}},
  \code{\link{charge,RefChoix-method}},
  \code{\link{charge,RefMsg-method}}
}

