% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefChoix.r
\docType{class}
\name{RefChoix-class}
\alias{RefChoix-class}
\title{Class "RefChoix"}
\description{
RefChoix referential class allows to choose within several values with
radiobuttons
}
\section{Slots}{

\describe{
\item{\code{listechoice}}{A character vector giving possible choices}

\item{\code{label}}{A character, title of the box giving the possible choices}

\item{\code{selected}}{An \code{Integer}  the initial selected value (as an index), first=1 used in gradio}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("RefChoix", listechoice=character() ,label=character()
,selected=integer())}.
}

\seealso{
Other Referential objects: \code{\link{RefAnnee-class}},
  \code{\link{RefCheckBox-class}},
  \code{\link{RefCoe-class}}, \code{\link{RefDC-class}},
  \code{\link{RefDF-class}},
  \code{\link{RefHorodate-class}},
  \code{\link{RefListe-class}},
  \code{\link{RefStades-class}},
  \code{\link{RefTaxon-class}}, \code{\link{Refpar-class}},
  \code{\link{Refparqual-class}},
  \code{\link{Refparquan-class}},
  \code{\link{Refperiode-class}},
  \code{\link{charge,RefChoix-method}}
}
\author{
cedric.briand"at"eptb-vilaine.fr
}
