% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefCoe.r
\docType{class}
\name{RefCoe-class}
\alias{RefCoe-class}
\title{Class "RefCoe"}
\description{
Enables to load conversion coefficients quantity-number. This class only exists to load
the data with its method charge. It is not used directly as component of the graphical interface,
as the year is already loaded in the different Bilan objects
}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{data.frame}}

\item{\code{datedebut}}{A "POSIXlt"}

\item{\code{datefin}}{A "POSIXlt"}
}}
\note{
Class loading coefficient of conversion between quantity (weights or
volumes of glass eel) and numbers between a starting and finishing date
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("RefCoe")}.
}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\seealso{
Other Referential.objects: \code{\link{RefAnnee-class}},
  \code{\link{RefCheckBox-class}},
  \code{\link{RefChoix-class}}, \code{\link{RefDC-class}},
  \code{\link{RefDF-class}},
  \code{\link{RefHorodate-class}},
  \code{\link{RefListe-class}},
  \code{\link{RefStades-class}},
  \code{\link{RefTaxon-class}}, \code{\link{Refpar-class}},
  \code{\link{Refparqual-class}},
  \code{\link{Refparquan-class}},
  \code{\link{Refperiode-class}},
  \code{\link{charge,RefChoix-method}}
}
\keyword{classes}

