% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefDC.r
\docType{class}
\name{RefDC-class}
\alias{RefDC-class}
\title{Class "RefDC"}
\description{
Description of a control device.
}
\section{Slots}{

\describe{
\item{\code{dc_selectionne}}{Object of class \code{"integer"}, The selected device}

\item{\code{ouvrage}}{Object of class \code{"integer"}, the attached dam}

\item{\code{station}}{Object of class \code{"character"}, the attached migration monitoring station, this is necessary to join the
table of escapments calculated at the station level.}

\item{\code{data}}{Object of class \code{"data.frame"} data pertaining to the control device}
}}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("RefDC", dc_selectionne=integer(), ouvrage=integer(),
data=data.frame())}.
}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\seealso{
Other referential classes \code{\linkS4class{RefAnnee}}
\code{\linkS4class{RefCheckBox}} \code{\linkS4class{RefChoix}}
\code{\linkS4class{RefCoe}} \code{\linkS4class{RefDC}}
\code{\linkS4class{RefDF}} \code{\linkS4class{RefListe}}
\code{\linkS4class{Refpar}} \code{\linkS4class{Refparqual}}
\code{\linkS4class{Refparquan}} \code{\linkS4class{RefPoidsMoyenPeche}}
\code{\linkS4class{RefStades}} \code{\linkS4class{RefStationMesure}}
\code{\linkS4class{RefTaxon}}

Other Referential.objects: \code{\link{RefAnnee-class}},
  \code{\link{RefCheckBox-class}},
  \code{\link{RefChoix-class}}, \code{\link{RefCoe-class}},
  \code{\link{RefDF-class}},
  \code{\link{RefHorodate-class}},
  \code{\link{RefListe-class}},
  \code{\link{RefStades-class}},
  \code{\link{Refpar-class}},
  \code{\link{Refparqual-class}},
  \code{\link{Refparquan-class}},
  \code{\link{Refperiode-class}},
  \code{\link{charge,RefChoix-method}},
  \code{\link{charge,RefMsg-method}}
}
\keyword{classes}

