% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefDF.r
\docType{class}
\name{RefDF-class}
\alias{RefDF-class}
\title{Class "RefDF"}
\description{
Representation of a fishway, Contains description data of all fishways from
the database along with the selected fishway(df) (integer)
Objects from the Class: Objects can be created by calls of the form
\code{new("RefDF", df_selectionne=integer(), ouvrage=integer(),
data=data.frame())}.
}
\section{Slots}{

\describe{
\item{\code{df_selectionne}}{Object of class \code{"integer"} The identifier of the fishway}

\item{\code{ouvrage}}{Object of class \code{"integer"} The attached dam}

\item{\code{data}}{Object of class \code{"data.frame"} Data concerning the fishway}
}}
\examples{

showClass("RefDF")

}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\seealso{
Other referential classes 
\code{\linkS4class{RefAnnee}}
\code{\linkS4class{RefCheckBox}} 
\code{\linkS4class{RefChoix}}
\code{\linkS4class{RefCoe}} 
\code{\linkS4class{RefDC}}
\code{\linkS4class{RefDF}} 
\code{\linkS4class{RefListe}}
\code{\linkS4class{Refpar}} 
\code{\linkS4class{Refparqual}}
\code{\linkS4class{Refparquan}} 
\code{\linkS4class{RefPoidsMoyenPeche}}
\code{\linkS4class{RefStades}} 
\code{\linkS4class{RefStationMesure}}
\code{\linkS4class{RefTaxon}}

Other Referential objects: \code{\link{RefAnnee-class}},
  \code{\link{RefCheckBox-class}},
  \code{\link{RefChoix-class}}, \code{\link{RefCoe-class}},
  \code{\link{RefDC-class}},
  \code{\link{RefHorodate-class}},
  \code{\link{RefListe-class}},
  \code{\link{RefPoidsMoyenPeche-class}},
  \code{\link{RefStades-class}},
  \code{\link{Refpar-class}},
  \code{\link{Refparqual-class}},
  \code{\link{Refparquan-class}},
  \code{\link{Refperiode-class}},
  \code{\link{charge,RefChoix-method}},
  \code{\link{charge,RefMsg-method}}
}

