% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefListe.r
\docType{class}
\name{RefListe-class}
\alias{RefListe-class}
\title{Class "RefListe"}
\arguments{
\item{liste}{choice="character" A vector of character to choose within a droplist}

\item{label="character"}{the title of the box}
}
\description{
Enables to load a "RefChoix" object fom a list given by a "Bilan" object
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("RefListe", listechoice, label)}.  \describe{
\item{list("listechoice")}{Object of class \code{"character"}}\item{:}{Object
of class \code{"character"}} \item{list("label")}{Object of class
\code{"character"}}\item{:}{Object of class \code{"character"}} }
}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\seealso{
Other Referential.objects: \code{\link{RefAnnee-class}},
  \code{\link{RefCheckBox-class}},
  \code{\link{RefChoix-class}}, \code{\link{RefCoe-class}},
  \code{\link{RefDC-class}}, \code{\link{RefDF-class}},
  \code{\link{RefHorodate-class}},
  \code{\link{RefStades-class}},
  \code{\link{RefTaxon-class}}, \code{\link{Refpar-class}},
  \code{\link{Refparqual-class}},
  \code{\link{Refparquan-class}},
  \code{\link{Refperiode-class}},
  \code{\link{charge,RefChoix-method}},
  \code{\link{charge,RefMsg-method}}
}
\keyword{classes}

