\name{RefMsg-class}
\docType{class}
\alias{RefMsg-class}
\alias{RefMsg}
\alias{createmessage,RefMsg-method}
\title{RefMsg referential class to load message according to the language chosen}
\description{
  RefMsg referential class to load message according to
  the language chosen
}
\section{Methods}{
  \describe{
    \item{createmessage}{\code{signature(objet = "RefMsg")}: creates a message }
}}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

