\name{RefMsg-class}
\docType{class}
\alias{RefMsg-class}
\alias{RefMsg}
\title{RefMsg referential class to load message according to the language chosen}
\description{
  RefMsg referential class to load message according to
  the language chosen
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

