% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refpar.r
\docType{class}
\name{Refpar-class}
\alias{Refpar-class}
\title{Class "Refpar"}
\description{
Class enabling to load the list of parameters and select one of them
}
\section{Slots}{

\describe{
\item{\code{data}}{A data.frame}

\item{\code{par_selectionne}}{A character vector corresponding to par_code}

\item{\code{data="data.frame"}}{the list of parameters}
}}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("Refpar", data)}.  \describe{ \item{list("data")}{Object of class
\code{"data.frame"} ~ All the parameters stored in the
database}\item{:}{Object of class \code{"data.frame"} ~ All the parameters
stored in the database} }
}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\seealso{
Other Referential.objects: \code{\link{RefAnnee-class}},
  \code{\link{RefCheckBox-class}},
  \code{\link{RefChoix-class}}, \code{\link{RefCoe-class}},
  \code{\link{RefDC-class}}, \code{\link{RefDF-class}},
  \code{\link{RefHorodate-class}},
  \code{\link{RefListe-class}},
  \code{\link{RefStades-class}},
  \code{\link{RefTaxon-class}},
  \code{\link{Refparqual-class}},
  \code{\link{Refparquan-class}},
  \code{\link{Refperiode-class}},
  \code{\link{charge,RefChoix-method}}
}
\keyword{classes}

