% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refparquan.r
\docType{class}
\name{Refparquan-class}
\alias{Refparquan-class}
\title{Class "Refparquan"}
\description{
Class enabling to load the list of quantitative parameters and to select one
of them. It inherits from 'Refpar', uses its 'choice' method
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("Refparquan", data=data.frame())}.  \describe{
\item{list("data")}{Object of class \code{"data.frame"} ~ Quantitative
parameters }\item{:}{Object of class \code{"data.frame"} ~ Quantitative
parameters } }
}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\seealso{
Other Referential.objects: \code{\link{Ref_year-class}},
  \code{\link{RefCheckBox-class}},
  \code{\link{RefChoix-class}}, \code{\link{RefCoe-class}},
  \code{\link{RefDC-class}}, \code{\link{RefDF-class}},
  \code{\link{RefHorodate-class}},
  \code{\link{RefListe-class}},
  \code{\link{RefStades-class}},
  \code{\link{RefTaxon-class}}, \code{\link{Refpar-class}},
  \code{\link{Refparqual-class}},
  \code{\link{Refperiode-class}},
  \code{\link{charge,RefChoix-method}}
}
\keyword{classes}

