% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refperiode.r
\docType{class}
\name{Refperiode-class}
\alias{Refperiode-class}
\title{Class "Refperiode" referential class}
\description{
Refperiode referential class to choose a period
}
\section{Slots}{

\describe{
\item{\code{data="data.frame"}}{providing correspondance between period and their english names}
}}
\note{
pgval are used by seq.POSIXT
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("Refperiode", ...)}.
}
\seealso{
Other Referential.objects: \code{\link{RefAnnee-class}},
  \code{\link{RefCheckBox-class}},
  \code{\link{RefChoix-class}}, \code{\link{RefCoe-class}},
  \code{\link{RefDC-class}}, \code{\link{RefDF-class}},
  \code{\link{RefHorodate-class}},
  \code{\link{RefListe-class}},
  \code{\link{RefStades-class}},
  \code{\link{RefTaxon-class}}, \code{\link{Refpar-class}},
  \code{\link{Refparqual-class}},
  \code{\link{Refparquan-class}},
  \code{\link{charge,RefChoix-method}}
}
\keyword{classes}

