% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefStationMesure.r
\docType{class}
\name{RefStationMesure-class}
\alias{RefStationMesure-class}
\title{Class "RefStationMesure"}
\description{
Enables to load measure stations and to select one of them
}
\section{Slots}{

\describe{
\item{\code{dataframe}}{Data concerning the
measure station}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("RefStationMesure", ...)}.
}

\author{
cedric.briand"at"eptb-vilaine.fr
}
\keyword{classes}
