% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefTaxon.r
\docType{class}
\name{RefTaxon-class}
\alias{RefTaxon-class}
\title{Class "RefTaxon"}
\description{
Loading and selection of fish species. This class is a referential class, and it is 
integrated into refBilan objects.
}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{"data.frame"} of species available in the database}
}}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("RefTaxon", ...)}.
}
\author{
cedric.briand"at"eptb-vilaine.fr
}
\seealso{
Other Referential objects: \code{\link{RefAnnee-class}},
  \code{\link{RefCheckBox-class}},
  \code{\link{RefChoix-class}}, \code{\link{RefCoe-class}},
  \code{\link{RefDC-class}}, \code{\link{RefDF-class}},
  \code{\link{RefHorodate-class}},
  \code{\link{RefListe-class}},
  \code{\link{RefStades-class}},
  \code{\link{Refpar-class}},
  \code{\link{Refparqual-class}},
  \code{\link{Refparquan-class}},
  \code{\link{Refperiode-class}},
  \code{\link{charge,RefChoix-method}},
  \code{\link{charge,RefMsg-method}}
}

