\name{round.POSIXt}
\alias{round.POSIXt}
\title{round...}
\usage{round.POSIXt(x, digits=c("secs", "mins", "hours", "days", "months",
    "years"))}
\description{round.POSIXT function imported from depuis Hmisc that cannot be loaded for reasons
of compatibility with xtable}
\value{a rounded time value}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{x}{}
\item{digits}{}}
