\name{round.POSIXt}
\alias{round.POSIXt}
\title{round.POSIXT function imported from depuis Hmisc that cannot be loaded for reasons
of compatibility with xtable}
\usage{
  round.POSIXt(x,
    digits = c("secs", "mins", "hours", "days", "months", "years"))
}
\arguments{
  \item{x}{}

  \item{digits}{}
}
\value{
  a rounded time value
}
\description{
  round.POSIXT function imported from depuis Hmisc that
  cannot be loaded for reasons of compatibility with xtable
}
\author{
 Charles Dupont 
}

