% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefHorodate.r
\docType{methods}
\name{setRefHorodate,RefHorodate-method}
\alias{setRefHorodate,RefHorodate-method}
\title{Method to set the Horodate}
\usage{
\S4method{setRefHorodate}{RefHorodate}(object, string)
}
\arguments{
\item{object}{An object of class \link{RefHorodate-class}}

\item{string}{A string representing an horodate in the format "\%Y-\%m-\%d \%H:\%M:\%S"}
}
\value{
An Object of class "RefHorodate"
}
\description{
Method to set the Horodate
}
\author{
cedric.briand
}
