\name{stacomi}
\alias{stacomi}
\title{Function that loads the loginwindow, tests connection, and then destroys the window...}
\description{Function that loads the loginwindow, tests connection, and then destroys the window}
\author{Cedric Briand \email{cedric.briand@lavilaine.com}}
\arguments{\item{baseODBC}{the ODBC connection chain contains user and password}}
