\name{stacomi}
\alias{stacomi}
\title{Function that loads the loginwindow, tests connection, and then destroys the window}
\usage{
  stacomi(gr_interface = TRUE)
}
\arguments{
  \item{baseODBC}{the ODBC connection chain contains user
  and password}
}
\description{
  Function that loads the loginwindow, tests connection,
  and then destroys the window
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

