% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacomi.r
\name{stacomi}
\alias{stacomi}
\title{Function that loads the loginwindow, tests connection, and then destroys the
window}
\usage{
stacomi(gr_interface = TRUE)
}
\arguments{
\item{gr_interface}{Will be used to launch the program as graphical
interface or in command line}
}
\description{
Function that loads the loginwindow, tests connection, and then destroys the
window
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

