\name{stacomiR-package}
\alias{stacomiR-package}
\alias{stacomiR}
\docType{package}
\title{
\packageTitle{stacomiR}
}
\description{
\packageDescription{stacomiR}
}
\details{
\packageDESCRIPTION{stacomiR}
\packageIndices{stacomiR}
To use the stacomiR package, simply load the library using \code{library(stacomiR)}. Once done, type \code{stacomi()} in R terminal.
The use of stacomi(gr_interface=FALSE) allows to launch stacomi without graphical interface from the command line. The program is currently
translated to French, to switch to English or French type \code{Sys.setenv(LANG = "en")} \code{Sys.setenv(LANG = "fr")}
}

\author{
\packageAuthor{stacomiR}
Maintainer: \packageMaintainer{stacomiR}
}
\keyword{ package }
\keyword{ dynamic }
\keyword{ database }
\seealso{
\code{\link[stacomirtools:stacomirtools-package]{stacomirtools}}
}

\examples{
  #stacomi()
}
