\name{stacomiR-package}
\Rdversion{1.1}
\alias{stacomi-package}
\docType{package}
\title{package stacomiR Stations de Controle des Migrations, Fish Migration control}
\description{
Stacomi enables to read informations in a database about fish migrations to build overviews of fish migrations and creating charts with datas.
}
\details{
\tabular{ll}{
Package: \tab stacomiR\cr
Type: \tab Package\cr
Version: \tab 0.4.1\cr
Date: \tab 2011-12-24\cr
License: \tab GPL license\cr
LazyLoad: \tab yes\cr
Depends: \tab methods\cr
}
To use the stacomiR package, simply load the library using \code{library(stacomiR)}. Once done, type \code{stacomi()} in R terminal
}
\author{
Author and Maintainer : \email{cedric.briand00@gmail.com}
}
\references{
\url{http://trac.eptb-vilaine.fr:8066/tracstacomi}
}

\keyword{package}
\keyword{dynamic}
\keyword{database}
\seealso{
\code{\link[pkg]{stacomiRtools}}
}

\examples{
  stacomi()
}
