\name{stacomiR-package}
\Rdversion{1.1}
\alias{stacomi-package}
\alias{stacomi}
\docType{package}
\title{package stacomiR Stations de Controle des Migrations, Fish Migration control}
\description{
Stacomi enables to read informations in a database about fish migrations to build overviews of fish migrations and creating charts with datas.
}
\details{
\tabular{ll}{
Package: \tab stacomiR\cr
Type: \tab Package\cr
Version: \tab 0.4\cr
Date: \tab 2011-02-09\cr
License: \tab GPL license\cr
LazyLoad: \tab yes\cr
Depends: \tab methods\cr
}

To use the stacomiR package, simply load the library using \code{library(stacomiR)}. Once done, type \code{stacomi()} in R terminal
}
\author{
Author and Maintainer : \email{cedric.briand00@gmail.com}
}
\references{
\url{http://trac.eptb-vilaine.fr:8066/tracstacomi}
}

\keyword{ package }
\keyword{ dynamic }
\keyword{ database }
\seealso{
\code{\link[<pkg>XML<pkg>]{XML}}
\code{\link[<pkg>methods<pkg>]{methods}}
\code{\link[<pkg>proto<pkg>]{proto}}
\code{\link[<pkg>tcltk2<pkg>]{tcltk2}}
\code{\link[<pkg>RODBC<pkg>]{RODBC}}
\code{\link[<pkg>ggplot2<pkg>]{ggplot2}}
\code{\link[<pkg>gWidgets<pkg>]{gWidgets}}
\code{\link[<pkg>R2HTML<pkg>]{R2HTML}}
\code{\link[<pkg>gWidgetsRGtk2<pkg>]{gWidgetsRGtk2}}
\code{\link[<pkg>lattice<pkg>]{lattice}}
\code{\link[<pkg>RColorBrewer<pkg>]{RColorBrewer}}
\code{\link[<pkg>Hmisc<pkg>]{Hmisc}}
}
}

\examples{
  stacomi()
}
