% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanAgedemer.r
\docType{methods}
\name{summary,BilanAgedemer-method}
\alias{summary,BilanAgedemer-method}
\title{summary for BilanAgedemer}
\usage{
\S4method{summary}{BilanAgedemer}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{An object of class \link{BilanAgedemer-class}}

\item{silent}{Default FALSE, if TRUE the program should no display messages.}

\item{...}{Additional parameters}
}
\description{
summary for BilanAgedemer
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

