% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_char.R
\docType{methods}
\name{summary,report_mig_char-method}
\alias{summary,report_mig_char-method}
\alias{summary.report_mig_char}
\title{summary for report_mig_char}
\usage{
\S4method{summary}{report_mig_char}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{report_mig_char-class}}}

\item{silent}{Should the program stay silent or display messages, default FALSE}

\item{...}{Additional parameters}
}
\description{
summary for report_mig_char
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
