% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_interannual.R
\docType{methods}
\name{summary,report_mig_interannual-method}
\alias{summary,report_mig_interannual-method}
\alias{summary.report_mig_interannual}
\title{summary for report_mig_interannual
provides summary statistics for the latest year (if silent=TRUE), or the year selected in the interface,
if silent=FALSE. Mean, min and max are historical statistics with the selected year excluded from the 
historical dataset.}
\usage{
\S4method{summary}{report_mig_interannual}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{report_mig_interannual-class}}}

\item{silent}{Should the program stay silent or display messages, default FALSE}

\item{...}{Additional parameters (not used there)}
}
\value{
A list, one element per DC
}
\description{
summary for report_mig_interannual
provides summary statistics for the latest year (if silent=TRUE), or the year selected in the interface,
if silent=FALSE. Mean, min and max are historical statistics with the selected year excluded from the 
historical dataset.
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

