% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_silver_eel.R
\name{summary,report_silver_eel-method}
\alias{summary,report_silver_eel-method}
\alias{summary.report_silver_eel}
\title{summary for report_silver_eel}
\usage{
\S4method{summary}{report_silver_eel}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{report_silver_eel-class}}}

\item{silent}{Should the program stay silent or display messages, default FALSE}

\item{...}{Additional parameters}
}
\value{
A list per DC with statistic for Durif stages, Pankhurst, MD Eye diameter, BL body length and weight W
}
\description{
summary for report_silver_eel
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
