% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_species.R
\name{summary,report_species-method}
\alias{summary,report_species-method}
\alias{summary.report_species}
\title{summary for report_species}
\usage{
\S4method{summary}{report_species}(object, silent = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{report_species-class}}}

\item{silent}{Should the program stay silent or display messages, default FALSE}
}
\value{
nothing, but writes summary in \code{get("datawd", envir = envir_stacomi)}, and prints output
}
\description{
generate csv and html output in the user data directory
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
