\name{supprime-methods}
\docType{methods}
\alias{supprime-methods}
\alias{supprime}
\alias{supprime,BilanMigrationInterAnnuelle-method}
\title{ Methods for Function supprime}
\description{
No idea what this one does
}
\section{Methods}{
\describe{

\item{\code{signature(objet = "BilanMigrationInterAnnuelle")}}{ Delete a record in the database}
}}
\keyword{methods}

