% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_coe.R
\docType{methods}
\name{supprime,ref_coe-method}
\alias{supprime,ref_coe-method}
\title{supprime method for "ref_coe" class}
\usage{
\S4method{supprime}{ref_coe}(object, tax, std, silent = FALSE)
}
\arguments{
\item{object}{An object of class \link{ref_coe-class}}

\item{tax}{'2038=Anguilla anguilla'}

\item{std}{'CIV=civelle'}

\item{silent}{Default FALSE, if TRUE the program should no display messages}
}
\value{
nothing
}
\description{
supprime method for "ref_coe" class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

