% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mig_interannual.R
\docType{methods}
\name{supprime,report_mig_interannual-method}
\alias{supprime,report_mig_interannual-method}
\alias{supprime.report_mig_interannual}
\title{supprime method for report_mig_interannual class}
\usage{
\S4method{supprime}{report_mig_interannual}(object)
}
\arguments{
\item{object}{An object of class \link{report_mig_interannual-class}}
}
\value{
nothing
}
\description{
supprime method for report_mig_interannual class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
