% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generic.R
\name{supprime}
\alias{supprime}
\title{Generic method to delete entires from the database}
\usage{
supprime(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Additional parms}
}
\description{
Generic method to delete entires from the database
}
\seealso{
\link{calcule.report_ge_weight}, \link{calcule.report_mig_char}, \link{calcule.report_mig_env},
\link{calcule.report_mig_interannual},\link{calcule.report_mig_mult},\link{calcule.report_mig_mult},
\link{calcule.report_sample_char},  \link{calcule.report_sea_age},  \link{calcule.report_silver_eel},
 \link{calcule.report_species}
}
\author{
cedric.briand
}
