% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_stades_pigm.r
\name{surface}
\alias{surface}
\title{plots polygons}
\usage{
surface(xmat, ymat, ordre = 1:dim(ymat)[2], couleur = 1:dim(ymat)[2], ...)
}
\arguments{
\item{xmat}{a matrix of x values of the polygons}

\item{ymat}{a matrix where the number of rows is of the same length as xmat}

\item{ordre}{the order in which the polygons will be drawn}

\item{couleur}{a color vector}

\item{\dots}{additional arguments passed to the function}
}
\value{
Bilan_stades_pigm with slots filled with user choice
}
\description{
plots polygons
}
\author{
Laurent Beaulaton \email{laurent.beaulaton"at"onema.fr}
}

