\name{surface}
\alias{surface}
\title{plots polygons}
\usage{
  surface(xmat, ymat, ordre = 1:dim(ymat)[2],
    couleur = 1:dim(ymat)[2], ...)
}
\value{
  Bilan_stades_pigm with slots filled with user choice
}
\description{
  plots polygons
}
\author{
  Laurent Beaulaton \email{laurent.beaulaton@onema.fr}
}

