% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitaires.r
\name{trunc.POSIXt}
\alias{trunc.POSIXt}
\title{trunc.POSIXt function copied from Hmisc that cannot be
loaded in Rforge}
\usage{
\method{trunc}{POSIXt}(x, digits = c("secs", "mins", "hours", "days",
  "months", "years"), ...)
}
\arguments{
\item{x}{Date to be floored, ceilinged, truncated, or rounded}

\item{digits}{Unit to that is is rounded up or down to, called digit for
compatibility with round or trunc generic method}

\item{\dots}{Further arguments to be passed to or from other methods}
}
\value{
An object of class POSIXlt.
}
\description{
trunc.POSIXt function copied from Hmisc that cannot be
loaded in Rforge
}
\author{
Charles Dupont
}

