\name{trunc.POSIXt}
\alias{ceil.POSIXt}
\title{trunc.POSIXT and ceil.POSIXt function imported from depuis Hmisc that cannot be loaded for reasons
of compatibility with xtable}
\usage{
  trunc.POSIXt(x,
    units = c("secs", "mins", "hours", "days", "months", "years"),
    ...)
  ceil.POSIXt(x,
    units = c("secs", "mins", "hours", "days", "months", "years"),
    ...)
}
\arguments{
  \item{x}{}

  \item{units}{}
}
\value{
  a truncated time value
}
\description{
  trunc.POSIXT function imported from depuis Hmisc that
  cannot be loaded for reasons of compatibility with xtable
}
\author{
  Charles Dupont
}

