\name{trunc.POSIXt}
\alias{trunc.POSIXt}
\title{trunc.POSIXT function imported from depuis Hmisc that cannot be loaded for reasons
of compatibility with xtable}
\usage{
  trunc.POSIXt(x,
    units = c("secs", "mins", "hours", "days", "months", "years"),
    ...)
}
\arguments{
  \item{x}{}

  \item{digits}{}
}
\value{
  a truncated time value
}
\description{
  trunc.POSIXT function imported from depuis Hmisc that
  cannot be loaded for reasons of compatibility with xtable
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

