% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefAnnee.r
\name{validite_Annee}
\alias{validite_Annee}
\title{Validity check for RefAnnee}
\usage{
validite_Annee(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{RefAnnee}}}
}
\value{
boolean The test for the object Refannee
}
\description{
validite_Annee tests for validity within the class
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

