\name{validite_Annee}
\alias{validite_Annee}
\title{validite_Annee tests for validity within the class}
\usage{
  validite_Annee(object)
}
\arguments{
  \item{object}{An object of class \code{\linkS4class{RefAnnee}}}
}
\value{
  the test for the object Refannee
}
\description{
  validite_Annee tests for validity within the class
}
\author{
  Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

