% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{vector_to_listsql}
\alias{vector_to_listsql}
\title{Transforms a vector into a string called within an sql command e.g.
c('A','B','C') => in ('A','B','C')}
\usage{
vector_to_listsql(vect)
}
\arguments{
\item{vect}{a character vector}
}
\value{
A list of value
}
\description{
Transforms a vector into a string called within an sql command e.g. c(A,B,C)
=> in ('A','B','C')
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
