% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bilan_poids_moyen.r
\docType{methods}
\name{write_database,Bilan_poids_moyen-method}
\alias{write_database,Bilan_poids_moyen-method}
\title{Function to write data to the stacomi database for \link{Bilan_poids_moyen-class}}
\usage{
\S4method{write_database}{Bilan_poids_moyen}(object, silent = FALSE,
  dbname = "bd_contmig_nat")
}
\arguments{
\item{object}{An object of class \link{Bilan_poids_moyen-class}}

\item{silent}{Boolean, if TRUE, information messages are not displayed}

\item{dbname}{default "bd_contmig_nat"}
}
\value{
An object of class \link{Bilan_poids_moyen-class}
}
\description{
Data will be written in tj_coefficientconversion_coe table, if the class retrieves some data
from the database, those will be deleted first.
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

