% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generic.R
\name{write_database}
\alias{write_database}
\title{Generic method write_database}
\usage{
write_database(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Additional parms}
}
\description{
Generic method write_database
}
\author{
cedric.briand
}
