% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generic.r
\name{write_database}
\alias{write_database}
\title{Generic method getvalue}
\usage{
write_database(object, ...)

write_database(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Additional parms}

\item{object}{Object}

\item{...}{Additional parms}
}
\description{
Generic method getvalue

Generic method getvalue
}
\author{
cedric.briand

cedric.briand
}

