% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BilanMigrationCar.r
\docType{methods}
\name{xtable,BilanMigrationCar-method}
\alias{xtable,BilanMigrationCar-method}
\title{xtable funciton for \link{BilanMigrationCar-class}
create an xtable objet to be later used by the print.xtable method.}
\usage{
\S4method{xtable}{BilanMigrationCar}(x, caption = NULL, label = NULL,
  align = NULL, digits = NULL, display = NULL, auto = FALSE, ...)
}
\arguments{
\item{x, }{an object of class "BilanMigrationCar"}

\item{caption, }{see xtable}

\item{label, }{see xtable}

\item{align, }{see xtable, overidden if NULL}

\item{digits}{default 0}

\item{display}{see xtable}
}
\description{
xtable funciton for \link{BilanMigrationCar-class}
create an xtable objet to be later used by the print.xtable method.
}

